polars::chunked_array
Module collect

Methods for collecting into a ChunkedArray.

For types that don’t have dtype parameters: iter.(try_)collect_ca(_trusted) (name)

For all types: iter.(try_)collect_ca(_trusted)like (other_df) Copies name/dtype from other_df iter.(try)collect_ca(_trusted)_with_dtype (name, df)

The try variants work on iterators of Results, the trusted variants do not check the length of the iterator.

Traits

ChunkedCollectInferIterExt

ChunkedCollectIterExt