Module polars::chunked_array::iterator

Modules

par

Structs

BooleanIterManyChunk
Iterator for chunked arrays with many chunks. The chunks cannot have null values so it does NOT perform null checks.

BooleanIterManyChunkNullCheck
Iterator for chunked arrays with many chunks. The chunks have null values so it DOES perform null checks.

BooleanIterSingleChunk
Iterator for chunked arrays with just one chunk. The chunk cannot have null values so it does NOT perform null checks.

BooleanIterSingleChunkNullCheck
Iterator for chunked arrays with just one chunk. The chunk have null values so it DOES perform null checks.

ListIterManyChunk
Iterator for chunked arrays with many chunks. The chunks cannot have null values so it does NOT perform null checks.

ListIterManyChunkNullCheck
Iterator for chunked arrays with many chunks. The chunks have null values so it DOES perform null checks.

ListIterSingleChunk
Iterator for chunked arrays with just one chunk. The chunk cannot have null values so it does NOT perform null checks.

ListIterSingleChunkNullCheck
Iterator for chunked arrays with just one chunk. The chunk have null values so it DOES perform null checks.

NumIterManyChunk
Iterator for chunked arrays with many chunks. The chunks cannot have null values so it does NOT perform null checks.

NumIterManyChunkNullCheck
Iterator for chunked arrays with many chunks. The chunks have null values so it DOES perform null checks.

NumIterSingleChunk
Iterator for chunked arrays with just one chunk. The chunk cannot have null values so it does NOT perform null checks.

NumIterSingleChunkNullCheck
Iterator for chunked arrays with just one chunk. The chunk have null values so it DOES perform null checks.

SomeIterator
Wrapper strunct to convert an iterator of type T into one of type Option<T>. It is useful to make the IntoIterator trait, in which every iterator shall return an Option<T>.

Utf8IterManyChunk
Iterator for chunked arrays with many chunks. The chunks cannot have null values so it does NOT perform null checks.

Utf8IterManyChunkNullCheck
Iterator for chunked arrays with many chunks. The chunks have null values so it DOES perform null checks.

Utf8IterSingleChunk
Iterator for chunked arrays with just one chunk. The chunk cannot have null values so it does NOT perform null checks.

Utf8IterSingleChunkNullCheck
Iterator for chunked arrays with just one chunk. The chunk have null values so it DOES perform null checks.

Traits

IntoNoNullIterator
Trait for ChunkedArrays that don't have null values. The result is the most efficient implementation Iterator, according to the number of chunks.

PolarsIterator
A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.