Module polars::chunked_array::iterator::par::boolean

Structs

BooleanParIterManyChunkNullCheckReturnOption
Parallel Iterator for chunked arrays with more than one chunk. It DOES perform null check, then, it is appropriated for chunks whose contents can be null.

BooleanParIterManyChunkReturnOption
Parallel Iterator for chunked arrays with more than one chunk. It does NOT perform null check, then, it is appropriated for chunks whose contents are never null.

BooleanParIterManyChunkReturnUnwrapped
Parallel Iterator for chunked arrays with many chunk. The chunks cannot have null values so it does NOT perform null checks.

BooleanParIterSingleChunkNullCheckReturnOption
Parallel Iterator for chunked arrays with just one chunk. It DOES perform null check, then, it is appropriated for chunks whose contents can be null.

BooleanParIterSingleChunkReturnOption
Parallel Iterator for chunked arrays with just one chunk. It does NOT perform null check, then, it is appropriated for chunks whose contents are never null.

BooleanParIterSingleChunkReturnUnwrapped
Parallel Iterator for chunked arrays with just one chunk. The chunks cannot have null values so it does NOT perform null checks.

Enums

BooleanNoNullParIterDispatcher
Static dispatching structure to allow static polymorphism of non-nullable chunked parallel iterators.

BooleanParIterDispatcher
Static dispatching structure to allow static polymorphism of chunked parallel iterators.