Module polars::chunked_array::iterator::par::utf8

This is supported on crate feature parallel only.

Structs

Utf8ParIterManyChunkNullCheckReturnOption
Parallel Iterator for chunked arrays with more than one chunk. It DOES perform null check, then, it is appropriated for chunks whose contents can be null.

Utf8ParIterManyChunkReturnOption
Parallel Iterator for chunked arrays with more than one chunk. It does NOT perform null check, then, it is appropriated for chunks whose contents are never null.

Utf8ParIterManyChunkReturnUnwrapped
Parallel Iterator for chunked arrays with many chunk. The chunks cannot have null values so it does NOT perform null checks.

Utf8ParIterSingleChunkNullCheckReturnOption
Parallel Iterator for chunked arrays with just one chunk. It DOES perform null check, then, it is appropriated for chunks whose contents can be null.

Utf8ParIterSingleChunkReturnOption
Parallel Iterator for chunked arrays with just one chunk. It does NOT perform null check, then, it is appropriated for chunks whose contents are never null.

Utf8ParIterSingleChunkReturnUnwrapped
Parallel Iterator for chunked arrays with just one chunk. The chunks cannot have null values so it does NOT perform null checks.

Enums

Utf8ChunkParIterReturnOptionDispatch
Static dispatching structure to allow static polymorphism of chunked parallel iterators.

Utf8ChunkParIterReturnUnwrapppedDispatch
Static dispatching structure to allow static polymorphism of non-nullable chunked parallel iterators.