Struct polars::chunked_array::metadata::Metadata

pub struct Metadata<T>
where
    T: PolarsDataType,
{ /* private fields */ }

Implementations

impl<T> Metadata<T>
where T: PolarsDataType,

pub const DEFAULT: Metadata<T> = _

pub fn sorted(self, is_sorted: IsSorted) -> Metadata<T>

pub fn fast_explode_list(self, fast_explode_list: bool) -> Metadata<T>

pub fn flags(self, flags: MetadataFlags) -> Metadata<T>

pub fn min_value( self, min_value: <T as PolarsDataType>::OwnedPhysical ) -> Metadata<T>

pub fn max_value( self, max_value: <T as PolarsDataType>::OwnedPhysical ) -> Metadata<T>

pub fn distinct_count(self, distinct_count: u32) -> Metadata<T>

pub fn sorted_opt(self, is_sorted: Option<IsSorted>) -> Metadata<T>

pub fn fast_explode_list_opt( self, fast_explode_list: Option<bool> ) -> Metadata<T>

pub fn flags_opt(self, flags: Option<MetadataFlags>) -> Metadata<T>

pub fn min_value_opt( self, min_value: Option<<T as PolarsDataType>::OwnedPhysical> ) -> Metadata<T>

pub fn max_value_opt( self, max_value: Option<<T as PolarsDataType>::OwnedPhysical> ) -> Metadata<T>

pub fn distinct_count_opt(self, distinct_count: Option<u32>) -> Metadata<T>

pub fn filter_props_cast<O>(&self, props: MetadataProperties) -> Metadata<O>
where O: PolarsDataType,

Create a Metadata with only the properties set in props.

pub fn filter_props(&self, props: MetadataProperties) -> Metadata<T>

Create a Metadata with only the properties set in props.

pub fn merge(&self, other: Metadata<T>) -> MetadataMerge<T>

Merge the maximum information from both Metadatas into one Metadata.

It returns

MetadataMerge::Keep if the self already contains all the information

MetadataMerge::New(md) if we have learned new information

MetadataMerge::Conflict if the two structures contain conflicting metadata

pub fn is_empty(&self) -> bool

pub fn is_sorted_ascending(&self) -> bool

pub fn set_sorted_ascending(&mut self, value: bool)

pub fn is_sorted_descending(&self) -> bool

pub fn set_sorted_descending(&mut self, value: bool)

pub fn get_fast_explode_list(&self) -> bool

pub fn set_fast_explode_list(&mut self, value: bool)

pub fn is_sorted(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, is_sorted: IsSorted)

pub fn set_min_value( &mut self, min_value: Option<<T as PolarsDataType>::OwnedPhysical> )

pub fn set_max_value( &mut self, max_value: Option<<T as PolarsDataType>::OwnedPhysical> )

pub fn set_distinct_count(&mut self, distinct_count: Option<u32>)

pub fn set_flags(&mut self, flags: MetadataFlags)

pub fn get_min_value(&self) -> Option<&<T as PolarsDataType>::OwnedPhysical>

pub fn get_max_value(&self) -> Option<&<T as PolarsDataType>::OwnedPhysical>

pub fn get_distinct_count(&self) -> Option<u32>

pub fn get_flags(&self) -> MetadataFlags