Module polars::chunked_array::object::datatypes

This is supported on crate feature object only.

Data types supported by Polars.

At the moment Polars doesn’t include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyValue variants for the data types that are currently supported.

Structs

ArrowField

Contains the meta-data for a single relative type.

ArrowSchema

Describes the meta-data of an ordered sequence of relative types.

BooleanType

CategoricalType

Date32Type

Date64Type

DurationMicrosecondType

DurationMillisecondType

DurationNanosecondType

DurationSecondType

Field

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

IntervalDayTimeType

IntervalYearMonthType

ListType

ObjectType

Schema

Time32MillisecondType

Time32SecondType

Time64MicrosecondType

Time64NanosecondType

TimestampMicrosecondType

TimestampMillisecondType

TimestampNanosecondType

TimestampSecondType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Utf8Type

Enums

AnyValue

ArrowDataType

The set of datatypes that are supported by this implementation of Apache Arrow.

DataType

IntervalUnit

YEAR_MONTH or DAY_TIME interval in SQL style.

TimeUnit

An absolute length of time in seconds, milliseconds, microseconds or nanoseconds.

Traits

ArrowNumericType

A subtype of primitive type that represents numeric values.

ArrowPrimitiveType

Trait bridging the dynamic-typed nature of Arrow (via DataType) with the static-typed nature of rust types (ArrowNativeType) for all types that implement ArrowNativeType.

PolarsDataType

PolarsFloatType

PolarsIntegerType

PolarsNumericType

PolarsPrimitiveType

PolarsSingleType

Any type that is not nested

Type Definitions

BooleanChunked

CategoricalChunked

Date32Chunked

Date64Chunked

DurationMillisecondChunked

DurationNanosecondChunked

Float32Chunked

Float64Chunked

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

ListChunked

ObjectChunked

PlHashMap

PlHashSet

SchemaRef

Time64NanosecondChunked

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked