Module polars::chunked_array::object::groupby

This is supported on crate feature object only.

Modules

aggregations

Structs

GroupBy

Returned by a groupby operation on a DataFrame. This struct supports several aggregations.

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsProxyIter

GroupsProxyParIter

Enums

GroupByMethod

GroupsIndicator

GroupsProxy

PivotAgg

Traits

IntoGroupsProxy

Used to create the tuples for a groupby operation.

Functions

fmt_groupby_column

Type Definitions

BorrowIdxItem

GroupedMap

GroupsSlice

Every group is indicated by an array where the

IdxItem