polars::chunked_array::object::registry

Trait AnonymousObjectBuilder

pub trait AnonymousObjectBuilder: ArrayBuilder {
    // Required methods
    fn as_array_builder(self: Box<Self>) -> Box<dyn ArrayBuilder>;
    unsafe fn from_chunks(
        self: Box<Self>,
        chunks: Vec<Box<dyn Array>>,
    ) -> Series;
    fn append_null(&mut self);
    fn append_value(&mut self, value: &(dyn Any + 'static));
    fn to_series(&mut self) -> Series;
    fn get_list_builder(
        &self,
        name: PlSmallStr,
        values_capacity: usize,
        list_capacity: usize,
    ) -> Box<dyn ListBuilderTrait>;

    // Provided method
    fn append_option(&mut self, value: Option<&(dyn Any + 'static)>) { ... }
}

This trait can be registered, after which that global registration can be used to materialize object types

Required Methods

fn as_array_builder(self: Box<Self>) -> Box<dyn ArrayBuilder>

unsafe fn from_chunks(self: Box<Self>, chunks: Vec<Box<dyn Array>>) -> Series

Safety

Expect ObjectArray<T> arrays.

fn append_null(&mut self)

Append a null value.

fn append_value(&mut self, value: &(dyn Any + 'static))

Append a T of ObjectChunked<T> made generic via the Any trait.

fn to_series(&mut self) -> Series

Take the current state and materialize as a Series the builder should not be used after that.

fn get_list_builder( &self, name: PlSmallStr, values_capacity: usize, list_capacity: usize, ) -> Box<dyn ListBuilderTrait>

Provided Methods

fn append_option(&mut self, value: Option<&(dyn Any + 'static)>)