Trait polars::chunked_array::object::ChunkAgg

pub trait ChunkAgg<T> {
    pub fn sum(&self) -> Option<T>;

    pub fn min(&self) -> Option<T>;

    pub fn max(&self) -> Option<T>;

    pub fn mean(&self) -> Option<T>;

    pub fn median(&self) -> Option<T>;

    pub fn quantile(&self, quantile: f64) -> Result<Option<T>, PolarsError>;
}

Aggregation operations

Required methods

pub fn sum(&self) -> Option<T>

Aggregate the sum of the ChunkedArray. Returns None if the array is empty or only contains null values.

pub fn min(&self) -> Option<T>

pub fn max(&self) -> Option<T>

Returns the maximum value in the array, according to the natural order. Returns None if the array is empty or only contains null values.

pub fn mean(&self) -> Option<T>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.

pub fn median(&self) -> Option<T>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.

pub fn quantile(&self, quantile: f64) -> Result<Option<T>, PolarsError>

Aggregate a given quantile of the ChunkedArray. Returns None if the array is empty or only contains null values.

Loading content...