Trait polars::chunked_array::object::ChunkCompare

pub trait ChunkCompare<Rhs> {
    pub fn eq_missing(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;

    pub fn eq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;

    pub fn neq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;

    pub fn gt(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;

    pub fn gt_eq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;

    pub fn lt(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;

    pub fn lt_eq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>;
}

Compare Series and ChunkedArray's and get a boolean mask that can be used to filter rows.

Example

use polars_core::prelude::*;
fn filter_all_ones(df: &DataFrame) -> Result<DataFrame> {
    let mask = df
    .column("column_a")?
    .eq(1);

    df.filter(&mask)
}

Required methods

pub fn eq_missing(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Check for equality and regard missing values as equal.

pub fn eq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Check for equality.

pub fn neq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Check for inequality.

pub fn gt(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Greater than comparison.

pub fn gt_eq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Greater than or equal comparison.

pub fn lt(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Less than comparison.

pub fn lt_eq(&self, rhs: Rhs) -> ChunkedArray<BooleanType>

Less than or equal comparison

Loading content...