Trait polars::chunked_array::object::ChunkCumAgg

pub trait ChunkCumAgg<T> {
    fn cummax(&self, _reverse: bool) -> ChunkedArray<T> { ... }

    fn cummin(&self, _reverse: bool) -> ChunkedArray<T> { ... }

    fn cumsum(&self, _reverse: bool) -> ChunkedArray<T> { ... }

    fn cumprod(&self, _reverse: bool) -> ChunkedArray<T> { ... }
}

This is supported on crate feature object only.

Provided methods

fn cummax(&self, _reverse: bool) -> ChunkedArray<T>

Get an array with the cumulative max computed at every element

fn cummin(&self, _reverse: bool) -> ChunkedArray<T>

Get an array with the cumulative min computed at every element

fn cumsum(&self, _reverse: bool) -> ChunkedArray<T>

Get an array with the cumulative sum computed at every element

fn cumprod(&self, _reverse: bool) -> ChunkedArray<T>

Get an array with the cumulative product computed at every element