Trait polars::chunked_array::object::StrConcat

pub trait StrConcat {
    fn str_concat(&self, delimiter: &str) -> ChunkedArray<Utf8Type>;
}

This is supported on crate features object and concat_str only.

Concat the values into a string array.

Required methods

fn str_concat(&self, delimiter: &str) -> ChunkedArray<Utf8Type>

Concat the values into a string array.

Arguments

delimiter - A string that will act as delimiter between values.