Trait polars::chunked_array::ops::ChunkAgg

pub trait ChunkAgg<T> {
    fn sum(&self) -> Option<T>;

    fn min(&self) -> Option<T>;

    fn max(&self) -> Option<T>;

    fn mean(&self) -> Option<T>;

    fn median(&self) -> Option<T>;
}

Aggregation operations

Required methods

fn sum(&self) -> Option<T>

Returns None if the array is empty or only contains null values.

fn min(&self) -> Option<T>

fn max(&self) -> Option<T>

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

fn mean(&self) -> Option<T>

Returns the mean value in the array. Returns an option because the array is nullable.

fn median(&self) -> Option<T>

Returns the mean value in the array. Returns an option because the array is nullable.

Loading content...