Trait polars::chunked_array::ops::ChunkApply

pub trait ChunkApply<'a, A, B> {
    fn apply<F>(&'a self, f: F) -> Self
    where
        F: Fn(A) -> B + Copy;

    fn apply_with_idx<F>(&'a self, f: F) -> Self
    where
        F: Fn((usize, A)) -> B + Copy;

    fn apply_with_idx_on_opt<F>(&'a self, f: F) -> Self
    where
        F: Fn((usize, Option<A>)) -> Option<B> + Copy;
}

Fastest way to do elementwise operations on a ChunkedArray

Required methods

fn apply<F>(&'a self, f: F) -> Self where
F: Fn(A) -> B + Copy,

Apply a closure F elementwise.

fn apply_with_idx<F>(&'a self, f: F) -> Self where
F: Fn((usize, A)) -> B + Copy,

Apply a closure elementwise. The closure gets the index of the element as first argument.

fn apply_with_idx_on_opt<F>(&'a self, f: F) -> Self where
F: Fn((usize, Option<A>)) -> Option<B> + Copy,

Apply a closure elementwise. The closure gets the index of the element as first argument.

Loading content...