Trait polars::chunked_array::ops::ChunkApplyKernel

pub trait ChunkApplyKernel<A>where
    A: Array,{
    // Required methods
    fn apply_kernel(
        &self,
        f: &dyn Fn(&A) -> Box<dyn Array + 'static, Global>
    ) -> Self;
    fn apply_kernel_cast<S>(
        &self,
        f: &dyn Fn(&A) -> Box<dyn Array + 'static, Global>
    ) -> ChunkedArray<S>
       where S: PolarsDataType;
}

Apply kernels on the arrow array chunks in a ChunkedArray.

Required Methods

fn apply_kernel( &self, f: &dyn Fn(&A) -> Box<dyn Array + 'static, Global> ) -> Self

Apply kernel and return result as a new ChunkedArray.

fn apply_kernel_cast<S>( &self, f: &dyn Fn(&A) -> Box<dyn Array + 'static, Global> ) -> ChunkedArray<S>where S: PolarsDataType,

Apply a kernel that outputs an array of different type.