polars::chunked_array::ops

Trait ChunkCast

pub trait ChunkCast {
    // Required methods
    fn cast_with_options(
        &self,
        dtype: &DataType,
        options: CastOptions,
    ) -> Result<Series, PolarsError>;
    unsafe fn cast_unchecked(
        &self,
        dtype: &DataType,
    ) -> Result<Series, PolarsError>;

    // Provided method
    fn cast(&self, dtype: &DataType) -> Result<Series, PolarsError> { ... }
}

Cast ChunkedArray<T> to ChunkedArray<N>

Required Methods

fn cast_with_options( &self, dtype: &DataType, options: CastOptions, ) -> Result<Series, PolarsError>

Cast a ChunkedArray to DataType

unsafe fn cast_unchecked(&self, dtype: &DataType) -> Result<Series, PolarsError>

Does not check if the cast is a valid one and may over/underflow

Safety

This doesn’t do utf8 validation checking when casting from binary

This doesn’t do categorical bound checking when casting from UInt32

Provided Methods

fn cast(&self, dtype: &DataType) -> Result<Series, PolarsError>

Cast a ChunkedArray to DataType