Trait polars::chunked_array::ops::ChunkFillNone

pub trait ChunkFillNone<T> {
    fn fill_none(&self, strategy: FillNoneStrategy) -> Result<Self>
    where
        Self: Sized;

    fn fill_none_with_value(&self, value: T) -> Result<Self>
    where
        Self: Sized;
}

Replace None values with various strategies

Required methods

fn fill_none(&self, strategy: FillNoneStrategy) -> Result<Self> where
Self: Sized,

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)

fn fill_none_with_value(&self, value: T) -> Result<Self> where
Self: Sized,

Replace None values with a give value T.

Loading content...