polars::chunked_array::ops

Trait ChunkNestingUtils

pub trait ChunkNestingUtils: Sized {
    // Required methods
    fn propagate_nulls(&self) -> Option<Self>;
    fn trim_lists_to_normalized_offsets(&self) -> Option<Self>;
    fn find_validity_mismatch(&self, other: &Series, idxs: &mut Vec<u32>);
}

Utility methods for dealing with nested chunked arrays.

Required Methods

fn propagate_nulls(&self) -> Option<Self>

Propagate nulls of nested datatype to all levels of nesting.

fn trim_lists_to_normalized_offsets(&self) -> Option<Self>

Trim all lists of unused start and end elements recursively.

fn find_validity_mismatch(&self, other: &Series, idxs: &mut Vec<u32>)

Find the indices of the values where the validity mismatches.

This is done recursively.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.