polars::chunked_array::ops

Trait ChunkQuantile

pub trait ChunkQuantile<T> {
    // Provided methods
    fn median(&self) -> Option<T> { ... }
    fn quantile(
        &self,
        _quantile: f64,
        _method: QuantileMethod,
    ) -> Result<Option<T>, PolarsError> { ... }
}

Quantile and median aggregation.

Provided Methods

fn median(&self) -> Option<T>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.

fn quantile( &self, _quantile: f64, _method: QuantileMethod, ) -> Result<Option<T>, PolarsError>

Aggregate a given quantile of the ChunkedArray. Returns None if the array is empty or only contains null values.