Trait polars::chunked_array::ops::ChunkShift

pub trait ChunkShift<T, V> {
    fn shift(
        &self, 
        periods: i32, 
        fill_value: &Option<V>
    ) -> Result<ChunkedArray<T>>;
}

Shift the values of a ChunkedArray by a number of periods.

Required methods

fn shift(&self, periods: i32, fill_value: &Option<V>) -> Result<ChunkedArray<T>>

Shift the values by a given period and fill the parts that will be empty due to this operation with fill_value.

Loading content...