Trait polars::chunked_array::ops::ChunkShiftFill

pub trait ChunkShiftFill<T, V> {
    pub fn shift_and_fill(&self, periods: i64, fill_value: V) -> ChunkedArray<T>;
}

Shift the values of a ChunkedArray by a number of periods.

Required methods

pub fn shift_and_fill(&self, periods: i64, fill_value: V) -> ChunkedArray<T>

Shift the values by a given period and fill the parts that will be empty due to this operation with fill_value.

Loading content...