Trait polars::chunked_array::ops::ChunkZip

pub trait ChunkZip<T> {
    fn zip_with(
        &self, 
        mask: &BooleanChunked, 
        other: &ChunkedArray<T>
    ) -> Result<ChunkedArray<T>>;

    fn zip_with_series(
        &self, 
        mask: &BooleanChunked, 
        other: &Series
    ) -> Result<ChunkedArray<T>>;
}

Combine 2 ChunkedArrays based on some predicate.

Required methods

fn zip_with(
&self,
mask: &BooleanChunked,
other: &ChunkedArray<T>
) -> Result<ChunkedArray<T>>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

fn zip_with_series(
&self,
mask: &BooleanChunked,
other: &Series
) -> Result<ChunkedArray<T>>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

Loading content...