Trait polars::chunked_array::ops::TakeRandom

pub trait TakeRandom {
    type Item;
    fn get(&self, index: usize) -> Option<Self::Item>;

    unsafe fn get_unchecked(&self, index: usize) -> Self::Item;
}

Random access

Associated Types

type Item

Loading content...
Required methods

fn get(&self, index: usize) -> Option<Self::Item>

Get a nullable value by index.

Safety

Out of bounds access doesn't Error but will return a Null value

unsafe fn get_unchecked(&self, index: usize) -> Self::Item

Get a value by index and ignore the null bit.

Safety

This doesn't check if the underlying type is null or not and may return an uninitialized value.

Loading content...