Struct polars::datatypes::categorical::CategoricalChunked

pub struct CategoricalChunked { /* private fields */ }

Available on crate feature dtype-categorical only.

Implementations

impl CategoricalChunked

pub fn sort_with(&self, options: SortOptions) -> CategoricalChunked

pub fn sort(&self, reverse: bool) -> CategoricalChunked

Returned a sorted ChunkedArray.

pub fn arg_sort(&self, options: SortOptions) -> ChunkedArray<UInt32Type>

Retrieve the indexes needed to sort this array.

impl CategoricalChunked

pub unsafe fn from_global_indices_unchecked( cats: ChunkedArray<UInt32Type> ) -> CategoricalChunked

Create a CategoricalChunked from a categorical indices. The indices will probe the global string cache.

Safety

This does not do any bound checks

impl CategoricalChunked

pub fn append(&mut self, other: &CategoricalChunked) -> Result<(), PolarsError>

impl CategoricalChunked

pub fn full_null(name: &str, length: usize) -> CategoricalChunked

impl CategoricalChunked

pub fn unique(&self) -> Result<CategoricalChunked, PolarsError>

pub fn n_unique(&self) -> Result<usize, PolarsError>

pub fn value_counts(&self) -> Result<DataFrame, PolarsError>

impl CategoricalChunked

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

pub fn name(&self) -> &str

pub fn logical(&self) -> &ChunkedArray<UInt32Type>

Get a reference to the logical array (the categories).

pub fn set_lexical_sorted(&mut self, toggle: bool)

pub unsafe fn from_cats_and_rev_map_unchecked( idx: ChunkedArray<UInt32Type>, rev_map: Arc<RevMapping> ) -> CategoricalChunked

Create a CategoricalChunked from an array of idx and an existing RevMapping: rev_map.

Safety

Invariant in v < rev_map.len() for v in idx must be hold.

pub fn get_rev_map(&self) -> &Arc<RevMapping>

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> CatIter<'_>

Create an [Iterator] that iterates over the &str values of the [CategoricalChunked].