Trait polars::datatypes::ArrowNumericType

pub trait ArrowNumericType: ArrowPrimitiveType {
    type Simd: Add<Self::Simd> + Sub<Self::Simd> + Mul<Self::Simd> + Div<Self::Simd> + Copy;
    type SimdMask;
    fn lanes() -> usize;

    fn init(value: Self::Native) -> Self::Simd;

    fn load(slice: &[Self::Native]) -> Self::Simd;

    fn mask_init(value: bool) -> Self::SimdMask;

    fn mask_from_u64(mask: u64) -> Self::SimdMask;

    fn mask_get(mask: &Self::SimdMask, idx: usize) -> bool;

    fn bitmask<T>(mask: &Self::SimdMask, action: T)
    where
        T: FnMut(&[u8]);

    fn mask_set(mask: Self::SimdMask, idx: usize, value: bool) -> Self::SimdMask;

    fn mask_select(
        mask: Self::SimdMask, 
        a: Self::Simd, 
        b: Self::Simd
    ) -> Self::Simd;

    fn mask_any(mask: Self::SimdMask) -> bool;

    fn bin_op<F>(left: Self::Simd, right: Self::Simd, op: F) -> Self::Simd
    where
        F: Fn(Self::Simd, Self::Simd) -> Self::Simd;

    fn eq(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn ne(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn lt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn le(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn gt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn ge(left: Self::Simd, right: Self::Simd) -> Self::SimdMask;

    fn write(simd_result: Self::Simd, slice: &mut [Self::Native]);
}

A subtype of primitive type that represents numeric values.

SIMD operations are defined in this trait if available on the target system.

Associated Types

type Simd: Add<Self::Simd> + Sub<Self::Simd> + Mul<Self::Simd> + Div<Self::Simd> + Copy

Defines the SIMD type that should be used for this numeric type

type SimdMask

Defines the SIMD Mask type that should be used for this numeric type

Loading content...
Required methods

fn lanes() -> usize

The number of SIMD lanes available

fn init(value: Self::Native) -> Self::Simd

Initializes a SIMD register to a constant value

fn load(slice: &[Self::Native]) -> Self::Simd

Loads a slice into a SIMD register

fn mask_init(value: bool) -> Self::SimdMask

Creates a new SIMD mask for this SIMD type filling it with value

fn mask_from_u64(mask: u64) -> Self::SimdMask

Creates a new SIMD mask for this SIMD type from the lower-most bits of the given mask

fn mask_get(mask: &Self::SimdMask, idx: usize) -> bool

Gets the value of a single lane in a SIMD mask

fn bitmask<T>(mask: &Self::SimdMask, action: T) where
T: FnMut(&[u8]),

Gets the bitmask for a SimdMask as a byte slice and passes it to the closure used as the action parameter

fn mask_set(mask: Self::SimdMask, idx: usize, value: bool) -> Self::SimdMask

Sets the value of a single lane of a SIMD mask

fn mask_select(mask: Self::SimdMask, a: Self::Simd, b: Self::Simd) -> Self::Simd

Selects elements of a and b using mask

fn mask_any(mask: Self::SimdMask) -> bool

Returns true if any of the lanes in the mask are true

fn bin_op<F>(left: Self::Simd, right: Self::Simd, op: F) -> Self::Simd where
F: Fn(Self::Simd, Self::Simd) -> Self::Simd,

Performs a SIMD binary operation

fn eq(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of equal

fn ne(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of not equal

fn lt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of less than

fn le(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of less than or equal to

fn gt(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of greater than

fn ge(left: Self::Simd, right: Self::Simd) -> Self::SimdMask

SIMD version of greater than or equal to

fn write(simd_result: Self::Simd, slice: &mut [Self::Native])

Writes a SIMD result back to a slice

Loading content...