Trait polars::datatypes::ArrowPrimitiveType

pub trait ArrowPrimitiveType: 'static {
    type Native: ArrowNativeType;
    fn get_data_type() -> DataType;

    fn get_bit_width() -> usize { ... }

    fn default_value() -> Self::Native { ... }

    unsafe fn index(raw_ptr: *const Self::Native, i: usize) -> Self::Native { ... }
}

Trait indicating a primitive fixed-width type (bool, ints and floats).

Associated Types

type Native: ArrowNativeType

Corresponding Rust native type for the primitive type.

Loading content...
Required methods

fn get_data_type() -> DataType

Returns the corresponding Arrow data type of this primitive type.

Loading content...
Provided methods

fn get_bit_width() -> usize

Returns the bit width of this primitive type.

fn default_value() -> Self::Native

Returns a default value of this primitive type.

This is useful for aggregate array ops like sum(), mean().

unsafe fn index(raw_ptr: *const Self::Native, i: usize) -> Self::Native

Returns a value offset from the given pointer by the given index. The default implementation (used for all non-boolean types) is simply equivalent to pointer-arithmetic.

Safety

Just like array-access in C: the raw_ptr must be the start of a valid array, and the index must be less than the size of the array.

Loading content...