Trait polars::datatypes::StaticallyMatchesPolarsType

pub unsafe trait StaticallyMatchesPolarsType<T>where
    T: PolarsDataType,{ }

Used to safely match the underlying type of Polars data structures.

Safety

The underlying physical type of the data structure on which this is implemented must always match the given PolarsDataType.