Type Alias polars::datatypes::IdxArr

pub type IdxArr = PrimitiveArray<u32>;

Available on non-crate feature bigidx only.

Aliased Type

struct IdxArr { /* private fields */ }

Implementations

impl<T> PrimitiveArray<T>where T: NativeType,

pub fn try_new( data_type: DataType, values: Buffer<T>, validity: Option<Bitmap> ) -> Result<PrimitiveArray<T>, PolarsError>

The canonical method to create a [PrimitiveArray] out of its internal components.

Implementation

This function is O(1).

Errors

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s [PhysicalType] is not equal to [PhysicalType::Primitive(T::PRIMITIVE)]

pub fn to(self, data_type: DataType) -> PrimitiveArray<T>

Returns a new [PrimitiveArray] with a different logical type.

This function is useful to assign a different DataType to the array. Used to change the arrays’ logical type (see example).

Example

use polars_arrow::array::Int32Array;
use polars_arrow::datatypes::DataType;

let array = Int32Array::from(&[Some(1), None, Some(2)]).to(DataType::Date32);
assert_eq!(
   format!("{:?}", array),
   "Date32[1970-01-02, None, 1970-01-03]"
);

Panics

Panics iff the data_type’s [PhysicalType] is not equal to [PhysicalType::Primitive(T::PRIMITIVE)]

pub fn from_vec(values: Vec<T, Global>) -> PrimitiveArray<T>

Creates a (non-null) [PrimitiveArray] from a vector of values. This function is O(1).

Examples

use polars_arrow::array::PrimitiveArray;

let array = PrimitiveArray::from_vec(vec![1, 2, 3]);
assert_eq!(format!("{:?}", array), "Int32[1, 2, 3]");

pub fn iter(&self) -> ZipValidity<&T, Iter<'_, T>, BitmapIter<'_>>

Returns an iterator over the values and validity, Option<&T>.

pub fn values_iter(&self) -> Iter<'_, T>

Returns an iterator of the values, &T, ignoring the arrays’ validity.

pub fn len(&self) -> usize

Returns the length of this array

pub fn values(&self) -> &Buffer<T>

The values [Buffer]. Values on null slots are undetermined (they can be anything).

pub fn validity(&self) -> Option<&Bitmap>

Returns the optional validity.

pub fn data_type(&self) -> &DataType

Returns the arrays’ DataType.

pub fn value(&self, i: usize) -> T

Returns the value at slot i.

Equivalent to self.values()[i]. The value of a null slot is undetermined (it can be anything).

Panic

This function panics iff i >= self.len.

pub unsafe fn value_unchecked(&self, i: usize) -> T

Returns the value at index i. The value on null slots is undetermined (it can be anything).

Safety

Caller must be sure that i < self.len()

pub fn get(&self, i: usize) -> Option<T>

Returns the element at index i or None if it is null

Panics

iff i >= self.len()

pub fn slice(&mut self, offset: usize, length: usize)

Slices this [PrimitiveArray] by an offset and length.

Implementation

This operation is O(1).

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices this [PrimitiveArray] by an offset and length.

Implementation

This operation is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn sliced(self, offset: usize, length: usize) -> PrimitiveArray<T>

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn sliced_unchecked( self, offset: usize, length: usize ) -> PrimitiveArray<T>

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> PrimitiveArray<T>

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn boxed(self) -> Box<dyn Array, Global>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array, Global>

Arcs this array into a std::sync::Arc<dyn Array>.

pub fn with_values(self, values: Buffer<T>) -> PrimitiveArray<T>

Returns this [PrimitiveArray] with new values.

Panics

This function panics iff values.len() != self.len().

pub fn set_values(&mut self, values: Buffer<T>)

Update the values of this [PrimitiveArray].

Panics

This function panics iff values.len() != self.len().

pub fn apply_validity<F>(&mut self, f: F)where F: FnOnce(Bitmap) -> Bitmap,

Applies a function f to the validity of this array.

This is an API to leverage clone-on-write

Panics

This function panics if the function f modifies the length of the [Bitmap].

pub fn get_mut_values(&mut self) -> Option<&mut [T]>

Returns an option of a mutable reference to the values of this [PrimitiveArray].

pub fn into_inner(self) -> (DataType, Buffer<T>, Option<Bitmap>)

Returns its internal representation

pub fn from_inner( data_type: DataType, values: Buffer<T>, validity: Option<Bitmap> ) -> Result<PrimitiveArray<T>, PolarsError>

Creates a [PrimitiveArray] from its internal representation. This is the inverted from [PrimitiveArray::into_inner]

pub unsafe fn from_inner_unchecked( data_type: DataType, values: Buffer<T>, validity: Option<Bitmap> ) -> PrimitiveArray<T>

Creates a [PrimitiveArray] from its internal representation. This is the inverted from [PrimitiveArray::into_inner]

Safety

Callers must ensure all invariants of this struct are upheld.

pub fn into_mut(self) -> Either<PrimitiveArray<T>, MutablePrimitiveArray<T>>

Try to convert this [PrimitiveArray] to a [MutablePrimitiveArray] via copy-on-write semantics.

A [PrimitiveArray] is backed by a [Buffer] and [Bitmap] which are essentially Arc<Vec<_>>. This function returns a [MutablePrimitiveArray] (via std::sync::Arc::get_mut) iff both values and validity have not been cloned / are unique references to their underlying vectors.

This function is primarily used to re-use memory regions.

pub fn new_empty(data_type: DataType) -> PrimitiveArray<T>

Returns a new empty (zero-length) [PrimitiveArray].

pub fn new_null(data_type: DataType, length: usize) -> PrimitiveArray<T>

Returns a new [PrimitiveArray] where all slots are null / None.

pub fn from_values<I>(iter: I) -> PrimitiveArray<T>where I: IntoIterator<Item = T>,

Creates a (non-null) [PrimitiveArray] from an iterator of values.

Implementation

This does not assume that the iterator has a known length.

pub fn from_slice<P>(slice: P) -> PrimitiveArray<T>where P: AsRef<[T]>,

Creates a (non-null) [PrimitiveArray] from a slice of values.

Implementation

This is essentially a memcopy and is thus O(N)

pub fn from_trusted_len_values_iter<I>(iter: I) -> PrimitiveArray<T>where I: TrustedLen<Item = T>,

Creates a (non-null) [PrimitiveArray] from a [TrustedLen] of values.

Implementation

This does not assume that the iterator has a known length.

pub unsafe fn from_trusted_len_values_iter_unchecked<I>( iter: I ) -> PrimitiveArray<T>where I: Iterator<Item = T>,

Creates a new [PrimitiveArray] from an iterator over values

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I>(iter: I) -> PrimitiveArray<T>where I: TrustedLen<Item = Option<T>>,

Creates a [PrimitiveArray] from a [TrustedLen] of optional values.

pub unsafe fn from_trusted_len_iter_unchecked<I>(iter: I) -> PrimitiveArray<T>where I: Iterator<Item = Option<T>>,

Creates a [PrimitiveArray] from an iterator of optional values.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn new( data_type: DataType, values: Buffer<T>, validity: Option<Bitmap> ) -> PrimitiveArray<T>

Alias for Self::try_new(..).unwrap().

Panics

This function errors iff:

The validity is not None and its length is different from values’s length

The data_type’s [PhysicalType] is not equal to [PhysicalType::Primitive].