Module polars::error

Modules

constants

Constant that help with creating error messages dependent on the host language.

Macros

feature_gated

polars_bail

polars_ensure

polars_err

polars_warn

Structs

ErrString

Enums

PolarsError

PolarsWarning

Functions

get_warning_function

map_err

set_warning_function⚠

Set the function that will be called by the polars_warn! macro. You can use this to set logging in polars.

to_compute_err

Type Aliases

PolarsResult