polars::error

Module abort

Enums

QueryAborted

Python hooks SIGINT to instead generate a KeyboardInterrupt exception. So we do the same to try and abort long-running computations and return to Python so that the Python exception can be generated.

Functions

catch_polars_abort

Runs the passed function, catching any query abortions if they occur while running the function.

polars_abort_ooc_out_of_disk

register_polars_abort_mechanism

try_raise_polars_abort

Checks if the abort flag is set, and if yes panics. This function is very cheap.