Enum polars::error::ArrowError

#[non_exhaustive]
pub enum ArrowError {
    NotYetImplemented(String),
    External(String, Box<dyn Error + Send + Sync + 'static, Global>),
    Io(Error),
    InvalidArgumentError(String),
    ExternalFormat(String),
    Overflow,
    OutOfSpec(String),
}

Enum with all errors in this crate.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NotYetImplemented(String)

Returned when functionality is not yet available.

External(String, Box<dyn Error + Send + Sync + 'static, Global>)

Wrapper for an error triggered by a dependency

Io(Error)

Wrapper for IO errors

InvalidArgumentError(String)

When an invalid argument is passed to a function.

ExternalFormat(String)

Error during import or export to/from a format

Overflow

Whenever pushing to a container fails because it does not support more entries. The solution is usually to use a higher-capacity container-backing type.

OutOfSpec(String)

Whenever incoming data from the C data interface, IPC or Flight does not fulfil the Arrow specification.

Implementations

impl Error

pub fn from_external_error(error: impl Error + Send + Sync + 'static) -> Error

Wraps an external error in an Error.