polars::error
Function set_warning_function

pub fn set_warning_function(function: fn(_: &str, _: PolarsWarning))

Set the function that will be called by the polars_warn! macro. You can use this to set logging in polars.