Struct polars::export::arrow::array::MutableListArray

pub struct MutableListArray<O, M>where
    O: Offset,
    M: MutableArray,{ /* private fields */ }

The mutable version of ListArray.

Implementations

impl<O, M> MutableListArray<O, M>where
O: Offset,
M: MutableArray + Default,

pub fn new() -> MutableListArray<O, M>

Creates a new empty MutableListArray.

pub fn with_capacity(capacity: usize) -> MutableListArray<O, M>

Creates a new MutableListArray with a capacity.

impl<O, M> MutableListArray<O, M>where
O: Offset,
M: MutableArray,

pub fn new_from(
values: M,
data_type: DataType,
capacity: usize
) -> MutableListArray<O, M>

Creates a new MutableListArray from a MutableArray and capacity.

pub fn new_with_field(
values: M,
name: &str,
nullable: bool
) -> MutableListArray<O, M>

Creates a new MutableListArray from a MutableArray.

pub fn new_with_capacity(values: M, capacity: usize) -> MutableListArray<O, M>

Creates a new MutableListArray from a MutableArray and capacity.

pub fn try_push_valid(&mut self) -> Result<(), Error>

Needs to be called when a valid value was extended to this array. This is a relatively low level function, prefer try_push when you can.

pub fn len(&self) -> usize

Returns the length of this array

pub fn mut_values(&mut self) -> &mut M

The values

pub fn offsets(&self) -> &Offsets<O>

The offsets

pub fn values(&self) -> &M

The values

pub fn into_arc(self) -> Arc<dyn Array + 'static>

Converts itself into an Array.

pub fn into_box(self) -> Box<dyn Array + 'static, Global>

converts itself into Box<dyn Array>

pub fn reserve(&mut self, additional: usize)

Reserves additional slots.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutableListArray to fit its current length.