Trait polars::export::arrow::array::TryExtend

pub trait TryExtend<A> {
    fn try_extend<I>(&mut self, iter: I) -> Result<(), Error>
    where
        I: IntoIterator<Item = A>;
}

A trait describing the ability of a struct to create itself from a iterator. This is similar to Extend, but accepted the creation to error.

Required Methods

fn try_extend<I>(&mut self, iter: I) -> Result<(), Error>where
I: IntoIterator<Item = A>,

Fallible version of Extend::extend.