Function polars::export::arrow::compute::aggregate::max

pub fn max(
    array: &(dyn Array + 'static)
) -> Result<Box<dyn Scalar + 'static, Global>, Error>

Returns the maximum of Array. The scalar is null when all elements are null.

Error

Errors iff the type does not support this operation.