Module polars::export::arrow::io::json::read

Available on crate feature io_json only.

APIs to read and deserialize from JSON

Modules

json_deserializer

Library to parse JSON

Functions

deserialize

Deserializes a json Value into an Array of DataType This is CPU-bounded.

deserialize_records

Deserializes a json Value serialized in Pandas record format into a Chunk.

infer

Infers DataType from Value.

infer_records_schema

Infers Schema from JSON Value in (pandas-compatible) records format.