Function polars::export::arrow::io::parquet::read::column_iter_to_arrays

pub fn column_iter_to_arrays<'a, I>(
    columns: Vec<I, Global>,
    types: Vec<&PrimitiveType, Global>,
    field: Field,
    chunk_size: Option<usize>,
    num_rows: usize
) -> Result<Box<dyn Iterator<Item = Result<Box<dyn Array + 'static, Global>, Error>> + Send + Sync + 'a, Global>, Error>where
    I: 'a + Pages,

An iterator adapter that maps multiple iterators of Pages into an iterator of Arrays.

For a non-nested datatypes such as DataType::Int32, this function requires a single element in columns and types. For nested types, columns must be composed by all parquet columns with associated types types.

The arrays are guaranteed to be at most of size chunk_size and data type field.data_type.