Function polars::export::arrow::io::parquet::read::get_column_iterator

pub fn get_column_iterator<R>(
    reader: R,
    row_group: &RowGroupMetaData,
    field_name: &str,
    page_filter: Option<Arc<dyn Fn(&Descriptor, &DataPageHeader) + Send + Sync + 'static>>,
    scratch: Vec<u8, Global>,
    max_page_size: usize
) -> ColumnIterator<R>where
    R: Read + Seek,

Returns a ColumnIterator of column chunks corresponding to field.

Contrarily to get_page_iterator that returns a single iterator of pages, this iterator iterates over columns, one by one, and returns a PageReader per column. For primitive fields (e.g. i64), ColumnIterator yields exactly one column. For complex fields, it yields multiple columns. max_page_size is the maximum number of bytes allowed.