Function polars::export::arrow::io::parquet::read::read_columns_many

pub fn read_columns_many<'a, R>(
    reader: &mut R,
    row_group: &RowGroupMetaData,
    fields: Vec<Field, Global>,
    chunk_size: Option<usize>,
    limit: Option<usize>,
    pages: Option<Vec<Vec<Vec<FilteredPage, Global>, Global>, Global>>
) -> Result<Vec<Box<dyn Iterator<Item = Result<Box<dyn Array + 'static, Global>, Error>> + Send + Sync + 'a, Global>, Global>, Error>where
    R: Read + Seek,

Returns a vector of iterators of Array (ArrayIter) corresponding to the top level parquet fields whose name matches fields’s names.

Implementation

This operation is IO-bounded O(C) where C is the number of columns in the row group - it reads all the columns to memory from the row group associated to the requested fields.

This operation is single-threaded. For readers with stronger invariants (e.g. implement Clone) you can use read_columns to read multiple columns at once and convert them to ArrayIter via to_deserializer.