Enum polars::export::arrow::io::parquet::write::ParquetType

pub enum ParquetType {
    PrimitiveType(PrimitiveType),
    GroupType {
        field_info: FieldInfo,
        logical_type: Option<GroupLogicalType>,
        converted_type: Option<GroupConvertedType>,
        fields: Vec<ParquetType, Global>,
    },
}

Representation of a Parquet type describing primitive and nested fields, including the top-level schema of the parquet file.

Variants

PrimitiveType(PrimitiveType)

GroupType

Fields

field_info: FieldInfo

logical_type: Option<GroupLogicalType>

converted_type: Option<GroupConvertedType>

fields: Vec<ParquetType, Global>

Implementations

impl ParquetType

pub fn try_from_thrift(elements: &[SchemaElement]) -> Result<ParquetType, Error>

Method to convert from Thrift.

impl ParquetType

Accessors

pub fn get_field_info(&self) -> &FieldInfo

Returns FieldInfo information about the type.

pub fn name(&self) -> &str

Returns this type’s field name.

pub fn check_contains(&self, sub_type: &ParquetType) -> bool

Checks if sub_type schema is part of current schema. This method can be used to check if projected columns are part of the root schema.

impl ParquetType

Constructors

pub fn from_converted(
name: String,
fields: Vec<ParquetType, Global>,
repetition: Repetition,
converted_type: Option<GroupConvertedType>,
id: Option<i32>
) -> ParquetType

pub fn try_from_primitive(
name: String,
physical_type: PhysicalType,
repetition: Repetition,
converted_type: Option<PrimitiveConvertedType>,
logical_type: Option<PrimitiveLogicalType>,
id: Option<i32>
) -> Result<ParquetType, Error>

Error

Errors iff the combination of physical, logical and coverted type is not valid.

pub fn from_physical(name: String, physical_type: PhysicalType) -> ParquetType

Helper method to create a ParquetType::PrimitiveType optional field with no logical or converted types.

pub fn from_group(
name: String,
repetition: Repetition,
converted_type: Option<GroupConvertedType>,
logical_type: Option<GroupLogicalType>,
fields: Vec<ParquetType, Global>,
id: Option<i32>
) -> ParquetType