Struct polars::export::arrow::io::parquet::write::FileSink

pub struct FileSink<'a, W>where
    W: AsyncWrite + Send + Unpin,{
    pub metadata: AHashMap<String, Option<String>, RandomState>,
    /* private fields */
}

Sink that writes array chunks as a Parquet file.

Any values in the sink’s metadata field will be written to the file’s footer when the sink is closed.

Examples

use futures::SinkExt;
use arrow2::array::{Array, Int32Array};
use arrow2::datatypes::{DataType, Field, Schema};
use arrow2::chunk::Chunk;
use arrow2::io::parquet::write::{Encoding, WriteOptions, CompressionOptions, Version};

let schema = Schema::from(vec![
    Field::new("values", DataType::Int32, true),
]);
let encoding = vec![vec![Encoding::Plain]];
let options = WriteOptions {
    write_statistics: true,
    compression: CompressionOptions::Uncompressed,
    version: Version::V2,
};

let mut buffer = vec![];
let mut sink = FileSink::try_new(
    &mut buffer,
    schema,
    encoding,
    options,
)?;

for i in 0..3 {
    let values = Int32Array::from(&[Some(i), None]);
    let chunk = Chunk::new(vec![values.boxed()]);
    sink.feed(chunk).await?;
}
sink.metadata.insert(String::from("key"), Some(String::from("value")));
sink.close().await?;

Fields
metadata: AHashMap<String, Option<String>, RandomState>

Key-value metadata that will be written to the file on close.

Implementations

impl<'a, W> FileSink<'a, W>where
W: 'a + AsyncWrite + Send + Unpin,

pub fn try_new(
writer: W,
schema: Schema,
encodings: Vec<Vec<Encoding, Global>, Global>,
options: WriteOptions
) -> Result<FileSink<'a, W>, Error>

Create a new sink that writes arrays to the provided writer.

Error

Iff

the Arrow schema can’t be converted to a valid Parquet schema.

the length of the encodings is different from the number of fields in schema

pub fn schema(&self) -> &Schema

The Arrow Schema for the file.

pub fn parquet_schema(&self) -> &SchemaDescriptor

The Parquet SchemaDescriptor for the file.

pub fn options(&self) -> &WriteOptions

The write options for the file.