Struct polars::export::arrow::io::parquet::write::FileWriter

pub struct FileWriter<W>where
    W: Write,{ /* private fields */ }

An interface to write a parquet to a Write

Implementations

impl<W> FileWriter<W>where
W: Write,

pub fn options(&self) -> WriteOptions

The options assigned to the file

pub fn parquet_schema(&self) -> &SchemaDescriptor

The SchemaDescriptor assigned to this file

pub fn schema(&self) -> &Schema

The Schema assigned to this file

impl<W> FileWriter<W>where
W: Write,

pub fn try_new(
writer: W,
schema: Schema,
options: WriteOptions
) -> Result<FileWriter<W>, Error>

Returns a new FileWriter.

Error

If it is unable to derive a parquet schema from Schema.

pub fn write(
&mut self,
row_group: DynIter<'_, Result<DynStreamingIterator<'_, CompressedPage, Error>, Error>>
) -> Result<(), Error>

Writes a row group to the file.

pub fn end(
&mut self,
key_value_metadata: Option<Vec<KeyValue, Global>>
) -> Result<u64, Error>

Writes the footer of the parquet file. Returns the total size of the file.

pub fn into_inner(self) -> W

Consumes this writer and returns the inner writer

pub fn into_inner_and_metadata(self) -> (W, FileMetaData)

Returns the underlying writer and ThriftFileMetaData

Panics

This function panics if Self::end has not yet been called