polars::frame::column

Struct ScalarColumn

pub struct ScalarColumn { /* private fields */ }

A Column that consists of a repeated Scalar

This is lazily materialized into a Series.

Implementations

impl ScalarColumn

pub fn new(name: PlSmallStr, scalar: Scalar, length: usize) -> ScalarColumn

pub fn new_empty(name: PlSmallStr, dtype: DataType) -> ScalarColumn

pub fn full_null( name: PlSmallStr, length: usize, dtype: DataType, ) -> ScalarColumn

pub fn name(&self) -> &PlSmallStr

pub fn scalar(&self) -> &Scalar

pub fn dtype(&self) -> &DataType

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn to_series(&self) -> Series

Materialize the ScalarColumn into a Series.

pub fn lazy_as_materialized_series(&self) -> Option<&Series>

Get the ScalarColumn as Series if it was already materialized.

pub fn as_materialized_series(&self) -> &Series

Get the ScalarColumn as Series

This needs to materialize upon the first call. Afterwards, this is cached.

pub fn take_materialized_series(self) -> Series

Take the ScalarColumn and materialize as a Series if not already done.

pub fn as_single_value_series(&self) -> Series

Take the ScalarColumn as a series with a single value.

If the ScalarColumn has length=0 the resulting Series will also have length=0.

pub fn as_n_values_series(&self, n: usize) -> Series

Take the ScalarColumn as a series with a n values.

If the ScalarColumn has length=0 the resulting Series will also have length=0.

pub fn unit_scalar_from_series(series: Series) -> ScalarColumn

Create a new ScalarColumn from a length=1 Series and expand it length.

This will panic if the value cannot be made static or if the series has length 0.

pub fn from_single_value_series(series: Series, length: usize) -> ScalarColumn

Create a new ScalarColumn from a length<=1 Series and expand it length.

If series is empty and length is non-zero, a full-NULL column of length will be returned.

This will panic if the value cannot be made static.

pub fn resize(&self, length: usize) -> ScalarColumn

Resize the ScalarColumn to new length.

This reuses the materialized Series, if length <= self.length.

pub fn cast_with_options( &self, dtype: &DataType, options: CastOptions, ) -> Result<ScalarColumn, PolarsError>

pub fn strict_cast(&self, dtype: &DataType) -> Result<ScalarColumn, PolarsError>

pub fn cast(&self, dtype: &DataType) -> Result<ScalarColumn, PolarsError>

pub unsafe fn cast_unchecked( &self, dtype: &DataType, ) -> Result<ScalarColumn, PolarsError>

Safety

This can lead to invalid memory access in downstream code.

pub fn rename(&mut self, name: PlSmallStr) -> &mut ScalarColumn

pub fn has_nulls(&self) -> bool

pub fn drop_nulls(&self) -> ScalarColumn

pub fn into_nulls(self) -> ScalarColumn

pub fn map_scalar(&mut self, map_scalar: impl Fn(Scalar) -> Scalar)

pub fn with_value(&mut self, value: AnyValue<'static>) -> &mut ScalarColumn