Module polars::functions

Functions

Functions that might be useful.

Functions

argsort_by

Find the indexes that would sort these series in order of appearance. That means that the first Series will be used to determine the ordering until duplicates are found. Once duplicates are found, the next Series will be used and so on.

concat_strconcat_str

Casts all series to string data and will concat them in linear time. The concatenated strings are separated by a delimiter. If no delimiter is needed, an empty &str should be passed as argument.

cov

Compute the covariance between two columns.

diag_concat_dfdiagonal_concat

Concat [DataFrame]s diagonally. Concat diagonally thereby combining different schemas.

hor_concat_dfhorizontal_concat

Concat [DataFrame]s horizontally. Concat horizontally and extend with null values if lengths don’t match

pearson_corr

Compute the pearson correlation between two columns.