Enum polars::lazy::dsl::Expr

pub enum Expr {
    Alias(Box<Expr>, Arc<String>),
    Column(Arc<String>),
    Literal(ScalarValue),
    BinaryExpr {
        left: Box<Expr>,
        op: Operator,
        right: Box<Expr>,
    },
    Not(Box<Expr>),
    IsNotNull(Box<Expr>),
    IsNull(Box<Expr>),
    Cast {
        expr: Box<Expr>,
        data_type: ArrowDataType,
    },
    Sort {
        expr: Box<Expr>,
        reverse: bool,
    },
    AggMin(Box<Expr>),
    AggMax(Box<Expr>),
    AggMedian(Box<Expr>),
    AggNUnique(Box<Expr>),
    AggFirst(Box<Expr>),
    AggLast(Box<Expr>),
    AggMean(Box<Expr>),
    AggQuantile {
        expr: Box<Expr>,
        quantile: f64,
    },
    AggSum(Box<Expr>),
    AggGroups(Box<Expr>),
    Ternary {
        predicate: Box<Expr>,
        truthy: Box<Expr>,
        falsy: Box<Expr>,
    },
}

Queries consists of multiple expressions.

Variants

Alias(Box<Expr>, Arc<String>)

Column(Arc<String>)

Literal(ScalarValue)

BinaryExpr

Fields of BinaryExpr

left: Box<Expr>op: Operatorright: Box<Expr>

Not(Box<Expr>)

IsNotNull(Box<Expr>)

IsNull(Box<Expr>)

Cast

Fields of Cast

expr: Box<Expr>data_type: ArrowDataType

Sort

Fields of Sort

expr: Box<Expr>reverse: bool

AggMin(Box<Expr>)

AggMax(Box<Expr>)

AggMedian(Box<Expr>)

AggNUnique(Box<Expr>)

AggFirst(Box<Expr>)

AggLast(Box<Expr>)

AggMean(Box<Expr>)

AggQuantile

Fields of AggQuantile

expr: Box<Expr>quantile: f64

AggSum(Box<Expr>)

AggGroups(Box<Expr>)

Ternary

Fields of Ternary

predicate: Box<Expr>truthy: Box<Expr>falsy: Box<Expr>

Methods

impl Expr

pub fn get_type(&self, schema: &Schema) -> Result<ArrowDataType>

Get DataType result of the expression. The schema is the input data.

impl Expr

pub fn eq(self, other: Expr) -> Expr

Compare Expr with other Expr on equality

pub fn neq(self, other: Expr) -> Expr

Compare Expr with other Expr on non-equality

pub fn lt(self, other: Expr) -> Expr

Check if Expr < Expr

pub fn gt(self, other: Expr) -> Expr

Check if Expr > Expr

pub fn gt_eq(self, other: Expr) -> Expr

Check if Expr >= Expr

pub fn lt_eq(self, other: Expr) -> Expr

Check if Expr <= Expr

pub fn not(self) -> Expr

Negate Expr

pub fn alias(self, name: &str) -> Expr

Rename Column.

pub fn is_null(self) -> Self

Run is_null operation on Expr.

pub fn is_not_null(self) -> Self

Run is_not_null operation on Expr.

pub fn agg_min(self) -> Self

Reduce groups to minimal value.

pub fn agg_max(self) -> Self

Reduce groups to maximum value.

pub fn agg_mean(self) -> Self

Reduce groups to the mean value.

pub fn agg_median(self) -> Self

Reduce groups to the median value.

pub fn agg_sum(self) -> Self

Reduce groups to the sum of all the values.

pub fn agg_n_unique(self) -> Self

Get the number of unique values in the groups.

pub fn agg_first(self) -> Self

Get the first value in the group.

pub fn agg_last(self) -> Self

Get the last value in the group.

pub fn agg_quantile(self, quantile: f64) -> Self

Compute the quantile per group.

pub fn agg_groups(self) -> Self

Get the group indexes of the group by operation.

pub fn cast(self, data_type: ArrowDataType) -> Self

Cast expression to another data type.

pub fn sort(self, reverse: bool) -> Self

Sort expression.