Struct polars::lazy::frame::LazyFrame

pub struct LazyFrame { /* fields omitted */ }

Lazy abstraction over an eager DataFrame. It really is an abstraction over a logical plan. The methods of this struct will incrementally modify a logical plan until output is requested (via collect)

Implementations

impl LazyFrame

pub fn with_projection_pushdown_optimization(self, toggle: bool) -> Self

Toggle projection pushdown optimization on or off.

pub fn with_predicate_pushdown_optimization(self, toggle: bool) -> Self

Toggle predicate pushdown optimization on or off.

pub fn with_type_coercion_optimization(self, toggle: bool) -> Self

Toggle type coercion optimization on or off.

pub fn with_simplify_expr_optimization(self, toggle: bool) -> Self

Toggle expression simplification optimization on or off

pub fn describe_plan(&self) -> String

Describe the logical plan.

pub fn describe_optimized_plan(&self) -> Result<String>

Describe the optimized logical plan.

pub fn sort(self, by_column: &str, reverse: bool) -> Self

Add a sort operation to the logical plan.

Example

use polars::prelude::*;

/// Sort DataFrame by 'sepal.width' column
fn example(df: DataFrame) -> LazyFrame {
      df.lazy()
        .sort("sepal.width", false)
}

pub fn reverse(self) -> Self

Reverse the DataFrame

Example

use polars::prelude::*;

fn example(df: DataFrame) -> LazyFrame {
      df.lazy()
        .reverse()
}

pub fn shift(self, periods: i32) -> Self

Shift the values by a given period and fill the parts that will be empty due to this operation with Nones.

See the method on Series for more info on the shift operation.

pub fn collect(self) -> Result<DataFrame>

Execute all the lazy operations and collect them into a DataFrame. Before execution the query is being optimized.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;

fn example(df: DataFrame) -> Result<DataFrame> {
      df.lazy()
        .groupby("foo")
        .agg(vec!(col("bar").agg_sum(),
                  col("ham").agg_mean().alias("avg_ham")))
        .collect()
}

pub fn filter(self, predicate: Expr) -> Self

Filter by some predicate expression.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;

fn example(df: DataFrame) -> LazyFrame {
      df.lazy()
        .filter(col("sepal.width").is_not_null())
        .select(&[col("sepal.width"), col("sepal.length")])
}

pub fn select<E: AsRef<[Expr]>>(self, exprs: E) -> Self

Select (and rename) columns from the query.

Columns can be selected with col; If you want to select all columns use col("*").

Example

use polars::prelude::*;
use polars::lazy::dsl::*;

fn example(df: DataFrame) -> LazyFrame {
      df.lazy()
        .select(&[col("foo"),
                  col("bar").alias("ham")])
}

pub fn groupby<'g, J, S: Selection<'g, J>>(self, by: S) -> LazyGroupBy

Group by and aggregate.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;

fn example(df: DataFrame) -> LazyFrame {
      df.lazy()
       .groupby("date")
       .agg(vec![
           col("rain").agg_min(),
           col("rain").agg_sum(),
           col("rain").agg_quantile(0.5).alias("median_rain"),
       ])
       .sort("date", false)
}

pub fn left_join(
self,
other: LazyFrame,
left_on: Expr,
right_on: Expr
) -> LazyFrame

Join query with other lazy query.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;
fn join_dataframes(ldf: LazyFrame, other: LazyFrame) -> LazyFrame {
        ldf
        .left_join(other, col("foo"), col("bar"))
}

pub fn outer_join(
self,
other: LazyFrame,
left_on: Expr,
right_on: Expr
) -> LazyFrame

Join query with other lazy query.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;
fn join_dataframes(ldf: LazyFrame, other: LazyFrame) -> LazyFrame {
        ldf
        .outer_join(other, col("foo"), col("bar"))
}

pub fn inner_join(
self,
other: LazyFrame,
left_on: Expr,
right_on: Expr
) -> LazyFrame

Join query with other lazy query.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;
fn join_dataframes(ldf: LazyFrame, other: LazyFrame) -> LazyFrame {
        ldf
        .inner_join(other, col("foo"), col("bar").cast(ArrowDataType::Utf8))
}

pub fn with_column(self, expr: Expr) -> LazyFrame

Add a column to a DataFrame

Example

use polars::prelude::*;
use polars::lazy::dsl::*;
fn add_column(df: DataFrame) -> LazyFrame {
    df.lazy()
        .with_column(
            when(col("sepal.length").lt(lit(5.0)))
            .then(lit(10))
            .otherwise(lit(1))
            .alias("new_column_name"),
            )
}

pub fn with_columns(self, exprs: Vec<Expr>) -> LazyFrame

Add multiple columns to a DataFrame.

Example

use polars::prelude::*;
use polars::lazy::dsl::*;
fn add_columns(df: DataFrame) -> LazyFrame {
    df.lazy()
        .with_columns(
            vec![lit(10).alias("foo"), lit(100).alias("bar")]
         )
}

pub fn max(self) -> LazyFrame

Aggregate all the columns as their maximum values.

pub fn min(self) -> LazyFrame

Aggregate all the columns as their minimum values.

pub fn sum(self) -> LazyFrame

Aggregate all the columns as their sum values.

pub fn mean(self) -> LazyFrame

Aggregate all the columns as their mean values.

pub fn median(self) -> LazyFrame

Aggregate all the columns as their median values.

pub fn quantile(self, quantile: f64) -> LazyFrame

Aggregate all the columns as their quantile values.

pub fn explode(self, column: &str) -> LazyFrame

Apply explode operation. See eager explode.

pub fn drop_duplicates(
self,
maintain_order: bool,
subset: Option<Vec<String>>
) -> LazyFrame

Drop duplicate rows. See eager.

pub fn drop_nulls(self, subset: Option<Vec<String>>) -> LazyFrame

Drop null rows. See eager.