polars

Module prelude

Modules

_csv_read_internal

aggregations

arity

binary

buffer

chunkedarray

Traits and utilities for temporal data.

cloud

Interface with cloud storage through the object_store crate.

compression

datatypes

Data types supported by Polars.

default_arrays

dt

expr

file

fill_null

float_sorted_arg_max

full

function_expr

gather

null

replace

row_encode

schema_inference

search_sorted

series

sort

string

strings

udf

utf8

zip

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

Structs

AnonymousScanArgs

AnonymousScanOptions

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

BatchedCsvReader

BatchedWriter

BinaryOffsetType

BinaryType

BooleanChunkedBuilder

BooleanType

Bounds

BoundsIter

CategoricalType

ChainedThen

Utility struct for the when-then-otherwise expression.

ChainedWhen

Utility struct for the when-then-otherwise expression.

ChunkId

ChunkedArray

ChunkedArray

CompatLevel

CrossJoinOptions

CsvParseOptions

CsvReadOptions

CsvReader

Create a new DataFrame by reading a csv file.

CsvWriter

Write a DataFrame to csv.

CsvWriterOptions

Options for writing CSV files.

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeArgs

Arguments used by datetime in order to produce an Expr of Datetime

DatetimeType

Dimension

Duration

DurationArgs

Arguments used by duration in order to produce an Expr of Duration

DurationType

DynamicGroupOptions

ExprNameNameSpace

Specialized expressions for modifying the name of existing expressions.

FalseT

Field

Characterizes the name and the DataType of a column.

FieldsMapper

Float32Type

Float64Type

GroupBy

Returned by a group_by operation on a DataFrame. This struct supports several aggregations.

GroupPositions

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsTypeIter

GroupsTypeParIter

InProcessQuery

Int8Type

Int16Type

Int32Type

Int64Type

JoinArgs

JoinBuilder

JoinOptions

LazyCsvReader

LazyFrame

Lazy abstraction over an eager DataFrame.

LazyGroupBy

Utility struct for lazy group_by operation.

ListBinaryChunkedBuilder

ListBooleanChunkedBuilder

ListNameSpace

Specialized expressions for Series of DataType::List.

ListPrimitiveChunkedBuilder

ListStringChunkedBuilder

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

NoNull

Just a wrapper structure which is useful for certain impl specializations.

Null

The literal Null

NullableIdxSize

OptFlags

Allowed optimizations.

OwnedBatchedCsvReader

PlSmallStr

String type that inlines small strings.

PrimitiveChunkedBuilder

RollingCovOptions

RollingGroupOptions

RollingQuantileParams

RollingVarParams

Scalar

ScanArgsAnonymous

SerializeOptions

Options to serialize logical types to CSV.

Series

Series

SortMultipleOptions

Sort options for multi-series sorting.

SortOptions

Options for single series sorting.

SpecialEq

Wrapper type that has special equality properties depending on the inner type specialization

StringType

StrptimeOptions

Then

Utility struct for the when-then-otherwise expression.

TimeType

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

UnionArgs

UnpivotArgsDSL

UnpivotArgsIR

Arguments for LazyFrame::unpivot function

UserDefinedFunction

Represents a user-defined function

When

Utility struct for the when-then-otherwise expression.

Window

Represents a window in time

Enums

AggExpr

Ambiguous

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

BooleanFunction

CategoricalOrdering

ClosedInterval

ClosedWindow

Column

A column within a DataFrame.

CommentPrefix

CsvEncoding

DataType

DslPlan

Excluded

Expr

Expressions that can be used in various contexts.

FillNullStrategy

FunctionExpr

GroupByMethod

GroupsIndicator

GroupsType

JoinCoalesce

JoinType

JoinTypeOptions

JoinTypeOptionsIR

JoinValidation

Label

LazySerde

LiteralValue

MaintainOrderJoin

NestedType

NonExistent

NullStrategy

NullValues

Operator

PolarsError

PowFunction

QuantileMethod

QuoteStyle

Quote style indicating when to insert quotes around a field.

ReshapeDimension

A dimension in a reshape.

RollingFnParams

SearchSortedSide

Selector

StartBy

TemporalFunction

TimeUnit

UniqueKeepStrategy

UnknownKind

WindowMapping

WindowType

Constants

IDX_DTYPE

NULL

URL_ENCODE_CHAR_SET

Statics

BOOLEAN_RE

EXTENSION_NAME

FLOAT_RE

FLOAT_RE_DECIMAL

INTEGER_RE

POLARS_TEMP_DIR_BASE_PATH

Traits

AnonymousScan

ArgAgg

Argmin/ Argmax

ArithmeticChunked

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsBinary

AsList

AsRefDataType

AsString

BinaryNameSpaceImpl

BinaryUdfOutputField

ChunkAgg

Aggregation operations.

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompareEq

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkCompareIneq

Compare Series and ChunkedArray’s using inequality operators (<, >=, etc.) and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/flatten a List or String Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

ChunkedBuilder

ChunkedCollectInferIterExt

ChunkedCollectIterExt

ChunkedSet

ColumnBinaryUdf

A wrapper trait for any binary closure Fn(Column, Column) -> PolarsResult<Column>

ColumnsUdf

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

CrossJoin

CrossJoinFilter

DataFrameJoinOps

DataFrameOps

DateMethods

DatetimeMethods

DurationMethods

FromData

FromDataBinary

FromDataUtf8

FunctionOutputField

GetAnyValue

IndexToUsize

InitHashMaps

InitHashMaps2

IntoColumn

Convert Self into a Column

IntoGroupsType

Used to create the tuples for a group_by operation.

IntoLazy

IntoMetadata

IntoScalar

IntoSeries

Used to convert a ChunkedArray, &dyn SeriesTrait and Series into a Series.

IntoVec

Convenience for x.into_iter().map(Into::into).collect() using an into_vec() function.

JoinDispatch

LazyFileListReader

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

LhsNumOps

ListBuilderTrait

ListFromIter

ListNameSpaceImpl

Literal

LogicalType

MetaDataExt

MinMaxHorizontal

NamedFrom

NamedFromOwned

NewChunkedArray

NumOpsDispatch

NumericNative

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

PolarsRound

PolarsTemporalGroupby

PolarsTruncate

PolarsUpsample

QuantileAggSeries

RenameAliasFn

SchemaExt

SchemaExtPl

SchemaNamesAndDtypes

SerReader

SerWriter

SeriesJoin

SeriesMethods

SeriesSealed

SeriesTrait

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

StaticArray

StringMethods

SumMeanHorizontal

TakeChunked

Gather by ChunkId

TakeChunkedHorPar

TemporalMethods

TimeMethods

UdfSchema

VarAggSeries

VecHash

Functions

_coalesce_full_join

_join_suffix_name

_set_check_length⚠

Meant for internal use. In very rare conditions this can be turned off.

all

Selects all columns. Shorthand for col("*").

all_horizontal

Create a new column with the bitwise-and of the elements in each row.

any_horizontal

Create a new column with the bitwise-or of the elements in each row.

apply_binary

Like map_binary, but used in a group_by-aggregation context.

apply_multiple

Apply a function/closure over the groups of multiple columns. This should only be used in a group_by aggregation.

avg

Find the mean of all the values in the column named name. Alias for mean.

binary_expr

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

cast

Casts the column given by Expr to a different type.

clip

Set values outside the given boundaries to the boundary value.

clip_max

Set values above the given maximum to the maximum value.

clip_min

Set values below the given minimum to the minimum value.

coalesce

Folds the expressions from left to right keeping the first non-null values.

coalesce_columns

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name.

concat

Concat multiple LazyFrames vertically.

concat_arr

Horizontally concatenate columns into a single array-type column.

concat_expr

concat_lf_horizontal

Concat LazyFrames horizontally.

concat_list

Concat lists entries.

convert_inner_type

Cast null arrays to inner type and ensure that all offsets remain correct

convert_to_unsigned_index

count_ones

count_rows

Read the number of rows without parsing columns useful for count(*) queries

count_rows_from_slice

Read the number of rows without parsing columns useful for count(*) queries

count_zeros

date_range

Create a DatetimeChunked from a given start and end date and a given interval.

datetime

Construct a column of Datetime from the provided DatetimeArgs.

datetime_to_timestamp_ms

datetime_to_timestamp_ns

datetime_to_timestamp_us

default_join_ids

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

Construct a column of Duration from the provided DurationArgs

ensure_duration_matches_dtype

ensure_is_constant_duration

ensure_matching_schema

expand_paths

Recursively traverses directories and expands globs if glob is true.

expand_paths_hive

Recursively traverses directories and expands globs if glob is true. Returns the expanded paths and the index at which to start parsing hive partitions from the path.

expanded_from_single_directory

Returns true if expanded_paths were expanded from a single directory

first

First column in a DataFrame.

fmt_group_by_column

fold_exprs

Accumulate over multiple columns horizontally / row wise.

get_glob_start_idx

Get the index of the first occurrence of a glob symbol.

get_reader_bytes

get_strftime_format

group_by_values

Different from group_by_windows, where define window buckets and search which values fit that pre-defined bucket.

group_by_windows

Window boundaries are created based on the given Window, which is defined by:

impl_duration

in_nanoseconds_window

index_cols

Select multiple columns by index.

indexes_to_usizes

infer_file_schema

Infer the schema of a CSV file by reading through the first n rows of the file, with max_read_rows controlling the maximum number of rows to read.

is_cloud_url

Check if the path is a cloud url.

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

is_positive_idx_uncertain

May give false negatives because it ignores the null values.

is_positive_idx_uncertain_col

May give false negatives because it ignores the null values.

last

Last column in a DataFrame.

leading_ones

leading_zeros

len

Return the number of rows in the context.

lit

Create a Literal Expression from L. A literal expression behaves like a column that contains a single distinct value.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

map_list_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

materialize_projection

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

merge_dtypes

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

negate

negate_bitwise

new_int_range

new_linear_space_f32

new_linear_space_f64

not

Negates a boolean column.

nth

Nth column in a DataFrame.

private_left_join_multiple_keys

quantile

Find a specific quantile of all the values in the column named name.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value.

replace_date

Replace specific time component of a DateChunked with a specified value.

replace_datetime

Replace specific time component of a DatetimeChunked with a specified value.

resolve_homedir

Replaces a “~” in the Path with the home directory.

sum

Sum all the values in the column named name. Shorthand for col(name).sum().

ternary_expr

time_range

Create a TimeChunked from a given start and end date and a given interval.

trailing_ones

trailing_zeros

when

Start a when-then-otherwise expression.

Type Aliases

AllowedOptimizations

AllowedOptimizations

ArrayRef

ArrowSchema

An ordered sequence of Fields

BinaryChunked

BinaryChunkedBuilder

BinaryOffsetChunked

BooleanChunked

BorrowIdxItem

ChunkJoinOptIds

DateChunked

DatetimeChunked

DurationChunked

FieldRef

FillNullLimit

Float32Chunked

Float64Chunked

GetOutput

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxCa

IdxItem

IdxSize

IdxType

InnerJoinIds

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

LargeBinaryArray

LargeListArray

LargeStringArray

LeftJoinIds

ListChunked

OpaqueColumnUdf

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PlRandomState

PolarsResult

QuantileInterpolOptionsDeprecated

Schema

SchemaRef

StringChunked

StringChunkedBuilder

TimeChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked