polars::prelude
Module _internal

Available on crate feature polars-io only.

Enums

PrefilterMaskSetting

Functions

calc_prefilter_cost

collect_statistics_with_live_columns

Collect the statistics in a row-group

ensure_matching_dtypes_if_found

Utility to ensure the dtype of the column in current_schema matches the dtype in schema if that column exists in schema.

read_this_row_group

to_deserializer