polars::prelude::chunkedarray

Trait DurationMethods

pub trait DurationMethods {

    // Required methods
    fn days(&self) -> ChunkedArray<Int64Type>;
    fn days_fractional(&self) -> ChunkedArray<Float64Type>;
    fn hours(&self) -> ChunkedArray<Int64Type>;
    fn hours_fractional(&self) -> ChunkedArray<Float64Type>;
    fn minutes(&self) -> ChunkedArray<Int64Type>;
    fn minutes_fractional(&self) -> ChunkedArray<Float64Type>;
    fn seconds(&self) -> ChunkedArray<Int64Type>;
    fn seconds_fractional(&self) -> ChunkedArray<Float64Type>;
    fn milliseconds(&self) -> ChunkedArray<Int64Type>;
    fn milliseconds_fractional(&self) -> ChunkedArray<Float64Type>;
    fn microseconds(&self) -> ChunkedArray<Int64Type>;
    fn microseconds_fractional(&self) -> ChunkedArray<Float64Type>;
    fn nanoseconds(&self) -> ChunkedArray<Int64Type>;
    fn nanoseconds_fractional(&self) -> ChunkedArray<Float64Type>;

}

Available on crate feature temporal only.

Required Methods

fn days(&self) -> ChunkedArray<Int64Type>

Extract the days from a Duration

fn days_fractional(&self) -> ChunkedArray<Float64Type>

Extract the days from a Duration as a fractional value

fn hours(&self) -> ChunkedArray<Int64Type>

Extract the hours from a Duration

fn hours_fractional(&self) -> ChunkedArray<Float64Type>

Extract the hours from a Duration as a fractional value

fn minutes(&self) -> ChunkedArray<Int64Type>

Extract the minutes from a Duration

fn minutes_fractional(&self) -> ChunkedArray<Float64Type>

Extract the minutes from a Duration as a fractional value

fn seconds(&self) -> ChunkedArray<Int64Type>

Extract the seconds from a Duration

fn seconds_fractional(&self) -> ChunkedArray<Float64Type>

Extract the seconds from a Duration as a fractional value

fn milliseconds(&self) -> ChunkedArray<Int64Type>

Extract the milliseconds from a Duration

fn milliseconds_fractional(&self) -> ChunkedArray<Float64Type>

Extract the milliseconds from a Duration as a fractional value

fn microseconds(&self) -> ChunkedArray<Int64Type>

Extract the microseconds from a Duration

fn microseconds_fractional(&self) -> ChunkedArray<Float64Type>

Extract the microseconds from a Duration as a fractional value

fn nanoseconds(&self) -> ChunkedArray<Int64Type>

Extract the nanoseconds from a Duration

fn nanoseconds_fractional(&self) -> ChunkedArray<Float64Type>

Extract the nanoseconds from a Duration as a fractional value