polars::prelude::chunkedarray
Trait SeriesOpsTime

pub trait SeriesOpsTime: AsSeries {

    // Provided methods
    fn rolling_mean_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_mean(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_sum_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_sum(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_quantile_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_quantile(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_min_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_min(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_max_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_max(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_var_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_var(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_std_by(
        &self,
        by: &Series,
        options: RollingOptionsDynamicWindow,
    ) -> Result<Series, PolarsError> { ... }
    fn rolling_std(
        &self,
        options: RollingOptionsFixedWindow,
    ) -> Result<Series, PolarsError> { ... }

}

Available on crate feature temporal only.

Provided Methods

fn rolling_mean_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling mean to a Series based on another Series.

fn rolling_mean( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling mean to a Series.

See: [RollingAgg::rolling_mean]

fn rolling_sum_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling sum to a Series based on another Series.

fn rolling_sum( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling sum to a Series.

fn rolling_quantile_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling quantile to a Series based on another Series.

fn rolling_quantile( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling quantile to a Series.

fn rolling_min_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling min to a Series based on another Series.

fn rolling_min( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling min to a Series.

fn rolling_max_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling max to a Series based on another Series.

fn rolling_max( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling max to a Series.

fn rolling_var_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling variance to a Series based on another Series.

fn rolling_var( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling variance to a Series.

fn rolling_std_by( &self, by: &Series, options: RollingOptionsDynamicWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window_by only.

Apply a rolling std_dev to a Series based on another Series.

fn rolling_std( &self, options: RollingOptionsFixedWindow, ) -> Result<Series, PolarsError>

Available on crate feature rolling_window only.

Apply a rolling std_dev to a Series.