polars::prelude
Module cloud

Available on crate feature polars-io only.

Interface with cloud storage through the object_store crate.

Modules

credential_providercloud

options

Structs

CloudLocation

A location on cloud storage, may have wildcards.

CloudOptions

Options to connect to various cloud providers.

CloudWriter

Adaptor which wraps the interface of ObjectStore::BufWriter exposing a synchronous interface which implements std::io::Write.

PolarsObjectStore

Polars wrapper around [ObjectStore] functionality. This struct is cheaply cloneable.

Enums

CloudType

Functions

build_object_store

Build an [ObjectStore] based on the URL and passed in url. Return the cloud location and an implementation of the object store.

glob

List files with a prefix derived from the pattern.

object_path_from_str

Construct an object_store Path from a string without any encoding/decoding.

Type Aliases

ObjectStorePath