polars::prelude::cloud

Struct BlockingCloudWriter

pub struct BlockingCloudWriter { /* private fields */ }

Available on crate feature polars-io only.

Adaptor which wraps the interface of ObjectStore::BufWriter exposing a synchronous interface which implements std::io::Write.

This allows it to be used in sync code which would otherwise write to a simple File or byte stream, such as with polars::prelude::CsvWriter.

Implementations

impl BlockingCloudWriter

pub fn new_with_object_store( object_store: Arc<dyn ObjectStore>, path: Path, ) -> Result<BlockingCloudWriter, PolarsError>

Construct a new BlockingCloudWriter, re-using the given object_store

Creates a new (current-thread) Tokio runtime which bridges the sync writing process with the async ObjectStore multipart uploading. TODO: Naming?

pub async fn new( uri: &str, cloud_options: Option<&CloudOptions>, ) -> Result<BlockingCloudWriter, PolarsError>

Constructs a new BlockingCloudWriter from a path and an optional set of CloudOptions.

Wrapper around BlockingCloudWriter::new_with_object_store that is useful if you only have a single write task. TODO: Naming?

pub fn try_into_inner(self) -> Result<BufWriter, Error>

Returns the underlying [object_store::buffered::BufWriter]

pub fn close(&mut self) -> Result<(), Error>

Closes the writer, or returns the existing error if it exists. After this function is called the writer is guaranteed to be in an error state.