polars::prelude

Module compression

Available on crate feature polars-io only.

Enums

SupportedCompression

Represents the compression algorithms that we have decoders for

Functions

maybe_decompress_bytes

Decompress bytes if compression is detected, otherwise simply return it. An out vec must be given for ownership of the decompressed data.