polars::prelude::datatypes

Struct CategoricalMapping

pub struct CategoricalMapping { /* private fields */ }

Implementations

impl CategoricalMapping

pub fn new(max_categories: usize) -> CategoricalMapping

pub fn with_hasher( max_categories: usize, hasher: SeedableRandomState, ) -> CategoricalMapping

pub fn hasher(&self) -> &SeedableRandomState

pub fn set_max_categories(&mut self, max_categories: usize)

pub fn get_cat(&self, s: &str) -> Option<u32>

Try to convert a string to a categorical id, but don’t insert it if it is missing.

pub fn get_cat_with_hash(&self, s: &str, hash: u64) -> Option<u32>

Same as get_cat, but with the hash pre-computed.

pub fn insert_cat(&self, s: &str) -> Result<u32, PolarsError>

Convert a string to a categorical id.

pub fn insert_cat_with_hash( &self, s: &str, hash: u64, ) -> Result<u32, PolarsError>

Same as to_cat, but with the hash pre-computed.

pub fn cat_to_str(&self, cat: u32) -> Option<&str>

Try to convert a categorical id to its corresponding string, returning None if the string is not in the data structure.

pub unsafe fn cat_to_str_unchecked(&self, cat: u32) -> &str

Get the string corresponding to a categorical id.

Safety

The categorical id must have been returned from to_cat, and you must have synchronized with the call which inserted it.

pub fn num_cats_upper_bound(&self) -> usize

Returns an upper bound such that all strings inserted into the CategoricalMapping have a categorical id less than it. Note that due to parallel inserts which you have not synchronized with, there may be gaps when using from_cat.

pub fn len(&mut self) -> usize

Returns the number of categories in this mapping.

pub fn is_empty(&mut self) -> bool

pub fn to_arrow(&self, as_views: bool) -> Box<dyn Array>