Type Alias polars::prelude::datatypes::Int64Chunked

pub type Int64Chunked = ChunkedArray<Int64Type>;

Aliased Type

struct Int64Chunked { /* private fields */ }

Implementations

impl<T> ChunkedArray<T>where T: PolarsDataType<Structure = Flat>, <T as PolarsDataType>::Physical<'a>: for<'a> TotalOrd,

pub fn append(&mut self, other: &ChunkedArray<T>)

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn apply_values_generic<'a, U, K, F>(&'a self, op: F) -> ChunkedArray<U>where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> K, <U as PolarsDataType>::Array: ArrayFromIter<K>,

pub fn try_apply_values_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> Result<K, E>, <U as PolarsDataType>::Array: ArrayFromIter<K>,

Applies a function to all elements, regardless of whether they are null or not, after which the null mask is copied from the original array.

pub fn apply_nonnull_values_generic<'a, U, K, F>( &'a self, dtype: DataType, op: F ) -> ChunkedArray<U>where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> K, <U as PolarsDataType>::Array: ArrayFromIterDtype<K> + ArrayFromIterDtype<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn try_apply_nonnull_values_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> Result<K, E>, <U as PolarsDataType>::Array: ArrayFromIter<K> + ArrayFromIter<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn apply_generic<'a, U, K, F>(&'a self, op: F) -> ChunkedArray<U>where U: PolarsDataType, F: FnMut(Option<<T as PolarsDataType>::Physical<'a>>) -> Option<K>, <U as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

pub fn try_apply_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>where U: PolarsDataType, F: FnMut(Option<<T as PolarsDataType>::Physical<'a>>) -> Result<Option<K>, E> + Copy, <U as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn cast_and_apply_in_place<F, S>(&self, f: F) -> ChunkedArray<S>where F: Fn(<S as PolarsNumericType>::Native) -> <S as PolarsNumericType>::Native + Copy, S: PolarsNumericType,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

pub fn apply_in_place<F>(self, f: F) -> ChunkedArray<T>where F: Fn(<T as PolarsNumericType>::Native) -> <T as PolarsNumericType>::Native + Copy,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn apply_mut<F>(&mut self, f: F)where F: Fn(<T as PolarsNumericType>::Native) -> <T as PolarsNumericType>::Native + Copy,

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn len(&self) -> usize

Get the length of the ChunkedArray

pub fn null_count(&self) -> usize

Count the null values.

pub fn is_empty(&self) -> bool

Check if ChunkedArray is empty.

pub fn rechunk(&self) -> ChunkedArray<T>

pub fn slice(&self, offset: i64, length: usize) -> ChunkedArray<T>

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

pub fn limit(&self, num_elements: usize) -> ChunkedArray<T>where ChunkedArray<T>: Sized,

Take a view of top n elements

pub fn head(&self, length: Option<usize>) -> ChunkedArray<T>where ChunkedArray<T>: Sized,

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> ChunkedArray<T>where ChunkedArray<T>: Sized,

Get the tail of the ChunkedArray

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn extend(&mut self, other: &ChunkedArray<T>)

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extend appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn for_each<'a, F>(&'a self, op: F)where F: FnMut(Option<<T as PolarsDataType>::Physical<'a>>),

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn is_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the null values.

pub fn is_not_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the valid values.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn tile(&self, n: usize) -> ChunkedArray<T>

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn not_equal_and_validity( &self, rhs: &ChunkedArray<T> ) -> ChunkedArray<BooleanType>

impl<T> ChunkedArray<T>where T: PolarsFloatType, <T as PolarsNumericType>::Native: Float,

pub fn is_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_not_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_finite(&self) -> ChunkedArray<BooleanType>

pub fn is_infinite(&self) -> ChunkedArray<BooleanType>

pub fn none_to_nan(&self) -> ChunkedArray<T>

Convert missing values to NaN values.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn with_chunk<A>(name: &str, arr: A) -> ChunkedArray<T>where A: Array, T: PolarsDataType<Array = A>,

pub fn from_chunk_iter<I>(name: &str, iter: I) -> ChunkedArray<T>where I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn from_chunk_iter_like<I>(ca: &ChunkedArray<T>, iter: I) -> ChunkedArray<T>where I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn try_from_chunk_iter<I, A, E>( name: &str, iter: I ) -> Result<ChunkedArray<T>, E>where I: IntoIterator<Item = Result<A, E>>, T: PolarsDataType<Array = A>, A: Array,

pub unsafe fn from_chunks( name: &str, chunks: Vec<Box<dyn Array, Global>, Global> ) -> ChunkedArray<T>

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn with_chunks( &self, chunks: Vec<Box<dyn Array, Global>, Global> ) -> ChunkedArray<T>

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn from_chunks_and_dtype( name: &str, chunks: Vec<Box<dyn Array, Global>, Global>, dtype: DataType ) -> ChunkedArray<T>

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn from_vec( name: &str, v: Vec<<T as PolarsNumericType>::Native, Global> ) -> ChunkedArray<T>

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn from_vec_validity( name: &str, values: Vec<<T as PolarsNumericType>::Native, Global>, buffer: Option<Bitmap> ) -> ChunkedArray<T>

Create a new ChunkedArray from a Vec and a validity mask.

pub unsafe fn mmap_slice( name: &str, values: &[<T as PolarsNumericType>::Native] ) -> ChunkedArray<T>

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl ChunkedArray<Int64Type>

pub fn into_datetime( self, timeunit: TimeUnit, tz: Option<String> ) -> Logical<DatetimeType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_duration( self, timeunit: TimeUnit ) -> Logical<DurationType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_time(self) -> Logical<TimeType, Int64Type>

impl<T> ChunkedArray<T>where T: PolarsNumericType, Standard: Distribution<<T as PolarsNumericType>::Native>,

pub fn init_rand( size: usize, null_density: f32, seed: Option<u64> ) -> ChunkedArray<T>

impl<T> ChunkedArray<T>where T: PolarsDataType, ChunkedArray<T>: ChunkTake<ChunkedArray<UInt32Type>>,

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> Result<ChunkedArray<T>, PolarsError>

Sample n datapoints from this ChunkedArray.

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> Result<ChunkedArray<T>, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T>where T: PolarsNumericType, <T as PolarsNumericType>::Native: Float,

pub fn rand_normal( name: &str, length: usize, mean: f64, std_dev: f64 ) -> Result<ChunkedArray<T>, PolarsError>

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> ChunkedArray<T>

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform( name: &str, length: usize, low: f64, high: f64 ) -> ChunkedArray<T>

Create ChunkedArray with samples from a Uniform distribution.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn to_vec(&self) -> Vec<Option<<T as PolarsNumericType>::Native>, Global>

Convert to a Vec of Option<T::Native>.

pub fn to_vec_null_aware( &self ) -> Either<Vec<<T as PolarsNumericType>::Native, Global>, Vec<Option<<T as PolarsNumericType>::Native>, Global>>

Convert to a Vec but don’t return Option<T::Native> if there are no null values

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn unset_fast_explode_list(&mut self)

pub fn get_flags(&self) -> Settings

pub fn is_sorted_flag(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

Set the ‘sorted’ bit meta info.

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn last_non_null(&self) -> Option<usize>

Get the index of the last non null value in this ChunkedArray.

pub fn iter_validities( &self ) -> Map<Iter<'_, Box<dyn Array, Global>>, fn(_: &Box<dyn Array, Global>) -> Option<&Bitmap>>

Get the buffer of bits representing null values

pub fn has_validity(&self) -> bool

Return if any the chunks in this ChunkedArray have a validity bitmap. no bitmap means no null values.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn clear(&self) -> ChunkedArray<T>

pub fn unpack_series_matching_type( &self, series: &Series ) -> Result<&ChunkedArray<T>, PolarsError>

Series to ChunkedArray<T>

pub fn chunk_id( &self ) -> Map<Iter<'_, Box<dyn Array, Global>>, fn(_: &Box<dyn Array, Global>) -> usize>

Unique id representing the number of chunks

pub fn chunks(&self) -> &Vec<Box<dyn Array, Global>, Global>

A reference to the chunks

pub unsafe fn chunks_mut(&mut self) -> &mut Vec<Box<dyn Array, Global>, Global>

A mutable reference to the chunks

Safety

The caller must ensure to not change the DataType or length of any of the chunks. And the null_count remains correct.

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

pub fn with_name(self, name: &str) -> ChunkedArray<T>

Return this ChunkedArray with a new name.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn get(&self, idx: usize) -> Option<<T as PolarsDataType>::Physical<'_>>

pub unsafe fn get_unchecked( &self, idx: usize ) -> Option<<T as PolarsDataType>::Physical<'_>>

Safety

It is the callers responsibility that the idx < self.len().

pub unsafe fn value_unchecked( &self, idx: usize ) -> <T as PolarsDataType>::Physical<'_>

Safety

It is the callers responsibility that the idx < self.len().

pub fn last(&self) -> Option<<T as PolarsDataType>::Physical<'_>>

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn layout(&self) -> ChunkedArrayLayout<'_, T>

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn cont_slice( &self ) -> Result<&[<T as PolarsNumericType>::Native], PolarsError>

Contiguous slice

pub fn data_views( &self ) -> impl Iterator<Item = &[<T as PolarsNumericType>::Native]> + DoubleEndedIterator

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter( &self ) -> impl Iterator<Item = <T as PolarsNumericType>::Native> + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl<T> ChunkedArray<T>where T: PolarsIntegerType, <T as PolarsNumericType>::Native: ToPrimitive + FromPrimitive + Debug,

pub fn group_tuples_perfect( &self, max: usize, multithreaded: bool, group_capacity: usize ) -> GroupsProxy

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn new_vec( name: &str, v: Vec<<T as PolarsNumericType>::Native, Global> ) -> ChunkedArray<T>

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T> ChunkedArray<T>where T: PolarsNumericType, ChunkedArray<T>: IntoSeries,

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N>(&self, lhs: N) -> ChunkedArray<T>where N: Num + NumCast,

Apply lhs - self

pub fn lhs_div<N>(&self, lhs: N) -> ChunkedArray<T>where N: Num + NumCast,

Apply lhs / self

pub fn lhs_rem<N>(&self, lhs: N) -> ChunkedArray<T>where N: Num + NumCast,

Apply lhs % self