Struct polars::prelude::dt::DateLikeNameSpace

pub struct DateLikeNameSpace(_);

Available on crate feature temporal only.

Specialized expressions for Series with dates/datetimes.

Implementations

impl DateLikeNameSpace

pub fn strftime(self, fmt: &str) -> Expr

Format Date/datetime with a formatting rule See chrono strftime/strptime.

pub fn cast_time_unit(self, tu: TimeUnit) -> Expr

Change the underlying TimeUnit. And update the data accordingly.

pub fn with_time_unit(self, tu: TimeUnit) -> Expr

Change the underlying TimeUnit of the Series. This does not modify the data.

pub fn convert_time_zone(self, time_zone: String) -> Expr

Available on crate feature timezones only.

Change the underlying TimeZone of the Series. This does not modify the data.

pub fn tz_localize(self, tz: String) -> Expr

๐Deprecated: use replace_time_zone

Available on crate feature timezones only.

Localize tz-naive Datetime Series to tz-aware Datetime Series.

pub fn year(self) -> Expr

Get the year of a Date/Datetime

pub fn is_leap_year(self) -> Expr

pub fn iso_year(self) -> Expr

Get the iso-year of a Date/Datetime. This may not correspond with a calendar year.

pub fn month(self) -> Expr

Get the month of a Date/Datetime

pub fn quarter(self) -> Expr

Extract quarter from underlying NaiveDateTime representation. Quarters range from 1 to 4.

pub fn week(self) -> Expr

Extract the week from the underlying Date representation. Can be performed on Date and Datetime Returns the ISO week number starting from 1. The return value ranges from 1 to 53. (The last week of year differs by years.)

pub fn weekday(self) -> Expr

Extract the week day from the underlying Date representation. Can be performed on Date and Datetime. Returns the weekday number where monday = 0 and sunday = 6

pub fn day(self) -> Expr

Get the month of a Date/Datetime

pub fn ordinal_day(self) -> Expr

Get the ordinal_day of a Date/Datetime

pub fn hour(self) -> Expr

Get the hour of a Datetime/Time64

pub fn minute(self) -> Expr

Get the minute of a Datetime/Time64

pub fn second(self) -> Expr

Get the second of a Datetime/Time64

pub fn millisecond(self) -> Expr

Get the millisecond of a Time64 (scaled from nanosecs)

pub fn microsecond(self) -> Expr

Get the microsecond of a Time64 (scaled from nanosecs)

pub fn nanosecond(self) -> Expr

Get the nanosecond part of a Time64

pub fn timestamp(self, tu: TimeUnit) -> Expr

pub fn truncate<S>(self, every: S, offset: S) -> Exprwhere S: AsRef<str>,

pub fn round<S>(self, every: S, offset: S) -> Exprwhere S: AsRef<str>,

pub fn replace_time_zone(self, time_zone: Option<String>) -> Expr

Available on crate feature timezones only.

pub fn combine(self, time: Expr, tu: TimeUnit) -> Expr