polars::prelude

Enum DataType

pub enum DataType {

    Boolean,
    UInt8,
    UInt16,
    UInt32,
    UInt64,
    Int8,
    Int16,
    Int32,
    Int64,
    Int128,
    Float32,
    Float64,
    Decimal(Option<usize>, Option<usize>),
    String,
    Binary,
    BinaryOffset,
    Date,
    Datetime(TimeUnit, Option<TimeZone>),
    Duration(TimeUnit),
    Time,
    Array(Box<DataType>, usize),
    List(Box<DataType>),
    Object(&'static str),
    Null,
    Categorical(Arc<Categories>, Arc<CategoricalMapping>),
    Enum(Arc<FrozenCategories>, Arc<CategoricalMapping>),
    Struct(Vec<Field>),
    Unknown(UnknownKind),

}

Variants

Boolean

UInt8

UInt16

UInt32

UInt64

Int8

Int16

Int32

Int64

Int128

Float32

Float64

Decimal(Option<usize>, Option<usize>)

Fixed point decimal type optional precision and non-negative scale. This is backed by a signed 128-bit integer which allows for up to 38 significant digits. Meaning max precision is 38.

String

String data

Binary

BinaryOffset

Date

A 32-bit date representing the elapsed time since UNIX epoch (1970-01-01) in days (32 bits).

Datetime(TimeUnit, Option<TimeZone>)

A 64-bit date representing the elapsed time since UNIX epoch (1970-01-01) in the given timeunit (64 bits).

Duration(TimeUnit)

64-bit integer representing difference between times in milliseconds or nanoseconds

Time

A 64-bit time representing the elapsed time since midnight in nanoseconds

Array(Box<DataType>, usize)

A nested list with a fixed size in each row

List(Box<DataType>)

A nested list with a variable size in each row

Object(&'static str)

A generic type that can be used in a Series &’static str can be used to determine/set inner type

Null

Categorical(Arc<Categories>, Arc<CategoricalMapping>)

Enum(Arc<FrozenCategories>, Arc<CategoricalMapping>)

Struct(Vec<Field>)

Unknown(UnknownKind)

Implementations

impl DataType

pub const IDX_DTYPE: DataType = { { DataType::UInt32 } }

pub fn value_within_range(&self, other: AnyValue<'_>) -> bool

pub fn is_known(&self) -> bool

Check if the whole dtype is known.

pub fn materialize_unknown( self, allow_unknown: bool, ) -> Result<DataType, PolarsError>

Materialize this datatype if it is unknown. All other datatypes are left unchanged.

pub fn get_shape(&self) -> Option<Vec<usize>>

Get the full shape of a multidimensional array.

pub fn inner_dtype(&self) -> Option<&DataType>

Get the inner data type of a nested type.

pub fn leaf_dtype(&self) -> &DataType

Get the absolute inner data type of a nested type.

pub fn array_leaf_dtype(&self) -> Option<&DataType>

Get the inner data type of a multidimensional array.

pub fn cast_leaf(&self, to: DataType) -> DataType

Cast the leaf types of Lists/Arrays and keep the nesting.

pub fn can_cast_to(&self, to: &DataType) -> Option<bool>

Return whether the cast to to makes sense.

If it None, we are not sure.

pub fn implode(self) -> DataType

pub fn to_physical(&self) -> DataType

Convert to the physical data type

pub fn is_supported_list_arithmetic_input(&self) -> bool

pub fn is_logical(&self) -> bool

Check if this DataType is a logical type

pub fn is_temporal(&self) -> bool

Check if this DataType is a temporal type

pub fn is_primitive(&self) -> bool

Check if datatype is a primitive type. By that we mean that it is not a nested or logical type.

pub fn is_primitive_numeric(&self) -> bool

Check if this DataType is a primitive numeric type (excludes Decimal).

pub fn is_bool(&self) -> bool

Check if this DataType is a boolean.

pub fn is_list(&self) -> bool

Check if this DataType is a list.

pub fn is_array(&self) -> bool

Check if this DataType is an array.

pub fn is_nested(&self) -> bool

pub fn is_struct(&self) -> bool

Check if this DataType is a struct

pub fn is_binary(&self) -> bool

pub fn is_date(&self) -> bool

pub fn is_datetime(&self) -> bool

pub fn is_duration(&self) -> bool

pub fn is_object(&self) -> bool

pub fn is_null(&self) -> bool

pub fn contains_views(&self) -> bool

pub fn contains_categoricals(&self) -> bool

pub fn contains_objects(&self) -> bool

pub fn contains_list_recursive(&self) -> bool

pub fn contains_unknown(&self) -> bool

pub fn is_ord(&self) -> bool

Check if type is sortable

pub fn is_decimal(&self) -> bool

Check if this DataType is a Decimal type (of any scale/precision).

pub fn is_float(&self) -> bool

Check if this DataType is a basic floating point type (excludes Decimal). Note, this also includes Unknown(UnknownKind::Float).

pub fn is_integer(&self) -> bool

Check if this DataType is an integer. Note, this also includes Unknown(UnknownKind::Int).

pub fn is_signed_integer(&self) -> bool

pub fn is_unsigned_integer(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_categorical(&self) -> bool

pub fn is_enum(&self) -> bool

pub fn to_arrow_field( &self, name: PlSmallStr, compat_level: CompatLevel, ) -> Field

Convert to an Arrow Field.

pub fn max(&self) -> Result<Scalar, PolarsError>

Try to get the maximum value for this datatype.

pub fn min(&self) -> Result<Scalar, PolarsError>

Try to get the minimum value for this datatype.

pub fn to_arrow(&self, compat_level: CompatLevel) -> ArrowDataType

Convert to an Arrow data type.

pub fn try_to_arrow( &self, compat_level: CompatLevel, ) -> Result<ArrowDataType, PolarsError>

pub fn is_nested_null(&self) -> bool

pub fn matches_schema_type( &self, schema_type: &DataType, ) -> Result<bool, PolarsError>

Answers if this type matches the given type of a schema.

Allows (nested) Null types in this type to match any type in the schema, but not vice versa. In such a case Ok(true) is returned, because a cast is necessary. If no cast is necessary Ok(false) is returned, and an error is returned if the types are incompatible.

pub fn is_unknown(&self) -> bool

pub fn nesting_level(&self) -> usize

pub fn cat_physical(&self) -> Result<CategoricalPhysical, PolarsError>

If this dtype is a Categorical or Enum, returns the physical backing type.

pub fn cat_mapping(&self) -> Result<&Arc<CategoricalMapping>, PolarsError>

If this dtype is a Categorical or Enum, returns the underlying mapping.

pub fn from_categories(cats: Arc<Categories>) -> DataType

pub fn from_frozen_categories(fcats: Arc<FrozenCategories>) -> DataType

pub fn is_numeric(&self) -> bool

impl DataType

pub fn boxed(self) -> Box<DataType>

pub fn from_arrow_field(field: &Field) -> DataType

pub fn from_arrow_dtype(dt: &ArrowDataType) -> DataType

pub fn from_arrow( dt: &ArrowDataType, md: Option<&BTreeMap<PlSmallStr, PlSmallStr>>, ) -> DataType