polars::prelude

Enum DslPlan

pub enum DslPlan {

    Filter {
        input: Arc<DslPlan>,
        predicate: Expr,
    },
    Cache {
        input: Arc<DslPlan>,
    },
    Scan {
        sources: ScanSources,
        unified_scan_args: Box<UnifiedScanArgs>,
        scan_type: Box<FileScanDsl>,
        cached_ir: Arc<Mutex<Option<IR>>>,
    },
    DataFrameScan {
        df: Arc<DataFrame>,
        schema: Arc<Schema<DataType>>,
    },
    Select {
        expr: Vec<Expr>,
        input: Arc<DslPlan>,
        options: ProjectionOptions,
    },
    GroupBy {
        input: Arc<DslPlan>,
        keys: Vec<Expr>,
        aggs: Vec<Expr>,
        maintain_order: bool,
        options: Arc<GroupbyOptions>,
        apply: Option<(Arc<dyn DataFrameUdf>, Arc<Schema<DataType>>)>,
    },
    Join {
        input_left: Arc<DslPlan>,
        input_right: Arc<DslPlan>,
        left_on: Vec<Expr>,
        right_on: Vec<Expr>,
        predicates: Vec<Expr>,
        options: Arc<JoinOptions>,
    },
    HStack {
        input: Arc<DslPlan>,
        exprs: Vec<Expr>,
        options: ProjectionOptions,
    },
    MatchToSchema {
        input: Arc<DslPlan>,
        match_schema: Arc<Schema<DataType>>,
        per_column: Arc<[MatchToSchemaPerColumn]>,
        extra_columns: ExtraColumnsPolicy,
    },
    Distinct {
        input: Arc<DslPlan>,
        options: DistinctOptionsDSL,
    },
    Sort {
        input: Arc<DslPlan>,
        by_column: Vec<Expr>,
        slice: Option<(i64, usize)>,
        sort_options: SortMultipleOptions,
    },
    Slice {
        input: Arc<DslPlan>,
        offset: i64,
        len: u32,
    },
    MapFunction {
        input: Arc<DslPlan>,
        function: DslFunction,
    },
    Union {
        inputs: Vec<DslPlan>,
        args: UnionArgs,
    },
    HConcat {
        inputs: Vec<DslPlan>,
        options: HConcatOptions,
    },
    ExtContext {
        input: Arc<DslPlan>,
        contexts: Vec<DslPlan>,
    },
    Sink {
        input: Arc<DslPlan>,
        payload: SinkType,
    },
    SinkMultiple {
        inputs: Vec<DslPlan>,
    },
    IR {
        dsl: Arc<DslPlan>,
        version: u32,
        node: Option<Node>,
    },

}

Available on crate feature lazy only.

Variants

Filter

Filter on a boolean mask

Fields

input: Arc<DslPlan>

predicate: Expr

Cache

Cache the input at this point in the LP

Fields

input: Arc<DslPlan>

Scan

Fields

sources: ScanSources

unified_scan_args: Box<UnifiedScanArgs>

scan_type: Box<FileScanDsl>

cached_ir: Arc<Mutex<Option<IR>>>

Local use cases often repeatedly collect the same LazyFrame (e.g. in interactive notebook use-cases), so we cache the IR conversion here, as the path expansion can be quite slow (especially for cloud paths). We don’t have the arena, as this is always a source node.

DataFrameScan

In memory DataFrame

Fields

df: Arc<DataFrame>

schema: Arc<Schema<DataType>>

Select

Polars’ select operation, this can mean projection, but also full data access.

Fields

expr: Vec<Expr>

input: Arc<DslPlan>

options: ProjectionOptions

GroupBy

Groupby aggregation

Fields

input: Arc<DslPlan>

keys: Vec<Expr>

aggs: Vec<Expr>

maintain_order: bool

options: Arc<GroupbyOptions>

apply: Option<(Arc<dyn DataFrameUdf>, Arc<Schema<DataType>>)>

Join

Join operation

Fields

input_left: Arc<DslPlan>

input_right: Arc<DslPlan>

left_on: Vec<Expr>

right_on: Vec<Expr>

predicates: Vec<Expr>

options: Arc<JoinOptions>

HStack

Adding columns to the table without a Join

Fields

input: Arc<DslPlan>

exprs: Vec<Expr>

options: ProjectionOptions

MatchToSchema

Match / Evolve into a schema

Fields

input: Arc<DslPlan>

match_schema: Arc<Schema<DataType>>

The schema to match to.

This is also always the output schema.

per_column: Arc<[MatchToSchemaPerColumn]>

extra_columns: ExtraColumnsPolicy

Distinct

Remove duplicates from the table

Fields

input: Arc<DslPlan>

options: DistinctOptionsDSL

Sort

Sort the table

Fields

input: Arc<DslPlan>

by_column: Vec<Expr>

slice: Option<(i64, usize)>

sort_options: SortMultipleOptions

Slice

Slice the table

Fields

input: Arc<DslPlan>

offset: i64

len: u32

MapFunction

A (User Defined) Function

Fields

input: Arc<DslPlan>

function: DslFunction

Union

Vertical concatenation

Fields

inputs: Vec<DslPlan>

args: UnionArgs

HConcat

Horizontal concatenation of multiple plans

Fields

inputs: Vec<DslPlan>

options: HConcatOptions

ExtContext

This allows expressions to access other tables

Fields

input: Arc<DslPlan>

contexts: Vec<DslPlan>

Sink

Fields

input: Arc<DslPlan>

payload: SinkType

SinkMultiple

Fields

inputs: Vec<DslPlan>

IR

Fields

dsl: Arc<DslPlan>

version: u32

node: Option<Node>

Implementations

impl DslPlan

pub fn describe(&self) -> Result<String, PolarsError>

pub fn describe_tree_format(&self) -> Result<String, PolarsError>

pub fn display(&self) -> Result<impl Display, PolarsError>

pub fn to_alp(self) -> Result<IRPlan, PolarsError>

pub fn serialize_versioned<W>( &self, writer: W, ctx: PlanSerializationContext, ) -> Result<(), PolarsError>
where W: Write,

pub fn deserialize_versioned<R>(reader: R) -> Result<DslPlan, PolarsError>
where R: Read,

impl DslPlan

pub fn compute_schema(&self) -> Result<Arc<Schema<DataType>>, PolarsError>

Compute the schema. This requires conversion to [IR] and type-resolving.