polars::prelude::file

Enum Writeable

pub enum Writeable {
    Dyn(Box<dyn WriteableTrait + Send>),
    Local(File),
    Cloud(CloudWriterIoTraitWrap),
}

Available on crate feature polars-io only.

Holds a non-async writeable file, abstracted over local files or cloud files.

This implements DerefMut to a trait object implementing std::io::Write.

Also see: Writeable::try_into_async_writeable and AsyncWriteable.

Variants

Dyn(Box<dyn WriteableTrait + Send>)

An abstract implementation for writable.

This is used to implement writing to in-memory and arbitrary file descriptors.

Local(File)

Cloud(CloudWriterIoTraitWrap)

Available on crate feature cloud only.

Implementations

impl Writeable

pub fn try_new( path: PlRefPath, cloud_options: Option<&CloudOptions>, cloud_upload_chunk_size: usize, cloud_upload_concurrency: usize, io_metrics: Option<Arc<IOMetrics>>, ) -> Result<Writeable, PolarsError>

pub fn try_into_async_writeable(self) -> Result<AsyncWriteable, PolarsError>

Available on crate feature cloud only.

This returns Result<> - if a write was performed before calling this, CloudWriter can be in an Err(_) state.

pub fn as_buffered(&mut self) -> BufferedWriteable<'_>

pub fn sync_all(&self) -> Result<(), Error>

pub fn sync_data(&self) -> Result<(), Error>

pub fn close(self, sync: SyncOnCloseType) -> Result<(), Error>