polars::prelude
Function deserialize

pub fn deserialize<'a>(
    field: &Field,
    field_md: impl ExactSizeIterator<Item = &'a ColumnChunkMetadata>,
) -> Result<Statistics, PolarsError>

Available on crate feature polars-io only.

Deserializes the statistics in the column chunks from a single row_group into Statistics associated from field’s name.

Errors

This function errors if the deserialization of the statistics fails (e.g. invalid utf8)