Function polars::prelude::map_binary

pub fn map_binary<F>(
    a: Expr,
    b: Expr,
    f: F,
    output_type: SpecialEq<Arc<dyn FunctionOutputField>>
) -> Expr
where
    F: 'static + Fn(Series, Series) -> Result<Option<Series>, PolarsError> + Send + Sync,

Available on crate feature lazy only.

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

The closure takes two arguments, each a Series. output_type must be the output dtype of the resulting Series.