Trait polars::prelude::groupby::IntoGroupsProxy

pub trait IntoGroupsProxy {
    fn group_tuples(&self, _multithreaded: bool, _sorted: bool) -> GroupsProxy { ... }
}

Used to create the tuples for a groupby operation.

Provided methods

fn group_tuples(&self, _multithreaded: bool, _sorted: bool) -> GroupsProxy

Create the tuples need for a groupby operation. * The first value in the tuple is the first index of the group. * The second value in the tuple is are the indexes of the groups including the first value.