Function polars::prelude::json::read::json_deserializer::parse

pub fn parse(json: &[u8]) -> Result<Value<'_>, Error>

Available on crate feature io_json only.

Parses JSON-compliant bytes into Value

Errors

If and only if json is not valid JSON.

Panics

If and only if there is not enough memory to allocate.