polars::prelude::row_encode

Function get_row_encoding_context

pub fn get_row_encoding_context(dtype: &DataType) -> Option<RowEncodingContext>

Get the [RowEncodingContext] for a certain DataType.

This should be given the logical type in order to communicate Polars datatype information down into the row encoding / decoding.