polars::prelude
Module search_sorted

Enums

SearchSortedSide

Functions

binary_search_ca

lower_bound_chunks

Search through a series of chunks for the first position where f(x) is true, assuming it is first always false and then always true.