Struct polars::prelude::string::StringNameSpace

pub struct StringNameSpace(_);

Available on crate feature strings only.

Specialized expressions for Series of DataType::Utf8.

Implementations

impl StringNameSpace

pub fn contains_literal(self, pat: Expr) -> Expr

Available on crate feature regex only.

Check if a string value contains a literal substring.

pub fn contains(self, pat: Expr, strict: bool) -> Expr

Available on crate feature regex only.

Check if a string value contains a Regex substring.

pub fn ends_with(self, sub: Expr) -> Expr

Check if a string value ends with the sub string.

pub fn starts_with(self, sub: Expr) -> Expr

Check if a string value starts with the sub string.

pub fn extract(self, pat: &str, group_index: usize) -> Expr

Extract a regex pattern from the a string value.

pub fn extract_all(self, pat: Expr) -> Expr

Extract each successive non-overlapping match in an individual string as an array

pub fn count_match(self, pat: &str) -> Expr

Count all successive non-overlapping regex matches.

pub fn strptime(self, options: StrpTimeOptions) -> Expr

Available on crate feature temporal only.

pub fn concat(self, delimiter: &str) -> Expr

Available on crate feature concat_str only.

Concat the values into a string array.

Arguments

delimiter - A string that will act as delimiter between values.

pub fn split(self, by: &str) -> Expr

Split the string by a substring. The resulting dtype is List<Utf8>.

pub fn split_inclusive(self, by: &str) -> Expr

Split the string by a substring and keep the substring. The resulting dtype is List<Utf8>.

pub fn split_exact(self, by: &str, n: usize) -> Expr

Available on crate feature dtype-struct only.

Split exactly n times by a given substring. The resulting dtype is DataType::Struct.

pub fn split_exact_inclusive(self, by: &str, n: usize) -> Expr

Available on crate feature dtype-struct only.

Split exactly n times by a given substring and keep the substring. The resulting dtype is DataType::Struct.

pub fn splitn(self, by: &str, n: usize) -> Expr

Available on crate feature dtype-struct only.

Split by a given substring, returning exactly n items. If there are more possible splits, keeps the remainder of the string intact. The resulting dtype is DataType::Struct.

pub fn replace(self, pat: Expr, value: Expr, literal: bool) -> Expr

Available on crate feature regex only.

Replace values that match a regex pat with a value.

pub fn replace_all(self, pat: Expr, value: Expr, literal: bool) -> Expr

Available on crate feature regex only.

Replace all values that match a regex pat with a value.

pub fn strip(self, matches: Option<String>) -> Expr

Remove leading and trailing characters, or whitespace if matches is None.

pub fn lstrip(self, matches: Option<String>) -> Expr

Remove leading characters, or whitespace if matches is None.

pub fn rstrip(self, matches: Option<String>) -> Expr

Remove trailing characters, or whitespace if matches is None..

pub fn to_lowercase(self) -> Expr

Convert all characters to lowercase.

pub fn to_uppercase(self) -> Expr

Convert all characters to uppercase.

pub fn from_radix(self, radix: Option<u32>) -> Expr

Available on crate feature string_from_radix only.

Parse string in base radix into decimal