polars::prelude
Struct CategoricalChunked

pub struct CategoricalChunked { /* private fields */ }

Implementations

impl CategoricalChunked

pub fn sort_with(&self, options: SortOptions) -> CategoricalChunked

Available on crate feature dtype-categorical only.

pub fn sort(&self, descending: bool) -> CategoricalChunked

Available on crate feature dtype-categorical only.

Returned a sorted ChunkedArray.

pub fn arg_sort(&self, options: SortOptions) -> ChunkedArray<UInt32Type>

Available on crate feature dtype-categorical only.

Retrieve the indexes needed to sort this array.

impl CategoricalChunked

pub unsafe fn from_global_indices_unchecked( cats: ChunkedArray<UInt32Type>, ordering: CategoricalOrdering, ) -> CategoricalChunked

Available on crate feature dtype-categorical only.

Create a CategoricalChunked from a categorical indices. The indices will probe the global string cache.

Safety

This does not do any bound checks

pub fn from_string_to_enum( values: &ChunkedArray<StringType>, categories: &BinaryViewArrayGeneric<str>, ordering: CategoricalOrdering, ) -> Result<CategoricalChunked, PolarsError>

Available on crate feature dtype-categorical only.

Create a CategoricalChunked from a fixed list of categories and a List of strings. This will error if a string is not in the fixed list of categories

impl CategoricalChunked

pub fn to_arrow( &self, compat_level: CompatLevel, as_i64: bool, ) -> Box<dyn Array>

Available on crate feature dtype-categorical only.

impl CategoricalChunked

pub fn append(&mut self, other: &CategoricalChunked) -> Result<(), PolarsError>

Available on crate feature dtype-categorical only.

pub fn append_owned( &mut self, other: CategoricalChunked, ) -> Result<(), PolarsError>

Available on crate feature dtype-categorical only.

impl CategoricalChunked

pub fn full_null( name: PlSmallStr, is_enum: bool, length: usize, ordering: CategoricalOrdering, ) -> CategoricalChunked

Available on crate feature dtype-categorical only.

impl CategoricalChunked

pub fn unique(&self) -> Result<CategoricalChunked, PolarsError>

Available on crate feature dtype-categorical only.

pub fn n_unique(&self) -> Result<usize, PolarsError>

Available on crate feature dtype-categorical only.

pub fn value_counts(&self) -> Result<DataFrame, PolarsError>

Available on crate feature dtype-categorical only.

impl CategoricalChunked

pub fn is_empty(&self) -> bool

Available on crate feature dtype-categorical only.

pub fn len(&self) -> usize

Available on crate feature dtype-categorical only.

pub fn null_count(&self) -> usize

Available on crate feature dtype-categorical only.

pub fn name(&self) -> &PlSmallStr

Available on crate feature dtype-categorical only.

pub fn physical(&self) -> &ChunkedArray<UInt32Type>

Available on crate feature dtype-categorical only.

Get a reference to the physical array (the categories).

pub fn is_enum(&self) -> bool

Available on crate feature dtype-categorical only.

pub fn to_local(&self) -> CategoricalChunked

Available on crate feature dtype-categorical only.

Convert a categorical column to its local representation.

pub fn to_global(&self) -> Result<CategoricalChunked, PolarsError>

Available on crate feature dtype-categorical only.

pub fn to_enum( &self, categories: &BinaryViewArrayGeneric<str>, hash: u128, ) -> CategoricalChunked

Available on crate feature dtype-categorical only.

pub fn uses_lexical_ordering(&self) -> bool

Available on crate feature dtype-categorical only.

Return whether or not the CategoricalChunked uses the lexical order of the string values when sorting.

pub unsafe fn from_cats_and_dtype_unchecked( idx: ChunkedArray<UInt32Type>, dtype: DataType, ) -> CategoricalChunked

Available on crate feature dtype-categorical only.

Create a CategoricalChunked from a physical array and dtype.

Safety

It’s not checked that the indices are in-bounds or that the dtype is correct.

pub unsafe fn from_cats_and_rev_map_unchecked( idx: ChunkedArray<UInt32Type>, rev_map: Arc<RevMapping>, is_enum: bool, ordering: CategoricalOrdering, ) -> CategoricalChunked

Available on crate feature dtype-categorical only.

Create a CategoricalChunked from an array of idx and an existing RevMapping: rev_map.

Safety

Invariant in v < rev_map.len() for v in idx must hold.

pub fn _can_fast_unique(&self) -> bool

Available on crate feature dtype-categorical only.

True if all categories are represented in this array. When this is the case, the unique values of the array are the categories.

pub unsafe fn _with_fast_unique(self, toggle: bool) -> CategoricalChunked

Available on crate feature dtype-categorical only.

Set FAST_UNIQUE metadata

Safety

This invariant must hold unique(categories) == unique(self)

pub fn get_rev_map(&self) -> &Arc<RevMapping>

Available on crate feature dtype-categorical only.

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> CatIter<'_>

Available on crate feature dtype-categorical only.

Create an Iterator that iterates over the &str values of the CategoricalChunked.