polars::prelude

Struct DslBuilder

pub struct DslBuilder(pub DslPlan);

Available on crate feature lazy only.

Tuple Fields
0: DslPlan
Implementations

impl DslBuilder

pub fn anonymous_scan( function: Arc<dyn AnonymousScan>, options: AnonymousScanOptions, unified_scan_args: UnifiedScanArgs, ) -> Result<DslBuilder, PolarsError>

pub fn scan_parquet( sources: ScanSources, options: ParquetOptions, unified_scan_args: UnifiedScanArgs, ) -> Result<DslBuilder, PolarsError>

pub fn scan_ipc( sources: ScanSources, options: IpcScanOptions, unified_scan_args: UnifiedScanArgs, ) -> Result<DslBuilder, PolarsError>

pub fn scan_csv( sources: ScanSources, options: CsvReadOptions, unified_scan_args: UnifiedScanArgs, ) -> Result<DslBuilder, PolarsError>

pub fn cache(self) -> DslBuilder

pub fn drop(self, columns: Selector) -> DslBuilder

pub fn project(self, exprs: Vec<Expr>, options: ProjectionOptions) -> DslBuilder

pub fn fill_null(self, fill_value: Expr) -> DslBuilder

pub fn drop_nans(self, subset: Option<Selector>) -> DslBuilder

pub fn drop_nulls(self, subset: Option<Selector>) -> DslBuilder

pub fn fill_nan(self, fill_value: Expr) -> DslBuilder

pub fn with_columns( self, exprs: Vec<Expr>, options: ProjectionOptions, ) -> DslBuilder

pub fn match_to_schema( self, match_schema: Arc<Schema<DataType>>, per_column: Arc<[MatchToSchemaPerColumn]>, extra_columns: ExtraColumnsPolicy, ) -> DslBuilder

pub fn with_context(self, contexts: Vec<DslPlan>) -> DslBuilder

pub fn filter(self, predicate: Expr) -> DslBuilder

Apply a filter predicate, keeping the rows that match it.

pub fn remove(self, predicate: Expr) -> DslBuilder

Remove rows matching a filter predicate (note that rows where the predicate resolves to null are not removed).

pub fn group_by<E>( self, keys: Vec<Expr>, aggs: E, apply: Option<(Arc<dyn DataFrameUdf>, Arc<Schema<DataType>>)>, maintain_order: bool, dynamic_options: Option<DynamicGroupOptions>, rolling_options: Option<RollingGroupOptions>, ) -> DslBuilder
where E: AsRef<[Expr]>,

pub fn build(self) -> DslPlan

pub fn from_existing_df(df: DataFrame) -> DslBuilder

pub fn sort( self, by_column: Vec<Expr>, sort_options: SortMultipleOptions, ) -> DslBuilder

pub fn explode(self, columns: Selector, allow_empty: bool) -> DslBuilder

pub fn row_index(self, name: PlSmallStr, offset: Option<u32>) -> DslBuilder

pub fn distinct(self, options: DistinctOptionsDSL) -> DslBuilder

pub fn slice(self, offset: i64, len: u32) -> DslBuilder

pub fn join( self, other: DslPlan, left_on: Vec<Expr>, right_on: Vec<Expr>, options: Arc<JoinOptions>, ) -> DslBuilder

pub fn map_private(self, function: DslFunction) -> DslBuilder

pub fn map<F>( self, function: F, optimizations: OptFlags, schema: Option<Arc<dyn UdfSchema>>, name: PlSmallStr, ) -> DslBuilder
where F: DataFrameUdf + 'static,