Struct polars::prelude::DurationArgs

pub struct DurationArgs {
    pub weeks: Expr,
    pub days: Expr,
    pub hours: Expr,
    pub minutes: Expr,
    pub seconds: Expr,
    pub milliseconds: Expr,
    pub microseconds: Expr,
    pub nanoseconds: Expr,
    pub time_unit: TimeUnit,
}

Available on crate feature lazy only.

Arguments used by duration in order to produce an Expr of Duration

To construct a DurationArgs, use struct literal syntax with ..Default::default() to leave unspecified fields at their default value of lit(0), as demonstrated below.

let args = DurationArgs {
    days: lit(5),
    hours: col("num_hours"),
    minutes: col("num_minutes"),
    ..Default::default()  // other fields are lit(0)
};

If you prefer builder syntax, with_* methods are also available.

let args = DurationArgs::new().with_weeks(lit(42)).with_hours(lit(84));

Fields
weeks: Exprdays: Exprhours: Exprminutes: Exprseconds: Exprmilliseconds: Exprmicroseconds: Exprnanoseconds: Exprtime_unit: TimeUnit
Implementations

impl DurationArgs

pub fn new() -> DurationArgs

Create a new DurationArgs with all fields set to lit(0). Use the with_* methods to set the fields.

pub fn with_hms(self, hours: Expr, minutes: Expr, seconds: Expr) -> DurationArgs

Set hours, minutes, and seconds

Equivalent to:

self.with_hours(hours)
    .with_minutes(minutes)
    .with_seconds(seconds)

pub fn with_fractional_seconds( self, milliseconds: Expr, microseconds: Expr, nanoseconds: Expr ) -> DurationArgs

Set milliseconds, microseconds, and nanoseconds

Equivalent to

self.with_milliseconds(milliseconds)
    .with_microseconds(microseconds)
    .with_nanoseconds(nanoseconds)

pub fn with_weeks(self, n: Expr) -> DurationArgs

Set the weeks

pub fn with_days(self, n: Expr) -> DurationArgs

Set the days

pub fn with_hours(self, n: Expr) -> DurationArgs

Set the hours

pub fn with_minutes(self, n: Expr) -> DurationArgs

Set the minutes

pub fn with_seconds(self, n: Expr) -> DurationArgs

Set the seconds

pub fn with_milliseconds(self, n: Expr) -> DurationArgs

Set the milliseconds

pub fn with_microseconds(self, n: Expr) -> DurationArgs

Set the microseconds

pub fn with_nanoseconds(self, n: Expr) -> DurationArgs

Set the nanoseconds