polars::prelude

Struct PlRefPath

pub struct PlRefPath { /* private fields */ }

Reference-counted PlPath.

Windows paths invariant

Windows paths will have leading \\?\ prefix stripped, and all backslashes normalized to forward slashes.

Implementations

impl PlRefPath

pub fn empty() -> PlRefPath

pub fn new(path: impl AsRef<str> + Into<PlRefStr>) -> PlRefPath

Normalizes Windows paths.

pub fn try_from_path(path: &Path) -> Result<PlRefPath, PolarsError>

pub fn try_from_pathbuf(path: PathBuf) -> Result<PlRefPath, PolarsError>

pub fn as_str(&self) -> &str

pub fn as_ref_str(&self) -> &PlRefStr

pub fn into_ref_str(self) -> PlRefStr

pub fn sliced(&self, range: Range<usize>) -> PlRefPath

Slices the path.

pub fn to_absolute_path(&self) -> Result<Cow<'_, PlRefPath>, PolarsError>

Returns

Returns an absolute local path if this path ref is a relative local path, otherwise returns None.

pub fn ptr_eq(this: &PlRefPath, other: &PlRefPath) -> bool

Checks if references point to the same allocation.