polars::prelude

Struct SortMultipleOptions

pub struct SortMultipleOptions {
    pub descending: Vec<bool>,
    pub nulls_last: Vec<bool>,
    pub multithreaded: bool,
    pub maintain_order: bool,
    pub limit: Option<u32>,
}

Sort options for multi-series sorting.

Indicating the order of sorting, nulls position, multithreading, and maintaining order.

Example

let df = df! {
    "a" => [Some(1), Some(2), None, Some(4), None],
    "b" => [Some(5), None, Some(3), Some(2), Some(1)]
}?;

let out = df
    .sort(
        ["a", "b"],
        SortMultipleOptions::default()
            .with_maintain_order(true)
            .with_multithreaded(false)
            .with_order_descending_multi([false, true])
            .with_nulls_last(true),
    )?;

let expected = df! {
    "a" => [Some(1), Some(2), Some(4), None, None],
    "b" => [Some(5), None, Some(2), Some(3), Some(1)]
}?;

assert_eq!(out, expected);

Fields
descending: Vec<bool>

Order of the columns. Default all `false``.

If only one value is given, it will broadcast to all columns.

Use SortMultipleOptions::with_order_descending_multi or SortMultipleOptions::with_order_descending to modify.

Safety

Len must match the number of columns, or equal 1.

nulls_last: Vec<bool>

Whether place null values last. Default false.

multithreaded: bool

Whether sort in multiple threads. Default true.

maintain_order: bool

Whether maintain the order of equal elements. Default false.

limit: Option<u32>

Limit a sort output, this is for optimization purposes and might be ignored.

Implementations

impl SortMultipleOptions

pub fn new() -> SortMultipleOptions

Create SortMultipleOptions with default values.

pub fn with_order_descending_multi( self, descending: impl IntoIterator<Item = bool>, ) -> SortMultipleOptions

Specify order for each column. Defaults all false.

Safety

Len must match the number of columns, or be equal to 1.

pub fn with_order_descending(self, descending: bool) -> SortMultipleOptions

Sort order for all columns. Default false which is ascending.

pub fn with_nulls_last_multi( self, nulls_last: impl IntoIterator<Item = bool>, ) -> SortMultipleOptions

Specify whether to place nulls last, per-column. Defaults all false.

Safety

Len must match the number of columns, or be equal to 1.

pub fn with_nulls_last(self, enabled: bool) -> SortMultipleOptions

Whether to place null values last. Default false.

pub fn with_multithreaded(self, enabled: bool) -> SortMultipleOptions

Whether to sort in multiple threads. Default true.

pub fn with_maintain_order(self, enabled: bool) -> SortMultipleOptions

Whether to maintain the order of equal elements. Default false.

pub fn with_order_reversed(self) -> SortMultipleOptions

Reverse the order of sorting for each column.