Struct polars::prelude::StructChunked

pub struct StructChunked { /* private fields */ }

This is logical type StructChunked that dispatches most logic to the fields implementations

Different from StructArray, this type does not have its own validity. That means some operations will be a bit less efficient because we need to check validity of all fields. However this does save a lot of code and compile times.

Implementations

impl StructChunked

pub fn new(name: &str, fields: &[Series]) -> Result<StructChunked, PolarsError>

pub fn rechunk(&mut self)

pub fn field_by_name(&self, name: &str) -> Result<Series, PolarsError>

Get access to one of this [StructChunked]’s fields

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn ref_field(&self) -> &Field

Get a reference to the Field of array.

pub fn name(&self) -> &String

pub fn fields(&self) -> &[Series]

pub fn fields_mut(&mut self) -> &mut Vec<Series, Global>

pub fn rename(&mut self, name: &str)