Trait polars::prelude::ArrowPrimitiveType

pub trait ArrowPrimitiveType: 'static {
    type Native: ArrowNativeType;

    pub const DATA_TYPE: DataType;
    pub fn get_byte_width() -> usize { ... }

    pub fn default_value() -> Self::Native { ... }
}

Trait bridging the dynamic-typed nature of Arrow (via DataType) with the static-typed nature of rust types (ArrowNativeType) for all types that implement ArrowNativeType.

Associated Types

type Native: ArrowNativeType

Corresponding Rust native type for the primitive type.

Loading content...
Associated Constants

pub const DATA_TYPE: DataType

the corresponding Arrow data type of this primitive type.

Loading content...
Provided methods

pub fn get_byte_width() -> usize

Returns the byte width of this primitive type.

pub fn default_value() -> Self::Native

Returns a default value of this primitive type.

This is useful for aggregate array ops like sum(), mean().

Loading content...