Trait polars::prelude::AsofJoinBy

pub trait AsofJoinBy: IntoDf {
    // Provided method
    fn join_asof_by<I, S>(
        &self,
        other: &DataFrame,
        left_on: &str,
        right_on: &str,
        left_by: I,
        right_by: I,
        strategy: AsofStrategy,
        tolerance: Option<AnyValue<'static>>
    ) -> Result<DataFrame, PolarsError>
       where I: IntoIterator<Item = S>,
             S: AsRef<str> { ... }
}

Available on crate feature polars-ops only.

Provided Methods

fn join_asof_by<I, S>( &self, other: &DataFrame, left_on: &str, right_on: &str, left_by: I, right_by: I, strategy: AsofStrategy, tolerance: Option<AnyValue<'static>> ) -> Result<DataFrame, PolarsError>
where I: IntoIterator<Item = S>, S: AsRef<str>,

This is similar to a left-join except that we match on nearest key rather than equal keys. The keys must be sorted to perform an asof join. This is a special implementation of an asof join that searches for the nearest keys within a subgroup set by by.

Object Safety

This trait is not object safe.