Trait polars::prelude::ChunkAgg

pub trait ChunkAgg<T> {
    fn sum(&self) -> Option<T> { ... }

    fn min(&self) -> Option<T> { ... }

    fn max(&self) -> Option<T> { ... }

    fn mean(&self) -> Option<f64> { ... }
}

Aggregation operations

Provided methods

fn sum(&self) -> Option<T>

Aggregate the sum of the ChunkedArray. Returns None if the array is empty or only contains null values.

fn min(&self) -> Option<T>

fn max(&self) -> Option<T>

Returns the maximum value in the array, according to the natural order. Returns None if the array is empty or only contains null values.

fn mean(&self) -> Option<f64>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.