Trait polars::prelude::ChunkApplyKernel

pub trait ChunkApplyKernel<A>where
    A: Array,{
    fn apply_kernel(&self, f: &dyn Fn(&A)) -> Self;
    fn apply_kernel_cast<S>(&self, f: &dyn Fn(&A)) -> ChunkedArray<S>
    where
        S: PolarsDataType;
}

Apply kernels on the arrow array chunks in a ChunkedArray.

Required Methods

fn apply_kernel(&self, f: &dyn Fn(&A)) -> Self

Apply kernel and return result as a new ChunkedArray.

fn apply_kernel_cast<S>(&self, f: &dyn Fn(&A)) -> ChunkedArray<S>where
S: PolarsDataType,

Apply a kernel that outputs an array of different type.