Trait polars::prelude::ChunkFillNone

pub trait ChunkFillNone {
    fn fill_none(&self, strategy: FillNoneStrategy) -> Result<Self, PolarsError>;
}

Replace None values with various strategies

Required methods

fn fill_none(&self, strategy: FillNoneStrategy) -> Result<Self, PolarsError>

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)