Trait polars::prelude::ChunkFillNull

pub trait ChunkFillNull {
    fn fill_null(&self, strategy: FillNullStrategy) -> Result<Self, PolarsError>
    where
        Self: Sized;
}

Replace None values with various strategies

Required Methods

fn fill_null(&self, strategy: FillNullStrategy) -> Result<Self, PolarsError>where
Self: Sized,

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)