Trait polars::prelude::ChunkPeaks

pub trait ChunkPeaks {
    fn peak_max(&self) -> ChunkedArray<BooleanType> { ... }

    fn peak_min(&self) -> ChunkedArray<BooleanType> { ... }
}

Find local minima/ maxima

Provided methods

fn peak_max(&self) -> ChunkedArray<BooleanType>

Get a boolean mask of the local maximum peaks.

fn peak_min(&self) -> ChunkedArray<BooleanType>

Get a boolean mask of the local minimum peaks.