Trait polars::prelude::ChunkShiftFill

pub trait ChunkShiftFill<T, V> {
    pub fn shift_and_fill(
        &self, 
        periods: i32, 
        fill_value: V
    ) -> Result<ChunkedArray<T>, PolarsError>;
}

Shift the values of a ChunkedArray by a number of periods.

Required methods

pub fn shift_and_fill(
&self,
periods: i32,
fill_value: V
) -> Result<ChunkedArray<T>, PolarsError>

Shift the values by a given period and fill the parts that will be empty due to this operation with fill_value.

Loading content...