Trait polars::prelude::ChunkTake

pub trait ChunkTake {
    // Required methods
    unsafe fn take_unchecked<I, INulls>(
        &self,
        indices: TakeIdx<'_, I, INulls>
    ) -> Self
       where Self: Sized,
             I: TakeIterator,
             INulls: TakeIteratorNulls;
    fn take<I, INulls>(
        &self,
        indices: TakeIdx<'_, I, INulls>
    ) -> Result<Self, PolarsError>
       where Self: Sized,
             I: TakeIterator,
             INulls: TakeIteratorNulls;
}

Fast access by index.

Required Methods

unsafe fn take_unchecked<I, INulls>( &self, indices: TakeIdx<'_, I, INulls> ) -> Selfwhere Self: Sized, I: TakeIterator, INulls: TakeIteratorNulls,

Take values from ChunkedArray by index.

Safety

Doesn’t do any bound checking.

fn take<I, INulls>( &self, indices: TakeIdx<'_, I, INulls> ) -> Result<Self, PolarsError>where Self: Sized, I: TakeIterator, INulls: TakeIteratorNulls,

Take values from ChunkedArray by index. Note that the iterator will be cloned, so prefer an iterator that takes the owned memory by reference.