Trait polars::prelude::LogSeries

pub trait LogSeries: SeriesSealed {
    // Provided methods
    fn log(&self, base: f64) -> Series { ... }
    fn log1p(&self) -> Series { ... }
    fn exp(&self) -> Series { ... }
    fn entropy(&self, base: f64, normalize: bool) -> Option<f64> { ... }
}

Provided Methods

fn log(&self, base: f64) -> Series

Compute the logarithm to a given base

fn log1p(&self) -> Series

Compute the natural logarithm of all elements plus one in the input array

fn exp(&self) -> Series

Calculate the exponential of all elements in the input array.

fn entropy(&self, base: f64, normalize: bool) -> Option<f64>

Compute the entropy as -sum(pk * log(pk). where pk are discrete probabilities.