polars::prelude

Trait MomentSeries

pub trait MomentSeries: SeriesSealed {
    // Provided methods
    fn skew(&self, bias: bool) -> Result<Option<f64>, PolarsError> { ... }
    fn kurtosis(
        &self,
        fisher: bool,
        bias: bool,
    ) -> Result<Option<f64>, PolarsError> { ... }
}

Available on crate feature polars-ops only.

Provided Methods

fn skew(&self, bias: bool) -> Result<Option<f64>, PolarsError>

Compute the sample skewness of a data set.

For normally distributed data, the skewness should be about zero. For uni-modal continuous distributions, a skewness value greater than zero means that there is more weight in the right tail of the distribution. The function skewtest can be used to determine if the skewness value is close enough to zero, statistically speaking.

see: scipy

fn kurtosis(&self, fisher: bool, bias: bool) -> Result<Option<f64>, PolarsError>

Compute the kurtosis (Fisher or Pearson) of a dataset.

Kurtosis is the fourth central moment divided by the square of the variance. If Fisher’s definition is used, then 3.0 is subtracted from the result to give 0.0 for a normal distribution. If bias is false then the kurtosis is calculated using k statistics to eliminate bias coming from biased moment estimators

see: scipy