polars::prelude

Trait TakeChunked

pub trait TakeChunked {
    // Required methods
    unsafe fn take_chunked_unchecked<const B: u64>(
        &self,
        by: &[ChunkId<B>],
        sorted: IsSorted,
        avoid_sharing: bool,
    ) -> Self;
    unsafe fn take_opt_chunked_unchecked<const B: u64>(
        &self,
        by: &[ChunkId<B>],
        avoid_sharing: bool,
    ) -> Self;
}

Available on crate feature polars-ops only.

Gather by ChunkId

Required Methods

unsafe fn take_chunked_unchecked<const B: u64>( &self, by: &[ChunkId<B>], sorted: IsSorted, avoid_sharing: bool, ) -> Self

Gathers elements from a ChunkedArray, specifying for each element a chunk index and index within that chunk through ChunkId. If avoid_sharing is true the returned data should not share references with the original array (like shared buffers in views).

Safety

This function doesn’t do any bound checks.

unsafe fn take_opt_chunked_unchecked<const B: u64>( &self, by: &[ChunkId<B>], avoid_sharing: bool, ) -> Self

Safety

This function doesn’t do any bound checks.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.