Module polars::chunked_array::iterator

Modules

par	parallel

Structs

BooleanIterCont
No null checks and dont return Option but T directly. So this struct is not return by the IntoIterator trait

BooleanIterManyChunk
Many chunks no nulls

BooleanIterManyChunkNullCheck
Many chunks no nulls

BooleanIterSingleChunk
No null checks

BooleanIterSingleChunkNullCheck

ListIterCont
No null checks and dont return Option but T directly. So this struct is not return by the IntoIterator trait

ListIterManyChunk
Many chunks no nulls

ListIterManyChunkNullCheck
Many chunks no nulls

ListIterSingleChunk
No null checks

ListIterSingleChunkNullCheck

NumIterManyChunk
Many chunks no null checks Both used as iterator with null checks and without. We later map Some on it for the iter with null checks

NumIterManyChunkNullCheck
Many chunks with null checks

NumIterSingleChunk
Single chunk no null values

NumIterSingleChunkNullCheck
Single chunk with null values

Utf8IterManyChunkNullCheckReturnOption
Iterator for chunked arrays with many chunks. The chunks have null values so it DOES perform null checks.

Utf8IterManyChunkReturnOption
Iterator for chunked arrays with many chunks. The chunks cannot have null values so it does NOT perform null checks.

Utf8IterManyChunkReturnUnwrapped
Iterator for chunked arrays with many chunks. The chunks cannot have null values so it does NOT perform null checks.

Utf8IterSingleChunkNullCheckReturnOption
Iterator for chunked arrays with just one chunk. The chunk have null values so it DOES perform null checks.

Utf8IterSingleChunkReturnOption
Iterator for chunked arrays with just one chunk. The chunk cannot have null values so it does NOT perform null checks.

Utf8IterSingleChunkReturnUnwrapped
Iterator for chunked arrays with just one chunk. The chunks cannot have null values so it does NOT perform null checks.

Enums

BooleanIterDispatch

ListIterDispatch

NumericChunkIterDispatch

Utf8ChunkIterReturnOptionDispatch
Static dispatching structure to allow static polymorphism of chunked iterators.

Utf8ChunkIterReturnUnwrappedDispatch
Static dispatching structure to allow static polymorphism of non-nullables chunked iterators.

Traits

IntoNoNullIterator
Trait for ChunkedArrays that don't have null values.