Module polars::chunked_array::ops

Traits for miscellaneous operations on ChunkedArray

Modules

arity

fill_null

float_sorted_arg_max

full

gather

search_sorted

sort

zipzip_with

Structs

SortMultipleOptions

Sort options for multi-series sorting.

SortOptions

Options for single series sorting.

Enums

FillNullStrategy

Traits

ChunkAgg

Aggregation operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/flatten a List or String Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkRollApplyrolling_window

This differs from ChunkWindowCustom and ChunkWindow by not using a fold aggregator, but reusing a Series wrapper and calling Series aggregators. This likely is a bit slower than ChunkWindow

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

IsFirstDistinctis_first_distinct

Mask the first unique values as true

IsLastDistinctis_last_distinct

Mask the last unique values as true

Reinterpretreinterpret

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

Type Aliases

FillNullLimit