polars::chunked_array

Module ops

Traits for miscellaneous operations on ChunkedArray

Modules

arity

fill_null

float_sorted_arg_max

full

gather

row_encode

search_sorted

sort

zip

Structs

SortMultipleOptions

Sort options for multi-series sorting.

SortOptions

Options for single series sorting.

Enums

FillNullStrategy

Traits

ChunkAgg

Aggregation operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkApproxNUnique

ChunkBitwiseReduce

Bitwise Reduction Operations.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompareEq

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkCompareIneq

Compare Series and ChunkedArray’s using inequality operators (<, >=, etc.) and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/flatten a List or String Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkNestingUtils

Utility methods for dealing with nested chunked arrays.

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkRollApply

This differs from ChunkWindowCustom and ChunkWindow by not using a fold aggregator, but reusing a Series wrapper and calling Series aggregators. This likely is a bit slower than ChunkWindow

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

IsFirstDistinct

Mask the first unique values as true

IsLastDistinct

Mask the last unique values as true

Reinterpret

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

Functions

_set_check_length⚠

Meant for internal use. In very rare conditions this can be turned off.

Type Aliases

FillNullLimit