Module polars::chunked_array::ops

Traits for miscellaneous operations on ChunkedArray

Enums

FillNoneStrategy

Traits

ChunkAgg
Aggregation operations

ChunkAggSeries
Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkApply
Fastest way to do elementwise operations on a ChunkedArray

ChunkApplyKernel
Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast
Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare
Compare Series and ChunkedArray's and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex
Create a new ChunkedArray filled with values at that index.

ChunkFillNone
Replace None values with various strategies

ChunkFilter
Filter values by a boolean mask.

ChunkFull
Fill a ChunkedArray with one value.

ChunkReverse
Reverse a ChunkedArray

ChunkSet
Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift
Shift the values of a ChunkedArray by a number of periods.

ChunkSort
Sort operations on ChunkedArray.

ChunkTake
Fast access by index.

ChunkUnique
Get unique values in a ChunkedArray

ChunkVar
Variance and standard deviation aggregation.

ChunkWindow

ChunkZip
Combine 2 ChunkedArrays based on some predicate.

TakeRandom
Random access

TakeRandomUtf8

ToDummies

VarAggSeries