polars::chunked_array::ops
Trait ChunkZip

pub trait ChunkZip<T>
where
    T: PolarsDataType,
{
    // Required method
    fn zip_with(
        &self,
        mask: &ChunkedArray<BooleanType>,
        other: &ChunkedArray<T>,
    ) -> Result<ChunkedArray<T>, PolarsError>;
}

Combine two ChunkedArray based on some predicate.

Required Methods

fn zip_with( &self, mask: &ChunkedArray<BooleanType>, other: &ChunkedArray<T>, ) -> Result<ChunkedArray<T>, PolarsError>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false