Struct polars::chunked_array::ChunkedArray

pub struct ChunkedArray<T> { /* fields omitted */ }

Implementations

impl<T> ChunkedArray<T> where
T: PolarsDataType,
ChunkedArray<T>: ChunkOps,

pub fn chunk_id(&self) -> &Vec<usize>

pub fn chunks(&self) -> &Vec<ArrayRef>

A reference to the chunks

pub fn null_count(&self) -> usize

Count the null values.

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn u32(self) -> Result<UInt32Chunked>

Downcast

pub fn i32(self) -> Result<Int32Chunked>

Downcast

pub fn i64(self) -> Result<Int64Chunked>

Downcast

pub fn f32(self) -> Result<Float32Chunked>

Downcast

pub fn f64(self) -> Result<Float64Chunked>

Downcast

pub fn bool(self) -> Result<BooleanChunked>

Downcast

pub fn utf8(self) -> Result<Utf8Chunked>

Downcast

pub fn date32(self) -> Result<Date32Chunked>

Downcast

pub fn date64(self) -> Result<Date64Chunked>

Downcast

pub fn time64ns(self) -> Result<Time64NsChunked>

Downcast

pub fn duration_ns(self) -> Result<DurationNsChunked>

Downcast

pub fn limit(&self, num_elements: usize) -> Result<Self>

Take a view of top n elements

pub fn filter(&self, filter: &BooleanChunked) -> Result<Self>

Chunk sizes should match or rhs should have one chunk

pub fn append_array(&mut self, other: ArrayRef) -> Result<()>

Append arrow array in place.

pub fn len(&self) -> usize

Combined length of all the chunks.

pub fn get(&self, index: usize) -> AnyType

Get a single value. Beware this is slow.

pub fn slice(&self, offset: usize, length: usize) -> Result<Self>

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

pub fn head(&self, length: Option<usize>) -> Result<Self>

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> Result<Self>

Get the tail of the ChunkedArray

pub fn append(&mut self, other: &Self) where
Self: Sized,

Append in place.

impl ChunkedArray<Utf8Type>

pub fn new_utf8_from_slice<S: AsRef<str>>(name: &str, v: &[S]) -> Self

pub fn new_utf8_from_opt_slice<S: AsRef<str>>(
name: &str,
opt_v: &[Option<S>]
) -> Self

impl<T> ChunkedArray<T> where
T: PolarsDataType,
ChunkedArray<T>: ChunkOps,

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

pub fn new_from_chunks(name: &str, chunks: Vec<ArrayRef>) -> Self

Create a new ChunkedArray from existing chunks.

pub fn cast<N>(&self) -> Result<ChunkedArray<N>> where
N: PolarsNumericType,

Cast the underlying data type.

impl<T> ChunkedArray<T> where
T: ArrowPrimitiveType,

pub fn new_from_slice(name: &str, v: &[T::Native]) -> Self

pub fn new_from_opt_slice(name: &str, opt_v: &[Option<T::Native>]) -> Self

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn cont_slice(&self) -> Result<&[T::Native]>

Contiguous slice

pub fn map<B, F>(&self, f: F) -> Result<Map<Copied<Iter<T::Native>>, F>> where
F: Fn(T::Native) -> B,

If cont_slice is successful a closure is mapped over the elements.

Example

use polars::prelude::*;
fn multiply(ca: &UInt32Chunked) -> Result<Series> {
    let mapped = ca.map(|v| v * 2)?;
    Ok(mapped.collect())
}

pub fn map_null_checks<'a, B, F>(
&'a self,
f: F
) -> Map<Box<dyn Iterator<Item = Option<T::Native>> + 'a>, F> where
F: Fn(Option<T::Native>) -> B,

If cont_slice fails we can fallback on an iterator with null checks and map a closure over the elements.

Example

use polars::prelude::*;
use itertools::Itertools;
fn multiply(ca: &UInt32Chunked) -> Series {
    let mapped_result = ca.map(|v| v * 2);

    if let Ok(mapped) = mapped_result {
        mapped.collect()
    } else {
        ca
        .map_null_checks(|opt_v| opt_v.map(|v |v * 2)).collect()
    }
}

pub fn fold<F, B>(&self, init: B, f: F) -> Result<B> where
F: Fn(B, T::Native) -> B,

If cont_slice is successful a closure can be applied as aggregation

Example

use polars::prelude::*;
fn compute_sum(ca: &UInt32Chunked) -> Result<u32> {
    ca.fold(0, |acc, value| acc + value)
}

pub fn fold_null_checks<F, B>(&self, init: B, f: F) -> B where
F: Fn(B, Option<T::Native>) -> B,

If cont_slice fails we can fallback on an iterator with null checks and a closure for aggregation

Example

use polars::prelude::*;
fn compute_sum(ca: &UInt32Chunked) -> u32 {
    match ca.fold(0, |acc, value| acc + value) {
        // faster sum without null checks was successful
        Ok(sum) => sum,
        // Null values or multiple chunks in ChunkedArray, we need to do more bounds checking
        Err(_) => ca.fold_null_checks(0, |acc, opt_value| {
            match opt_value {
                Some(v) => acc + v,
                None => acc
            }
        })
    }
}

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
T::Native: PartialOrd,

pub fn sort(&self) -> Self

pub fn sort_in_place(&mut self)

pub fn argsort(&self) -> UInt32Chunked