Struct polars::chunked_array::ChunkedArray

pub struct ChunkedArray<T> { /* fields omitted */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArray's are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of te same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine(ca: &Float32Chunked) -> Float32Chunked {
    ca.apply(|v| v.cos())
}

If we would like to cast the result we could use a Rust Iterator instead of an apply method. Note that Iterators are slightly slower as the null values aren't ignored implicitly.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.into_iter()
        .map(|opt_v| {
        opt_v.map(|v| v.cos() as f64)
    }).collect()
}

Another option is to first cast and then use an apply.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.cast::<Float64Type>()
        .unwrap()
        .apply(|v| v.cos())
}

Conversion between Series and ChunkedArray's

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> Result<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excelent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.into_iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.into_iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArray's use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make mutliple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkArray will slow down the Iterators, arithmetic and other operations. When multiplying two ChunkArray's with different chunk sizes they cannot utilize SIMD for instance. However, when chunk size don't match, Iterators will be used to do the operation (instead of arrows upstream implementation, which may utilize SIMD) and the result will be a single chunked array.

The key takeaway is that by applying operations on a ChunkArray of multiple chunks, the results will converge to a ChunkArray of a single chunk! It is recommended to leave them as is. If you want to have predictable performance (no unexpected re-allocation of memory), it is adviced to call the rechunk after multiple append operations.

Methods

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn to_ndarray(&self) -> Result<ArrayView1<T::Native>>

This is supported on feature="ndarray" only.

If data is aligned in a single chunk and has no Null values a zero copy view is returned as an ndarray

impl ChunkedArray<LargeListType>

pub fn to_ndarray<N>(&self) -> Result<Array2<N::Native>> where
N: PolarsNumericType,

This is supported on feature="ndarray" only.

If all nested Series have the same length, a 2 dimensional ndarray::Array is returned.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
T::Native: Float + NumCast,

pub fn rand_normal(
name: &str,
length: usize,
mean: f64,
std_dev: f64
) -> Result<Self>

This is supported on feature="random" only.

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> Self

This is supported on feature="random" only.

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform(name: &str, length: usize, low: f64, high: f64) -> Self

This is supported on feature="random" only.

Create ChunkedArray with samples from a Uniform distribution.

impl ChunkedArray<BooleanType>

pub fn rand_bernoulli(name: &str, length: usize, p: f64) -> Result<Self>

This is supported on feature="random" only.

Create ChunkedArray with samples from a Bernoulli distribution.

impl<T> ChunkedArray<T>

pub fn array_data(&self) -> Vec<ArrayDataRef>

Get Arrow ArrayData

pub fn null_bits(&self) -> Vec<(usize, Option<Buffer>)>

Get the null count and the buffer of bits representing null values

pub fn unpack_series_matching_type(
&self,
series: &Series
) -> Result<&ChunkedArray<T>>

Series to ChunkedArray

pub fn len(&self) -> usize

Combined length of all the chunks.

pub fn chunk_id(&self) -> &Vec<usize>

Unique id representing the number of chunks

pub fn chunks(&self) -> &Vec<ArrayRef>

A reference to the chunks

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn null_count(&self) -> usize

Count the null values.

pub fn limit(&self, num_elements: usize) -> Result<Self>

Take a view of top n elements

pub fn append_array(&mut self, other: ArrayRef) -> Result<()>

Append arrow array in place.

let mut array = Int32Chunked::new_from_slice("array", &[1, 2]);
let array_2 = Int32Chunked::new_from_slice("2nd", &[3]);

array.append(&array_2);
assert_eq!(Vec::from(&array), [Some(1), Some(2), Some(3)])

pub fn slice(&self, offset: usize, length: usize) -> Result<Self>

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn dtype(&self) -> &ArrowDataType

Get data type of ChunkedArray.

pub fn head(&self, length: Option<usize>) -> Self

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> Self

Get the tail of the ChunkedArray

pub fn append(&mut self, other: &Self) where
Self: Sized,

Append in place.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

impl<T> ChunkedArray<T> where
T: PolarsDataType,

pub fn new_from_chunks(name: &str, chunks: Vec<ArrayRef>) -> Self

Create a new ChunkedArray from existing chunks.

impl ChunkedArray<Utf8Type>

pub fn new_utf8_from_slice<S: AsRef<str>>(name: &str, v: &[S]) -> Self

Deprecated since 3.1:
Use new_from_slice

pub fn new_utf8_from_opt_slice<S: AsRef<str>>(
name: &str,
opt_v: &[Option<S>]
) -> Self

Deprecated since 3.1:
Use new_from_opt_slice

impl<T> ChunkedArray<T> where
T: ArrowPrimitiveType,

pub fn new_from_aligned_vec(name: &str, v: AlignedVec<T::Native>) -> Self

Create a new ChunkedArray by taking ownershipt of the AlignedVec. This operation is zero copy.

pub fn new_with_null_bitmap(
name: &str,
values: &[T::Native],
buffer: Option<Buffer>,
null_count: usize
) -> Self

Nullify values in slice with an existing null bitmap

pub fn new_from_owned_with_null_bitmap(
name: &str,
values: AlignedVec<T::Native>,
buffer: Option<Buffer>,
null_count: usize
) -> Self

Nullify values in slice with an existing null bitmap

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn cont_slice(&self) -> Result<&[T::Native]>

Contiguous slice

pub fn data_views(&self) -> Vec<&[T::Native]>

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn as_single_ptr(&mut self) -> usize

Rechunk and return a ptr to the start of the array

pub fn map<B, F>(&self, f: F) -> Result<Map<Copied<Iter<T::Native>>, F>> where
F: Fn(T::Native) -> B,

If cont_slice is successful a closure is mapped over the elements.

Example

use polars::prelude::*;
fn multiply(ca: &UInt32Chunked) -> Result<Series> {
    let mapped = ca.map(|v| v * 2)?;
    Ok(mapped.collect())
}

pub fn map_null_checks<B, F>(&self, f: F) -> Map<NumericChunkIterDispatch<T>, F> where
F: Fn(Option<T::Native>) -> B,

If cont_slice fails we can fallback on an iterator with null checks and map a closure over the elements.

Example

use polars::prelude::*;
use itertools::Itertools;
fn multiply(ca: &UInt32Chunked) -> Series {
    let mapped_result = ca.map(|v| v * 2);

    if let Ok(mapped) = mapped_result {
        mapped.collect()
    } else {
        ca
        .map_null_checks(|opt_v| opt_v.map(|v |v * 2)).collect()
    }
}

pub fn fold<F, B>(&self, init: B, f: F) -> Result<B> where
F: Fn(B, T::Native) -> B,

If cont_slice is successful a closure can be applied as aggregation

Example

use polars::prelude::*;
fn compute_sum(ca: &UInt32Chunked) -> Result<u32> {
    ca.fold(0, |acc, value| acc + value)
}

pub fn fold_null_checks<F, B>(&self, init: B, f: F) -> B where
F: Fn(B, Option<T::Native>) -> B,

If cont_slice fails we can fallback on an iterator with null checks and a closure for aggregation

Example

use polars::prelude::*;
fn compute_sum(ca: &UInt32Chunked) -> u32 {
    match ca.fold(0, |acc, value| acc + value) {
        // faster sum without null checks was successful
        Ok(sum) => sum,
        // Null values or multiple chunks in ChunkedArray, we need to do more bounds checking
        Err(_) => ca.fold_null_checks(0, |acc, opt_value| {
            match opt_value {
                Some(v) => acc + v,
                None => acc
            }
        })
    }
}

impl ChunkedArray<LargeListType>

pub fn explode(&self) -> Result<(Series, Vec<usize>)>