Struct polars::chunked_array::ChunkedArray

pub struct ChunkedArray<T>
where
    T: PolarsDataType,
{ /* private fields */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArrays are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of the same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.apply_values_generic(|v| v.cos() as f64)
}

Conversion between Series and ChunkedArrays

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> PolarsResult<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excellent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArrays use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make multiple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkedArray will slow down many operations that need random access because we have an extra indirection and indexes need to be mapped to the proper chunk. Arithmetic may also be slowed down by this. When multiplying two ChunkedArrays with different chunk sizes they cannot utilize SIMD for instance.

If you want to have predictable performance (no unexpected re-allocation of memory), it is advised to call the ChunkedArray::rechunk after multiple append operations.

See also ChunkedArray::extend for appends within a chunk.

Invariants

A ChunkedArray should always have at least a single ArrayRef.

The PolarsDataType T should always map to the correct ArrowDataType in the ArrayRef chunks.

Nested datatypes such as List and [Array] store the physical types instead of the logical type given by the datatype.

Implementations

impl ChunkedArray<BooleanType>

Booleans are cast to 1 or 0.

pub fn sum(&self) -> Option<u32>

pub fn min(&self) -> Option<bool>

pub fn max(&self) -> Option<bool>

pub fn mean(&self) -> Option<f64>

impl<T> ChunkedArray<T>
where T: PolarsDataType<IsNested = FalseT>, <T as PolarsDataType>::Physical<'a>: for<'a> TotalOrd,

pub fn append(&mut self, other: &ChunkedArray<T>)

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn apply_values_generic<'a, U, K, F>(&'a self, op: F) -> ChunkedArray<U>
where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> K, <U as PolarsDataType>::Array: ArrayFromIter<K>,

pub fn apply_nonnull_values_generic<'a, U, K, F>( &'a self, dtype: DataType, op: F ) -> ChunkedArray<U>
where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> K, <U as PolarsDataType>::Array: ArrayFromIterDtype<K> + ArrayFromIterDtype<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn try_apply_nonnull_values_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>
where U: PolarsDataType, F: FnMut(<T as PolarsDataType>::Physical<'a>) -> Result<K, E>, <U as PolarsDataType>::Array: ArrayFromIter<K> + ArrayFromIter<Option<K>>,

Applies a function only to the non-null elements, propagating nulls.

pub fn apply_generic<'a, U, K, F>(&'a self, op: F) -> ChunkedArray<U>
where U: PolarsDataType, F: FnMut(Option<<T as PolarsDataType>::Physical<'a>>) -> Option<K>, <U as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

pub fn try_apply_generic<'a, U, K, F, E>( &'a self, op: F ) -> Result<ChunkedArray<U>, E>
where U: PolarsDataType, F: FnMut(Option<<T as PolarsDataType>::Physical<'a>>) -> Result<Option<K>, E> + Copy, <U as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn cast_and_apply_in_place<F, S>(&self, f: F) -> ChunkedArray<S>
where F: Fn(<S as PolarsNumericType>::Native) -> <S as PolarsNumericType>::Native + Copy, S: PolarsNumericType,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

pub fn apply_in_place<F>(self, f: F) -> ChunkedArray<T>
where F: Fn(<T as PolarsNumericType>::Native) -> <T as PolarsNumericType>::Native + Copy,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn apply_mut<F>(&mut self, f: F)
where F: Fn(<T as PolarsNumericType>::Native) -> <T as PolarsNumericType>::Native + Copy,

impl ChunkedArray<StringType>

pub fn apply_mut<'a, F>(&'a self, f: F) -> ChunkedArray<StringType>
where F: FnMut(&'a str) -> &'a str,

pub fn apply_to_buffer<'a, F>(&'a self, f: F) -> ChunkedArray<StringType>
where F: FnMut(&'a str, &mut String),

Utility that reuses an string buffer to amortize allocations. Prefer this over an apply that returns an owned String.

impl ChunkedArray<BinaryType>

pub fn apply_mut<'a, F>(&'a self, f: F) -> ChunkedArray<BinaryType>
where F: FnMut(&'a [u8]) -> &'a [u8],

impl ChunkedArray<Float32Type>

Used to save compilation paths. Use carefully. Although this is safe, if misused it can lead to incorrect results.

pub fn apply_as_ints<F>(&self, f: F) -> Series
where F: Fn(&Series) -> Series,

impl ChunkedArray<Float64Type>

pub fn apply_as_ints<F>(&self, f: F) -> Series
where F: Fn(&Series) -> Series,

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn len(&self) -> usize

Get the length of the ChunkedArray

pub fn null_count(&self) -> usize

Return the number of null values in the ChunkedArray.

pub unsafe fn set_null_count(&mut self, null_count: u32)

Set the null count directly.

This can be useful after mutably adjusting the validity of the underlying arrays.

Safety

The new null count must match the total null count of the underlying arrays.

pub fn is_empty(&self) -> bool

Check if ChunkedArray is empty.

pub fn rechunk(&self) -> ChunkedArray<T>

pub fn slice(&self, offset: i64, length: usize) -> ChunkedArray<T>

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

pub fn limit(&self, num_elements: usize) -> ChunkedArray<T>
where ChunkedArray<T>: Sized,

Take a view of top n elements

pub fn head(&self, length: Option<usize>) -> ChunkedArray<T>
where ChunkedArray<T>: Sized,

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> ChunkedArray<T>
where ChunkedArray<T>: Sized,

Get the tail of the ChunkedArray

pub fn prune_empty_chunks(&mut self)

Remove empty chunks.

impl ChunkedArray<StringType>

pub fn to_decimal(&self, infer_length: usize) -> Result<Series, PolarsError>

Convert an StringChunked to a Series of DataType::Decimal. Scale needed for the decimal type are inferred. Parsing is not strict.
Scale inference assumes that all tested strings are well-formed numbers, and may produce unexpected results for scale if this is not the case.

If the decimal precision and scale are already known, consider using the cast method.

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn extend(&mut self, other: &ChunkedArray<T>)

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extend appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn for_each<'a, F>(&'a self, op: F)
where F: FnMut(Option<<T as PolarsDataType>::Physical<'a>>),

impl ChunkedArray<FixedSizeListType>

pub fn full_null_with_dtype( name: &str, length: usize, inner_dtype: &DataType, width: usize ) -> ChunkedArray<FixedSizeListType>

Available on crate feature dtype-array only.

impl ChunkedArray<ListType>

pub fn full_null_with_dtype( name: &str, length: usize, inner_dtype: &DataType ) -> ChunkedArray<ListType>

impl ChunkedArray<UInt32Type>

pub fn with_nullable_idx<T, F>(idx: &[NullableIdxSize], f: F) -> T
where F: FnOnce(&ChunkedArray<UInt32Type>) -> T,

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn is_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the null values.

pub fn is_not_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the valid values.

impl<T> ChunkedArray<T>
where ChunkedArray<T>: IntoSeries, T: PolarsFloatType, <T as PolarsNumericType>::Native: Float + IsFloat + SubAssign + Pow<<T as PolarsNumericType>::Native, Output = <T as PolarsNumericType>::Native>,

pub fn rolling_map_float<F>( &self, window_size: usize, f: F ) -> Result<ChunkedArray<T>, PolarsError>
where F: FnMut(&mut ChunkedArray<T>) -> Option<<T as PolarsNumericType>::Native>,

Apply a rolling custom function. This is pretty slow because of dynamic dispatch.

impl ChunkedArray<BinaryType>

pub unsafe fn to_string_unchecked(&self) -> ChunkedArray<StringType>

Safety

String is not validated

impl ChunkedArray<StringType>

pub fn as_binary(&self) -> ChunkedArray<BinaryType>

impl ChunkedArray<BooleanType>

pub fn any(&self) -> bool

Returns whether any of the values in the column are true.

Null values are ignored.

pub fn all(&self) -> bool

Returns whether all values in the array are true.

Null values are ignored.

pub fn any_kleene(&self) -> Option<bool>

Returns whether any of the values in the column are true.

The output is unknown (None) if the array contains any null values and no true values.

pub fn all_kleene(&self) -> Option<bool>

Returns whether all values in the column are true.

The output is unknown (None) if the array contains any null values and no false values.

impl<T> ChunkedArray<T>
where T: PolarsFloatType, <T as PolarsNumericType>::Native: Float,

pub fn is_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_not_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_finite(&self) -> ChunkedArray<BooleanType>

pub fn is_infinite(&self) -> ChunkedArray<BooleanType>

pub fn none_to_nan(&self) -> ChunkedArray<T>

Convert missing values to NaN values.

impl<T> ChunkedArray<T>
where T: PolarsFloatType, <T as PolarsNumericType>::Native: Float + Canonical,

pub fn to_canonical(&self) -> ChunkedArray<T>

impl ChunkedArray<ListType>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<Series>>

pub fn par_iter_indexed(&mut self) -> impl IndexedParallelIterator

impl ChunkedArray<StringType>

pub fn par_iter_indexed(&self) -> impl IndexedParallelIterator

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>>

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn iter(&self) -> impl PolarsIterator

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn to_ndarray( &self ) -> Result<ArrayBase<ViewRepr<&<T as PolarsNumericType>::Native>, Dim<[usize; 1]>>, PolarsError>

If data is aligned in a single chunk and has no Null values a zero copy view is returned as an ndarray

impl ChunkedArray<ListType>

pub fn to_ndarray<N>( &self ) -> Result<ArrayBase<OwnedRepr<<N as PolarsNumericType>::Native>, Dim<[usize; 2]>>, PolarsError>
where N: PolarsNumericType,

If all nested Series have the same length, a 2 dimensional ndarray::Array is returned.

impl ChunkedArray<FixedSizeListType>

pub fn amortized_iter( &self ) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array>>>>

This is an iterator over a ArrayChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

Safety

The lifetime of AmortSeries is bound to the iterator. Keeping it alive longer than the iterator is UB.

pub fn amortized_iter_with_name( &self, name: &str ) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array>>>>

This is an iterator over a ArrayChunked that save allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<>

If the returned AmortSeries is cloned, the local copy will be replaced and a new container will be set.

pub fn try_apply_amortized_to_list<F>( &self, f: F ) -> Result<ChunkedArray<ListType>, PolarsError>
where F: FnMut(AmortSeries) -> Result<Series, PolarsError>,

pub unsafe fn apply_amortized_same_type<F>( &self, f: F ) -> ChunkedArray<FixedSizeListType>
where F: FnMut(AmortSeries) -> Series,

Apply a closure F to each array.

Safety

Return series of F must has the same dtype and number of elements as input.

pub unsafe fn try_apply_amortized_same_type<F>( &self, f: F ) -> Result<ChunkedArray<FixedSizeListType>, PolarsError>
where F: FnMut(AmortSeries) -> Result<Series, PolarsError>,

Try apply a closure F to each array.

Safety

Return series of F must has the same dtype and number of elements as input if it is Ok.

pub unsafe fn zip_and_apply_amortized_same_type<'a, T, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> ChunkedArray<FixedSizeListType>
where T: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<<T as PolarsDataType>::Physical<'a>>) -> Option<Series>,

Zip with a ChunkedArray then apply a binary function F elementwise.

Safety

pub fn apply_amortized_generic<F, K, V>(&self, f: F) -> ChunkedArray<V>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Option<K> + Copy, <V as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

Apply a closure F elementwise.

pub fn try_apply_amortized_generic<F, K, V>( &self, f: F ) -> Result<ChunkedArray<V>, PolarsError>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Result<Option<K>, PolarsError> + Copy, <V as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

Try apply a closure F elementwise.

pub fn for_each_amortized<F>(&self, f: F)
where F: FnMut(Option<AmortSeries>),

impl ChunkedArray<FixedSizeListType>

pub fn inner_dtype(&self) -> &DataType

Get the inner data type of the fixed size list.

pub fn width(&self) -> usize

pub unsafe fn to_logical(&mut self, inner_dtype: DataType)

Safety

The caller must ensure that the logical type given fits the physical type of the array.

pub fn get_inner(&self) -> Series

Get the inner values as Series

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> Result<Series, PolarsError> ) -> Result<ChunkedArray<FixedSizeListType>, PolarsError>

Ignore the list indices and apply func to the inner type as Series.

impl<T> ChunkedArray<T>
where T: PolarsDataType, <<T as PolarsDataType>::Array as StaticArray>::ValueT<'a>: for<'a> AsRef<[u8]>,

pub fn to_bytes_hashes<'a>( &'a self, multithreaded: bool, hb: RandomState ) -> Vec<Vec<BytesHash<'a>>>

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn with_chunk<A>(name: &str, arr: A) -> ChunkedArray<T>
where A: Array, T: PolarsDataType<Array = A>,

pub fn with_chunk_like<A>(ca: &ChunkedArray<T>, arr: A) -> ChunkedArray<T>
where A: Array, T: PolarsDataType<Array = A>,

pub fn from_chunk_iter<I>(name: &str, iter: I) -> ChunkedArray<T>
where I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn from_chunk_iter_like<I>(ca: &ChunkedArray<T>, iter: I) -> ChunkedArray<T>
where I: IntoIterator, T: PolarsDataType<Array = <I as IntoIterator>::Item>, <I as IntoIterator>::Item: Array,

pub fn try_from_chunk_iter<I, A, E>( name: &str, iter: I ) -> Result<ChunkedArray<T>, E>
where I: IntoIterator<Item = Result<A, E>>, T: PolarsDataType<Array = A>, A: Array,

pub unsafe fn from_chunks( name: &str, chunks: Vec<Box<dyn Array>> ) -> ChunkedArray<T>

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn with_chunks(&self, chunks: Vec<Box<dyn Array>>) -> ChunkedArray<T>

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn from_chunks_and_dtype( name: &str, chunks: Vec<Box<dyn Array>>, dtype: DataType ) -> ChunkedArray<T>

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub fn full_null_like(ca: &ChunkedArray<T>, length: usize) -> ChunkedArray<T>

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn from_vec( name: &str, v: Vec<<T as PolarsNumericType>::Native> ) -> ChunkedArray<T>

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn from_vec_validity( name: &str, values: Vec<<T as PolarsNumericType>::Native>, buffer: Option<Bitmap> ) -> ChunkedArray<T>

Create a new ChunkedArray from a Vec and a validity mask.

pub unsafe fn mmap_slice( name: &str, values: &[<T as PolarsNumericType>::Native] ) -> ChunkedArray<T>

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl ChunkedArray<BooleanType>

pub unsafe fn mmap_slice( name: &str, values: &[u8], offset: usize, len: usize ) -> ChunkedArray<BooleanType>

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl ChunkedArray<ListType>

pub fn amortized_iter( &self ) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array>>>>

This is an iterator over a ListChunked that saves allocations. A Series is: 1. Arc<ChunkedArray> ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<>

If the returned AmortSeries is cloned, the local copy will be replaced and a new container will be set.

pub fn amortized_iter_with_name( &self, name: &str ) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array>>>>

See amortized_iter.

pub fn apply_amortized_generic<F, K, V>(&self, f: F) -> ChunkedArray<V>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Option<K> + Copy, <V as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

Apply a closure F elementwise.

pub fn try_apply_amortized_generic<F, K, V>( &self, f: F ) -> Result<ChunkedArray<V>, PolarsError>
where V: PolarsDataType, F: FnMut(Option<AmortSeries>) -> Result<Option<K>, PolarsError> + Copy, <V as PolarsDataType>::Array: ArrayFromIter<Option<K>>,

pub fn for_each_amortized<F>(&self, f: F)
where F: FnMut(Option<AmortSeries>),

pub fn zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> ChunkedArray<ListType>
where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<<T as PolarsDataType>::Physical<'a>>>, F: FnMut(Option<AmortSeries>, Option<<T as PolarsDataType>::Physical<'a>>) -> Option<Series>,

Zip with a ChunkedArray then apply a binary function F elementwise.

pub fn binary_zip_and_apply_amortized<'a, T, U, F>( &'a self, ca1: &'a ChunkedArray<T>, ca2: &'a ChunkedArray<U>, f: F ) -> ChunkedArray<ListType>
where T: PolarsDataType, U: PolarsDataType, F: FnMut(Option<AmortSeries>, Option<<T as PolarsDataType>::Physical<'a>>, Option<<U as PolarsDataType>::Physical<'a>>) -> Option<Series>,

pub fn try_zip_and_apply_amortized<'a, T, I, F>( &'a self, ca: &'a ChunkedArray<T>, f: F ) -> Result<ChunkedArray<ListType>, PolarsError>
where T: PolarsDataType, &'a ChunkedArray<T>: IntoIterator<IntoIter = I>, I: TrustedLen<Item = Option<<T as PolarsDataType>::Physical<'a>>>, F: FnMut(Option<AmortSeries>, Option<<T as PolarsDataType>::Physical<'a>>) -> Result<Option<Series>, PolarsError>,

pub fn apply_amortized<F>(&self, f: F) -> ChunkedArray<ListType>
where F: FnMut(AmortSeries) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<F>( &self, f: F ) -> Result<ChunkedArray<ListType>, PolarsError>
where F: FnMut(AmortSeries) -> Result<Series, PolarsError>,

impl ChunkedArray<ListType>

pub fn inner_dtype(&self) -> &DataType

Get the inner data type of the list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

pub fn set_fast_explode(&mut self)

pub fn _can_fast_explode(&self) -> bool

pub unsafe fn to_logical(&mut self, inner_dtype: DataType)

Set the logical type of the ListChunked.

Safety

The caller must ensure that the logical type given fits the physical type of the array.

pub fn get_inner(&self) -> Series

Get the inner values as Series, ignoring the list offsets.

pub fn iter_offsets(&self) -> impl Iterator<Item = i64>

Returns an iterator over the offsets of this chunked array.

The offsets are returned as though the array consisted of a single chunk.

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> Result<Series, PolarsError> ) -> Result<ChunkedArray<ListType>, PolarsError>

Ignore the list indices and apply func to the inner type as Series.

impl ChunkedArray<Int32Type>

pub fn into_date(self) -> Logical<DateType, Int32Type>

impl ChunkedArray<Int64Type>

pub fn into_datetime( self, timeunit: TimeUnit, tz: Option<String> ) -> Logical<DatetimeType, Int64Type>

impl ChunkedArray<Int128Type>

pub fn into_decimal_unchecked( self, precision: Option<usize>, scale: usize ) -> Logical<DecimalType, Int128Type>

pub fn into_decimal( self, precision: Option<usize>, scale: usize ) -> Result<Logical<DecimalType, Int128Type>, PolarsError>

impl ChunkedArray<Int64Type>

pub fn into_duration( self, timeunit: TimeUnit ) -> Logical<DurationType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_time(self) -> Logical<TimeType, Int64Type>

impl<T> ChunkedArray<ObjectType<T>>
where T: PolarsObject,

pub fn new_from_vec(name: &str, v: Vec<T>) -> ChunkedArray<ObjectType<T>>

pub fn new_from_vec_and_validity( name: &str, v: Vec<T>, validity: Bitmap ) -> ChunkedArray<ObjectType<T>>

pub fn new_empty(name: &str) -> ChunkedArray<ObjectType<T>>

impl<T> ChunkedArray<ObjectType<T>>
where T: PolarsObject,

pub unsafe fn get_object_unchecked( &self, index: usize ) -> Option<&(dyn PolarsObjectSafe + 'static)>

Get a hold to an object that can be formatted or downcasted via the Any trait.

Safety

No bounds checks

pub fn get_object( &self, index: usize ) -> Option<&(dyn PolarsObjectSafe + 'static)>

Get a hold to an object that can be formatted or downcasted via the Any trait.

impl<T> ChunkedArray<T>
where T: PolarsNumericType, Standard: Distribution<<T as PolarsNumericType>::Native>,

pub fn init_rand( size: usize, null_density: f32, seed: Option<u64> ) -> ChunkedArray<T>

impl<T> ChunkedArray<T>
where T: PolarsDataType, ChunkedArray<T>: ChunkTake<ChunkedArray<UInt32Type>>,

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> Result<ChunkedArray<T>, PolarsError>

Sample n datapoints from this ChunkedArray.

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> Result<ChunkedArray<T>, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T>
where T: PolarsNumericType, <T as PolarsNumericType>::Native: Float,

pub fn rand_normal( name: &str, length: usize, mean: f64, std_dev: f64 ) -> Result<ChunkedArray<T>, PolarsError>

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> ChunkedArray<T>

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform( name: &str, length: usize, low: f64, high: f64 ) -> ChunkedArray<T>

Create ChunkedArray with samples from a Uniform distribution.

impl ChunkedArray<BooleanType>

pub fn rand_bernoulli( name: &str, length: usize, p: f64 ) -> Result<ChunkedArray<BooleanType>, PolarsError>

Create ChunkedArray with samples from a Bernoulli distribution.

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn to_vec(&self) -> Vec<Option<<T as PolarsNumericType>::Native>>

Convert to a Vec of Option<T::Native>.

pub fn to_vec_null_aware( &self ) -> Either<Vec<<T as PolarsNumericType>::Native>, Vec<Option<<T as PolarsNumericType>::Native>>>

Convert to a Vec but dont return Option<T::Native> if there are no null values

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn new_with_compute_len( field: Arc<Field>, chunks: Vec<Box<dyn Array>> ) -> ChunkedArray<T>

Create a new ChunkedArray and compute its length and null_count.

If you want to explicitly the length and null_count, look at ChunkedArray::new_with_dims

pub unsafe fn new_with_dims( field: Arc<Field>, chunks: Vec<Box<dyn Array>>, length: u32, null_count: u32 ) -> ChunkedArray<T>

Create a new ChunkedArray and explicitly set its length and null_count.

If you want to compute the length and null_count, look at ChunkedArray::new_with_compute_len

Safety

The length and null_count must be correct.

pub fn effective_metadata(&self) -> &Metadata<T>

Get a reference to the used Metadata

This results a reference to an empty Metadata if its unset for this ChunkedArray.

pub fn metadata(&self) -> Option<&Metadata<T>>

Get a reference to the ChunkedArrays Metadata

pub fn metadata_arc(&self) -> Option<&Arc<Metadata<T>>>

Get a reference to Arc that contains the ChunkedArrays Metadata

pub fn metadata_owned_arc(&self) -> Arc<Metadata<T>>

Get a Arc that contains the ChunkedArrays Metadata

pub fn metadata_mut(&mut self) -> &mut Arc<Metadata<T>>

Get a mutable reference to the Arc that contains the ChunkedArrays Metadata

pub fn unset_fast_explode_list(&mut self)

pub fn set_fast_explode_list(&mut self, value: bool)

pub fn get_fast_explode_list(&self) -> bool

pub fn get_flags(&self) -> MetadataFlags

pub fn is_sorted_flag(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

Set the sorted bit meta info.

pub fn with_sorted_flag(&self, sorted: IsSorted) -> ChunkedArray<T>

Set the sorted bit meta info.

pub fn get_min_value(&self) -> Option<&<T as PolarsDataType>::OwnedPhysical>

pub fn get_max_value(&self) -> Option<&<T as PolarsDataType>::OwnedPhysical>

pub fn get_distinct_count(&self) -> Option<u32>

pub fn merge_metadata(&mut self, md: Metadata<T>)

pub fn copy_metadata_cast<O>( &mut self, other: &ChunkedArray<O>, props: MetadataProperties )
where O: PolarsDataType,

Copies Metadata properties specified by props from other with different underlying PolarsDataType into self.

This does not copy the properties with a different type between the Metadatas (e.g. min_value and max_value) and will panic on debug builds if that is attempted.

pub fn copy_metadata( &mut self, other: &ChunkedArray<T>, props: MetadataProperties )

Copies Metadata properties specified by props from other into self.

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn last_non_null(&self) -> Option<usize>

Get the index of the last non null value in this ChunkedArray.

pub fn drop_nulls(&self) -> ChunkedArray<T>

pub fn iter_validities( &self ) -> Map<Iter<'_, Box<dyn Array>>, fn(_: &Box<dyn Array>) -> Option<&Bitmap>>

Get the buffer of bits representing null values

pub fn has_validity(&self) -> bool

Return if any the chunks in this ChunkedArray have a validity bitmap. no bitmap means no null values.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn clear(&self) -> ChunkedArray<T>

pub fn unpack_series_matching_type( &self, series: &Series ) -> Result<&ChunkedArray<T>, PolarsError>

Series to ChunkedArray<T>

pub fn chunk_lengths( &self ) -> Map<Iter<'_, Box<dyn Array>>, fn(_: &Box<dyn Array>) -> usize>

Returns an iterator over the lengths of the chunks of the array.

pub fn chunks(&self) -> &Vec<Box<dyn Array>>

A reference to the chunks

pub unsafe fn chunks_mut(&mut self) -> &mut Vec<Box<dyn Array>>

A mutable reference to the chunks

Safety

The caller must ensure to not change the DataType or length of any of the chunks. And the null_count remains correct.

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

pub fn with_name(self, name: &str) -> ChunkedArray<T>

Return this ChunkedArray with a new name.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn get(&self, idx: usize) -> Option<<T as PolarsDataType>::Physical<'_>>

Get a single value from this ChunkedArray. If the return values is None this indicates a NULL value.

Panics

This function will panic if idx is out of bounds.

pub unsafe fn get_unchecked( &self, idx: usize ) -> Option<<T as PolarsDataType>::Physical<'_>>

Get a single value from this ChunkedArray. If the return values is None this indicates a NULL value.

Safety

It is the callers responsibility that the idx < self.len().

pub unsafe fn value_unchecked( &self, idx: usize ) -> <T as PolarsDataType>::Physical<'_>

Get a single value from this ChunkedArray. Null values are ignored and the returned value could be garbage if it was masked out by NULL. Note that the value always is initialized.

Safety

It is the callers responsibility that the idx < self.len().

pub fn last(&self) -> Option<<T as PolarsDataType>::Physical<'_>>

impl ChunkedArray<ListType>

pub fn get_as_series(&self, idx: usize) -> Option<Series>

impl ChunkedArray<FixedSizeListType>

pub fn get_as_series(&self, idx: usize) -> Option<Series>

Available on crate feature dtype-array only.

impl<T> ChunkedArray<T>
where T: PolarsDataType,

pub fn layout(&self) -> ChunkedArrayLayout<'_, T>

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn cont_slice( &self ) -> Result<&[<T as PolarsNumericType>::Native], PolarsError>

Returns the values of the array as a contiguous slice.

pub fn data_views(&self) -> impl DoubleEndedIterator

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter( &self ) -> impl Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl<T> ChunkedArray<T>
where T: PolarsIntegerType, <T as PolarsNumericType>::Native: ToPrimitive + FromPrimitive + Debug,

pub fn group_tuples_perfect( &self, max: usize, multithreaded: bool, group_capacity: usize ) -> GroupsProxy

impl<T> ChunkedArray<T>
where T: PolarsNumericType,

pub fn new_vec( name: &str, v: Vec<<T as PolarsNumericType>::Native> ) -> ChunkedArray<T>

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T> ChunkedArray<T>
where T: PolarsNumericType, ChunkedArray<T>: IntoSeries,

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N>(&self, lhs: N) -> ChunkedArray<T>
where N: Num + NumCast,

Apply lhs - self

pub fn lhs_div<N>(&self, lhs: N) -> ChunkedArray<T>
where N: Num + NumCast,

Apply lhs / self

pub fn lhs_rem<N>(&self, lhs: N) -> ChunkedArray<T>
where N: Num + NumCast,

Apply lhs % self