Struct polars::chunked_array::ChunkedArray

pub struct ChunkedArray<T> { /* fields omitted */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArray's are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of te same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine(ca: &Float32Chunked) -> Float32Chunked {
    ca.apply(|v| v.cos())
}

If we would like to cast the result we could use a Rust Iterator instead of an apply method. Note that Iterators are slightly slower as the null values aren't ignored implicitly.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.into_iter()
        .map(|opt_v| {
        opt_v.map(|v| v.cos() as f64)
    }).collect()
}

Another option is to first cast and then use an apply.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.cast::<Float64Type>()
        .unwrap()
        .apply(|v| v.cos())
}

Conversion between Series and ChunkedArray's

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> Result<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excelent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.into_iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.into_iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArray's use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make mutliple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkArray will slow down the Iterators, arithmetic and other operations. When multiplying two ChunkArray's with different chunk sizes they cannot utilize SIMD for instance. However, when chunk size don't match, Iterators will be used to do the operation (instead of arrows upstream implementation, which may utilize SIMD) and the result will be a single chunked array.

The key takeaway is that by applying operations on a ChunkArray of multiple chunks, the results will converge to a ChunkArray of a single chunk! It is recommended to leave them as is. If you want to have predictable performance (no unexpected re-allocation of memory), it is adviced to call the rechunk after multiple append operations.

Implementations

impl<T> ChunkedArray<T> where
T: PolarsDataType,

pub fn new_from_chunks(name: &str, chunks: Vec<ArrayRef>) -> Self

Create a new ChunkedArray from existing chunks.

pub fn null_bits(&self) -> Vec<(usize, Option<Buffer>)>

Get the null count and the buffer of bits representing null values

pub fn into_series(self) -> Series

Wrap as Series

impl<T> ChunkedArray<T> where
T: PolarsDataType,
ChunkedArray<T>: ChunkOps,

pub fn chunk_id(&self) -> &Vec<usize>

pub fn chunks(&self) -> &Vec<ArrayRef>

A reference to the chunks

pub fn null_count(&self) -> usize

Count the null values.

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn u8(self) -> Result<UInt8Chunked>

Downcast generic ChunkedArray<T> to u8.

pub fn u16(self) -> Result<UInt16Chunked>

Downcast generic ChunkedArray<T> to u16.

pub fn u32(self) -> Result<UInt32Chunked>

Downcast generic ChunkedArray<T> to u32.

pub fn u64(self) -> Result<UInt64Chunked>

Downcast generic ChunkedArray<T> to u64.

pub fn i8(self) -> Result<Int8Chunked>

Downcast generic ChunkedArray<T> to i8.

pub fn i16(self) -> Result<Int16Chunked>

Downcast generic ChunkedArray<T> to i16.

pub fn i32(self) -> Result<Int32Chunked>

Downcast generic ChunkedArray<T> to i32.

pub fn i64(self) -> Result<Int64Chunked>

Downcast generic ChunkedArray<T> to i64.

pub fn f32(self) -> Result<Float32Chunked>

Downcast generic ChunkedArray<T> to f32.

pub fn f64(self) -> Result<Float64Chunked>

Downcast generic ChunkedArray<T> to f64.

pub fn bool(self) -> Result<BooleanChunked>

Downcast generic ChunkedArray<T> to bool.

pub fn utf8(self) -> Result<Utf8Chunked>

Downcast generic ChunkedArray<T> to UTF-8 encoded string.

pub fn date32(self) -> Result<Date32Chunked>

Downcast generic ChunkedArray<T> to date32.

pub fn date64(self) -> Result<Date64Chunked>

Downcast generic ChunkedArray<T> to date32.

pub fn time32_second(self) -> Result<Time32SecondChunked>

Downcast generic ChunkedArray<T> to time32 with milliseconds unit.

pub fn time32_millisecond(self) -> Result<Time32MillisecondChunked>

Downcast generic ChunkedArray<T> to time32 with milliseconds unit.

pub fn time64_nanosecond(self) -> Result<Time64NanosecondChunked>

Downcast generic ChunkedArray<T> to time64 with nanoseconds unit.

pub fn time64_microsecond(self) -> Result<Time64MicrosecondChunked>

Downcast generic ChunkedArray<T> to time64 with microseconds unit.

pub fn duration_nanosecond(self) -> Result<DurationNanosecondChunked>

Downcast generic ChunkedArray<T> to duration with nanoseconds unit.

pub fn duration_microsecond(self) -> Result<DurationMicrosecondChunked>

Downcast generic ChunkedArray<T> to duration with microseconds unit.

pub fn duration_second(self) -> Result<DurationSecondChunked>

Downcast generic ChunkedArray<T> to duration with seconds unit.

pub fn timestamp_nanosecond(self) -> Result<TimestampNanosecondChunked>

Downcast generic ChunkedArray<T> to timestamp with nanoseconds unit.

pub fn timestamp_microsecond(self) -> Result<TimestampMicrosecondChunked>

Downcast generic ChunkedArray<T> to timestamp with microseconds unit.

pub fn timestamp_millisecond(self) -> Result<TimestampMillisecondChunked>

Downcast generic ChunkedArray<T> to timestamp with milliseconds unit.

pub fn timestamp_second(self) -> Result<TimestampSecondChunked>

Downcast generic ChunkedArray<T> to timestamp with seconds unit.

pub fn limit(&self, num_elements: usize) -> Result<Self>

Take a view of top n elements

pub fn filter(&self, filter: &BooleanChunked) -> Result<Self>

Filter values in the ChunkedArray with a boolean mask.

let array = Int32Chunked::new_from_slice("array", &[1, 2, 3]);
let mask = BooleanChunked::new_from_slice("mask", &[true, false, true]);

let filtered = array.filter(&mask).unwrap();
assert_eq!(Vec::from(&filtered), [Some(1), Some(3)])

pub fn append_array(&mut self, other: ArrayRef) -> Result<()>

Append arrow array in place.

let mut array = Int32Chunked::new_from_slice("array", &[1, 2]);
let array_2 = Int32Chunked::new_from_slice("2nd", &[3]);

array.append(&array_2);
assert_eq!(Vec::from(&array), [Some(1), Some(2), Some(3)])

pub fn len(&self) -> usize

Combined length of all the chunks.

pub fn get(&self, index: usize) -> AnyType

Get a single value. Beware this is slow.

pub fn slice(&self, offset: usize, length: usize) -> Result<Self>

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

pub fn head(&self, length: Option<usize>) -> Self

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> Self

Get the tail of the ChunkedArray

pub fn append(&mut self, other: &Self) where
Self: Sized,

Append in place.

impl ChunkedArray<Utf8Type>

pub fn new_utf8_from_slice<S: AsRef<str>>(name: &str, v: &[S]) -> Self

pub fn new_utf8_from_opt_slice<S: AsRef<str>>(
name: &str,
opt_v: &[Option<S>]
) -> Self

impl<T> ChunkedArray<T> where
T: PolarsDataType,
ChunkedArray<T>: ChunkOps,

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

impl<T> ChunkedArray<T> where
T: ArrowPrimitiveType,

pub fn new_from_slice(name: &str, v: &[T::Native]) -> Self

pub fn new_from_opt_slice(name: &str, opt_v: &[Option<T::Native>]) -> Self

pub fn new_with_null_bitmap(
name: &str,
values: &[T::Native],
buffer: Option<Buffer>,
null_count: usize
) -> Self

Nullify values in slice with an existing null bitmap

pub fn new_from_owned_with_null_bitmap(
name: &str,
values: AlignedVec<T::Native>,
buffer: Option<Buffer>,
null_count: usize
) -> Self

Nullify values in slice with an existing null bitmap

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn cont_slice(&self) -> Result<&[T::Native]>

Contiguous slice

pub fn data_views(&self) -> Vec<&[T::Native]>

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn as_single_ptr(&mut self) -> usize

Rechunk and return a ptr to the start of the array

pub fn map<B, F>(&self, f: F) -> Result<Map<Copied<Iter<T::Native>>, F>> where
F: Fn(T::Native) -> B,

If cont_slice is successful a closure is mapped over the elements.

Example

use polars::prelude::*;
fn multiply(ca: &UInt32Chunked) -> Result<Series> {
    let mapped = ca.map(|v| v * 2)?;
    Ok(mapped.collect())
}

pub fn map_null_checks<'a, B, F>(
&'a self,
f: F
) -> Map<Box<dyn ExactSizeDoubleEndedIterator<Item = Option<T::Native>> + 'a>, F> where
F: Fn(Option<T::Native>) -> B,

If cont_slice fails we can fallback on an iterator with null checks and map a closure over the elements.

Example

use polars::prelude::*;
use itertools::Itertools;
fn multiply(ca: &UInt32Chunked) -> Series {
    let mapped_result = ca.map(|v| v * 2);

    if let Ok(mapped) = mapped_result {
        mapped.collect()
    } else {
        ca
        .map_null_checks(|opt_v| opt_v.map(|v |v * 2)).collect()
    }
}

pub fn fold<F, B>(&self, init: B, f: F) -> Result<B> where
F: Fn(B, T::Native) -> B,

If cont_slice is successful a closure can be applied as aggregation

Example

use polars::prelude::*;
fn compute_sum(ca: &UInt32Chunked) -> Result<u32> {
    ca.fold(0, |acc, value| acc + value)
}

pub fn fold_null_checks<F, B>(&self, init: B, f: F) -> B where
F: Fn(B, Option<T::Native>) -> B,

If cont_slice fails we can fallback on an iterator with null checks and a closure for aggregation

Example

use polars::prelude::*;
fn compute_sum(ca: &UInt32Chunked) -> u32 {
    match ca.fold(0, |acc, value| acc + value) {
        // faster sum without null checks was successful
        Ok(sum) => sum,
        // Null values or multiple chunks in ChunkedArray, we need to do more bounds checking
        Err(_) => ca.fold_null_checks(0, |acc, opt_value| {
            match opt_value {
                Some(v) => acc + v,
                None => acc
            }
        })
    }
}