Struct polars::chunked_array::ChunkedArray

pub struct ChunkedArray<T> { /* fields omitted */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArray’s are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of the same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine(ca: &Float32Chunked) -> Float32Chunked {
    ca.apply(|v| v.cos())
}

If we would like to cast the result we could use a Rust Iterator instead of an apply method. Note that Iterators are slightly slower as the null values aren’t ignored implicitly.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.into_iter()
        .map(|opt_v| {
        opt_v.map(|v| v.cos() as f64)
    }).collect()
}

Another option is to first cast and then use an apply.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.apply_cast_numeric(|v| v.cos() as f64)
}

Conversion between Series and ChunkedArray’s

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> Result<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excellent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.into_iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.into_iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArray’s use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make multiple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkArray will slow down many operations that need random access because we have an extra indirection and indexes need to be mapped to the proper chunk. Arithmetic may also be slowed down by this. When multiplying two ChunkArray's with different chunk sizes they cannot utilize SIMD for instance.

If you want to have predictable performance (no unexpected re-allocation of memory), it is advised to call the rechunk after multiple append operations.

See also ChunkedArray::extend for appends within a chunk.

Implementations

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
<T as PolarsNumericType>::Native: Signed,

pub fn abs(&self) -> ChunkedArray<T>

Convert all values to their absolute/positive value.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn append(&mut self, other: &ChunkedArray<T>)

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn extend(&mut self, other: &ChunkedArray<T>)

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extent appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
<T as PolarsNumericType>::Native: Float,
ChunkedArray<T>: IntoSeries,

pub fn rolling_mean(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling mean (moving mean) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be multiplied with the weights given by the weights vector. The resulting values will be aggregated to their mean.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn rolling_sum(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling sum (moving sum) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be multiplied with the weights given by the weights vector. The resulting values will be aggregated to their sum.

pub fn rolling_median(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling median (moving median) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be weighted according to the weights vector.

pub fn rolling_quantile(
&self,
quantile: f64,
interpolation: QuantileInterpolOptions,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling quantile (moving quantile) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be weighted according to the weights vector.

pub fn rolling_min(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling min (moving min) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be multiplied with the weights given by the weights vector. The resulting values will be aggregated to their min.

pub fn rolling_max(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling max (moving max) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be multiplied with the weights given by the weights vector. The resulting values will be aggregated to their max.

impl<T> ChunkedArray<T> where
T: PolarsFloatType,
ChunkedArray<T>: IntoSeries,
<T as PolarsNumericType>::Native: Float,

pub fn rolling_apply_float<F>(
&self,
window_size: usize,
f: F
) -> Result<ChunkedArray<T>, PolarsError> where
F: Fn(&ChunkedArray<T>) -> Option<<T as PolarsNumericType>::Native>,

Apply a rolling custom function. This is pretty slow because of dynamic dispatch.

pub fn rolling_var(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling var (moving var) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be multiplied with the weights given by the weights vector. The resulting values will be aggregated to their var.

pub fn rolling_std(
&self,
options: RollingOptions
) -> Result<Series, PolarsError>

Apply a rolling std (moving std) over the values in this array. A window of length window_size will traverse the array. The values that fill this window will (optionally) be multiplied with the weights given by the weights vector. The resulting values will be aggregated to their std.

impl ChunkedArray<BooleanType>

pub fn arg_true(&self) -> ChunkedArray<UInt32Type>

impl ChunkedArray<BooleanType>

pub fn all(&self) -> bool

Check if all values are true

pub fn any(&self) -> bool

Check if any value is true

impl<T> ChunkedArray<T> where
T: PolarsFloatType,
<T as PolarsNumericType>::Native: Float,

pub fn is_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_not_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_finite(&self) -> ChunkedArray<BooleanType>

pub fn is_infinite(&self) -> ChunkedArray<BooleanType>

pub fn none_to_nan(&self) -> ChunkedArray<T>

Convert missing values to NaN values.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn to_ndarray(
&self
) -> Result<ArrayBase<ViewRepr<&<T as PolarsNumericType>::Native>, Dim<[usize; 1]>>, PolarsError>

This is supported on crate feature ndarray only.

If data is aligned in a single chunk and has no Null values a zero copy view is returned as an ndarray

impl ChunkedArray<ListType>

pub fn to_ndarray<N>(
&self
) -> Result<ArrayBase<OwnedRepr<<N as PolarsNumericType>::Native>, Dim<[usize; 2]>>, PolarsError> where
N: PolarsNumericType,

This is supported on crate feature ndarray only.

If all nested Series have the same length, a 2 dimensional ndarray::Array is returned.

impl ChunkedArray<CategoricalType>

pub fn get_categorical_map(&self) -> Option<&Arc<RevMapping>>

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> CatIter<'_>

Create an [Iterator] that iterates over the &str values of the [CategoricalChunked].

impl ChunkedArray<ListType>

pub fn amortized_iter(
&self
) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array + 'static, Global>>>>

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

pub fn apply_amortized<'a, F>(&'a self, f: F) -> ChunkedArray<ListType> where
F: FnMut(UnstableSeries<'a>) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<'a, F>(
&'a self,
f: F
) -> Result<ChunkedArray<ListType>, PolarsError> where
F: FnMut(UnstableSeries<'a>) -> Result<Series, PolarsError>,

impl ChunkedArray<ListType>

pub fn lst_join(
&self,
separator: &str
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

In case the inner dtype DataType::Utf8, the individual items will be joined into a single string separated by separator.

pub fn lst_max(&self) -> Series

pub fn lst_min(&self) -> Series

pub fn lst_sum(&self) -> Series

pub fn lst_mean(&self) -> ChunkedArray<Float64Type>

pub fn lst_sort(&self, reverse: bool) -> ChunkedArray<ListType>

pub fn lst_reverse(&self) -> ChunkedArray<ListType>

pub fn lst_unique(&self) -> Result<ChunkedArray<ListType>, PolarsError>

pub fn lst_lengths(&self) -> ChunkedArray<UInt32Type>

pub fn lst_get(&self, idx: i64) -> Result<Series, PolarsError>

Get the value by index in the sublists. So index 0 would return the first item of every sublist and index -1 would return the last item of every sublist if an index is out of bounds, it will return a None.

pub fn lst_concat(
&self,
other: &[Series]
) -> Result<ChunkedArray<ListType>, PolarsError>

impl ChunkedArray<ListType>

pub fn set_fast_explode(&mut self)

pub fn to_logical(&mut self, inner_dtype: DataType)

impl ChunkedArray<Int32Type>

pub fn into_date(self) -> Logical<DateType, Int32Type>

impl ChunkedArray<Int64Type>

pub fn into_datetime(
self,
timeunit: TimeUnit,
tz: Option<String>
) -> Logical<DatetimeType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_duration(
self,
timeunit: TimeUnit
) -> Logical<DurationType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_time(self) -> Logical<TimeType, Int64Type>

impl<T> ChunkedArray<ObjectType<T>> where
T: PolarsObject,

pub fn new_from_vec(
name: &str,
v: Vec<T, Global>
) -> ChunkedArray<ObjectType<T>>

impl<T> ChunkedArray<ObjectType<T>> where
T: PolarsObject,

pub unsafe fn get_object_unchecked(
&self,
index: usize
) -> Option<&(dyn PolarsObjectSafe + 'static)>

Get a hold to an object that can be formatted or downcasted via the Any trait.

Safety

No bounds checks

pub fn get_object(
&self,
index: usize
) -> Option<&(dyn PolarsObjectSafe + 'static)>

Get a hold to an object that can be formatted or downcasted via the Any trait.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
Standard: Distribution<<T as PolarsNumericType>::Native>,

pub fn init_rand(size: usize, null_density: f32, seed: u64) -> ChunkedArray<T>

impl<T> ChunkedArray<T> where
ChunkedArray<T>: ChunkTake,

pub fn sample_n(
&self,
n: usize,
with_replacement: bool,
seed: u64
) -> Result<ChunkedArray<T>, PolarsError>

Sample n datapoints from this ChunkedArray.

pub fn sample_frac(
&self,
frac: f64,
with_replacement: bool,
seed: u64
) -> Result<ChunkedArray<T>, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
<T as PolarsNumericType>::Native: Float,

pub fn rand_normal(
name: &str,
length: usize,
mean: f64,
std_dev: f64
) -> Result<ChunkedArray<T>, PolarsError>

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> ChunkedArray<T>

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform(
name: &str,
length: usize,
low: f64,
high: f64
) -> ChunkedArray<T>

Create ChunkedArray with samples from a Uniform distribution.

impl ChunkedArray<BooleanType>

pub fn rand_bernoulli(
name: &str,
length: usize,
p: f64
) -> Result<ChunkedArray<BooleanType>, PolarsError>

Create ChunkedArray with samples from a Bernoulli distribution.

impl ChunkedArray<Utf8Type>

pub fn json_path_match(
&self,
json_path: &str
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Extract json path, first match Refer to https://goessner.net/articles/JsonPath/

impl ChunkedArray<Utf8Type>

pub fn hex_decode(
&self,
strict: Option<bool>
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

pub fn hex_encode(&self) -> ChunkedArray<Utf8Type>

pub fn base64_decode(
&self,
strict: Option<bool>
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

pub fn base64_encode(&self) -> ChunkedArray<Utf8Type>

impl ChunkedArray<Utf8Type>

pub fn str_lengths(&self) -> ChunkedArray<UInt32Type>

Get the length of the string values.

pub fn contains(
&self,
pat: &str
) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if strings contain a regex pattern

pub fn replace(
&self,
pat: &str,
val: &str
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Replace the leftmost (sub)string by a regex pattern

pub fn replace_all(
&self,
pat: &str,
val: &str
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Replace all (sub)strings by a regex pattern

pub fn extract(
&self,
pat: &str,
group_index: usize
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Extract the nth capture group from pattern

pub fn to_lowercase(&self) -> ChunkedArray<Utf8Type>

Modify the strings to their lowercase equivalent

pub fn to_uppercase(&self) -> ChunkedArray<Utf8Type>

Modify the strings to their uppercase equivalent

pub fn concat(&self, other: &ChunkedArray<Utf8Type>) -> ChunkedArray<Utf8Type>

Concat with the values from a second Utf8Chunked

pub fn str_slice(
&self,
start: i64,
length: Option<u64>
) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Slice the string values Determines a substring starting from start and with optional length length of each of the elements in array. start can be negative, in which case the start counts from the end of the string.

impl ChunkedArray<Utf8Type>

pub fn as_time(
&self,
fmt: Option<&str>
) -> Result<Logical<TimeType, Int64Type>, PolarsError>

pub fn as_date_not_exact(
&self,
fmt: Option<&str>
) -> Result<Logical<DateType, Int32Type>, PolarsError>

pub fn as_datetime_not_exact(
&self,
fmt: Option<&str>,
tu: TimeUnit
) -> Result<Logical<DatetimeType, Int64Type>, PolarsError>

pub fn as_date(
&self,
fmt: Option<&str>
) -> Result<Logical<DateType, Int32Type>, PolarsError>

pub fn as_datetime(
&self,
fmt: Option<&str>,
tu: TimeUnit
) -> Result<Logical<DatetimeType, Int64Type>, PolarsError>

impl<T> ChunkedArray<T>

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn iter_validities(&self) -> impl Iterator<Item = Option<&Bitmap>>

Get the buffer of bits representing null values

pub fn has_validity(&self) -> bool

Return if any the chunks in this [ChunkedArray] have a validity bitmap. no bitmap means no null values.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit it’s length.

pub fn unpack_series_matching_type(
&self,
series: &Series
) -> Result<&ChunkedArray<T>, PolarsError>

Series to ChunkedArray

pub fn chunk_id(
&self
) -> Map<Iter<'_, Arc<dyn Array + 'static>>, fn(&Arc<dyn Array + 'static>) -> usize>

Unique id representing the number of chunks

pub fn chunks(&self) -> &Vec<Arc<dyn Array + 'static>, Global>

A reference to the chunks

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn null_count(&self) -> usize

Count the null values.

pub fn append_array(
&mut self,
other: Arc<dyn Array + 'static>
) -> Result<(), PolarsError>

Append arrow array in place.

let mut array = Int32Chunked::new("array", &[1, 2]);
let array_2 = Int32Chunked::new("2nd", &[3]);

array.append(&array_2);
assert_eq!(Vec::from(&array), [Some(1), Some(2), Some(3)])

pub fn is_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the null values.

pub fn is_not_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the valid values.

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

impl<T> ChunkedArray<T> where
T: PolarsDataType,

pub fn from_chunks(
name: &str,
chunks: Vec<Arc<dyn Array + 'static>, Global>
) -> ChunkedArray<T>

Create a new ChunkedArray from existing chunks.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn from_vec(
name: &str,
v: Vec<<T as PolarsNumericType>::Native, Global>
) -> ChunkedArray<T>

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn new_from_owned_with_null_bitmap(
name: &str,
values: Vec<<T as PolarsNumericType>::Native, Global>,
buffer: Option<Bitmap>
) -> ChunkedArray<T>

Nullify values in slice with an existing null bitmap

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn cont_slice(
&self
) -> Result<&[<T as PolarsNumericType>::Native], PolarsError>

Contiguous slice

pub fn data_views(
&self
) -> impl Iterator<Item = &[<T as PolarsNumericType>::Native]> + DoubleEndedIterator

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter(
&self
) -> impl Iterator<Item = <T as PolarsNumericType>::Native> + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl ChunkedArray<ListType>

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the list.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,

pub fn new_vec(
name: &str,
v: Vec<<T as PolarsNumericType>::Native, Global>
) -> ChunkedArray<T>

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T> ChunkedArray<T> where
T: PolarsNumericType,
ChunkedArray<T>: IntoSeries,

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N>(&self, lhs: N) -> ChunkedArray<T> where
N: Num + NumCast,

Apply lhs - self

pub fn lhs_div<N>(&self, lhs: N) -> ChunkedArray<T> where
N: Num + NumCast,

Apply lhs / self

pub fn lhs_rem<N>(&self, lhs: N) -> ChunkedArray<T> where
N: Num + NumCast,

Apply lhs % self