polars
Module datatypes

Data types supported by Polars.

At the moment Polars doesn’t include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyValue variants for the data types that are currently supported.

Modules

categoricaldtype-categorical

enum_dtype-categorical

Structs

BinaryOffsetType

BinaryType

BooleanType

CatIter

CategoricalChunked

CategoricalChunkedBuilder

CategoricalType

CompatLevel

DateType

DatetimeType

DecimalType

Dimension

DurationType

FalseT

Field

Characterizes the name and the DataType of a column.

FixedSizeListTypedtype-array

Float32Type

Float64Type

GlobalRevMapMerger

Int8Type

Int16Type

Int32Type

Int64Type

Int128Typedtype-decimal

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

ObjectTypeobject

OwnedObjectobject

PlSmallStr

String type that inlines small strings.

StringType

StructTypedtype-struct

TimeType

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Enums

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

CategoricalOrdering

DataType

ReshapeDimension

A dimension in a reshape.

RevMapping

TimeUnit

UnknownKind

Constants

IDX_DTYPENon-bigidx

Statics

DTYPE_ENUM_KEY

DTYPE_ENUM_VALUE

Traits

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsRefDataType

CategoricalMergeOperation

GetAnyValue

InitHashMaps

InitHashMaps2

IntoScalar

LogicalType

NumericNative

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsNumericType

StaticArray

Functions

call_categorical_merge_operation

create_enum_dtypedtype-categorical

make_categoricals_compatible

make_list_categoricals_compatible

merge_dtypes

Type Aliases

ArrayChunkeddtype-array

BinaryChunked

BinaryOffsetChunked

BooleanChunked

DateChunked

DatetimeChunked

DecimalChunked

DurationChunked

FieldRef

Float32Chunked

Float64Chunked

IdxArr

IdxCaNon-bigidx

IdxTypeNon-bigidx

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

Int128Chunkeddtype-decimal

ListChunked

ObjectChunkedobject

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PlRandomState

StringChunked

TimeChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked