polars

Module datatypes

Data types supported by Polars.

At the moment Polars doesn’t include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyValue variants for the data types that are currently supported.

Modules

categorical

time_unit

time_zone

Structs

BinaryOffsetType

BinaryType

BooleanType

Categorical8Type

Categorical16Type

Categorical32Type

CategoricalMapping

CategoricalType

Categories

A (named) object which is used to indicate which categorical data types have the same mapping.

CompatLevel

DateType

DatetimeType

DecimalType

Dimension

DurationType

FalseT

Field

Characterizes the name and the DataType of a column.

FixedSizeListType

Float32Type

Float64Type

FrozenCategories

An ordered collection of unique strings with an associated pre-computed mapping to go from string <-> index.

Int8Type

Int16Type

Int32Type

Int64Type

Int128Type

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

ObjectType

OwnedObject

PlSmallStr

String type that inlines small strings.

StringType

StructType

TimeType

TimeZone

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Enums

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

CategoricalPhysical

The physical datatype backing a categorical / enum.

DataType

ReshapeDimension

A dimension in a reshape.

TimeUnit

UnknownKind

Constants

IDX_DTYPE

Statics

EXTENSION_NAME

Traits

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsRefDataType

CatNative

CategoricalPhysicalDtypeExt

GetAnyValue

InitHashMaps

InitHashMaps2

IntoMetadata

IntoScalar

LogicalType

MetaDataExt

NumericNative

PolarsCategoricalType

Safety

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsNumericType

PolarsPhysicalType

SchemaExtPl

StaticArray

Functions

ensure_same_categories

ensure_same_frozen_categories

merge_dtypes

unpack_dtypes

Type Aliases

ArrayChunked

BinaryChunked

BinaryOffsetChunked

BooleanChunked

CatSize

Categorical8Chunked

Categorical16Chunked

Categorical32Chunked

CategoricalChunked

DateChunked

DatetimeChunked

DecimalChunked

DurationChunked

FieldRef

Float32Chunked

Float64Chunked

IdxArr

IdxCa

IdxType

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

Int128Chunked

ListChunked

ObjectChunked

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PlRandomState

StringChunked

TimeChunked

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked