Module polars::datatypes

Structs

BooleanType

Date32Type

Date64Type

DurationNanosecondType

Float32Type

Float64Type

Int32Type

Int64Type

Time64NanosecondType

UInt32Type

Utf8Type

Enums

AnyType

ArrowDataType
The set of datatypes that are supported by this implementation of Apache Arrow.

DateUnit
Date is either a 32-bit or 64-bit type representing elapsed time since UNIX epoch (1970-01-01) in days or milliseconds.

TimeUnit
An absolute length of time in seconds, milliseconds, microseconds or nanoseconds.

Traits

PolarsDataType

PolarsIntegerType

PolarsNumericType

ToStr

Type Definitions

BooleanChunked

Date32Chunked

Date64Chunked

DurationNsChunked

Float32Chunked

Float64Chunked

Int32Chunked

Int64Chunked

Time64NsChunked

UInt32Chunked

Utf8Chunked