Module polars::datatypes

Data types supported by Polars.

At the moment Polars doesn't include all data types available by Arrow. The goal is to incrementally support more data types and prioritize these by usability.

See the AnyType variants for the data types that are currently supported.

Structs

BooleanType

Date32Type

Date64Type

DurationMicrosecondType

DurationMillisecondType

DurationNanosecondType

DurationSecondType

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

IntervalDayTimeType

IntervalYearMonthType

ListType

ObjectType

Time32MillisecondType

Time32SecondType

Time64MicrosecondType

Time64NanosecondType

TimestampMicrosecondType

TimestampMillisecondType

TimestampNanosecondType

TimestampSecondType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Utf8Type

Enums

AnyType

ArrowDataType
The set of datatypes that are supported by this implementation of Apache Arrow.

DateUnit
Date is either a 32-bit or 64-bit type representing elapsed time since UNIX epoch (1970-01-01) in days or milliseconds.

IntervalUnit
YEAR_MONTH or DAY_TIME interval in SQL style.

TimeUnit
An absolute length of time in seconds, milliseconds, microseconds or nanoseconds.

Traits

ArrowNumericType
A subtype of primitive type that represents numeric values.

ArrowPrimitiveType
Trait indicating a primitive fixed-width type (bool, ints and floats).

PolarsDataType

PolarsFloatType

PolarsIntegerType

PolarsNumericType

PolarsPrimitiveType

PolarsSingleType
Any type that is not nested

ToStr

Type Definitions

BooleanChunked

Date32Chunked

Date64Chunked

DurationMicrosecondChunked

DurationMillisecondChunked

DurationNanosecondChunked

DurationSecondChunked

Float32Chunked

Float64Chunked

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

IntervalDayTimeChunked

IntervalYearMonthChunked

ListChunked

ObjectChunked

Time32MillisecondChunked

Time32SecondChunked

Time64MicrosecondChunked

Time64NanosecondChunked

TimestampMicrosecondChunked

TimestampMillisecondChunked

TimestampNanosecondChunked

TimestampSecondChunked

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked