Enum polars::datatypes::DataType

pub enum DataType {

    Boolean,
    UInt8,
    UInt16,
    UInt32,
    UInt64,
    Int8,
    Int16,
    Int32,
    Int64,
    Float32,
    Float64,
    Decimal(Option<usize>, Option<usize>),
    String,
    Binary,
    BinaryOffset,
    Date,
    Datetime(TimeUnit, Option<String>),
    Duration(TimeUnit),
    Time,
    Array(Box<DataType>, usize),
    List(Box<DataType>),
    Object(&'static str, Option<Arc<ObjectRegistry>>),
    Null,
    Categorical(Option<Arc<RevMapping>>, CategoricalOrdering),
    Enum(Option<Arc<RevMapping>>, CategoricalOrdering),
    Struct(Vec<Field>),
    Unknown(UnknownKind),

}

Variants

Boolean

UInt8

UInt16

UInt32

UInt64

Int8

Int16

Int32

Int64

Float32

Float64

Decimal(Option<usize>, Option<usize>)

Available on crate feature dtype-decimal only.

Fixed point decimal type optional precision and non-negative scale. This is backed by a signed 128-bit integer which allows for up to 38 significant digits.

String

String data

Binary

BinaryOffset

Date

A 32-bit date representing the elapsed time since UNIX epoch (1970-01-01) in days (32 bits).

Datetime(TimeUnit, Option<String>)

A 64-bit date representing the elapsed time since UNIX epoch (1970-01-01) in the given timeunit (64 bits).

Duration(TimeUnit)

Time

A 64-bit time representing the elapsed time since midnight in nanoseconds

Array(Box<DataType>, usize)

Available on crate feature dtype-array only.

A nested list with a fixed size in each row

List(Box<DataType>)

A nested list with a variable size in each row

Object(&'static str, Option<Arc<ObjectRegistry>>)

Available on crate feature object only.

A generic type that can be used in a Series &’static str can be used to determine/set inner type

Null

Categorical(Option<Arc<RevMapping>>, CategoricalOrdering)

Available on crate feature dtype-categorical only.

Enum(Option<Arc<RevMapping>>, CategoricalOrdering)

Available on crate feature dtype-categorical only.

Struct(Vec<Field>)

Available on crate feature dtype-struct only.

Unknown(UnknownKind)

Implementations

impl DataType

pub fn value_within_range(&self, other: AnyValue<'_>) -> bool

pub fn is_known(&self) -> bool

Check if the whole dtype is known.

pub fn get_shape(&self) -> Option<Vec<usize>>

Available on crate feature dtype-array only.

Get the full shape of a multidimensional array.

pub fn inner_dtype(&self) -> Option<&DataType>

Get the inner data type of a nested type.

pub fn leaf_dtype(&self) -> &DataType

Get the absolute inner data type of a nested type.

pub fn cast_leaf(&self, to: DataType) -> DataType

Cast the leaf types of Lists/Arrays and keep the nesting.

pub fn to_physical(&self) -> DataType

Convert to the physical data type

pub fn is_logical(&self) -> bool

Check if this DataType is a logical type

pub fn is_temporal(&self) -> bool

Check if this DataType is a temporal type

pub fn is_primitive(&self) -> bool

Check if datatype is a primitive type. By that we mean that it is not a container type.

pub fn is_numeric(&self) -> bool

Check if this DataType is a basic numeric type (excludes Decimal).

pub fn is_bool(&self) -> bool

Check if this DataType is a boolean.

pub fn is_list(&self) -> bool

Check if this DataType is a list.

pub fn is_array(&self) -> bool

Check if this DataType is an array.

pub fn is_nested(&self) -> bool

pub fn is_struct(&self) -> bool

Check if this DataType is a struct

pub fn is_binary(&self) -> bool

pub fn is_object(&self) -> bool

pub fn is_null(&self) -> bool

pub fn contains_views(&self) -> bool

pub fn contains_categoricals(&self) -> bool

pub fn contains_objects(&self) -> bool

pub fn is_ord(&self) -> bool

Check if type is sortable

pub fn is_decimal(&self) -> bool

Check if this DataType is a Decimal type (of any scale/precision).

pub fn is_float(&self) -> bool

Check if this DataType is a basic floating point type (excludes Decimal).

pub fn is_integer(&self) -> bool

Check if this DataType is an integer.

pub fn is_signed_integer(&self) -> bool

pub fn is_unsigned_integer(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_categorical(&self) -> bool

pub fn is_enum(&self) -> bool

pub fn to_arrow_field(&self, name: &str, pl_flavor: bool) -> Field

Convert to an Arrow Field

pub fn to_arrow(&self, pl_flavor: bool) -> ArrowDataType

Convert to an Arrow data type.

pub fn try_to_arrow( &self, pl_flavor: bool ) -> Result<ArrowDataType, PolarsError>

pub fn is_nested_null(&self) -> bool

pub fn matches_schema_type( &self, schema_type: &DataType ) -> Result<bool, PolarsError>

impl DataType

pub fn boxed(self) -> Box<DataType>

pub fn from_arrow(dt: &ArrowDataType, bin_to_view: bool) -> DataType