Module polars::export::arrow::compute

contains a wide range of compute operations (e.g. arithmetics, aggregate, filter, comparison, and [sort])

This module’s general design is that each operator has two interfaces, a statically-typed version and a dynamically-typed version. The statically-typed version expects concrete arrays (such as PrimitiveArray); the dynamically-typed version expects &dyn Array and errors if the the type is not supported. Some dynamically-typed operators have an auxiliary function, can_*, that returns true if the operator can be applied to the particular DataType.

Modules

aggregate

Contains different aggregation functions

arithmetics

Defines basic arithmetic kernels for PrimitiveArrays.

arity

Defines kernels suitable to perform operations to primitive arrays.

arity_assign

Defines generics suitable to perform operations to PrimitiveArray in-place.

boolean

null-preserving operators such as and, or and not.

boolean_kleene

Boolean operators of Kleene logic.

cast

Defines different casting operators such as cast or primitive_to_binary.

comparison

Contains comparison operators

concatenate

Contains the concatenate kernel

filter

Contains operators to filter arrays such as filter.

if_then_else

Contains the operator if_then_else.

take

Defines take kernel for Array

temporal

Defines temporal kernels for time and date related functions.