Module polars::export::rayon::prelude

The rayon prelude imports the various ParallelIterator traits. The intention is that one can include use rayon::prelude::* and have easy access to the various traits and methods you will need.

Traits

FromParallelIterator

FromParallelIterator implements the creation of a collection from a ParallelIterator. By implementing FromParallelIterator for a given type, you define how it will be created from an iterator.

IndexedParallelIterator

An iterator that supports “random access” to its data, meaning that you can split it at arbitrary indices and draw data from those points.

IntoParallelIterator

IntoParallelIterator implements the conversion to a ParallelIterator.

IntoParallelRefIterator

IntoParallelRefIterator implements the conversion to a ParallelIterator, providing shared references to the data.

IntoParallelRefMutIterator

IntoParallelRefMutIterator implements the conversion to a ParallelIterator, providing mutable references to the data.

ParallelBridge

Conversion trait to convert an Iterator to a ParallelIterator.

ParallelDrainFull

ParallelDrainFull creates a parallel iterator that moves all items from a collection while retaining the original capacity.

ParallelDrainRange

ParallelDrainRange creates a parallel iterator that moves a range of items from a collection while retaining the original capacity.

ParallelExtend

ParallelExtend extends an existing collection with items from a ParallelIterator.

ParallelIterator

Parallel version of the standard iterator trait.

ParallelSlice

Parallel extensions for slices.

ParallelSliceMut

Parallel extensions for mutable slices.

ParallelString

Parallel extensions for strings.