polars::frame

Module group_by

Modules

aggregations

expr

Structs

GroupBy

Returned by a group_by operation on a DataFrame. This struct supports several aggregations.

GroupPositions

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsTypeIter

GroupsTypeParIter

Enums

GroupByMethod

GroupsIndicator

GroupsType

Traits

IntoGroupsType

Used to create the tuples for a group_by operation.

Functions

fmt_group_by_column

Type Aliases

BorrowIdxItem

GroupsSlice

Every group is indicated by an array where the

IdxItem