Enum polars::frame::row::AnyValueBufferTrusted

pub enum AnyValueBufferTrusted<'a> {

    Boolean(BooleanChunkedBuilder),
    Int8(PrimitiveChunkedBuilder<Int8Type>),
    Int16(PrimitiveChunkedBuilder<Int16Type>),
    Int32(PrimitiveChunkedBuilder<Int32Type>),
    Int64(PrimitiveChunkedBuilder<Int64Type>),
    UInt8(PrimitiveChunkedBuilder<UInt8Type>),
    UInt16(PrimitiveChunkedBuilder<UInt16Type>),
    UInt32(PrimitiveChunkedBuilder<UInt32Type>),
    UInt64(PrimitiveChunkedBuilder<UInt64Type>),
    Float32(PrimitiveChunkedBuilder<Float32Type>),
    Float64(PrimitiveChunkedBuilder<Float64Type>),
    String(BinViewChunkedBuilder<str>),
    Struct(Vec<(AnyValueBuffer<'a>, SmartString<LazyCompact>)>),
    Null(NullChunkedBuilder),
    All(DataType, Vec<AnyValue<'a>>),

}

Available on crate features rows or object only.

An AnyValueBuffer that should be used when we trust the builder

Variants

Boolean(BooleanChunkedBuilder)

Int8(PrimitiveChunkedBuilder<Int8Type>)

Available on crate feature dtype-i8 only.

Int16(PrimitiveChunkedBuilder<Int16Type>)

Available on crate feature dtype-i16 only.

Int32(PrimitiveChunkedBuilder<Int32Type>)

Int64(PrimitiveChunkedBuilder<Int64Type>)

UInt8(PrimitiveChunkedBuilder<UInt8Type>)

Available on crate feature dtype-u8 only.

UInt16(PrimitiveChunkedBuilder<UInt16Type>)

Available on crate feature dtype-u16 only.

UInt32(PrimitiveChunkedBuilder<UInt32Type>)

UInt64(PrimitiveChunkedBuilder<UInt64Type>)

Float32(PrimitiveChunkedBuilder<Float32Type>)

Float64(PrimitiveChunkedBuilder<Float64Type>)

String(BinViewChunkedBuilder<str>)

Struct(Vec<(AnyValueBuffer<'a>, SmartString<LazyCompact>)>)

Available on crate feature dtype-struct only.

Null(NullChunkedBuilder)

All(DataType, Vec<AnyValue<'a>>)

Implementations

impl<'a> AnyValueBufferTrusted<'a>

pub fn new(dtype: &DataType, len: usize) -> AnyValueBufferTrusted<'a>

pub unsafe fn add_unchecked_owned_physical(&mut self, val: &AnyValue<'_>)

Will add the AnyValue into Self and unpack as the physical type belonging to Self. This should only be used with physical buffers

If a type is not primitive or String, the AnyValues will be converted to static

Safety

The caller must ensure that the AnyValue type exactly matches the Buffer type and is owned.

pub unsafe fn add_unchecked_borrowed_physical(&mut self, val: &AnyValue<'_>)

Safety

The caller must ensure that the AnyValue type exactly matches the Buffer type and is borrowed.

pub fn reset(&mut self, capacity: usize) -> Series

pub fn into_series(self) -> Series