Module polars::prelude

Modules

aggregations

binary

cloud

datatypes

Data types supported by Polars.

default_arrays

dt

expr

full

null

predicates

read_impl

slice

sort

string

utf8

utils

zip

Macros

df

polars_bail

polars_ensure

polars_err

Structs

AggregationContext

AnonymousScanOptions

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

ArrowSchema

An ordered sequence of Fields with associated Metadata.

BatchedParquetReader

BinaryChunkedBuilder

BinaryTakeRandom

BinaryTakeRandomSingleChunk

BinaryType

BoolTakeRandom

BoolTakeRandomSingleChunk

BooleanChunkedBuilder

BooleanType

BrotliLevel

CategoricalType

ChunkedArray

ChunkedArray

CsvReader

Create a new DataFrame by reading a csv file.

CsvWriter

Write a DataFrame to csv.

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeArgs

Arguments used by datetime in order to produce an Expr of Datetime

DatetimeType

DurationArgs

Arguments used by duration in order to produce an Expr of Duration

DurationType

Field

Characterizes the name and the DataType of a column.

Float32Type

Float64Type

GroupBy

Returned by a groupby operation on a DataFrame. This struct supports several aggregations.

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsProxyIter

GroupsProxyParIter

GzipLevel

Int8Type

Int16Type

Int32Type

Int64Type

IpcReader

Read Arrows IPC format into a DataFrame

IpcWriter

Write a DataFrame to Arrow’s IPC format

IpcWriterOption

JoinBuilder

JoinOptions

LazyCsvReader

LazyFrame

Lazy abstraction over an eager DataFrame. It really is an abstraction over a logical plan. The methods of this struct will incrementally modify a logical plan until output is requested (via collect)

LazyGroupBy

Utility struct for lazy groupby operation.

ListBinaryChunkedBuilder

ListBooleanChunkedBuilder

ListNameSpace

Specialized expressions for Series of DataType::List.

ListPrimitiveChunkedBuilder

ListTakeRandom

ListTakeRandomSingleChunk

ListType

ListUtf8ChunkedBuilder

Logical

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

MeltArgs

Arguments for [DataFrame::melt] function

NoNull

Just a wrapper structure. Useful for certain impl specializations This is for instance use to implement impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>> as Option<T::Native> was already implemented: impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>

Null

The literal Null

NumTakeRandomChunked

NumTakeRandomCont

NumTakeRandomSingleChunk

OptState

State of the allowed optimizations

ParquetReader

Read Apache parquet format into a DataFrame.

ParquetWriteOptions

ParquetWriter

Write a DataFrame to parquet format

PhysicalIoHelper

Wrapper struct that allow us to use a PhysicalExpr in polars-io.

PrimitiveChunkedBuilder

RollingCovOptions

ScanArgsAnonymous

ScanArgsParquet

Schema

A map from field/column name (String) to the type of that field/column (DataType)

Series

Series

SlicedGroups

SortMultipleOptions

SortOptions

SpecialEq

Wrapper type that has special equality properties depending on the inner type specialization

StrptimeOptions

TakeRandomBitmap

TimeType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Utf8ChunkedBuilder

Utf8TakeRandom

Utf8TakeRandomSingleChunk

Utf8Type

When

Intermediate state of when(..).then(..).otherwise(..) expr.

WhenThen

Intermediate state of when(..).then(..).otherwise(..) expr.

WhenThenThen

Intermediate state of chain when then exprs.

WindowOptions

ZstdLevel

Represents a valid zstd compression level.

Enums

AggExpr

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

BooleanFunction

CsvEncoding

DataType

Excluded

Expr

Expressions that can be used in various contexts. Queries consist of multiple expressions. When using the polars lazy API, don’t construct an Expr directly; instead, create one using the functions in the polars_lazy::dsl module. See that module’s docs for more info.

FillNullStrategy

FunctionExpr

GroupByMethod

GroupsIndicator

GroupsProxy

IpcCompression

Compression codec

JoinType

LiteralValue

LogicalPlan

NullValues

Operator

ParallelStrategy

ParquetCompression

PolarsError

QuantileInterpolOptions

SearchSortedSide

TakeIdx

One of the three arguments allowed in unchecked_take

TakeRandBranch2

TakeRandBranch3

TimeUnit

UniqueKeepStrategy

WindowMapping

Constants

IDX_DTYPE

NULL

Traits

AnonymousScan

ArgAgg

Argmin/ Argmax

ArrowGetItem

AsBinary

AsList

AsRefDataType

AsUtf8

BinaryNameSpaceImpl

BinaryUdfOutputField

ChunkAgg

Aggregation operations

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray when the operation is cheaper than branching due to null checking

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkCumAgg

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/ flatten a List or Utf8 Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkPeaks

Find local minima/ maxima

ChunkQuantile

Quantile and median aggregation

ChunkReverse

Reverse a ChunkedArray

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

Fast access by index.

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine 2 ChunkedArrays based on some predicate.

ChunkedBuilder

ChunkedSet

DataFrameJoinOps

DataFrameOps

FromData

FromDataBinary

FromDataUtf8

FunctionOutputField

FxHash

Ensure that the same hash is used as with VecHash.

GetAnyValue

IndexOfSchema

This trait exists to be unify the API of polars Schema and arrows Schema

IndexToUsize

InitHashMaps

IntoGroupsProxy

Used to create the tuples for a groupby operation.

IntoLazy

IntoSeries

Used to convert a ChunkedArray, &dyn SeriesTrait and Series into a Series.

IntoTakeRandom

Create a type that implements a faster TakeRandom.

IntoVec

IsFloat

Safety

IsIn

Check if element is member of list array

LazyFileListReader

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

LhsNumOps

ListBuilderTrait

ListFromIter

ListNameSpaceImpl

Literal

LogSeries

LogicalType

MutableBitmapExtension

NamedFrom

NamedFromOwned

NewChunkedArray

NumOpsDispatch

NumericNative

PartitionedAggregation

PhysicalExpr

Take a DataFrame and evaluate the expressions. Implement this for Column, lt, eq, etc

PolarsArray

PolarsDataType

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

PolarsSingleType

Any type that is not nested

QuantileAggSeries

Range

RenameAliasFn

SerReader

SerWriter

SeriesBinaryUdf

A wrapper trait for any binary closure Fn(Series, Series) -> PolarsResult<Series>

SeriesMethods

SeriesSealed

SeriesTrait

SeriesUdf

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

TakeIterator

TakeIteratorNulls

TakeRandom

Random access

TakeRandomUtf8

UdfSchema

Utf8NameSpaceImpl

ValueSize

VarAggSeries

VecHash

Functions

all

Selects all columns. Shorthand for col("*").

all_exprs

Create a new column with the the bitwise-and of the elements in each row.

any_exprs

Create a new column with the the bitwise-or of the elements in each row.

apply_binary

Like map_binary, but used in a groupby-aggregation context.

apply_multiple

Apply a function/closure over the groups of multiple columns. This should only be used in a groupby aggregation.

avg

Find the mean of all the values in the column named name. Alias for mean.

binary_expr

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

cast

Casts the column given by Expr to a different type.

coalesce

Folds the expressions from left to right keeping the first non-null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name.

concat

Concat multiple

concat_list

Concat lists entries.

count

Count expression

cov

Compute the covariance between two columns.

date_range

Create a date range from a start and stop expression.

datetime

Construct a column of Datetime from the provided DatetimeArgs.

datetime_to_timestamp_ms

datetime_to_timestamp_ns

datetime_to_timestamp_us

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

Construct a column of Duration from the provided DurationArgs

first

First column in DataFrame

floor_div_series

fmt_groupby_column

fold_exprs

Accumulate over multiple columns horizontally / row wise.

indexes_to_usizes

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

last

Last column in DataFrame

lit

Create a Literal Expression from L. A literal expression behaves like a column that contains a single distinct value.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

map_list_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

max_exprs

Create a new column with the the maximum value per row.

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

merge_dtypes

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

min_exprs

Create a new column with the the minimum value per row.

not

Negates a boolean column.

pearson_corr

Compute the pearson correlation between two columns.

quantile

Find a specific quantile of all the values in the column named name.

range

Create a range literal.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value. Generally you won’t need this function, as lit(value) already represents a column containing only value whose length is automatically set to the correct number of rows.

resolve_homedir

search_sorted

sum

Sum all the values in the column named name. Shorthand for col(name).sum().

sum_exprs

Create a new column with the the sum of the values in each row.

ternary_expr

time_range

Create a time range from a start and stop expression.

when

Start a when-then-otherwise expression

Type Definitions

AllowedOptimizations

AllowedOptimizations

ArrayRef

BinaryChunked

BooleanChunked

BorrowIdxItem

CloudOptions

DateChunked

DatetimeChunked

Dummy

Dummy type, we need to instantiate all generic types, so we fill one with a dummy.

DurationChunked

FillNullLimit

Float32Chunked

Float64Chunked

GetOutput

GlobIterator

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxCa

IdxItem

IdxSize

The type used by polars to index data.

IdxType

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

LargeBinaryArray

LargeListArray

LargeStringArray

ListChunked

PlHashMap

PlHashSet

PlIdHashMap

This hashmap has the uses an IdHasher

PlIndexMap

PlIndexSet

PolarsResult

SchemaRef

TimeChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked