Module polars::prelude

Modules

aggregations

arity

binary

cloud

Interface with cloud storage through the object_store crate.

consts

datatypes

Data types supported by Polars.

default_arrays

dt

expr

fill_null

full

gather

gather_skip_nulls

null

predicates

read_impl

slice

sort

sum_f64

udf

utf8

utils

zip

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

Structs

AggregationContext

AnonymousScanOptions

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

ArrowSchema

An ordered sequence of Fields with associated Metadata.

BatchedParquetReader

BinaryChunkedBuilder

BinaryType

BooleanChunkedBuilder

BooleanType

BrotliLevel

CategoricalType

ChainedThen

Utility struct for the when-then-otherwise expression.

ChainedWhen

Utility struct for the when-then-otherwise expression.

ChunkedArray

ChunkedArray

CsvReader

Create a new DataFrame by reading a csv file.

CsvWriter

Write a DataFrame to csv.

CsvWriterOptions

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeArgs

Arguments used by datetime in order to produce an Expr of Datetime

DatetimeType

DurationArgs

Arguments used by duration in order to produce an Expr of Duration

DurationType

Field

Characterizes the name and the DataType of a column.

FieldsMapper

FileMetaData

Metadata for a Parquet file.

Flat

Float32Type

Float64Type

GroupBy

Returned by a group_by operation on a DataFrame. This struct supports several aggregations.

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsProxyIter

GroupsProxyParIter

GzipLevel

Int8Type

Int16Type

Int32Type

Int64Type

JoinArgs

JoinBuilder

JoinOptions

LazyCsvReader

LazyFrame

Lazy abstraction over an eager DataFrame. It really is an abstraction over a logical plan. The methods of this struct will incrementally modify a logical plan until output is requested (via collect).

LazyGroupBy

Utility struct for lazy group_by operation.

ListBinaryChunkedBuilder

ListBooleanChunkedBuilder

ListNameSpace

Specialized expressions for Series of DataType::List.

ListPrimitiveChunkedBuilder

ListType

ListUtf8ChunkedBuilder

Logical

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

MeltArgs

Arguments for [DataFrame::melt] function

Nested

NoNull

Just a wrapper structure. Useful for certain impl specializations This is for instance use to implement impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>> as Option<T::Native> was already implemented: impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>

Null

The literal Null

OptState

State of the allowed optimizations

ParquetReader

Read Apache parquet format into a DataFrame.

ParquetWriteOptions

ParquetWriter

Write a DataFrame to parquet format

PhysicalIoHelper

Wrapper struct that allow us to use a PhysicalExpr in polars-io.

PrimitiveChunkedBuilder

RollingCovOptions

RollingQuantileParams

RollingVarParams

ScanArgsAnonymous

ScanArgsParquet

Schema

A map from field/column name (String) to the type of that field/column (DataType)

SerializeOptions

Options to serialize logical types to CSV.

Series

Series

SlicedGroups

SortMultipleOptions

SortOptions

SpecialEq

Wrapper type that has special equality properties depending on the inner type specialization

StrptimeOptions

Then

Utility struct for the when-then-otherwise expression.

TimeType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

UnionArgs

UserDefinedFunction

Represents a user-defined function

Utf8ChunkedBuilder

Utf8Type

When

Utility struct for the when-then-otherwise expression.

ZstdLevel

Represents a valid zstd compression level.

Enums

AggExpr

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

BooleanFunction

CsvEncoding

DataType

Excluded

Expr

Expressions that can be used in various contexts. Queries consist of multiple expressions. When using the polars lazy API, don’t construct an Expr directly; instead, create one using the functions in the polars_lazy::dsl module. See that module’s docs for more info.

FillNullStrategy

FunctionExpr

GroupByMethod

GroupsIndicator

GroupsProxy

JoinType

JoinValidation

LiteralValue

LogicalPlan

NullValues

Operator

ParallelStrategy

ParquetCompression

PolarsError

QuantileInterpolOptions

QuoteStyle

TimeUnit

UniqueKeepStrategy

WindowMapping

WindowType

Constants

IDX_DTYPE

NULL

Statics

BOOLEAN_RE

FLOAT_RE

INTEGER_RE

Traits

AnonymousScan

ArgAgg

Argmin/ Argmax

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsBinary

AsList

AsRefDataType

AsUtf8

BinaryNameSpaceImpl

BinaryUdfOutputField

ChunkAgg

Aggregation operations.

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/ flatten a List or Utf8 Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

ChunkedBuilder

ChunkedCollectInferIterExt

ChunkedCollectIterExt

ChunkedSet

CrossJoin

DataFrameJoinOps

DataFrameOps

FromData

FromDataBinary

FromDataUtf8

FunctionOutputField

GetAnyValue

IndexOfSchema

This trait exists to be unify the API of polars Schema and arrows Schema

IndexToUsize

InitHashMaps

InitHashMaps2

IntoGroupsProxy

Used to create the tuples for a group_by operation.

IntoLazy

IntoSeries

Used to convert a ChunkedArray, &dyn SeriesTrait and Series into a Series.

IntoVec

IsFloat

Safety

JoinDispatch

LazyFileListReader

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

LhsNumOps

ListBuilderTrait

ListFromIter

ListNameSpaceImpl

Literal

LogicalType

MutableBitmapExtension

NamedFrom

NamedFromOwned

NewChunkedArray

NumOpsDispatch

NumericNative

PartitionedAggregation

PhysicalExpr

Take a DataFrame and evaluate the expressions. Implement this for Column, lt, eq, etc

PolarsArray

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

QuantileAggSeries

RenameAliasFn

SerReader

SerWriter

SeriesBinaryUdf

A wrapper trait for any binary closure Fn(Series, Series) -> PolarsResult<Series>

SeriesJoin

SeriesMethods

SeriesSealed

SeriesTrait

SeriesUdf

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

StaticArray

UdfSchema

ValueSize

VarAggSeries

VecHash

Functions

_join_suffix_name

_sort_or_hash_inner

all

Selects all columns. Shorthand for col("*").

apply_binary

Like map_binary, but used in a group_by-aggregation context.

apply_multiple

Apply a function/closure over the groups of multiple columns. This should only be used in a group_by aggregation.

avg

Find the mean of all the values in the column named name. Alias for mean.

binary_expr

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

cast

Casts the column given by Expr to a different type.

check_projected_arrow_schema

Checks if the projected columns are equal

check_projected_schema

Checks if the projected columns are equal

check_projected_schema_impl

clip

Clamp underlying values to the min and max values.

clip_max

Clamp underlying values to the max value.

clip_min

Clamp underlying values to the min value.

coalesce

Folds the expressions from left to right keeping the first non-null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name.

concat

Concat multiple LazyFrames vertically.

concat_expr

concat_list

Concat lists entries.

count

Count expression.

datetime

Construct a column of Datetime from the provided DatetimeArgs.

datetime_to_timestamp_ms

datetime_to_timestamp_ns

datetime_to_timestamp_us

default_join_ids

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

Construct a column of Duration from the provided DurationArgs

first

First column in DataFrame.

fmt_group_by_column

fold_exprs

Accumulate over multiple columns horizontally / row wise.

get_reader_bytes

get_sequential_row_statistics

Compute remaining_rows_to_read to be taken per file up front, so we can actually read concurrently/parallel

indexes_to_usizes

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

last

Last column in DataFrame.

lit

Create a Literal Expression from L. A literal expression behaves like a column that contains a single distinct value.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

map_list_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

materialize_projection

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

merge_dtypes

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

not

Negates a boolean column.

private_left_join_multiple_keys

quantile

Find a specific quantile of all the values in the column named name.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value. Generally you won’t need this function, as lit(value) already represents a column containing only value whose length is automatically set to the correct number of rows.

resolve_homedir

sum

Sum all the values in the column named name. Shorthand for col(name).sum().

ternary_expr

when

Start a when-then-otherwise expression.

Type Aliases

AllowedOptimizations

AllowedOptimizations

ArrayRef

BinaryChunked

BooleanChunked

BorrowIdxItem

ChunkId

[ChunkIdx, DfIdx]

ChunkJoinIds

ChunkJoinOptIds

DateChunked

DatetimeChunked

DurationChunked

DynArgs

FileMetaDataRef

FillNullLimit

Float32Chunked

Float64Chunked

GetOutput

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxCa

IdxItem

IdxSize

The type used by polars to index data.

IdxType

InnerJoinIds

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

LargeBinaryArray

LargeListArray

LargeStringArray

LeftJoinIds

ListChunked

PathIterator

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PolarsResult

SchemaRef

TimeChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked