Module polars::prelude

Modules

aggregations

arity

arraypolars-ops and dtype-array

binarylazy

catlazy and dtype-categorical

chunkedarraytemporal

Traits and utilities for temporal data.

cloudpolars-io

Interface with cloud storage through the object_store crate.

datatypes

Data types supported by Polars.

datetimepolars-ops and timezones

default_arrays

dtlazy and temporal

expr

fill_null

fixed_size_listdtype-array

float_sorted_arg_max

full

function_exprlazy

gather

interpolatepolars-ops and interpolate

interpolate_bypolars-ops and interpolate_by

modepolars-ops and mode

nan_propagating_aggregatepolars-ops and propagate_nans

null

schema_inferencepolars-io

search_sorted

seriestemporal

slice

sort

stringspolars-ops

udflazy

utf8

zipzip_with

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

Structs

AnonymousScanArgslazy

AnonymousScanOptionslazy

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrayNameSpacelazy

Specialized expressions for Series of DataType::Array.

ArrowField

Represents Arrow’s metadata of a “column”.

ArrowSchema

An ordered sequence of Fields with associated [Metadata].

AsOfOptionspolars-ops

BatchedCsvReaderpolars-io

BatchedParquetReaderpolars-io

BinaryOffsetType

BinaryType

BooleanChunkedBuilder

BooleanType

Boundstemporal

BoundsItertemporal

BrotliLevelpolars-io

A valid Brotli compression level.

CatIter

CategoricalChunked

CategoricalChunkedBuilder

CategoricalNameSpacelazy

Specialized expressions for Categorical dtypes.

CategoricalType

ChainedThenlazy

Utility struct for the when-then-otherwise expression.

ChainedWhenlazy

Utility struct for the when-then-otherwise expression.

ChunkId

ChunkedArray

ChunkedArray

CsvParseOptionspolars-io

CsvReadOptionspolars-io

CsvReaderpolars-io

Create a new DataFrame by reading a csv file.

CsvWriterpolars-io

Write a DataFrame to csv.

CsvWriterOptionspolars-io

Options for writing CSV files.

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeArgslazy

Arguments used by datetime in order to produce an Expr of Datetime

DatetimeType

DecimalType

Durationlazy

DurationArgslazy

Arguments used by duration in order to produce an Expr of Duration

DurationType

DynamicGroupOptionslazy

ExprNameNameSpacelazy

Specialized expressions for modifying the name of existing expressions.

FalseT

Field

Characterizes the name and the DataType of a column.

FieldsMapperlazy

FileMetaDatapolars-io

Metadata for a Parquet file.

FixedSizeListTypedtype-array

Float32Type

Float64Type

GlobalRevMapMerger

GroupBy

Returned by a group_by operation on a DataFrame. This struct supports several aggregations.

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsProxyIter

GroupsProxyParIter

GzipLevelpolars-io

A valid Gzip compression level.

InProcessQuerylazy

Int8Type

Int16Type

Int32Type

Int64Type

Int128Typedtype-decimal

IpcReaderpolars-io

Read Arrows IPC format into a DataFrame

IpcScanOptionspolars-io

IpcStreamReaderpolars-io

Read Arrows Stream IPC format into a DataFrame

IpcStreamWriterpolars-io

Write a DataFrame to Arrow’s Streaming IPC format

IpcStreamWriterOptionpolars-io

IpcWriterpolars-io

Write a DataFrame to Arrow’s IPC format

IpcWriterOptionpolars-io

IpcWriterOptionspolars-io

JoinArgslazy

JoinBuilderlazy

JoinOptionslazy

JsonLineReaderpolars-io

JsonReaderpolars-io

Reads JSON in one of the formats in JsonFormat into a DataFrame.

JsonWriterpolars-io

Writes a DataFrame to JSON.

JsonWriterOptionspolars-io

LazyCsvReaderlazy and csv

LazyFramelazy

Lazy abstraction over an eager DataFrame. It really is an abstraction over a logical plan. The methods of this struct will incrementally modify a logical plan until output is requested (via collect).

LazyGroupBylazy

Utility struct for lazy group_by operation.

LazyJsonLineReaderlazy

ListBinaryChunkedBuilder

ListBooleanChunkedBuilder

ListNameSpacelazy

Specialized expressions for Series of DataType::List.

ListPrimitiveChunkedBuilder

ListStringChunkedBuilder

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

MeltArgs

Arguments for [DataFrame::melt] function

NoNull

Just a wrapper structure. Useful for certain impl specializations This is for instance use to implement impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>> as Option<T::Native> was already implemented: impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>

Nulllazy

The literal Null

NullableIdxSize

ObjectTypeobject

OptStatelazy

State of the allowed optimizations

OwnedBatchedCsvReaderpolars-io

OwnedObjectobject

ParquetOptionspolars-io

ParquetReaderpolars-io

Read Apache parquet format into a DataFrame.

ParquetStatisticspolars-io

Arrow-deserialized parquet Statistics of a file

ParquetWriteOptionspolars-io

ParquetWriterpolars-io

Write a DataFrame to Parquet format.

PrimitiveChunkedBuilder

RankOptionslazy

RollingCovOptionslazy

RollingGroupOptionslazy

RollingOptionsDynamicWindowtemporal

RollingOptionsFixedWindow

RollingQuantileParams

RollingVarParams

Scalar

ScanArgsAnonymouslazy

ScanArgsIpclazy

ScanArgsParquetlazy

Schema

A map from field/column name (String) to the type of that field/column (DataType)

SerializeOptionspolars-io

Options to serialize logical types to CSV.

Series

Series

SlicedGroups

SortMultipleOptions

Sort options for multi-series sorting.

SortOptions

Options for single series sorting.

SpecialEqlazy

Wrapper type that has special equality properties depending on the inner type specialization

SplitNCharspolars-ops

StatisticsOptionspolars-io

The statistics to write

StringCacheHolder

Enable the global string cache as long as the object is alive (RAII).

StringType

StrptimeOptionslazy

StructArraypolars-io

A StructArray is a nested [Array] with an optional validity representing multiple [Array] with the same number of rows.

StructChunked

This is logical type StructChunked that dispatches most logic to the fields implementations

StructNameSpacelazy

Specialized expressions for Struct dtypes.

Thenlazy

Utility struct for the when-then-otherwise expression.

TimeType

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

UnionArgslazy

UserDefinedFunctionlazy

Represents a user-defined function

Whenlazy

Utility struct for the when-then-otherwise expression.

Windowtemporal

Represents a window in time

ZstdLevelpolars-io

A valid Zstandard compression level.

Enums

AggExprlazy

Ambiguous

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

AsofStrategypolars-ops

BooleanFunctionlazy

CategoricalFunctionlazy

CategoricalOrdering

ClosedIntervalpolars-ops

ClosedWindowtemporal

CommentPrefixpolars-io

CsvEncodingpolars-io

DataType

DslPlanlazy

Excludedlazy

Exprlazy

Expressions that can be used in various contexts. Queries consist of multiple expressions. When using the polars lazy API, don’t construct an Expr directly; instead, create one using the functions in the polars_lazy::dsl module. See that module’s docs for more info.

FillNullStrategy

FunctionExprlazy

GroupByMethod

GroupsIndicator

GroupsProxy

IndexOrder

InterpolationMethodpolars-ops

IpcCompressionpolars-io

Compression codec

JoinCoalescepolars-ops

JoinTypelazy

JoinValidationlazy

JsonFormatpolars-io

The format to use to write the DataFrame to JSON: Json (a JSON array) or JsonLines (each row output on a separate line). In either case, each row is serialized as a JSON object whose keys are the column names and whose values are the row’s corresponding values.

Labeltemporal

LiteralValuelazy

NestedTypelazy

NonExistent

NullValuespolars-io

Operatorlazy

ParallelStrategypolars-io

ParquetCompressionpolars-io

The compression strategy to use for writing Parquet files.

PolarsError

PowFunctionlazy

QuantileInterpolOptions

QuoteStylepolars-io

Quote style indicating when to insert quotes around a field.

RankMethodlazy

RevMapping

SearchSortedSidepolars-ops

StartBytemporal

StringFunctionlazy

TemporalFunctionlazy

TimeUnit

UniqueKeepStrategy

UnknownKind

WindowMappinglazy

WindowTypelazy

Constants

IDX_DTYPENon-bigidx

NULLlazy

Statics

BOOLEAN_REpolars-io

DTYPE_ENUM_KEY

DTYPE_ENUM_VALUE

FLOAT_REpolars-io

FLOAT_RE_DECIMALpolars-io

INTEGER_REpolars-io

POLARS_TEMP_DIR_BASE_PATHpolars-io

Traits

AnonymousScanlazy

ArgAggpolars-ops

Argmin/ Argmax

ArithmeticChunked

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsBinarypolars-ops

AsListpolars-ops

AsRefDataType

AsStringpolars-ops

AsofJoinpolars-ops

AsofJoinBypolars-ops

BinaryNameSpaceImplpolars-ops

BinaryUdfOutputFieldlazy

CategoricalMergeOperation

ChunkAgg

Aggregation operations.

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/flatten a List or String Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkRollApplyrolling_window

This differs from ChunkWindowCustom and ChunkWindow by not using a fold aggregator, but reusing a Series wrapper and calling Series aggregators. This likely is a bit slower than ChunkWindow

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

ChunkedBuilder

ChunkedCollectInferIterExt

ChunkedCollectIterExt

ChunkedSetpolars-ops

CrossJoinpolars-ops

DataFrameJoinOpspolars-ops

DataFrameOpspolars-ops

DateMethodstemporal

DatetimeMethodstemporal

DfTakepolars-ops

DurationMethodstemporal

ExprEvalExtensionlazy

FromData

FromDataBinary

FromDataUtf8

FunctionOutputFieldlazy

GetAnyValue

IndexOfSchema

This trait exists to be unify the API of polars Schema and arrows Schema

IndexToUsize

InitHashMaps

InitHashMaps2

IntoGroupsProxy

Used to create the tuples for a group_by operation.

IntoLazylazy

IntoListNameSpacelazy

IntoSeries

Used to convert a ChunkedArray, &dyn SeriesTrait and Series into a Series.

IntoVec

IsFirstDistinctis_first_distinct

Mask the first unique values as true

IsLastDistinctis_last_distinct

Mask the last unique values as true

JoinDispatchpolars-ops

LazyFileListReaderlazy

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

LhsNumOps

ListBuilderTrait

ListFromIter

ListNameSpaceExtensionlazy

ListNameSpaceImplpolars-ops

Literallazy

LogicalType

NamedFrom

NamedFromOwned

NewChunkedArray

NumOpsDispatch

NumOpsDispatchChecked

NumericNative

PolarsArray

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

PolarsObject

Values need to implement this so that they can be stored into a Series and DataFrame

PolarsRoundtemporal

PolarsTemporalGroupbylazy

PolarsTruncatetemporal

PolarsUpsampletemporal

QuantileAggSeries

Reinterpretreinterpret

RenameAliasFnlazy

RollingSeriespolars-ops

RoundSeriespolars-ops

SerReaderpolars-io

SerWriterpolars-io

SeriesBinaryUdflazy

A wrapper trait for any binary closure Fn(Series, Series) -> PolarsResult<Series>

SeriesJoinpolars-ops

SeriesMethodspolars-ops

SeriesOpsTimetemporal

SeriesRankpolars-ops

SeriesSealedpolars-ops

SeriesTrait

SeriesUdflazy

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

StaticArray

StringMethodstemporal

StringNameSpaceImplpolars-ops

TakeChunkedpolars-ops

Gather by ChunkId

TemporalMethodstemporal

TimeMethodstemporal

ToDummiespolars-ops

UdfSchemalazy

VarAggSeries

VecHash

Functions

_coalesce_full_joinpolars-ops

_join_suffix_namepolars-ops

abspolars-ops

Convert numerical values to their absolute value.

alllazy

Selects all columns. Shorthand for col("*").

all_horizontallazy

Create a new column with the bitwise-and of the elements in each row.

any_horizontallazy

Create a new column with the bitwise-or of the elements in each row.

apply_binarylazy

Like map_binary, but used in a group_by-aggregation context.

apply_multiplelazy

Apply a function/closure over the groups of multiple columns. This should only be used in a group_by aggregation.

arangelazy

Generate a range of integers.

arg_sort_bylazy and range

Find the indexes that would sort these series in order of appearance. That means that the first Series will be used to determine the ordering until duplicates are found. Once duplicates are found, the next Series will be used and so on.

arg_wherelazy and arg_where

Get the indices where condition evaluates true.

as_structlazy

Take several expressions and collect them into a StructChunked.

avglazy

Find the mean of all the values in the column named name. Alias for mean.

base_utc_offsettemporal and timezones

binary_exprlazy

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

call_categorical_merge_operation

castlazy

Casts the column given by Expr to a different type.

check_projected_arrow_schemapolars-io

Checks if the projected columns are equal

check_projected_schemapolars-io

Checks if the projected columns are equal

check_projected_schema_implpolars-io

clippolars-ops

Set values outside the given boundaries to the boundary value.

clip_maxpolars-ops

Set values above the given maximum to the maximum value.

clip_minpolars-ops

Set values below the given minimum to the minimum value.

coalescelazy

Folds the expressions from left to right keeping the first non-null values.

coalesce_seriespolars-ops

collazy

Create a Column Expression based on a column name.

collect_alllazy

Collect all LazyFrame computations.

colslazy

Select multiple columns by name.

concatlazy

Concat multiple LazyFrames vertically.

concat_exprlazy

concat_lf_diagonallazy and diagonal_concat

Concat LazyFrames diagonally. Calls concat internally.

concat_lf_horizontallazy

Concat LazyFrames horizontally.

concat_listlazy

Concat lists entries.

concat_strconcat_str and strings and lazy

Horizontally concat string columns in linear time

convert_inner_type

Cast null arrays to inner type and ensure that all offsets remain correct

convert_to_unsigned_indexpolars-ops

count_rowspolars-io

Read the number of rows without parsing columns useful for count(*) queries

create_enum_data_typedtype-categorical

cum_countpolars-ops

cum_fold_exprslazy and dtype-struct

Accumulate over multiple columns horizontally / row wise.

cum_maxpolars-ops

Get an array with the cumulative max computed at every element.

cum_minpolars-ops

Get an array with the cumulative min computed at every element.

cum_prodpolars-ops

Get an array with the cumulative product computed at every element.

cum_reduce_exprslazy and dtype-struct

Accumulate over multiple columns horizontally / row wise.

cum_sumpolars-ops

Get an array with the cumulative sum computed at every element

date_rangeslazy and temporal

Create a column of date ranges from a start and stop expression.

datetimelazy and temporal

Construct a column of Datetime from the provided DatetimeArgs.

datetime_rangelazy and dtype-datetime

Create a datetime range from a start and stop expression.

datetime_rangeslazy and dtype-datetime

Create a column of datetime ranges from a start and stop expression.

datetime_to_timestamp_ms

datetime_to_timestamp_ns

datetime_to_timestamp_us

default_join_idspolars-ops

deserializepolars-io

Deserializes the statistics in the column chunks from a single row_group into Statistics associated from field’s name.

diffpolars-ops

dst_offsettemporal and timezones

dtype_collazy

Select multiple columns by dtype.

dtype_colslazy

Select multiple columns by dtype.

durationlazy and temporal

Construct a column of Duration from the provided DurationArgs

ensure_duration_matches_data_typetemporal

ensure_is_constant_durationtemporal

firstlazy

First column in a DataFrame.

floor_div_seriespolars-ops

fmt_group_by_column

fold_exprslazy

Accumulate over multiple columns horizontally / row wise.

format_strconcat_str and strings and lazy

Format the results of an array of expressions using a format string

get_reader_bytespolars-io

get_sequential_row_statisticspolars-io

Compute remaining_rows_to_read to be taken per file up front, so we can actually read concurrently/parallel

group_by_valuestemporal

Different from group_by_windows, where define window buckets and search which values fit that pre-defined bucket, this function defines every window based on the: - timestamp (lower bound) - timestamp + period (upper bound) where timestamps are the individual values in the array time

group_by_windowstemporal

Window boundaries are created based on the given Window, which is defined by:

hor_str_concatpolars-ops

Horizontally concatenate all strings.

impl_replace_time_zonepolars-ops

impl_replace_time_zone_fastpolars-ops

If ambiguous is length-1 and not equal to “null”, we can take a slightly faster path.

in_nanoseconds_windowtemporal

index_colslazy

Select multiple columns by index.

indexes_to_usizes

infer_file_schemapolars-io

Infer the schema of a CSV file by reading through the first n rows of the file, with max_read_rows controlling the maximum number of rows to read.

int_rangelazy

Generate a range of integers.

int_rangeslazy

Generate a range of integers for each row of the input columns.

interpolatepolars-ops

interpolate_bypolars-ops

is_betweenpolars-ops

is_cloud_urlpolars-io

Check if the path is a cloud url.

is_compressedpolars-io

check if csv file is compressed

is_first_distinctpolars-ops

is_inpolars-ops

is_last_distinctpolars-ops

is_not_nulllazy

A column which is false wherever expr is null, true elsewhere.

is_nulllazy

A column which is true wherever expr is null, false elsewhere.

is_positive_idx_uncertainpolars-ops

May give false negatives because it ignores the null values.

lastlazy

Last column in a DataFrame.

lenlazy

Return the number of rows in the context.

litlazy

Create a Literal Expression from L. A literal expression behaves like a column that contains a single distinct value.

make_categoricals_compatible

make_list_categoricals_compatible

map_binarylazy

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

map_list_multiplelazy

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiplelazy

Apply a function/closure over multiple columns once the logical plan get executed.

materialize_empty_dfpolars-io

materialize_projectionpolars-io

maxlazy

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

meanlazy

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

medianlazy

Find the median of all the values in the column named name. Shorthand for col(name).median().

merge_dtypes

minlazy

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

negatepolars-ops

negate_bitwisepolars-ops

new_int_rangepolars-ops

notlazy

Negates a boolean column.

nthlazy

Nth column in a DataFrame.

private_left_join_multiple_keyspolars-ops

quantilelazy

Find a specific quantile of all the values in the column named name.

reduce_exprslazy

Analogous to Iterator::reduce.

repeatlazy

Create a column of length n containing n copies of the literal value. Generally you won’t need this function, as lit(value) already represents a column containing only value whose length is automatically set to the correct number of rows.

repeat_bypolars-ops

replacepolars-ops

replace_time_zonepolars-ops

resolve_homedirpolars-io

split_helperpolars-ops

split_to_structpolars-ops and dtype-struct

str_joinpolars-ops

strip_charspolars-ops

strip_chars_endpolars-ops

strip_chars_startpolars-ops

strip_prefixpolars-ops

strip_suffixpolars-ops

sumlazy

Sum all the values in the column named name. Shorthand for col(name).sum().

ternary_exprlazy

time_rangeslazy and dtype-time

Create a column of time ranges from a start and stop expression.

whenlazy

Start a when-then-otherwise expression.

Type Aliases

AllowedOptimizationslazy

AllowedOptimizations

ArrayChunkeddtype-array

ArrayRef

BinaryChunked

BinaryChunkedBuilder

BinaryOffsetChunked

BooleanChunked

BorrowIdxItem

ChunkJoinOptIdspolars-ops and chunked_ids

DateChunked

DatetimeChunked

DecimalChunked

DurationChunked

DynArgs

FieldRef

FieldsNameMapperlazy and dtype-struct

FileMetaDataRefpolars-io

FillNullLimit

Float32Chunked

Float64Chunked

GetOutputlazy

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxCaNon-bigidx

IdxItem

IdxSizeNon-bigidx

IdxTypeNon-bigidx

InnerJoinIdspolars-ops

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

Int128Chunkeddtype-decimal

LargeBinaryArray

LargeListArray

LargeStringArray

LeftJoinIdspolars-ops

ListChunked

NullableChunkId

ObjectChunkedobject

PathIteratorlazy

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PolarsResult

RowGroupIterColumnspolars-io

SchemaRef

StringChunked

StringChunkedBuilder

TimeChunked

TimeZone

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked