polars

Module prelude

Modules

_csv_read_internalpolars-io

_internalpolars-io

aggregations

arity

arraypolars-ops

binarylazy

bufferpolars-io

byte_sourcepolars-io

catlazy

chunkedarraytemporal

Traits and utilities for temporal data.

cloudpolars-io

Interface with cloud storage through the object_store crate.

compressionpolars-io

concat_arrpolars-ops

datatypes

Data types supported by Polars.

datetimepolars-ops

default_arrays

deletionlazy

dtlazy

expr

filepolars-io

fill_null

fixed_size_list

float_sorted_arg_max

full

function_exprlazy

gather

interpolatepolars-ops

interpolate_bypolars-ops

mkdirpolars-io

modepolars-ops

named_serdelazy

nan_propagating_aggregatepolars-ops

null

replacetemporal

roundpolars-ops

row_encode

schema_inferencepolars-io

search_sorted

seriestemporal

sort

stringspolars-ops

sync_on_closepolars-io

udflazy

utf8

zip

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

with_match_categorical_physical_type

Structs

AnonymousScanArgslazy

AnonymousScanOptionslazy

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrayNameSpacelazy

Specialized expressions for Series of DataType::Array.

ArrowField

Represents Arrow’s metadata of a “column”.

AsOfOptionspolars-ops

BatchedCsvReaderpolars-io

BinaryOffsetType

BinaryType

BooleanChunkedBuilder

BooleanType

Boundstemporal

BoundsItertemporal

BrotliLevelpolars-io

A valid Brotli compression level.

CastColumnsPolicylazy

Used by scans.

Categorical8Type

Categorical16Type

Categorical32Type

CategoricalMapping

CategoricalNameSpacelazy

Specialized expressions for Categorical dtypes.

CategoricalType

Categories

A (named) object which is used to indicate which categorical data types have the same mapping.

ChainedThenlazy

Utility struct for the when-then-otherwise expression.

ChainedWhenlazy

Utility struct for the when-then-otherwise expression.

ChunkId

ChunkedArray

ChunkedArray

CompatLevel

CrossJoinOptionspolars-ops

CsvParseOptionspolars-io

CsvReadOptionspolars-io

CsvReaderpolars-io

Create a new DataFrame by reading a csv file.

CsvWriterpolars-io

Write a DataFrame to csv.

CsvWriterOptionspolars-io

Options for writing CSV files.

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeArgslazy

Arguments used by datetime in order to produce an Expr of Datetime

DatetimeType

DecimalType

Dimension

DistinctOptionsDSLlazy

DslBuilderlazy

Durationlazy

DurationArgslazy

Arguments used by duration in order to produce an Expr of Duration

DurationType

DynamicGroupOptionslazy

ExprNameNameSpacelazy

Specialized expressions for modifying the name of existing expressions.

FalseT

Field

Characterizes the name and the DataType of a column.

FileMetadatapolars-io

Metadata for a Parquet file.

FileSinkOptionslazy

FileSinkTypelazy

FixedSizeListType

Float32Type

Float64Type

FrozenCategories

An ordered collection of unique strings with an associated pre-computed mapping to go from string <-> index.

GroupBy

Returned by a group_by operation on a DataFrame. This struct supports several aggregations.

GroupPositions

GroupbyOptionslazy

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsTypeIter

GroupsTypeParIter

GzipLevelpolars-io

A valid Gzip compression level.

HConcatOptionslazy

IEJoinOptionspolars-ops

InProcessQuerylazy

Int8Type

Int16Type

Int32Type

Int64Type

Int128Type

IpcReadOptionspolars-io

IpcReaderpolars-io

Read Arrows IPC format into a DataFrame

IpcReaderAsyncpolars-io

An Arrow IPC reader implemented on top of PolarsObjectStore.

IpcScanOptionspolars-io

IpcStreamReaderpolars-io

Read Arrows Stream IPC format into a DataFrame

IpcStreamWriterpolars-io

Write a DataFrame to Arrow’s Streaming IPC format

IpcStreamWriterOptionpolars-io

IpcWriterpolars-io

Write a DataFrame to Arrow’s IPC format

IpcWriterOptionspolars-io

JoinArgslazy

JoinBuilderlazy

JoinOptionslazy

JoinOptionsIRlazy

JsonLineReaderpolars-io

JsonReaderpolars-io

Reads JSON in one of the formats in JsonFormat into a DataFrame.

JsonWriterpolars-io

Writes a DataFrame to JSON.

JsonWriterOptionspolars-io

LazyCsvReaderlazy

LazyFramelazy

Lazy abstraction over an eager DataFrame.

LazyGroupBylazy

Utility struct for lazy group_by operation.

LazyJsonLineReaderlazy

ListBinaryChunkedBuilder

ListBooleanChunkedBuilder

ListNameSpacelazy

Specialized expressions for Series of DataType::List.

ListPrimitiveChunkedBuilder

ListStringChunkedBuilder

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

LogicalPlanUdfOptionslazy

MatchToSchemaPerColumnlazy

MetadataKeyValuepolars-io

NDJsonReadOptionslazy

NoNull

Just a wrapper structure which is useful for certain impl specializations.

Nulllazy

The literal Null

NullableIdxSize

ObjectType

OptFlagslazy

Allowed optimizations.

OwnedBatchedCsvReaderpolars-io

OwnedObject

ParquetFieldOverwritespolars-io

ParquetMetadataContextpolars-io

Context that can be used to construct custom file-level key value metadata for a Parquet file.

ParquetObjectStorepolars-io

ParquetOptionspolars-io

ParquetReaderpolars-io

Read Apache parquet format into a DataFrame.

ParquetWriteOptionspolars-io

ParquetWriterpolars-io

Write a DataFrame to Parquet format.

PartitionSinkTypelazy

PartitionSinkTypeIRlazy

PartitionTargetContextlazy

PartitionTargetContextKeylazy

PlSmallStr

String type that inlines small strings.

PlanSerializationContextlazy

PrimitiveChunkedBuilder

RankOptionslazy

RollingCovOptionslazy

RollingGroupOptionslazy

RollingOptionsDynamicWindowtemporal

RollingOptionsFixedWindow

RollingVarParams

Scalar

ScanArgsAnonymouslazy

ScanArgsIpclazy

ScanArgsParquetlazy

ScanFlagslazy

ScanSourceIterlazy

An iterator for ScanSources

SerializeOptionspolars-io

Options to serialize logical types to CSV.

Series

Series

SinkFinishContextlazy

SinkOptionslazy

Options that apply to all sinks.

SinkWrittenlazy

SortColumnlazy

SortColumnIRlazy

SortMultipleOptions

Sort options for multi-series sorting.

SortOptions

Options for single series sorting.

SpecialEqlazy

Wrapper type that has special equality properties depending on the inner type specialization

SplitNCharspolars-ops

StatisticsOptionspolars-io

The statistics to write

StringType

StrptimeOptionslazy

StructArraypolars-io

A StructArray is a nested [Array] with an optional validity representing multiple [Array] with the same number of rows.

StructNameSpacelazy

Specialized expressions for Struct dtypes.

StructType

Thenlazy

Utility struct for the when-then-otherwise expression.

TimeType

TimeUnitSetlazy

TimeZone

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

UnifiedScanArgslazy

Scan arguments shared across different scan types.

UnionArgslazy

UnionOptionslazy

UnpivotArgsDSLlazy

UnpivotArgsIR

Arguments for LazyFrame::unpivot function

UserDefinedFunctionlazy

Represents a user-defined function

Whenlazy

Utility struct for the when-then-otherwise expression.

Windowtemporal

Represents a window in time

ZstdLevelpolars-io

A valid Zstandard compression level.

Enums

AggExprlazy

Ambiguous

AnyValue

ArrayFunctionlazy

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

AsofStrategypolars-ops

BinaryFunctionlazy

BitwiseFunctionlazy

BooleanFunctionlazy

CategoricalFunctionlazy

CategoricalPhysical

The physical datatype backing a categorical / enum.

ChildFieldOverwritespolars-io

ClosedIntervalpolars-ops

ClosedWindowtemporal

Column

A column within a DataFrame.

CommentPrefixpolars-io

CsvEncodingpolars-io

DataType

DataTypeExprlazy

DataTypeSelectorlazy

DslPlanlazy

Enginelazy

EvalVariantlazy

Excludedlazy

Exprlazy

Expressions that can be used in various contexts.

ExtraColumnsPolicylazy

FileScanDsllazy

FileScanIRlazy

FileTypelazy

FillNullStrategy

FunctionExprlazy

GroupByMethod

GroupsIndicator

GroupsType

IndexOrder

InequalityOperatorpolars-ops

InterpolationMethodpolars-ops

IpcCompressionpolars-io

Compression codec

JoinCoalescepolars-ops

JoinTypelazy

JoinTypeOptionspolars-ops

JoinTypeOptionsIRlazy

JoinValidationlazy

JsonFormatpolars-io

The format to use to write the DataFrame to JSON: Json (a JSON array) or JsonLines (each row output on a separate line).

KeyValueMetadatapolars-io

Key/value pairs that can be attached to a Parquet file as file-level metadtaa.

Labeltemporal

LazySerdelazy

ListFunctionlazy

LiteralValuelazy

MaintainOrderJoinpolars-ops

MissingColumnsPolicylazy

MissingColumnsPolicyOrExprlazy

NestedTypelazy

NonExistent

NullStrategypolars-ops

NullValuespolars-io

Operatorlazy

ParallelStrategypolars-io

ParquetCompressionpolars-io

The compression strategy to use for writing Parquet files.

ParquetStatisticspolars-io

Parquet statistics for a nesting level

PartitionTargetCallbacklazy

PartitionTargetCallbackResultlazy

PartitionVariantlazy

PartitionVariantIRlazy

PlPath

A Path or URI

PlanCallbacklazy

PolarsError

PowFunctionlazy

QuantileMethod

QuoteStylepolars-io

Quote style indicating when to insert quotes around a field.

RandomMethodlazy

RangeFunctionlazy

RankMethodlazy

RenameAliasFnlazy

ReshapeDimension

A dimension in a reshape.

RollingFnParams

RollingFunctionlazy

RollingFunctionBylazy

RoundModepolars-ops

ScanSourcelazy

A single source to scan from

ScanSourceReflazy

A reference to a single item in ScanSources

ScanSourceslazy

Set of sources to scan from

SearchSortedSidepolars-ops

Selectorlazy

SinkFinishCallbacklazy

SinkTargetlazy

SinkTypelazy

SinkTypeIRlazy

StartBytemporal

StringFunctionlazy

StructFunctionlazy

TemporalFunctionlazy

TimeUnit

TimeZoneSetlazy

UniqueKeepStrategy

UnknownKind

UpcastOrForbidlazy

WindowMappinglazy

WindowTypelazy

Constants

BUILD_STREAMING_EXECUTORlazy

IDX_DTYPE

NULLlazy

URL_ENCODE_CHAR_SETpolars-io

Statics

BOOLEAN_REpolars-io

DSL_VERSIONlazy

EXTENSION_NAME

FLOAT_REpolars-io

FLOAT_RE_DECIMALpolars-io

INTEGER_REpolars-io

POLARS_TEMP_DIR_BASE_PATHpolars-io

Traits

AnonymousScanlazy

ArgAggpolars-ops

Argmin/ Argmax

ArithmeticChunked

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsBinarypolars-ops

AsListpolars-ops

AsRefDataType

AsStringpolars-ops

AsofJoinpolars-ops

AsofJoinBypolars-ops

BinaryNameSpaceImplpolars-ops

BinaryUdfOutputFieldlazy

CatNative

CategoricalPhysicalDtypeExt

ChunkAgg

Aggregation operations.

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkApproxNUnique

ChunkBitwiseReduce

Bitwise Reduction Operations.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompareEq

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkCompareIneq

Compare Series and ChunkedArray’s using inequality operators (<, >=, etc.) and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/flatten a List or String Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkNestingUtils

Utility methods for dealing with nested chunked arrays.

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkRollApply

This differs from ChunkWindowCustom and ChunkWindow by not using a fold aggregator, but reusing a Series wrapper and calling Series aggregators. This likely is a bit slower than ChunkWindow

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

ChunkedBuilder

ChunkedCollectInferIterExt

ChunkedCollectIterExt

ChunkedSetpolars-ops

ColumnBinaryUdflazy

A wrapper trait for any binary closure Fn(Column, Column) -> PolarsResult<Column>

ColumnsUdflazy

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

CrossJoinpolars-ops

CrossJoinFilterpolars-ops

DataFrameJoinOpspolars-ops

DataFrameOpspolars-ops

DateMethodstemporal

DatetimeMethodstemporal

DurationMethodstemporal

FromData

FromDataBinary

FromDataUtf8

FunctionOutputFieldlazy

GetAnyValue

IndexToUsize

InitHashMaps

InitHashMaps2

IntoColumn

Convert Self into a Column

IntoGroupsType

Used to create the tuples for a group_by operation.

IntoLazylazy

IntoMetadata

IntoScalar

IntoSeries

Used to convert a ChunkedArray, &dyn SeriesTrait and Series into a Series.

IntoVec

Convenience for x.into_iter().map(Into::into).collect() using an into_vec() function.

IsFirstDistinct

Mask the first unique values as true

IsLastDistinct

Mask the last unique values as true

JoinDispatchpolars-ops

LazyFileListReaderlazy

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

LhsNumOps

ListBuilderTrait

ListFromIter

ListNameSpaceImplpolars-ops

Literallazy

LogicalType

MetaDataExt

MinMaxHorizontalpolars-ops

MomentSeriespolars-ops

NamedFrom

NamedFromOwned

NewChunkedArray

NumOpsDispatch

NumOpsDispatchChecked

NumericNative

PolarsCategoricalType

Safety

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

PolarsObject

Values need to implement this so that they can be stored into a Series and DataFrame

PolarsPhysicalType

PolarsRoundtemporal

PolarsTemporalGroupbylazy

PolarsTruncatetemporal

PolarsUpsampletemporal

QuantileAggSeries

Reinterpret

RoundSeriespolars-ops

SchemaExt

SchemaExtPl

SchemaNamesAndDtypes

SeedableFromU64SeedExt

SerReaderpolars-io

SerWriterpolars-io

SeriesJoinpolars-ops

SeriesMethodspolars-ops

SeriesOpsTimetemporal

SeriesRankpolars-ops

SeriesSealedpolars-ops

SeriesTrait

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

StaticArray

StringMethodstemporal

StringNameSpaceImplpolars-ops

SumMeanHorizontalpolars-ops

TakeChunkedpolars-ops

Gather by ChunkId

TakeChunkedHorParpolars-ops

TemporalMethodstemporal

TimeMethodstemporal

ToDummiespolars-ops

UdfSchemalazy

VarAggSeries

VecHash

Functions

_coalesce_full_joinpolars-ops

_join_suffix_namepolars-ops

_set_check_length⚠

Meant for internal use. In very rare conditions this can be turned off.

abspolars-ops

Convert numerical values to their absolute value.

alllazy

Selects all columns.

all_horizontallazy

Create a new column with the bitwise-and of the elements in each row.

any_horizontallazy

Create a new column with the bitwise-or of the elements in each row.

apply_binarylazy

Like map_binary, but used in a group_by-aggregation context.

apply_multiplelazy

Apply a function/closure over the groups of multiple columns. This should only be used in a group_by aggregation.

apply_projectionpolars-io

arangelazy

Generate a range of integers.

arg_sort_bylazy

Find the indexes that would sort these series in order of appearance.

arg_wherelazy

Get the indices where condition evaluates true.

as_structlazy

Take several expressions and collect them into a StructChunked.

avglazy

Find the mean of all the values in the column named name. Alias for mean.

base_utc_offsettemporal

binary_exprlazy

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

build_streaming_query_executorlazy

by_namelazy

Select multiple columns by dtype.

castlazy

Casts the column given by Expr to a different type.

clippolars-ops

Set values outside the given boundaries to the boundary value.

clip_maxpolars-ops

Set values above the given maximum to the maximum value.

clip_minpolars-ops

Set values below the given minimum to the minimum value.

coalescelazy

Folds the expressions from left to right keeping the first non-null values.

coalesce_columnspolars-ops

collazy

Create a Column Expression based on a column name.

collect_alllazy

Collect all LazyFrame computations.

colslazy

Select multiple columns by name.

columns_to_projectionpolars-io

concatlazy

Concat multiple LazyFrames vertically.

concat_arrlazy

Horizontally concatenate columns into a single array-type column.

concat_exprlazy

concat_lf_diagonallazy

Concat LazyFrames diagonally. Calls concat internally.

concat_lf_horizontallazy

Concat LazyFrames horizontally.

concat_listlazy

Concat lists entries.

concat_strlazy

Horizontally concat string columns in linear time

convert_inner_type

Cast null arrays to inner type and ensure that all offsets remain correct

convert_to_unsigned_indexpolars-ops

count_onespolars-ops

count_rowspolars-io

Read the number of rows without parsing columns useful for count(*) queries

count_rows_from_slicepolars-io

Read the number of rows without parsing columns

count_rows_from_slice_parpolars-io

Read the number of rows without parsing columns useful for count(*) queries

count_zerospolars-ops

create_sorting_mappolars-io

cum_countpolars-ops

cum_fold_exprslazy

Accumulate over multiple columns horizontally / row wise.

cum_maxpolars-ops

Get an array with the cumulative max computed at every element.

cum_minpolars-ops

Get an array with the cumulative min computed at every element.

cum_prodpolars-ops

Get an array with the cumulative product computed at every element.

cum_reduce_exprslazy

Accumulate over multiple columns horizontally / row wise.

cum_sumpolars-ops

Get an array with the cumulative sum computed at every element

date_rangeslazy

Create a column of date ranges from a start and stop expression.

datetimelazy

Construct a column of Datetime from the provided DatetimeArgs.

datetime_rangelazy

Create a datetime range from a start and stop expression.

datetime_rangeslazy

Create a column of datetime ranges from a start and stop expression.

datetime_to_timestamp_ms

datetime_to_timestamp_ns

datetime_to_timestamp_us

decode_json_responsepolars-io

Utility for decoding JSON that adds the response value to the error message if decoding fails. This makes it much easier to debug errors from parsing network responses.

default_join_idspolars-ops

deserializepolars-io

Deserializes the statistics in the column chunks from a single row_group into Statistics associated from field’s name.

diffpolars-ops

dst_offsettemporal

dtype_collazy

Select multiple columns by dtype.

dtype_colslazy

Select multiple columns by dtype.

durationlazy

Construct a column of Duration from the provided DurationArgs

emptylazy

Selects no columns.

ensure_duration_matches_dtypetemporal

ensure_is_constant_durationtemporal

ensure_matching_schema

ensure_same_categories

ensure_same_frozen_categories

escape_regexpolars-ops

escape_regex_strpolars-ops

estimate_n_lines_in_chunkpolars-io

Total len divided by max len of first and last non-empty lines. This is intended to be cheaper than estimate_n_lines_in_file.

estimate_n_lines_in_filepolars-io

expand_pathspolars-io

Recursively traverses directories and expands globs if glob is true.

expand_paths_hivepolars-io

Recursively traverses directories and expands globs if glob is true. Returns the expanded paths and the index at which to start parsing hive partitions from the path.

expanded_from_single_directorypolars-io

Returns true if expanded_paths were expanded from a single directory

firstlazy

First column in a DataFrame.

floor_div_seriespolars-ops

fmt_group_by_column

fold_exprslazy

Accumulate over multiple columns horizontally / row wise.

format_strlazy

Format the results of an array of expressions using a format string

get_column_write_optionspolars-io

get_glob_start_idxpolars-io

Get the index of the first occurrence of a glob symbol.

get_reader_bytespolars-io

get_strftime_format

group_by_valuestemporal

Different from group_by_windows, where define window buckets and search which values fit that pre-defined bucket.

group_by_windowstemporal

Window boundaries are created based on the given Window, which is defined by:

hor_str_concatpolars-ops

Horizontally concatenate all strings.

impl_durationpolars-ops

impl_replace_time_zonepolars-ops

impl_replace_time_zone_fastpolars-ops

If ambiguous is length-1 and not equal to “null”, we can take a slightly faster path.

in_nanoseconds_windowtemporal

index_colslazy

Select multiple columns by index.

indexes_to_usizes

infer_file_schemapolars-io

Infer the schema of a CSV file by reading through the first n rows of the file, with max_read_rows controlling the maximum number of rows to read.

infer_schemapolars-io

Infers a ArrowSchema from parquet’s FileMetadata.

int_rangelazy

Generate a range of integers.

int_rangeslazy

Generate a range of integers for each row of the input columns.

interpolatepolars-ops

interpolate_bypolars-ops

is_betweenpolars-ops

is_closepolars-ops

is_first_distinctpolars-ops

is_inpolars-ops

is_last_distinctpolars-ops

is_not_nulllazy

A column which is false wherever expr is null, true elsewhere.

is_nulllazy

A column which is true wherever expr is null, false elsewhere.

is_positive_idx_uncertainpolars-ops

May give false negatives because it ignores the null values.

is_positive_idx_uncertain_colpolars-ops

May give false negatives because it ignores the null values.

json_linespolars-io

known_timezonestemporal

lastlazy

Last column in a DataFrame.

leading_onespolars-ops

leading_zerospolars-ops

lenlazy

Return the number of rows in the context.

linear_spacelazy

Generate a series of equally-spaced points.

linear_spaceslazy

Create a column of linearly-spaced sequences from ‘start’, ‘end’, and ‘num_samples’ expressions.

litlazy

Create a Literal Expression from L. A literal expression behaves like a column that contains a single distinct value.

map_binarylazy

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

map_multiplelazy

Apply a function/closure over multiple columns once the logical plan get executed.

materialize_empty_dfpolars-io

materialize_projectionpolars-io

maxlazy

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

meanlazy

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

medianlazy

Find the median of all the values in the column named name. Shorthand for col(name).median().

merge_dtypes

minlazy

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

negatepolars-ops

negate_bitwisepolars-ops

new_int_rangepolars-ops

new_linear_space_f32polars-ops

new_linear_space_f64polars-ops

notlazy

Negates a boolean column.

nthlazy

Nth column in a DataFrame.

overwrite_schemapolars-io

parse_ndjsonpolars-io

prepare_cloud_planlazy

Prepare the given DslPlan for execution on Polars Cloud.

private_left_join_multiple_keyspolars-ops

quantilelazy

Find a specific quantile of all the values in the column named name.

reduce_exprslazy

Analogous to Iterator::reduce.

remove_bompolars-io

repeatlazy

Create a column of length n containing n copies of the literal value.

repeat_bypolars-ops

replacepolars-ops

Replace values by different values of the same data type.

replace_datetemporal

Replace specific time component of a DateChunked with a specified value.

replace_datetimetemporal

Replace specific time component of a DatetimeChunked with a specified value.

replace_or_defaultpolars-ops

Replace all values by different values.

replace_strictpolars-ops

Replace all values by different values.

replace_time_zonepolars-ops

resolve_homedirpolars-io

Replaces a “~” in the Path with the home directory.

rlepolars-ops

Get the lengths of runs of identical values.

rle_idpolars-ops

Similar to rle, but maps values to run IDs.

rle_lengthspolars-ops

Get the run-Lengths of values.

rolling_kurtosispolars-ops

rolling_skewpolars-ops

split_helperpolars-ops

split_to_structpolars-ops

str_joinpolars-ops

strip_charspolars-ops

strip_chars_endpolars-ops

strip_chars_startpolars-ops

strip_prefixpolars-ops

strip_suffixpolars-ops

substring_ternary_offsets_valuepolars-ops

sumlazy

Sum all the values in the column named name. Shorthand for col(name).sum().

ternary_exprlazy

time_rangeslazy

Create a column of time ranges from a start and stop expression.

trailing_onespolars-ops

trailing_zerospolars-ops

try_raise_keyboard_interrupt

Checks if the keyboard interrupt flag is set, and if yes panics as a keyboard interrupt. This function is very cheap.

try_set_sorted_flagpolars-io

unique_countspolars-ops

Returns a count of the unique values in the order of appearance.

unpack_dtypes

update_viewpolars-ops

whenlazy

Start a when-then-otherwise expression.

write_partitioned_datasetpolars-io

Write a partitioned parquet dataset. This functionality is unstable.

Type Aliases

AllowedOptimizationslazy

AllowedOptimizations

ArrayChunked

ArrayRef

ArrowSchema

An ordered sequence of Fields

BinaryChunked

BinaryChunkedBuilder

BinaryOffsetChunked

BooleanChunked

BorrowIdxItem

CatSize

Categorical8Chunked

Categorical16Chunked

Categorical32Chunked

CategoricalChunked

ChunkJoinOptIdspolars-ops

DateChunked

DatetimeChunked

DecimalChunked

DurationChunked

FieldRef

FieldsNameMapperlazy

FileMetadataRefpolars-io

FillNullLimit

Float32Chunked

Float64Chunked

GetOutputlazy

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxCa

IdxItem

IdxSize

IdxType

InnerJoinIdspolars-ops

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

Int128Chunked

LargeBinaryArray

LargeListArray

LargeStringArray

LeftJoinIdspolars-ops

ListChunked

ObjectChunked

OpaqueColumnUdflazy

PlFixedStateQuality

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PlRandomState

PlRandomStateQuality

PlSeedableRandomStateQuality

PolarsResult

RenameAliasRustFnlazy

RowGroupIterColumnspolars-io

Schema

SchemaRef

StringChunked

StringChunkedBuilder

StructChunked

TimeChunked

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked