Module polars::prelude

Modules

datatypes

Data types supported by Polars.

full

writeio_csv_write

APIs to write to CSV

zip

Macros

df

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

A logical DataType and its associated metadata per Arrow specification

ArrowSchema

An ordered sequence of Field with optional metadata.

BoolTakeRandom

BoolTakeRandomSingleChunk

BooleanChunkedBuilder

BooleanType

CategoricalType

ChunkedArray

ChunkedArray

CsvReader

Create a new DataFrame by reading a csv file.

CsvWriter

Write a DataFrame to csv.

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeType

Field

Float32Type

Float64Type

Int8Type

Int16Type

Int32Type

Int64Type

ListBooleanChunkedBuilder

ListPrimitiveChunkedBuilder

ListTakeRandom

ListTakeRandomSingleChunk

ListType

ListUtf8ChunkedBuilder

Logical

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

NoNull

Just a wrapper structure. Useful for certain impl specializations This is for instance use to implement impl<T> FromIterator<T::Native> for NoNull<ChunkedArray<T>> as Option<T::Native> was already implemented: impl<T> FromIterator<Option<T::Native>> for ChunkedArray<T>

NumTakeRandomChunked

NumTakeRandomCont

NumTakeRandomSingleChunk

PrimitiveChunkedBuilder

Schema

Series

Series

SortOptions

TimeType

UInt8Type

UInt16Type

UInt32Type

UInt64Type

Utf8ChunkedBuilder

Utf8TakeRandom

Utf8TakeRandomSingleChunk

Utf8Type

Enums

AnyValue

ArrowDataType

The set of supported logical types. Each variant uniquely identifies a logical type, which define specific semantics to the data (e.g. how it should be represented). Each variant has a corresponding PhysicalType, obtained via DataType::to_physical_type, which declares the in-memory representation of data. The DataType::Extension is special in that it augments a DataType with metadata to support custom types. Use to_logical_type to desugar such type and return its correspoding logical type.

CsvEncoding

DataType

FillNullStrategy

JoinType

NullValues

PolarsError

TakeIdx

One of the three arguments allowed in unchecked_take

TakeRandBranch2

TakeRandBranch3

TimeUnit

The time units defined in Arrow.

Traits

ArgAgg

Argmin/ Argmax

ChunkAgg

Aggregation operations

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray when the operation is cheaper than branching due to null checking

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompare

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/ flatten a

ChunkFillNull

Replace None values with various strategies

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkOps

ChunkPeaks

Find local minima/ maxima

ChunkReverse

Reverse a ChunkedArray

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask. Note that these operations clone data. This is however the only way we can modify at mask or index level as the underlying Arrow arrays are immutable.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

Fast access by index.

ChunkTakeEvery

Traverse and collect every nth element

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine 2 ChunkedArrays based on some predicate.

ChunkedBuilder

IntoNoNullIterator

Trait for ChunkedArrays that don’t have null values. The result is the most efficient implementation Iterator, according to the number of chunks.

IntoSeries

IntoTakeRandom

Create a type that implements a faster TakeRandom.

IntoVec

LhsNumOps

ListBuilderTrait

LogicalType

NamedFrom

NewChunkedArray

NumOpsDispatch

NumericNative

PolarsDataType

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

PolarsSingleType

Any type that is not nested

Pow

SerReader

SerWriter

SeriesTrait

TakeIterator

TakeIteratorNulls

TakeRandom

Random access

TakeRandomUtf8

ToDummies

VarAggSeries

VecHash

Functions

naive_datetime_to_datetime

resolve_homedir

Type Definitions

AlignedVec

BooleanChunked

CategoricalChunked

DateChunked

DatetimeChunked

Dummy

Dummy type, we need to instantiate all generic types, so we fill one with a dummy.

Float32Chunked

Float64Chunked

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

LargeListArray

LargeStringArray

ListChunked

PlHashMap

PlHashSet

Result

SchemaRef

TimeChunked

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked

Utf8Chunked