polars

Module prelude

Modules

_csv_read_internal

aggregations

arity

buffer

chunkedarray

Traits and utilities for temporal data.

cloud

Interface with cloud storage through the object_store crate.

compression

datatypes

Data types supported by Polars.

default_arrays

expr

file

fill_null

float_sorted_arg_max

full

gather

iceberg

TODO

mkdir

null

replace

row_encode

schema_inference

search_sorted

series

sort

string

strings

sync_on_close

utf8

zip

Macros

df

polars_bail

polars_ensure

polars_err

polars_warn

with_match_categorical_physical_type

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrowField

Represents Arrow’s metadata of a “column”.

BatchedCsvReader

BatchedWriter

BinaryOffsetType

BinaryType

BooleanChunkedBuilder

BooleanType

Bounds

BoundsIter

Categorical8Type

Categorical16Type

Categorical32Type

CategoricalMapping

CategoricalType

Categories

A (named) object which is used to indicate which categorical data types have the same mapping.

ChunkId

ChunkedArray

ChunkedArray

CompatLevel

CrossJoinOptions

CsvParseOptions

CsvReadOptions

CsvReader

Create a new DataFrame by reading a csv file.

CsvWriter

Write a DataFrame to csv.

CsvWriterOptions

Options for writing CSV files.

DataFrame

A contiguous growable collection of Series that have the same length.

DateType

DatetimeType

Dimension

Duration

DurationType

DynamicGroupOptions

FalseT

Field

Characterizes the name and the DataType of a column.

Float32Type

Float64Type

FrozenCategories

An ordered collection of unique strings with an associated pre-computed mapping to go from string <-> index.

GroupBy

Returned by a group_by operation on a DataFrame. This struct supports several aggregations.

GroupPositions

GroupsIdx

Indexes of the groups, the first index is stored separately. this make sorting fast.

GroupsTypeIter

GroupsTypeParIter

Int8Type

Int16Type

Int32Type

Int64Type

JoinArgs

ListBinaryChunkedBuilder

ListBooleanChunkedBuilder

ListPrimitiveChunkedBuilder

ListStringChunkedBuilder

ListType

Logical

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

NoNull

Just a wrapper structure which is useful for certain impl specializations.

NullChunked

NullableIdxSize

OwnedBatchedCsvReader

PlSmallStr

String type that inlines small strings.

PrimitiveChunkedBuilder

RollingGroupOptions

RollingVarParams

Scalar

SerializeOptions

Options to serialize logical types to CSV.

Series

Series

SortMultipleOptions

Sort options for multi-series sorting.

SortOptions

Options for single series sorting.

StringType

TimeType

TimeZone

TrueT

UInt8Type

UInt16Type

UInt32Type

UInt64Type

UnpivotArgsIR

Arguments for LazyFrame::unpivot function

Window

Represents a window in time

Enums

Ambiguous

AnyValue

ArrowDataType

The set of supported logical types in this crate.

ArrowTimeUnit

The time units defined in Arrow.

CategoricalPhysical

The physical datatype backing a categorical / enum.

ClosedInterval

ClosedWindow

Column

A column within a DataFrame.

CommentPrefix

CsvEncoding

DataType

FillNullStrategy

GroupByMethod

GroupsIndicator

GroupsType

JoinCoalesce

JoinType

JoinTypeOptions

JoinValidation

Label

MaintainOrderJoin

NonExistent

NullStrategy

NullValues

PlPath

A Path or URI

PlPathRef

A reference to a Path or URI

PolarsError

QuantileMethod

QuoteStyle

Quote style indicating when to insert quotes around a field.

ReshapeDimension

A dimension in a reshape.

RollingFnParams

SearchSortedSide

StartBy

TimeUnit

UniqueKeepStrategy

UnknownKind

Constants

IDX_DTYPE

URL_ENCODE_CHAR_SET

Statics

BOOLEAN_RE

EXTENSION_NAME

FLOAT_RE

FLOAT_RE_DECIMAL

INTEGER_RE

POLARS_TEMP_DIR_BASE_PATH

Traits

ArgAgg

Argmin/ Argmax

ArithmeticChunked

ArrayCollectIterExt

ArrayFromIter

ArrayFromIterDtype

AsBinary

AsList

AsRefDataType

AsString

BinaryNameSpaceImpl

CatNative

CategoricalPhysicalDtypeExt

ChunkAgg

Aggregation operations.

ChunkAggSeries

Aggregations that return Series of unit length. Those can be used in broadcasting operations.

ChunkAnyValue

ChunkApply

Fastest way to do elementwise operations on a ChunkedArray<T> when the operation is cheaper than branching due to null checking.

ChunkApplyKernel

Apply kernels on the arrow array chunks in a ChunkedArray.

ChunkBytes

ChunkCast

Cast ChunkedArray<T> to ChunkedArray<N>

ChunkCompareEq

Compare Series and ChunkedArray’s and get a boolean mask that can be used to filter rows.

ChunkCompareIneq

Compare Series and ChunkedArray’s using inequality operators (<, >=, etc.) and get a boolean mask that can be used to filter rows.

ChunkExpandAtIndex

Create a new ChunkedArray filled with values at that index.

ChunkExplode

Explode/flatten a List or String Series

ChunkFillNullValue

Replace None values with a value

ChunkFilter

Filter values by a boolean mask.

ChunkFull

Fill a ChunkedArray with one value.

ChunkFullNull

ChunkNestingUtils

Utility methods for dealing with nested chunked arrays.

ChunkQuantile

Quantile and median aggregation.

ChunkReverse

Reverse a ChunkedArray<T>

ChunkSet

Create a ChunkedArray with new values by index or by boolean mask.

ChunkShift

ChunkShiftFill

Shift the values of a ChunkedArray by a number of periods.

ChunkSort

Sort operations on ChunkedArray.

ChunkTake

ChunkTakeUnchecked

ChunkUnique

Get unique values in a ChunkedArray

ChunkVar

Variance and standard deviation aggregation.

ChunkZip

Combine two ChunkedArray based on some predicate.

ChunkedBuilder

ChunkedCollectInferIterExt

ChunkedCollectIterExt

ChunkedSet

CrossJoin

CrossJoinFilter

DataFrameJoinOps

DataFrameOps

DateMethods

DatetimeMethods

DurationMethods

FromData

FromDataBinary

FromDataUtf8

GetAnyValue

IndexToUsize

InitHashMaps

InitHashMaps2

IntoColumn

Convert Self into a Column

IntoGroupsType

Used to create the tuples for a group_by operation.

IntoMetadata

IntoScalar

IntoSeries

Used to convert a ChunkedArray, &dyn SeriesTrait and Series into a Series.

IntoVec

Convenience for x.into_iter().map(Into::into).collect() using an into_vec() function.

JoinDispatch

LhsNumOps

ListBuilderTrait

ListFromIter

ListNameSpaceImpl

LogicalType

MetaDataExt

MinMaxHorizontal

NamedFrom

NamedFromOwned

NewChunkedArray

NumOpsDispatch

NumericNative

PolarsCategoricalType

Safety

PolarsDataType

Safety

PolarsFloatType

PolarsIntegerType

PolarsIterator

A PolarsIterator is an iterator over a ChunkedArray which contains polars types. A PolarsIterator must implement ExactSizeIterator and DoubleEndedIterator.

PolarsNumericType

PolarsPhysicalType

PolarsRound

PolarsTemporalGroupby

PolarsTruncate

PolarsUpsample

QuantileAggSeries

SchemaExt

SchemaExtPl

SchemaNamesAndDtypes

SeedableFromU64SeedExt

SerReader

SerWriter

SeriesJoin

SeriesMethods

SeriesSealed

SeriesTrait

SlicedArray

Utility trait to slice concrete arrow arrays whilst keeping their concrete type. E.g. don’t return Box<dyn Array>.

StaticArray

StringMethods

SumMeanHorizontal

TemporalMethods

TimeMethods

VarAggSeries

VecHash

Functions

_coalesce_full_join

_join_suffix_name

_set_check_length⚠

Meant for internal use. In very rare conditions this can be turned off.

clip

Set values outside the given boundaries to the boundary value.

clip_max

Set values above the given maximum to the maximum value.

clip_min

Set values below the given minimum to the minimum value.

coalesce_columns

convert_inner_type

Cast null arrays to inner type and ensure that all offsets remain correct

convert_to_unsigned_index

count_ones

count_rows

Read the number of rows without parsing columns useful for count(*) queries

count_rows_from_slice

Read the number of rows without parsing columns

count_rows_from_slice_par

Read the number of rows without parsing columns useful for count(*) queries

count_zeros

date_range

Create a DatetimeChunked from a given start and end date and a given interval.

datetime_to_timestamp_ms

datetime_to_timestamp_ns

datetime_to_timestamp_us

default_join_ids

ensure_duration_matches_dtype

ensure_is_constant_duration

ensure_matching_schema

ensure_same_categories

ensure_same_frozen_categories

expand_paths

Recursively traverses directories and expands globs if glob is true.

expand_paths_hive

Recursively traverses directories and expands globs if glob is true. Returns the expanded paths and the index at which to start parsing hive partitions from the path.

expanded_from_single_directory

Returns true if expanded_paths were expanded from a single directory

fmt_group_by_column

get_glob_start_idx

Get the index of the first occurrence of a glob symbol.

get_reader_bytes

get_strftime_format

group_by_values

Different from group_by_windows, where define window buckets and search which values fit that pre-defined bucket.

group_by_windows

Window boundaries are created based on the given Window, which is defined by:

impl_duration

in_nanoseconds_window

indexes_to_usizes

infer_file_schema

Infer the schema of a CSV file by reading through the first n rows of the file, with max_read_rows controlling the maximum number of rows to read.

is_positive_idx_uncertain

May give false negatives because it ignores the null values.

is_positive_idx_uncertain_col

May give false negatives because it ignores the null values.

leading_ones

leading_zeros

materialize_projection

max_horizontal

mean_horizontal

merge_dtypes

min_horizontal

negate

negate_bitwise

new_int_range

new_linear_space_f32

new_linear_space_f64

private_left_join_multiple_keys

replace_date

Replace specific time component of a DateChunked with a specified value.

replace_datetime

Replace specific time component of a DatetimeChunked with a specified value.

resolve_homedir

Replaces a “~” in the Path with the home directory.

sum_horizontal

time_range

Create a TimeChunked from a given start and end date and a given interval.

trailing_ones

trailing_zeros

try_raise_keyboard_interrupt

Checks if the keyboard interrupt flag is set, and if yes panics as a keyboard interrupt. This function is very cheap.

unpack_dtypes

Type Aliases

ArrayRef

ArrowSchema

An ordered sequence of Fields

BinaryChunked

BinaryChunkedBuilder

BinaryOffsetChunked

BooleanChunked

BorrowIdxItem

CatSize

Categorical8Chunked

Categorical16Chunked

Categorical32Chunked

CategoricalChunked

ChunkJoinIds

ChunkJoinOptIds

DateChunked

DatetimeChunked

DurationChunked

FieldRef

FillNullLimit

Float32Chunked

Float64Chunked

GroupsSlice

Every group is indicated by an array where the

IdxArr

IdxCa

IdxItem

IdxSize

IdxType

InnerJoinIds

Int8Chunked

Int16Chunked

Int32Chunked

Int64Chunked

LargeBinaryArray

LargeListArray

LargeStringArray

LeftJoinIds

ListChunked

PlFixedStateQuality

PlHashMap

PlHashSet

PlIdHashMap

This hashmap uses an IdHasher

PlIndexMap

PlIndexSet

PlRandomState

PlRandomStateQuality

PlSeedableRandomStateQuality

PolarsResult

Schema

SchemaRef

StringChunked

StringChunkedBuilder

TimeChunked

UInt8Chunked

UInt16Chunked

UInt32Chunked

UInt64Chunked