polars::prelude::datatypes

Struct Logical

pub struct Logical<Logical, Physical>
where
    Logical: PolarsDataType,
    Physical: PolarsDataType,
{
    pub phys: ChunkedArray<Physical>,
    pub dtype: DataType,
    /* private fields */
}

Maps a logical type to a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Fields
phys: ChunkedArray<Physical>dtype: DataType
Implementations

impl<T> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>
where T: PolarsCategoricalType,

pub fn sort_with( &self, options: SortOptions, ) -> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

pub fn sort( &self, descending: bool, ) -> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

Returned a sorted ChunkedArray.

pub fn arg_sort(&self, options: SortOptions) -> ChunkedArray<UInt32Type>

Retrieve the indexes needed to sort this array.

impl Logical<DecimalType, Int128Type>

pub fn precision(&self) -> Option<usize>

pub fn scale(&self) -> usize

pub fn to_scale( &self, scale: usize, ) -> Result<Cow<'_, Logical<DecimalType, Int128Type>>, PolarsError>

impl<T> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>
where T: PolarsCategoricalType,

pub fn is_enum(&self) -> bool

pub fn uses_lexical_ordering(&self) -> bool

Return whether or not the CategoricalChunked uses the lexical order of the string values when sorting.

pub fn full_null_with_dtype( name: PlSmallStr, length: usize, dtype: DataType, ) -> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

pub fn from_cats_and_dtype( cat_ids: ChunkedArray<<T as PolarsCategoricalType>::PolarsPhysical>, dtype: DataType, ) -> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

Create a CategoricalChunked from a physical array and dtype.

Checks that all the category ids are valid, mapping invalid ones to nulls.

pub unsafe fn from_cats_and_dtype_unchecked( cat_ids: ChunkedArray<<T as PolarsCategoricalType>::PolarsPhysical>, dtype: DataType, ) -> Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>

Create a CategoricalChunked from a physical array and dtype.

Safety

It’s not checked that the indices are in-bounds or that the dtype is correct.

pub fn get_mapping(&self) -> &Arc<CategoricalMapping>

Get a reference to the mapping of categorical types to the string values.

pub fn iter_str(&self) -> impl PolarsIterator

Create an Iterator that iterates over the &str values of the CategoricalChunked.

pub fn from_str_iter<'a, I>( name: PlSmallStr, dtype: DataType, strings: I, ) -> Result<Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>, PolarsError>
where I: IntoIterator<Item = Option<&'a str>>,

Converts from strings to this CategoricalChunked.

If this dtype is an Enum any non-existing strings get mapped to null.

pub fn to_arrow( &self, compat_level: CompatLevel, ) -> DictionaryArray<<T as PolarsCategoricalType>::Native>

impl<K, T> Logical<K, T>
where K: PolarsDataType, T: PolarsDataType,

pub unsafe fn new_logical( phys: ChunkedArray<T>, dtype: DataType, ) -> Logical<K, T>

Safety

You must uphold the logical types’ invariants.

impl<K, T> Logical<K, T>
where K: PolarsDataType, T: PolarsDataType, Logical<K, T>: LogicalType,

pub fn name(&self) -> &PlSmallStr

pub fn rename(&mut self, name: PlSmallStr)

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn null_count(&self) -> usize

pub fn has_nulls(&self) -> bool

pub fn is_null(&self) -> ChunkedArray<BooleanType>

pub fn is_not_null(&self) -> ChunkedArray<BooleanType>

pub fn split_at(&self, offset: i64) -> (Logical<K, T>, Logical<K, T>)

pub fn slice(&self, offset: i64, length: usize) -> Logical<K, T>

pub fn field(&self) -> Field

pub fn physical(&self) -> &ChunkedArray<T>

pub fn physical_mut(&mut self) -> &mut ChunkedArray<T>

pub fn into_physical(self) -> ChunkedArray<T>

impl Logical<DateType, Int32Type>

pub fn as_date_iter(&self) -> impl TrustedLen

pub fn from_naive_date<I>( name: PlSmallStr, v: I, ) -> Logical<DateType, Int32Type>
where I: IntoIterator<Item = NaiveDate>,

Construct a new DateChunked from an iterator over NaiveDate.

pub fn to_string( &self, format: &str, ) -> Result<ChunkedArray<StringType>, PolarsError>

Convert from Date into String with the given format. See chrono strftime/strptime.

pub fn strftime( &self, format: &str, ) -> Result<ChunkedArray<StringType>, PolarsError>

Convert from Date into String with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_date_options<I>( name: PlSmallStr, v: I, ) -> Logical<DateType, Int32Type>
where I: IntoIterator<Item = Option<NaiveDate>>,

Construct a new DateChunked from an iterator over optional NaiveDate.

impl Logical<DatetimeType, Int64Type>

pub fn as_datetime_iter(&self) -> impl TrustedLen

pub fn time_unit(&self) -> TimeUnit

pub fn time_zone(&self) -> &Option<TimeZone>

pub fn to_string( &self, format: &str, ) -> Result<ChunkedArray<StringType>, PolarsError>

Convert from Datetime into String with the given format. See chrono strftime/strptime.

pub fn strftime( &self, format: &str, ) -> Result<ChunkedArray<StringType>, PolarsError>

Convert from Datetime into String with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_datetime<I>( name: PlSmallStr, v: I, tu: TimeUnit, ) -> Logical<DatetimeType, Int64Type>
where I: IntoIterator<Item = NaiveDateTime>,

Construct a new DatetimeChunked from an iterator over NaiveDateTime.

pub fn from_naive_datetime_options<I>( name: PlSmallStr, v: I, tu: TimeUnit, ) -> Logical<DatetimeType, Int64Type>
where I: IntoIterator<Item = Option<NaiveDateTime>>,

pub fn cast_time_unit(&self, tu: TimeUnit) -> Logical<DatetimeType, Int64Type>

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, time_unit: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.

pub fn set_time_zone(&mut self, time_zone: TimeZone) -> Result<(), PolarsError>

Change the underlying TimeZone. This does not modify the data. This does not validate the time zone - it’s up to the caller to verify that it’s already been validated.

pub fn set_time_unit_and_time_zone( &mut self, time_unit: TimeUnit, time_zone: TimeZone, ) -> Result<(), PolarsError>

Change the underlying TimeUnit and TimeZone. This does not modify the data. This does not validate the time zone - it’s up to the caller to verify that it’s already been validated.

impl Logical<DurationType, Int64Type>

pub fn time_unit(&self) -> TimeUnit

pub fn cast_time_unit(&self, tu: TimeUnit) -> Logical<DurationType, Int64Type>

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, tu: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.

pub fn to_string( &self, format: &str, ) -> Result<ChunkedArray<StringType>, PolarsError>

Convert from Duration to String; note that strftime format strings are not supported, only the specifiers ‘iso’ and ‘polars’.

pub fn from_duration<I>( name: PlSmallStr, v: I, tu: TimeUnit, ) -> Logical<DurationType, Int64Type>
where I: IntoIterator<Item = TimeDelta>,

Construct a new DurationChunked from an iterator over ChronoDuration.

pub fn from_duration_options<I>( name: PlSmallStr, v: I, tu: TimeUnit, ) -> Logical<DurationType, Int64Type>
where I: IntoIterator<Item = Option<TimeDelta>>,

Construct a new DurationChunked from an iterator over optional ChronoDuration.

impl Logical<TimeType, Int64Type>

pub fn to_string(&self, format: &str) -> ChunkedArray<StringType>

Convert from Time into String with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> ChunkedArray<StringType>

Convert from Time into String with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn as_time_iter(&self) -> impl TrustedLen

pub fn from_naive_time<I>( name: PlSmallStr, v: I, ) -> Logical<TimeType, Int64Type>
where I: IntoIterator<Item = NaiveTime>,

Construct a new TimeChunked from an iterator over NaiveTime.

pub fn from_naive_time_options<I>( name: PlSmallStr, v: I, ) -> Logical<TimeType, Int64Type>
where I: IntoIterator<Item = Option<NaiveTime>>,

Construct a new TimeChunked from an iterator over optional NaiveTime.