Trait polars::prelude::datatypes::StaticArray

pub trait StaticArray: Array + for<'a> ArrayFromIterDtype<Self::ValueT<'a>> + for<'a> ArrayFromIterDtype<Option<Self::ValueT<'a>>> {
    type ValueT<'a>
       where Self: 'a;
    type ValueIterT<'a>: Iterator<Item = Self::ValueT<'a>> + TrustedLen + TrustedLen
       where Self: 'a;

    // Required methods
    unsafe fn value_unchecked(&self, idx: usize) -> Self::ValueT<'_>;
    fn iter(
        &self
    ) -> ZipValidity<Self::ValueT<'_>, Self::ValueIterT<'_>, BitmapIter<'_>> ;
    fn values_iter(&self) -> Self::ValueIterT<'_>;
    fn with_validity_typed(self, validity: Option<Bitmap>) -> Self;

    // Provided methods
    fn get(&self, idx: usize) -> Option<Self::ValueT<'_>> { ... }
    unsafe fn get_unchecked(&self, idx: usize) -> Option<Self::ValueT<'_>> { ... }
    fn last(&self) -> Option<Self::ValueT<'_>> { ... }
    fn value(&self, idx: usize) -> Self::ValueT<'_> { ... }
}

Required Associated Types

type ValueT<'a> where Self: 'a

type ValueIterT<'a>: Iterator<Item = Self::ValueT<'a>> + TrustedLen + TrustedLen where Self: 'a

Required Methods

unsafe fn value_unchecked(&self, idx: usize) -> Self::ValueT<'_>

Safety

It is the callers responsibility that the idx < self.len().

fn iter( &self ) -> ZipValidity<Self::ValueT<'_>, Self::ValueIterT<'_>, BitmapIter<'_>>

fn values_iter(&self) -> Self::ValueIterT<'_>

fn with_validity_typed(self, validity: Option<Bitmap>) -> Self

Provided Methods

fn get(&self, idx: usize) -> Option<Self::ValueT<'_>>

unsafe fn get_unchecked(&self, idx: usize) -> Option<Self::ValueT<'_>>

Safety

It is the callers responsibility that the idx < self.len().

fn last(&self) -> Option<Self::ValueT<'_>>

fn value(&self, idx: usize) -> Self::ValueT<'_>