polars::prelude::dt

Struct DateLikeNameSpace

pub struct DateLikeNameSpace(/* private fields */);

Available on crate feature lazy only.

Specialized expressions for Series with dates/datetimes.

Implementations

impl DateLikeNameSpace

pub fn to_string(self, format: &str) -> Expr

Convert from Date/Time/Datetime into String with the given format. See chrono strftime/strptime.

pub fn strftime(self, format: &str) -> Expr

Convert from Date/Time/Datetime into String with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn cast_time_unit(self, tu: TimeUnit) -> Expr

Change the underlying TimeUnit. And update the data accordingly.

pub fn with_time_unit(self, tu: TimeUnit) -> Expr

Change the underlying TimeUnit of the Series. This does not modify the data.

pub fn convert_time_zone(self, time_zone: TimeZone) -> Expr

Change the underlying TimeZone of the Series. This does not modify the data.

pub fn millennium(self) -> Expr

Get the millennium of a Date/Datetime

pub fn century(self) -> Expr

Get the century of a Date/Datetime

pub fn year(self) -> Expr

Get the year of a Date/Datetime

pub fn is_leap_year(self) -> Expr

pub fn iso_year(self) -> Expr

Get the iso-year of a Date/Datetime. This may not correspond with a calendar year.

pub fn month(self) -> Expr

Get the month of a Date/Datetime.

pub fn quarter(self) -> Expr

Extract quarter from underlying NaiveDateTime representation. Quarters range from 1 to 4.

pub fn week(self) -> Expr

Extract the week from the underlying Date representation. Can be performed on Date and Datetime

Returns the ISO week number starting from 1. The return value ranges from 1 to 53. (The last week of year differs by years.)

pub fn weekday(self) -> Expr

Extract the ISO week day from the underlying Date representation. Can be performed on Date and Datetime.

Returns the weekday number where monday = 1 and sunday = 7

pub fn day(self) -> Expr

Get the month of a Date/Datetime.

pub fn ordinal_day(self) -> Expr

Get the ordinal_day of a Date/Datetime.

pub fn time(self) -> Expr

Get the (local) time of a Date/Datetime/Time.

pub fn date(self) -> Expr

Get the (local) date of a Date/Datetime.

pub fn datetime(self) -> Expr

Get the (local) datetime of a Datetime.

pub fn hour(self) -> Expr

Get the hour of a Datetime/Time64.

pub fn minute(self) -> Expr

Get the minute of a Datetime/Time64.

pub fn second(self) -> Expr

Get the second of a Datetime/Time64.

pub fn millisecond(self) -> Expr

Get the millisecond of a Time64 (scaled from nanosecs).

pub fn microsecond(self) -> Expr

Get the microsecond of a Time64 (scaled from nanosecs).

pub fn nanosecond(self) -> Expr

Get the nanosecond part of a Time64.

pub fn timestamp(self, tu: TimeUnit) -> Expr

Return the timestamp (UNIX epoch) of a Datetime/Date.

pub fn truncate(self, every: Expr) -> Expr

Truncate the Datetime/Date range into buckets.

pub fn base_utc_offset(self) -> Expr

Get the base offset from UTC.

pub fn dst_offset(self) -> Expr

Get the additional offset from UTC currently in effect (usually due to daylight saving time).

pub fn round(self, every: Expr) -> Expr

Round the Datetime/Date range into buckets.

pub fn replace_time_zone( self, time_zone: Option<TimeZone>, ambiguous: Expr, non_existent: NonExistent, ) -> Expr

pub fn combine(self, time: Expr, tu: TimeUnit) -> Expr

Combine an existing Date/Datetime with a Time, creating a new Datetime value.

pub fn total_days(self) -> Expr

Express a Duration in terms of its total number of integer days.

pub fn total_hours(self) -> Expr

Express a Duration in terms of its total number of integer hours.

pub fn total_minutes(self) -> Expr

Express a Duration in terms of its total number of integer minutes.

pub fn total_seconds(self) -> Expr

Express a Duration in terms of its total number of integer seconds.

pub fn total_milliseconds(self) -> Expr

Express a Duration in terms of its total number of milliseconds.

pub fn total_microseconds(self) -> Expr

Express a Duration in terms of its total number of microseconds.

pub fn total_nanoseconds(self) -> Expr

Express a Duration in terms of its total number of nanoseconds.

pub fn replace( self, year: Expr, month: Expr, day: Expr, hour: Expr, minute: Expr, second: Expr, microsecond: Expr, ambiguous: Expr, ) -> Expr

Replace the time units of a value