polars::prelude

Enum AnyValue

pub enum AnyValue<'a> {

    Null,
    Boolean(bool),
    String(&'a str),
    UInt8(u8),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Int128(i128),
    Float32(f32),
    Float64(f64),
    Date(i32),
    Datetime(i64, TimeUnit, Option<&'a TimeZone>),
    DatetimeOwned(i64, TimeUnit, Option<Arc<TimeZone>>),
    Duration(i64, TimeUnit),
    Time(i64),
    Categorical(u32, &'a Arc<CategoricalMapping>),
    CategoricalOwned(u32, Arc<CategoricalMapping>),
    Enum(u32, &'a Arc<CategoricalMapping>),
    EnumOwned(u32, Arc<CategoricalMapping>),
    List(Series),
    Array(Series, usize),
    Object(&'a (dyn PolarsObjectSafe + 'static)),
    ObjectOwned(OwnedObject),
    Struct(usize, &'a StructArray, &'a [Field]),
    StructOwned(Box<(Vec<AnyValue<'a>>, Vec<Field>)>),
    StringOwned(PlSmallStr),
    Binary(&'a [u8]),
    BinaryOwned(Vec<u8>),
    Decimal(i128, usize),

}

Variants

Null

Boolean(bool)

A binary true or false.

String(&'a str)

A UTF8 encoded string type.

UInt8(u8)

An unsigned 8-bit integer number.

UInt16(u16)

An unsigned 16-bit integer number.

UInt32(u32)

An unsigned 32-bit integer number.

UInt64(u64)

An unsigned 64-bit integer number.

Int8(i8)

An 8-bit integer number.

Int16(i16)

A 16-bit integer number.

Int32(i32)

A 32-bit integer number.

Int64(i64)

A 64-bit integer number.

Int128(i128)

A 128-bit integer number.

Float32(f32)

A 32-bit floating point number.

Float64(f64)

A 64-bit floating point number.

Date(i32)

A 32-bit date representing the elapsed time since UNIX epoch (1970-01-01) in days (32 bits).

Datetime(i64, TimeUnit, Option<&'a TimeZone>)

A 64-bit date representing the elapsed time since UNIX epoch (1970-01-01) in nanoseconds (64 bits).

DatetimeOwned(i64, TimeUnit, Option<Arc<TimeZone>>)

A 64-bit date representing the elapsed time since UNIX epoch (1970-01-01) in nanoseconds (64 bits).

Duration(i64, TimeUnit)

A 64-bit integer representing difference between date-times in TimeUnit

Time(i64)

A 64-bit time representing the elapsed time since midnight in nanoseconds

Categorical(u32, &'a Arc<CategoricalMapping>)

CategoricalOwned(u32, Arc<CategoricalMapping>)

Enum(u32, &'a Arc<CategoricalMapping>)

EnumOwned(u32, Arc<CategoricalMapping>)

List(Series)

Nested type, contains arrays that are filled with one of the datatypes.

Array(Series, usize)

Object(&'a (dyn PolarsObjectSafe + 'static))

Can be used to fmt and implements Any, so can be downcasted to the proper value type.

ObjectOwned(OwnedObject)

Struct(usize, &'a StructArray, &'a [Field])

StructOwned(Box<(Vec<AnyValue<'a>>, Vec<Field>)>)

StringOwned(PlSmallStr)

An UTF8 encoded string type.

Binary(&'a [u8])

BinaryOwned(Vec<u8>)

Decimal(i128, usize)

A 128-bit fixed point decimal number with a scale.

Implementations

impl<'a> AnyValue<'a>

pub fn _iter_struct_av(&self) -> impl Iterator<Item = AnyValue<'_>>

pub fn _materialize_struct_av(&'a self, buf: &mut Vec<AnyValue<'a>>)

impl AnyValue<'static>

pub fn zero_sum(dtype: &DataType) -> AnyValue<'static>

pub fn can_have_dtype(&self, dtype: &DataType) -> bool

Can the AnyValue exist as having dtype as its DataType.

impl<'a> AnyValue<'a>

pub fn dtype(&self) -> DataType

Get the matching DataType for this AnyValue`.

Note: For Categorical and Enum values, the exact mapping information is not preserved in the result for performance reasons.

pub fn try_extract<T>(&self) -> Result<T, PolarsError>
where T: NumCast,

pub fn is_boolean(&self) -> bool

pub fn is_primitive_numeric(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_integer(&self) -> bool

pub fn is_signed_integer(&self) -> bool

pub fn is_unsigned_integer(&self) -> bool

pub fn is_nan(&self) -> bool

pub fn is_null(&self) -> bool

pub fn is_nested_null(&self) -> bool

pub fn strict_cast(&self, dtype: &'a DataType) -> Option<AnyValue<'a>>

Cast AnyValue to the provided data type and return a new AnyValue with type dtype, if possible.

pub fn try_strict_cast( &self, dtype: &'a DataType, ) -> Result<AnyValue<'a>, PolarsError>

Cast AnyValue to the provided data type and return a new AnyValue with type dtype, if possible.

pub fn cast(&self, dtype: &'a DataType) -> AnyValue<'a>

pub fn idx(&self) -> u32

pub fn str_value(&self) -> Cow<'a, str>

pub fn to_physical(self) -> AnyValue<'a>

pub fn extract_bool(&self) -> Option<bool>

pub fn extract_str(&self) -> Option<&str>

pub fn extract_bytes(&self) -> Option<&[u8]>

impl AnyValue<'_>

pub fn hash_impl<H>(&self, state: &mut H, cheap: bool)
where H: Hasher,

impl<'a> AnyValue<'a>

pub fn add(&self, rhs: &AnyValue<'_>) -> AnyValue<'static>

pub fn as_borrowed(&self) -> AnyValue<'_>

pub fn into_static(self) -> AnyValue<'static>

Try to coerce to an AnyValue with static lifetime. This can be done if it does not borrow any values.

pub fn get_str(&self) -> Option<&str>

Get a reference to the &str contained within AnyValue.

impl AnyValue<'_>

pub fn eq_missing(&self, other: &AnyValue<'_>, null_equal: bool) -> bool