Enum polars::prelude::AnyValue

pub enum AnyValue<'a> {

    Null,
    Boolean(bool),
    Utf8(&'a str),
    UInt8(u8),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Float32(f32),
    Float64(f64),
    Date(i32),
    Datetime(i64, TimeUnit, &'a Option<String>),
    Duration(i64, TimeUnit),
    Time(i64),
    Categorical(u32, &'a RevMapping),
    List(Series),
    Object(&'a (dyn PolarsObjectSafe + 'static)),

}

Variants

Null

Boolean(bool)

Tuple Fields

0: bool

A binary true or false.

Utf8(&'a str)

Tuple Fields

0: &'a str

A UTF8 encoded string type.

UInt8(u8)

Tuple Fields

0: u8

An unsigned 8-bit integer number.

UInt16(u16)

Tuple Fields

0: u16

An unsigned 16-bit integer number.

UInt32(u32)

Tuple Fields

0: u32

An unsigned 32-bit integer number.

UInt64(u64)

Tuple Fields

0: u64

An unsigned 64-bit integer number.

Int8(i8)

Tuple Fields

0: i8

An 8-bit integer number.

Int16(i16)

Tuple Fields

0: i16

A 16-bit integer number.

Int32(i32)

Tuple Fields

0: i32

A 32-bit integer number.

Int64(i64)

Tuple Fields

0: i64

A 64-bit integer number.

Float32(f32)

Tuple Fields

0: f32

A 32-bit floating point number.

Float64(f64)

Tuple Fields

0: f64

A 64-bit floating point number.

Date(i32)

Tuple Fields

0: i32

A 32-bit date representing the elapsed time since UNIX epoch (1970-01-01) in days (32 bits).

Datetime(i64, TimeUnit, &'a Option<String>)

Tuple Fields

0: i64

1: TimeUnit

2: &'a Option<String>

A 64-bit date representing the elapsed time since UNIX epoch (1970-01-01) in nanoseconds (64 bits).

Duration(i64, TimeUnit)

Tuple Fields

0: i64

1: TimeUnit

Time(i64)

Tuple Fields

0: i64

A 64-bit time representing the elapsed time since midnight in nanoseconds

Categorical(u32, &'a RevMapping)

Tuple Fields

0: u32

1: &'a RevMapping

List(Series)

Tuple Fields

0: Series

Nested type, contains arrays that are filled with one of the datetypes.

Object(&'a (dyn PolarsObjectSafe + 'static))

Tuple Fields

0: &'a (dyn PolarsObjectSafe + 'static)

Can be used to fmt and implements Any, so can be downcasted to the proper value type.

Implementations

impl<'a> AnyValue<'a>

pub fn add(&self, rhs: &AnyValue<'b>) -> AnyValue<'a>

pub fn to_static(&self) -> Result<AnyValue<'static>, PolarsError>

Try to coerce to an AnyValue with static lifetime. This can be done if it does not borrow any values.