polars::prelude

Enum Expr

pub enum Expr {

    Alias(Arc<Expr>, PlSmallStr),
    Column(PlSmallStr),
    Selector(Selector),
    Literal(LiteralValue),
    BinaryExpr {
        left: Arc<Expr>,
        op: Operator,
        right: Arc<Expr>,
    },
    Cast {
        expr: Arc<Expr>,
        dtype: DataTypeExpr,
        options: CastOptions,
    },
    Sort {
        expr: Arc<Expr>,
        options: SortOptions,
    },
    Gather {
        expr: Arc<Expr>,
        idx: Arc<Expr>,
        returns_scalar: bool,
    },
    SortBy {
        expr: Arc<Expr>,
        by: Vec<Expr>,
        sort_options: SortMultipleOptions,
    },
    Agg(AggExpr),
    Ternary {
        predicate: Arc<Expr>,
        truthy: Arc<Expr>,
        falsy: Arc<Expr>,
    },
    Function {
        input: Vec<Expr>,
        function: FunctionExpr,
    },
    Explode {
        input: Arc<Expr>,
        skip_empty: bool,
    },
    Filter {
        input: Arc<Expr>,
        by: Arc<Expr>,
    },
    Window {
        function: Arc<Expr>,
        partition_by: Vec<Expr>,
        order_by: Option<(Arc<Expr>, SortOptions)>,
        options: WindowType,
    },
    Slice {
        input: Arc<Expr>,
        offset: Arc<Expr>,
        length: Arc<Expr>,
    },
    KeepName(Arc<Expr>),
    Len,
    Field(Arc<[PlSmallStr]>),
    AnonymousFunction {
        input: Vec<Expr>,
        function: LazySerde<SpecialEq<Arc<dyn ColumnsUdf>>>,
        output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>,
        options: FunctionOptions,
        fmt_str: Box<PlSmallStr>,
    },
    Eval {
        expr: Arc<Expr>,
        evaluation: Arc<Expr>,
        variant: EvalVariant,
    },
    SubPlan(SpecialEq<Arc<DslPlan>>, Vec<String>),
    RenameAlias {
        function: RenameAliasFn,
        expr: Arc<Expr>,
    },

}

Available on crate feature lazy only.

Expressions that can be used in various contexts.

Queries consist of multiple expressions. When using the polars lazy API, don’t construct an Expr directly; instead, create one using the functions in the polars_lazy::dsl module. See that module’s docs for more info.

Variants

Alias(Arc<Expr>, PlSmallStr)

Column(PlSmallStr)

Selector(Selector)

Literal(LiteralValue)

BinaryExpr

Fields

left: Arc<Expr>

op: Operator

right: Arc<Expr>

Cast

Fields

expr: Arc<Expr>

dtype: DataTypeExpr

options: CastOptions

Sort

Fields

expr: Arc<Expr>

options: SortOptions

Gather

Fields

expr: Arc<Expr>

idx: Arc<Expr>

returns_scalar: bool

SortBy

Fields

expr: Arc<Expr>

by: Vec<Expr>

sort_options: SortMultipleOptions

Agg(AggExpr)

Ternary

A ternary operation if true then “foo” else “bar”

Fields

predicate: Arc<Expr>

truthy: Arc<Expr>

falsy: Arc<Expr>

Function

Fields

input: Vec<Expr>

function arguments

function: FunctionExpr

function to apply

Explode

Fields

input: Arc<Expr>

skip_empty: bool

Filter

Fields

input: Arc<Expr>

by: Arc<Expr>

Window

Polars flavored window functions.

Fields

function: Arc<Expr>

Also has the input. i.e. avg(“foo”)

partition_by: Vec<Expr>

order_by: Option<(Arc<Expr>, SortOptions)>

options: WindowType

Slice

Fields

input: Arc<Expr>

offset: Arc<Expr>

length is not yet known so we accept negative offsets

length: Arc<Expr>

KeepName(Arc<Expr>)

Set root name as Alias

Len

Field(Arc<[PlSmallStr]>)

AnonymousFunction

Fields

input: Vec<Expr>

function arguments

function: LazySerde<SpecialEq<Arc<dyn ColumnsUdf>>>

function to apply

output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>

output dtype of the function

options: FunctionOptions

fmt_str: Box<PlSmallStr>

used for formatting

Eval

Evaluates the evaluation expression on the output of the expr.

Consequently, expr is an input and evaluation is not and needs a different schema.

Fields

expr: Arc<Expr>

evaluation: Arc<Expr>

variant: EvalVariant

SubPlan(SpecialEq<Arc<DslPlan>>, Vec<String>)

RenameAlias

Fields

function: RenameAliasFn

expr: Arc<Expr>

Implementations

impl Expr

pub fn floor_div(self, rhs: Expr) -> Expr

Floor divide self by rhs.

pub fn pow<E>(self, exponent: E) -> Expr
where E: Into<Expr>,

Raise expression to the power exponent

pub fn sqrt(self) -> Expr

Compute the square root of the given expression

pub fn cbrt(self) -> Expr

Compute the cube root of the given expression

impl Expr

pub fn bitwise_count_ones(self) -> Expr

Evaluate the number of set bits.

pub fn bitwise_count_zeros(self) -> Expr

Evaluate the number of unset bits.

pub fn bitwise_leading_ones(self) -> Expr

Evaluate the number most-significant set bits before seeing an unset bit.

pub fn bitwise_leading_zeros(self) -> Expr

Evaluate the number most-significant unset bits before seeing an set bit.

pub fn bitwise_trailing_ones(self) -> Expr

Evaluate the number least-significant set bits before seeing an unset bit.

pub fn bitwise_trailing_zeros(self) -> Expr

Evaluate the number least-significant unset bits before seeing an set bit.

pub fn bitwise_and(self) -> Expr

Perform an aggregation of bitwise ANDs

pub fn bitwise_or(self) -> Expr

Perform an aggregation of bitwise ORs

pub fn bitwise_xor(self) -> Expr

Perform an aggregation of bitwise XORs

impl Expr

pub fn to_field( &self, schema: &Schema<DataType>, ctxt: Context, ) -> Result<Field, PolarsError>

Get Field result of the expression. The schema is the input data.

pub fn into_selector(self) -> Option<Selector>

pub fn try_into_selector(self) -> Result<Selector, PolarsError>

pub fn extract_usize(&self) -> Result<usize, PolarsError>

Extract a constant usize from an expression.

pub fn extract_i64(&self) -> Result<i64, PolarsError>

pub fn map_unary(self, function: impl Into<FunctionExpr>) -> Expr

pub fn map_binary(self, function: impl Into<FunctionExpr>, rhs: Expr) -> Expr

pub fn map_ternary( self, function: impl Into<FunctionExpr>, arg1: Expr, arg2: Expr, ) -> Expr

pub fn try_map_n_ary( self, function: impl Into<FunctionExpr>, exprs: impl IntoIterator<Item = Result<Expr, PolarsError>>, ) -> Result<Expr, PolarsError>

pub fn map_n_ary( self, function: impl Into<FunctionExpr>, exprs: impl IntoIterator<Item = Expr>, ) -> Expr

pub fn n_ary(function: impl Into<FunctionExpr>, input: Vec<Expr>) -> Expr

impl Expr

pub fn shuffle(self, seed: Option<u64>) -> Expr

pub fn sample_n( self, n: Expr, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> Expr

pub fn sample_frac( self, frac: Expr, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> Expr

impl Expr

pub fn std(self, ddof: u8) -> Expr

Standard deviation of the values of the Series.

pub fn var(self, ddof: u8) -> Expr

Variance of the values of the Series.

pub fn min(self) -> Expr

Reduce groups to minimal value.

pub fn max(self) -> Expr

Reduce groups to maximum value.

pub fn nan_min(self) -> Expr

Reduce groups to minimal value.

pub fn nan_max(self) -> Expr

Reduce groups to maximum value.

pub fn mean(self) -> Expr

Reduce groups to the mean value.

pub fn median(self) -> Expr

Reduce groups to the median value.

pub fn sum(self) -> Expr

Reduce groups to the sum of all the values.

impl Expr

pub fn eq<E>(self, other: E) -> Expr
where E: Into<Expr>,

Compare Expr with other Expr on equality.

pub fn eq_missing<E>(self, other: E) -> Expr
where E: Into<Expr>,

Compare Expr with other Expr on equality where None == None.

pub fn neq<E>(self, other: E) -> Expr
where E: Into<Expr>,

Compare Expr with other Expr on non-equality.

pub fn neq_missing<E>(self, other: E) -> Expr
where E: Into<Expr>,

Compare Expr with other Expr on non-equality where None == None.

pub fn lt<E>(self, other: E) -> Expr
where E: Into<Expr>,

Check if Expr < Expr.

pub fn gt<E>(self, other: E) -> Expr
where E: Into<Expr>,

Check if Expr > Expr.

pub fn gt_eq<E>(self, other: E) -> Expr
where E: Into<Expr>,

Check if Expr >= Expr.

pub fn lt_eq<E>(self, other: E) -> Expr
where E: Into<Expr>,

Check if Expr <= Expr.

pub fn not(self) -> Expr

Negate Expr.

pub fn alias<S>(self, name: S) -> Expr
where S: Into<PlSmallStr>,

Rename Column.

pub fn is_null(self) -> Expr

Run is_null operation on Expr.

pub fn is_not_null(self) -> Expr

Run is_not_null operation on Expr.

pub fn drop_nulls(self) -> Expr

Drop null values.

pub fn drop_nans(self) -> Expr

Drop NaN values.

pub fn n_unique(self) -> Expr

Get the number of unique values in the groups.

pub fn first(self) -> Expr

Get the first value in the group.

pub fn last(self) -> Expr

Get the last value in the group.

pub fn implode(self) -> Expr

GroupBy the group to a Series.

pub fn quantile(self, quantile: Expr, method: QuantileMethod) -> Expr

Compute the quantile per group.

pub fn agg_groups(self) -> Expr

Get the group indexes of the group by operation.

pub fn flatten(self) -> Expr

Alias for explode.

pub fn explode(self) -> Expr

Explode the String/List column.

pub fn slice<E, F>(self, offset: E, length: F) -> Expr
where E: Into<Expr>, F: Into<Expr>,

Slice the Series. offset may be negative.

pub fn append<E>(self, other: E, upcast: bool) -> Expr
where E: Into<Expr>,

Append expressions. This is done by adding the chunks of other to this Series.

pub fn head(self, length: Option<usize>) -> Expr

Get the first n elements of the Expr result.

pub fn tail(self, length: Option<usize>) -> Expr

Get the last n elements of the Expr result.

pub fn unique(self) -> Expr

Get unique values of this expression.

pub fn unique_stable(self) -> Expr

Get unique values of this expression, while maintaining order. This requires more work than Expr::unique.

pub fn arg_unique(self) -> Expr

Get the first index of unique values of this expression.

pub fn arg_min(self) -> Expr

Get the index value that has the minimum value.

pub fn arg_max(self) -> Expr

Get the index value that has the maximum value.

pub fn arg_sort(self, descending: bool, nulls_last: bool) -> Expr

Get the index values that would sort this expression.

pub fn strict_cast(self, dtype: impl Into<DataTypeExpr>) -> Expr

Cast expression to another data type. Throws an error if conversion had overflows. Returns an Error if cast is invalid on rows after predicates are pushed down.

pub fn cast(self, dtype: impl Into<DataTypeExpr>) -> Expr

Cast expression to another data type.

pub fn cast_with_options( self, dtype: impl Into<DataTypeExpr>, cast_options: CastOptions, ) -> Expr

Cast expression to another data type.

pub fn gather<E>(self, idx: E) -> Expr
where E: Into<Expr>,

Take the values by idx.

pub fn get<E>(self, idx: E) -> Expr
where E: Into<Expr>,

Take the values by a single index.

pub fn sort(self, options: SortOptions) -> Expr

Sort with given options.

Example

let lf = df! {
   "a" => [Some(5), Some(4), Some(3), Some(2), None]
}?
.lazy();

let sorted = lf
    .select(
        vec![col("a").sort(SortOptions::default())],
    )
    .collect()?;

assert_eq!(
    sorted,
    df! {
        "a" => [None, Some(2), Some(3), Some(4), Some(5)]
    }?
);

See SortOptions for more options.

pub fn reverse(self) -> Expr

Reverse column

pub fn map<F>( self, function: F, output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, ) -> Expr
where F: Fn(Column) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

Apply a function/closure once the logical plan get executed.

This function is very similar to Expr::apply, but differs in how it handles aggregations.

map should be used for operations that are independent of groups, e.g. multiply * 2, or raise to the power

apply should be used for operations that work on a group of data. e.g. sum, count, etc.

It is the responsibility of the caller that the schema is correct by giving the correct output_type. If None given the output type of the input expr is used.

pub fn map_with_fmt_str<F>( self, function: F, output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, fmt_str: impl Into<PlSmallStr>, ) -> Expr
where F: Fn(Column) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

pub fn agg_with_fmt_str<F>( self, function: F, output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, fmt_str: impl Into<PlSmallStr>, ) -> Expr
where F: Fn(Column) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

pub fn apply_with_fmt_str<F>( self, function: F, output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, fmt_str: impl Into<PlSmallStr>, ) -> Expr
where F: Fn(Column) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

pub fn map_many<F>( self, function: F, arguments: &[Expr], output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, ) -> Expr
where F: Fn(&mut [Column]) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

Apply a function/closure once the logical plan get executed with many arguments.

See the Expr::map function for the differences between map and apply.

pub fn apply<F>( self, function: F, output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, ) -> Expr
where F: Fn(Column) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

Apply a function/closure over the groups. This should only be used in a group_by aggregation.

It is the responsibility of the caller that the schema is correct by giving the correct output_type. If None given the output type of the input expr is used.

This difference with map is that apply will create a separate Series per group.

map should be used for operations that are independent of groups, e.g. multiply * 2, or raise to the power

apply should be used for operations that work on a group of data. e.g. sum, count, etc.

pub fn apply_many<F>( self, function: F, arguments: &[Expr], output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, ) -> Expr
where F: Fn(&mut [Column]) -> Result<Option<Column>, PolarsError> + 'static + Send + Sync,

Apply a function/closure over the groups with many arguments. This should only be used in a group_by aggregation.

See the Expr::apply function for the differences between map and apply.

pub fn is_finite(self) -> Expr

Get mask of finite values if dtype is Float.

pub fn is_infinite(self) -> Expr

Get mask of infinite values if dtype is Float.

pub fn is_nan(self) -> Expr

Get mask of NaN values if dtype is Float.

pub fn is_not_nan(self) -> Expr

Get inverse mask of NaN values if dtype is Float.

pub fn shift(self, n: Expr) -> Expr

Shift the values in the array by some period. See the eager implementation.

pub fn shift_and_fill<E, IE>(self, n: E, fill_value: IE) -> Expr
where E: Into<Expr>, IE: Into<Expr>,

Shift the values in the array by some period and fill the resulting empty values.

pub fn cumulative_eval(self, evaluation: Expr, min_samples: usize) -> Expr

Cumulatively count values from 0 to len.

pub fn cum_count(self, reverse: bool) -> Expr

Cumulatively count values from 0 to len.

pub fn cum_sum(self, reverse: bool) -> Expr

Get an array with the cumulative sum computed at every element.

pub fn cum_prod(self, reverse: bool) -> Expr

Get an array with the cumulative product computed at every element.

pub fn cum_min(self, reverse: bool) -> Expr

Get an array with the cumulative min computed at every element.

pub fn cum_max(self, reverse: bool) -> Expr

Get an array with the cumulative max computed at every element.

pub fn product(self) -> Expr

Get the product aggregation of an expression.

pub fn round(self, decimals: u32, mode: RoundMode) -> Expr

Round underlying floating point array to given decimal numbers.

pub fn round_sig_figs(self, digits: i32) -> Expr

Round to a number of significant figures.

pub fn floor(self) -> Expr

Floor underlying floating point array to the lowest integers smaller or equal to the float value.

pub fn pi() -> Expr

Constant Pi

pub fn ceil(self) -> Expr

Ceil underlying floating point array to the highest integers smaller or equal to the float value.

pub fn clip(self, min: Expr, max: Expr) -> Expr

Clip underlying values to a set boundary.

pub fn clip_max(self, max: Expr) -> Expr

Clip underlying values to a set boundary.

pub fn clip_min(self, min: Expr) -> Expr

Clip underlying values to a set boundary.

pub fn abs(self) -> Expr

Convert all values to their absolute/positive value.

pub fn over<E, IE>(self, partition_by: E) -> Expr
where E: AsRef<[IE]>, IE: Into<Expr> + Clone,

Apply window function over a subgroup. This is similar to a group_by + aggregation + self join. Or similar to window functions in Postgres.

Example

#[macro_use] extern crate polars_core;
use polars_core::prelude::*;
use polars_lazy::prelude::*;

fn example() -> PolarsResult<()> {
    let df = df! {
            "groups" => &[1, 1, 2, 2, 1, 2, 3, 3, 1],
            "values" => &[1, 2, 3, 4, 5, 6, 7, 8, 8]
        }?;

    let out = df
     .lazy()
     .select(&[
         col("groups"),
         sum("values").over([col("groups")]),
     ])
     .collect()?;
    println!("{}", &out);
    Ok(())
}

Outputs:

╭────────┬────────╮
│ groups ┆ values │
│ ---    ┆ ---    │
│ i32    ┆ i32    │
╞════════╪════════╡
│ 1      ┆ 16     │
│ 1      ┆ 16     │
│ 2      ┆ 13     │
│ 2      ┆ 13     │
│ …      ┆ …      │
│ 1      ┆ 16     │
│ 2      ┆ 13     │
│ 3      ┆ 15     │
│ 3      ┆ 15     │
│ 1      ┆ 16     │
╰────────┴────────╯

pub fn over_with_options<E, IE>( self, partition_by: Option<E>, order_by: Option<(E, SortOptions)>, options: WindowMapping, ) -> Result<Expr, PolarsError>
where E: AsRef<[IE]>, IE: Into<Expr> + Clone,

pub fn rolling(self, options: RollingGroupOptions) -> Expr

pub fn fill_null<E>(self, fill_value: E) -> Expr
where E: Into<Expr>,

Replace the null values by a value.

pub fn fill_null_with_strategy(self, strategy: FillNullStrategy) -> Expr

pub fn fill_nan<E>(self, fill_value: E) -> Expr
where E: Into<Expr>,

Replace the floating point NaN values by a value.

pub fn count(self) -> Expr

Count the values of the Series or Get counts of the group by operation.

pub fn len(self) -> Expr

pub fn is_between<E>(self, lower: E, upper: E, closed: ClosedInterval) -> Expr
where E: Into<Expr>,

pub fn is_close<E>( self, expr: E, abs_tol: f64, rel_tol: f64, nans_equal: bool, ) -> Expr
where E: Into<Expr>,

Check whether floating point values are close to each other.

pub fn approx_n_unique(self) -> Expr

Get the approximate count of unique values.

pub fn and<E>(self, expr: E) -> Expr
where E: Into<Expr>,

Bitwise “and” operation.

pub fn xor<E>(self, expr: E) -> Expr
where E: Into<Expr>,

Bitwise “xor” operation.

pub fn or<E>(self, expr: E) -> Expr
where E: Into<Expr>,

Bitwise “or” operation.

pub fn logical_or<E>(self, expr: E) -> Expr
where E: Into<Expr>,

Logical “or” operation.

pub fn logical_and<E>(self, expr: E) -> Expr
where E: Into<Expr>,

Logical “and” operation.

pub fn filter<E>(self, predicate: E) -> Expr
where E: Into<Expr>,

Filter a single column.

Should be used in aggregation context. If you want to filter on a DataFrame level, use LazyFrame::filter.

pub fn is_in<E>(self, other: E, nulls_equal: bool) -> Expr
where E: Into<Expr>,

Check if the values of the left expression are in the lists of the right expr.

pub fn sort_by<E, IE>(self, by: E, sort_options: SortMultipleOptions) -> Expr
where E: AsRef<[IE]>, IE: Into<Expr> + Clone,

Sort this column by the ordering of another column evaluated from given expr. Can also be used in a group_by context to sort the groups.

Example

let lf = df! {
    "a" => [1, 2, 3, 4, 5],
    "b" => [5, 4, 3, 2, 1]
}?.lazy();

let sorted = lf
    .select(
        vec![col("a").sort_by(col("b"), SortOptions::default())],
    )
    .collect()?;

assert_eq!(
    sorted,
    df! { "a" => [5, 4, 3, 2, 1] }?
);

pub fn repeat_by<E>(self, by: E) -> Expr
where E: Into<Expr>,

Repeat the column n times, where n is determined by the values in by. This yields an Expr of dtype List.

pub fn is_first_distinct(self) -> Expr

Get a mask of the first unique value.

pub fn is_last_distinct(self) -> Expr

Get a mask of the last unique value.

pub fn dot<E>(self, other: E) -> Expr
where E: Into<Expr>,

Compute the dot/inner product between two expressions.

pub fn mode(self) -> Expr

Compute the mode(s) of this column. This is the most occurring value.

pub fn interpolate(self, method: InterpolationMethod) -> Expr

Interpolate intermediate values. Nulls at the beginning and end of the series remain null.

pub fn interpolate_by(self, by: Expr) -> Expr

Interpolate intermediate values. Nulls at the beginning and end of the series remain null. The by column provides the x-coordinates for interpolation and must not contain nulls.

pub fn rolling_min_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling minimum based on another column.

pub fn rolling_max_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling maximum based on another column.

pub fn rolling_mean_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling mean based on another column.

pub fn rolling_sum_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling sum based on another column.

pub fn rolling_quantile_by( self, by: Expr, method: QuantileMethod, quantile: f64, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling quantile based on another column.

pub fn rolling_var_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling variance based on another column.

pub fn rolling_std_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling std-dev based on another column.

pub fn rolling_median_by( self, by: Expr, options: RollingOptionsDynamicWindow, ) -> Expr

Apply a rolling median based on another column.

pub fn rolling_min(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling minimum.

See: [RollingAgg::rolling_min]

pub fn rolling_max(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling maximum.

See: [RollingAgg::rolling_max]

pub fn rolling_mean(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling mean.

See: [RollingAgg::rolling_mean]

pub fn rolling_sum(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling sum.

See: [RollingAgg::rolling_sum]

pub fn rolling_median(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling median.

See: [RollingAgg::rolling_median]

pub fn rolling_quantile( self, method: QuantileMethod, quantile: f64, options: RollingOptionsFixedWindow, ) -> Expr

Apply a rolling quantile.

See: [RollingAgg::rolling_quantile]

pub fn rolling_var(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling variance.

pub fn rolling_std(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling std-dev.

pub fn rolling_skew(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling skew.

pub fn rolling_kurtosis(self, options: RollingOptionsFixedWindow) -> Expr

Apply a rolling skew.

pub fn rolling_map( self, f: Arc<dyn Fn(&Series) -> Series + Sync + Send>, output_type: LazySerde<SpecialEq<Arc<dyn FunctionOutputField>>>, options: RollingOptionsFixedWindow, ) -> Expr

Apply a custom function over a rolling/ moving window of the array. This has quite some dynamic dispatch, so prefer rolling_min, max, mean, sum over this.

pub fn rolling_map_float<F>(self, window_size: usize, f: F) -> Expr
where F: 'static + FnMut(&mut ChunkedArray<Float64Type>) -> Option<f64> + Send + Sync + Copy,

Apply a custom function over a rolling/ moving window of the array. Prefer this over rolling_apply in case of floating point numbers as this is faster. This has quite some dynamic dispatch, so prefer rolling_min, max, mean, sum over this.

pub fn rank(self, options: RankOptions, seed: Option<u64>) -> Expr

Assign ranks to data, dealing with ties appropriately.

pub fn replace<E>(self, old: E, new: E) -> Expr
where E: Into<Expr>,

Replace the given values with other values.

pub fn replace_strict<E>( self, old: E, new: E, default: Option<E>, return_dtype: Option<impl Into<DataTypeExpr>>, ) -> Expr
where E: Into<Expr>,

Replace the given values with other values.

pub fn rle(self) -> Expr

Get the lengths of runs of identical values.

pub fn rle_id(self) -> Expr

Similar to rle, but maps values to run IDs.

pub fn diff(self, n: Expr, null_behavior: NullBehavior) -> Expr

Calculate the n-th discrete difference between values.

pub fn skew(self, bias: bool) -> Expr

Compute the sample skewness of a data set.

For normally distributed data, the skewness should be about zero. For uni-modal continuous distributions, a skewness value greater than zero means that there is more weight in the right tail of the distribution. The function skewtest can be used to determine if the skewness value is close enough to zero, statistically speaking.

see: scipy

pub fn kurtosis(self, fisher: bool, bias: bool) -> Expr

Compute the kurtosis (Fisher or Pearson).

Kurtosis is the fourth central moment divided by the square of the variance. If Fisher’s definition is used, then 3.0 is subtracted from the result to give 0.0 for a normal distribution. If bias is False then the kurtosis is calculated using k statistics to eliminate bias coming from biased moment estimators.

pub fn upper_bound(self) -> Expr

Get maximal value that could be hold by this dtype.

pub fn lower_bound(self) -> Expr

Get minimal value that could be hold by this dtype.

pub fn reshape(self, dimensions: &[i64]) -> Expr

pub fn any(self, ignore_nulls: bool) -> Expr

Returns whether any of the values in the column are true.

If ignore_nulls is False, Kleene logic is used to deal with nulls: if the column contains any null values and no true values, the output is null.

pub fn all(self, ignore_nulls: bool) -> Expr

Returns whether all values in the column are true.

If ignore_nulls is False, Kleene logic is used to deal with nulls: if the column contains any null values and no false values, the output is null.

pub fn shrink_dtype(self) -> Expr

Shrink numeric columns to the minimal required datatype needed to fit the extrema of this Series. This can be used to reduce memory pressure.

pub fn value_counts( self, sort: bool, parallel: bool, name: &str, normalize: bool, ) -> Expr

Count all unique values and create a struct mapping value to count. (Note that it is better to turn parallel off in the aggregation context). The name of the struct field with the counts is given by the parameter name.

pub fn unique_counts(self) -> Expr

Returns a count of the unique values in the order of appearance. This method differs from Expr::value_counts in that it does not return the values, only the counts and might be faster.

pub fn null_count(self) -> Expr

Get the null count of the column/group.

pub fn set_sorted_flag(self, sorted: IsSorted) -> Expr

Set this Series as sorted so that downstream code can use fast paths for sorted arrays.

Warning

This can lead to incorrect results if this Series is not sorted!! Use with care!

pub fn to_physical(self) -> Expr

pub fn gather_every(self, n: usize, offset: usize) -> Expr

pub fn extend_constant(self, value: Expr, n: Expr) -> Expr

pub fn str(self) -> StringNameSpace

Get the [string::StringNameSpace]

pub fn binary(self) -> BinaryNameSpace

Get the binary::BinaryNameSpace

pub fn dt(self) -> DateLikeNameSpace

Get the dt::DateLikeNameSpace

pub fn list(self) -> ListNameSpace

Get the list::ListNameSpace

pub fn name(self) -> ExprNameNameSpace

Get the name::ExprNameNameSpace

pub fn arr(self) -> ArrayNameSpace

Get the array::ArrayNameSpace.

pub fn cat(self) -> CategoricalNameSpace

Get the CategoricalNameSpace.

pub fn struct_(self) -> StructNameSpace

Get the struct_::StructNameSpace.

impl Expr

pub fn nodes<'a>(&'a self, container: &mut UnitVec<&'a Expr>)

pub fn nodes_owned(self, container: &mut UnitVec<Expr>)

pub fn map_expr<F>(self, f: F) -> Expr
where F: FnMut(Expr) -> Expr,

pub fn try_map_expr<F>(self, f: F) -> Result<Expr, PolarsError>
where F: FnMut(Expr) -> Result<Expr, PolarsError>,