polars::prelude::round

Trait RoundSeries

pub trait RoundSeries: SeriesSealed {
    // Provided methods
    fn round(
        &self,
        decimals: u32,
        mode: RoundMode,
    ) -> Result<Series, PolarsError> { ... }
    fn round_sig_figs(&self, digits: i32) -> Result<Series, PolarsError> { ... }
    fn floor(&self) -> Result<Series, PolarsError> { ... }
    fn ceil(&self) -> Result<Series, PolarsError> { ... }
}

Available on crate feature polars-ops only.

Provided Methods

fn round(&self, decimals: u32, mode: RoundMode) -> Result<Series, PolarsError>

Round underlying floating point array to given decimal.

fn round_sig_figs(&self, digits: i32) -> Result<Series, PolarsError>

fn floor(&self) -> Result<Series, PolarsError>

Floor underlying floating point array to the lowest integers smaller or equal to the float value.

fn ceil(&self) -> Result<Series, PolarsError>

Ceil underlying floating point array to the highest integers smaller or equal to the float value.