Struct polars::prelude::ChunkedArray

pub struct ChunkedArray<T>where
    T: PolarsDataType,{ /* private fields */ }

ChunkedArray

Every Series contains a ChunkedArray<T>. Unlike Series, ChunkedArray’s are typed. This allows us to apply closures to the data and collect the results to a ChunkedArray of the same type T. Below we use an apply to use the cosine function to the values of a ChunkedArray.

fn apply_cosine(ca: &Float32Chunked) -> Float32Chunked {
    ca.apply(|v| v.cos())
}

If we would like to cast the result we could use a Rust Iterator instead of an apply method. Note that Iterators are slightly slower as the null values aren’t ignored implicitly.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.into_iter()
        .map(|opt_v| {
        opt_v.map(|v| v.cos() as f64)
    }).collect()
}

Another option is to first cast and then use an apply.

fn apply_cosine_and_cast(ca: &Float32Chunked) -> Float64Chunked {
    ca.apply_cast_numeric(|v| v.cos() as f64)
}

Conversion between Series and ChunkedArray’s

Conversion from a Series to a ChunkedArray is effortless.

fn to_chunked_array(series: &Series) -> PolarsResult<&Int32Chunked>{
    series.i32()
}

fn to_series(ca: Int32Chunked) -> Series {
    ca.into_series()
}

Iterators

ChunkedArrays fully support Rust native Iterator and DoubleEndedIterator traits, thereby giving access to all the excellent methods available for Iterators.

fn iter_forward(ca: &Float32Chunked) {
    ca.into_iter()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

fn iter_backward(ca: &Float32Chunked) {
    ca.into_iter()
        .rev()
        .for_each(|opt_v| println!("{:?}", opt_v))
}

Memory layout

ChunkedArray’s use Apache Arrow as backend for the memory layout. Arrows memory is immutable which makes it possible to make multiple zero copy (sub)-views from a single array.

To be able to append data, Polars uses chunks to append new memory locations, hence the ChunkedArray<T> data structure. Appends are cheap, because it will not lead to a full reallocation of the whole array (as could be the case with a Rust Vec).

However, multiple chunks in a ChunkArray will slow down many operations that need random access because we have an extra indirection and indexes need to be mapped to the proper chunk. Arithmetic may also be slowed down by this. When multiplying two ChunkArray's with different chunk sizes they cannot utilize SIMD for instance.

If you want to have predictable performance (no unexpected re-allocation of memory), it is advised to call the ChunkedArray::rechunk after multiple append operations.

See also ChunkedArray::extend for appends within a chunk.

Implementations

impl<T> ChunkedArray<T>where T: PolarsNumericType, <T as PolarsNumericType>::Native: Signed,

pub fn abs(&self) -> ChunkedArray<T>

Convert all values to their absolute/positive value.

impl ChunkedArray<BooleanType>

Booleans are casted to 1 or 0.

pub fn sum(&self) -> Option<u32>

Returns None if the array is empty or only contains null values.

pub fn min(&self) -> Option<bool>

pub fn max(&self) -> Option<bool>

pub fn mean(&self) -> Option<f64>

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn append(&mut self, other: &ChunkedArray<T>)

Append in place. This is done by adding the chunks of other to this ChunkedArray.

See also extend for appends to the underlying memory

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn cast_and_apply_in_place<F, S>(&self, f: F) -> ChunkedArray<S>where F: Fn(<S as PolarsNumericType>::Native) -> <S as PolarsNumericType>::Native + Copy, S: PolarsNumericType,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

pub fn apply_in_place<F>(self, f: F) -> ChunkedArray<T>where F: Fn(<T as PolarsNumericType>::Native) -> <T as PolarsNumericType>::Native + Copy,

Cast a numeric array to another numeric data type and apply a function in place. This saves an allocation.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn apply_mut<F>(&mut self, f: F)where F: Fn(<T as PolarsNumericType>::Native) -> <T as PolarsNumericType>::Native + Copy,

impl ChunkedArray<Utf8Type>

pub fn apply_mut<'a, F>(&'a self, f: F) -> ChunkedArray<Utf8Type>where F: FnMut(&'a str) -> &'a str,

impl ChunkedArray<BinaryType>

pub fn apply_mut<'a, F>(&'a self, f: F) -> ChunkedArray<BinaryType>where F: FnMut(&'a [u8]) -> &'a [u8],

impl ChunkedArray<Float32Type>

This impl block contains no items.

Used to save compilation paths. Use carefully. Although this is safe, if misused it can lead to incorrect results.

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn len(&self) -> usize

Get the length of the ChunkedArray

pub fn is_empty(&self) -> bool

Check if ChunkedArray is empty.

pub fn rechunk(&self) -> ChunkedArray<T>

pub fn slice(&self, offset: i64, length: usize) -> ChunkedArray<T>

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

pub fn limit(&self, num_elements: usize) -> ChunkedArray<T>where ChunkedArray<T>: Sized,

Take a view of top n elements

pub fn head(&self, length: Option<usize>) -> ChunkedArray<T>where ChunkedArray<T>: Sized,

Get the head of the ChunkedArray

pub fn tail(&self, length: Option<usize>) -> ChunkedArray<T>where ChunkedArray<T>: Sized,

Get the tail of the ChunkedArray

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn extend(&mut self, other: &ChunkedArray<T>)

Extend the memory backed by this array with the values from other.

Different from ChunkedArray::append which adds chunks to this ChunkedArray extend appends the data from other to the underlying PrimitiveArray and thus may cause a reallocation.

However if this does not cause a reallocation, the resulting data structure will not have any extra chunks and thus will yield faster queries.

Prefer extend over append when you want to do a query after a single append. For instance during online operations where you add n rows and rerun a query.

Prefer append over extend when you want to append many times before doing a query. For instance when you read in multiple files and when to store them in a single DataFrame. In the latter case finish the sequence of append operations with a rechunk.

impl ChunkedArray<ListType>

pub fn full_null_with_dtype( name: &str, length: usize, inner_dtype: &DataType ) -> ChunkedArray<ListType>

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn is_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the null values.

pub fn is_not_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the valid values.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn tile(&self, n: usize) -> ChunkedArray<T>

impl ChunkedArray<BinaryType>

pub unsafe fn to_utf8(&self) -> ChunkedArray<Utf8Type>

Safety

Utf8 is not validated

impl ChunkedArray<Utf8Type>

pub fn as_binary(&self) -> ChunkedArray<BinaryType>

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn not_equal_and_validity( &self, rhs: &ChunkedArray<T> ) -> ChunkedArray<BooleanType>

impl ChunkedArray<BooleanType>

pub fn all(&self) -> bool

Check if all values are true

pub fn any(&self) -> bool

Check if any value is true

impl<T> ChunkedArray<T>where T: PolarsFloatType, <T as PolarsNumericType>::Native: Float,

pub fn is_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_not_nan(&self) -> ChunkedArray<BooleanType>

pub fn is_finite(&self) -> ChunkedArray<BooleanType>

pub fn is_infinite(&self) -> ChunkedArray<BooleanType>

pub fn none_to_nan(&self) -> ChunkedArray<T>

Convert missing values to NaN values.

impl ChunkedArray<ListType>

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<Series>>

pub fn par_iter_indexed(&mut self) -> impl IndexedParallelIterator

impl ChunkedArray<Utf8Type>

pub fn par_iter_indexed(&self) -> impl IndexedParallelIterator

pub fn par_iter(&self) -> impl ParallelIterator<Item = Option<&str>>

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub unsafe fn from_chunks( name: &str, chunks: Vec<Box<dyn Array + 'static, Global>, Global> ) -> ChunkedArray<T>

Create a new ChunkedArray from existing chunks.

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

pub unsafe fn with_chunks( &self, chunks: Vec<Box<dyn Array + 'static, Global>, Global> ) -> ChunkedArray<T>

Safety

The Arrow datatype of all chunks must match the PolarsDataType T.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn from_vec( name: &str, v: Vec<<T as PolarsNumericType>::Native, Global> ) -> ChunkedArray<T>

Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.

pub fn new_from_owned_with_null_bitmap( name: &str, values: Vec<<T as PolarsNumericType>::Native, Global>, buffer: Option<Bitmap> ) -> ChunkedArray<T>

Nullify values in slice with an existing null bitmap

pub unsafe fn mmap_slice( name: &str, values: &[<T as PolarsNumericType>::Native] ) -> ChunkedArray<T>

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl ChunkedArray<BooleanType>

pub unsafe fn mmap_slice( name: &str, values: &[u8], offset: usize, len: usize ) -> ChunkedArray<BooleanType>

Create a temporary ChunkedArray from a slice.

Safety

The lifetime will be bound to the lifetime of the slice. This will not be checked by the borrowchecker.

impl ChunkedArray<ListType>

pub fn amortized_iter( &self ) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array + 'static, Global>>>>

This is an iterator over a ListChunked that save allocations. A Series is: 1. Arc ChunkedArray is: 2. Vec< 3. ArrayRef>

The ArrayRef we indicated with 3. will be updated during iteration. The Series will be pinned in memory, saving an allocation for

Arc<..>

Vec<…>

Warning

Though memory safe in the sense that it will not read unowned memory, UB, or memory leaks this function still needs precautions. The returned should never be cloned or taken longer than a single iteration, as every call on next of the iterator will change the contents of that Series.

pub fn amortized_iter_with_name( &self, name: &str ) -> AmortizedListIter<'_, impl Iterator<Item = Option<Box<dyn Array + 'static, Global>>>>

pub fn apply_amortized<'a, F>(&'a self, f: F) -> ChunkedArray<ListType>where F: FnMut(UnstableSeries<'a>) -> Series,

Apply a closure F elementwise.

pub fn try_apply_amortized<'a, F>( &'a self, f: F ) -> Result<ChunkedArray<ListType>, PolarsError>where F: FnMut(UnstableSeries<'a>) -> Result<Series, PolarsError>,

impl ChunkedArray<ListType>

pub fn inner_dtype(&self) -> DataType

Get the inner data type of the list.

pub fn set_inner_dtype(&mut self, dtype: DataType)

pub fn set_fast_explode(&mut self)

pub fn _can_fast_explode(&self) -> bool

pub fn to_physical(&mut self, inner_dtype: DataType)

pub fn get_inner(&self) -> Series

Get the inner values as Series, ignoring the list offsets.

pub fn apply_to_inner( &self, func: &dyn Fn(Series) -> Result<Series, PolarsError> ) -> Result<ChunkedArray<ListType>, PolarsError>

Ignore the list indices and apply func to the inner type as Series.

impl ChunkedArray<Int32Type>

pub fn into_date(self) -> Logical<DateType, Int32Type>

impl ChunkedArray<Int64Type>

pub fn into_datetime( self, timeunit: TimeUnit, tz: Option<String> ) -> Logical<DatetimeType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_duration( self, timeunit: TimeUnit ) -> Logical<DurationType, Int64Type>

impl ChunkedArray<Int64Type>

pub fn into_time(self) -> Logical<TimeType, Int64Type>

impl<T> ChunkedArray<T>where T: PolarsNumericType, Standard: Distribution<<T as PolarsNumericType>::Native>,

pub fn init_rand( size: usize, null_density: f32, seed: Option<u64> ) -> ChunkedArray<T>

impl<T> ChunkedArray<T>where T: PolarsDataType, ChunkedArray<T>: ChunkTake,

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> Result<ChunkedArray<T>, PolarsError>

Sample n datapoints from this ChunkedArray.

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64> ) -> Result<ChunkedArray<T>, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

impl<T> ChunkedArray<T>where T: PolarsNumericType, <T as PolarsNumericType>::Native: Float,

pub fn rand_normal( name: &str, length: usize, mean: f64, std_dev: f64 ) -> Result<ChunkedArray<T>, PolarsError>

Create ChunkedArray with samples from a Normal distribution.

pub fn rand_standard_normal(name: &str, length: usize) -> ChunkedArray<T>

Create ChunkedArray with samples from a Standard Normal distribution.

pub fn rand_uniform( name: &str, length: usize, low: f64, high: f64 ) -> ChunkedArray<T>

Create ChunkedArray with samples from a Uniform distribution.

impl ChunkedArray<BooleanType>

pub fn rand_bernoulli( name: &str, length: usize, p: f64 ) -> Result<ChunkedArray<BooleanType>, PolarsError>

Create ChunkedArray with samples from a Bernoulli distribution.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn to_vec(&self) -> Vec<Option<<T as PolarsNumericType>::Native>, Global>

Convert to a Vec of Option<T::Native>.

pub fn to_vec_null_aware( &self ) -> Either<Vec<<T as PolarsNumericType>::Native, Global>, Vec<Option<<T as PolarsNumericType>::Native>, Global>>

Convert to a Vec but don’t return Option<T::Native> if there are no null values

impl<T> ChunkedArray<T>where T: PolarsDataType,

pub fn unset_fast_explode_list(&mut self)

pub fn is_sorted_flag(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

Set the ‘sorted’ bit meta info.

pub fn first_non_null(&self) -> Option<usize>

Get the index of the first non null value in this ChunkedArray.

pub fn last_non_null(&self) -> Option<usize>

Get the index of the last non null value in this ChunkedArray.

pub fn iter_validities( &self ) -> Map<Iter<'_, Box<dyn Array + 'static, Global>>, fn(_: &Box<dyn Array + 'static, Global>) -> Option<&Bitmap>>

Get the buffer of bits representing null values

pub fn has_validity(&self) -> bool

Return if any the chunks in this [ChunkedArray] have a validity bitmap. no bitmap means no null values.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn unpack_series_matching_type( &self, series: &Series ) -> Result<&ChunkedArray<T>, PolarsError>

Series to ChunkedArray

pub fn chunk_id( &self ) -> Map<Iter<'_, Box<dyn Array + 'static, Global>>, fn(_: &Box<dyn Array + 'static, Global>) -> usize>

Unique id representing the number of chunks

pub fn chunks(&self) -> &Vec<Box<dyn Array + 'static, Global>, Global>

A reference to the chunks

pub unsafe fn chunks_mut( &mut self ) -> &mut Vec<Box<dyn Array + 'static, Global>, Global>

A mutable reference to the chunks

Safety

The caller must ensure to not change the DataType or length of any of the chunks.

pub fn is_optimal_aligned(&self) -> bool

Returns true if contains a single chunk and has no null values

pub fn null_count(&self) -> usize

Count the null values.

pub fn dtype(&self) -> &DataType

Get data type of ChunkedArray.

pub fn name(&self) -> &str

Name of the ChunkedArray.

pub fn ref_field(&self) -> &Field

Get a reference to the field.

pub fn rename(&mut self, name: &str)

Rename this ChunkedArray.

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn cont_slice( &self ) -> Result<&[<T as PolarsNumericType>::Native], PolarsError>

Contiguous slice

pub fn data_views( &self ) -> impl Iterator<Item = &[<T as PolarsNumericType>::Native]> + DoubleEndedIterator

Get slices of the underlying arrow data. NOTE: null values should be taken into account by the user of these slices as they are handled separately

pub fn into_no_null_iter( &self ) -> impl Iterator<Item = <T as PolarsNumericType>::Native> + Send + Sync + ExactSizeIterator + DoubleEndedIterator + TrustedLen

impl<T> ChunkedArray<T>where T: PolarsIntegerType, <T as PolarsNumericType>::Native: AsU64 + FromPrimitive + Debug,

pub fn group_tuples_perfect( &self, max: usize, multithreaded: bool, group_capacity: usize ) -> GroupsProxy

impl<T> ChunkedArray<T>where T: PolarsNumericType,

pub fn new_vec( name: &str, v: Vec<<T as PolarsNumericType>::Native, Global> ) -> ChunkedArray<T>

Specialization that prevents an allocation prefer this over ChunkedArray::new when you have a Vec<T::Native> and no null values.

impl<T> ChunkedArray<T>where T: PolarsNumericType, ChunkedArray<T>: IntoSeries,

We cannot override the left hand side behaviour. So we create a trait LhsNumOps. This allows for 1.add(&Series)

pub fn lhs_sub<N>(&self, lhs: N) -> ChunkedArray<T>where N: Num + NumCast,

Apply lhs - self

pub fn lhs_div<N>(&self, lhs: N) -> ChunkedArray<T>where N: Num + NumCast,

Apply lhs / self

pub fn lhs_rem<N>(&self, lhs: N) -> ChunkedArray<T>where N: Num + NumCast,

Apply lhs % self