Struct polars::prelude::CsvReader

pub struct CsvReader<'a, R>
where
    R: MmapBytesReader,
{ /* private fields */ }

Available on crate feature polars-io only.

Create a new DataFrame by reading a csv file.

Example

use polars_core::prelude::*;
use polars_io::prelude::*;
use std::fs::File;

fn example() -> PolarsResult<DataFrame> {
    CsvReader::from_path("iris.csv")?
            .has_header(true)
            .finish()
}

Implementations

impl<'a, R> CsvReader<'a, R>
where R: 'a + MmapBytesReader,

pub fn with_skip_rows_after_header(self, offset: usize) -> CsvReader<'a, R>

Skip these rows after the header

pub fn with_row_index(self, row_index: Option<RowIndex>) -> CsvReader<'a, R>

Add a row index column.

pub fn with_chunk_size(self, chunk_size: usize) -> CsvReader<'a, R>

Sets the chunk size used by the parser. This influences performance

pub fn with_encoding(self, enc: CsvEncoding) -> CsvReader<'a, R>

Set CsvEncoding

pub fn with_n_rows(self, num_rows: Option<usize>) -> CsvReader<'a, R>

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

pub fn with_ignore_errors(self, ignore: bool) -> CsvReader<'a, R>

Continue with next batch when a ParserError is encountered.

pub fn with_schema(self, schema: Option<Arc<Schema>>) -> CsvReader<'a, R>

Set the CSV file’s schema. This only accepts datatypes that are implemented in the csv parser and expects a complete Schema.

It is recommended to use with_dtypes instead.

pub fn with_skip_rows(self, skip_rows: usize) -> CsvReader<'a, R>

Skip the first n rows during parsing. The header will be parsed at n lines.

pub fn with_rechunk(self, rechunk: bool) -> CsvReader<'a, R>

Rechunk the DataFrame to contiguous memory after the CSV is parsed.

pub fn has_header(self, has_header: bool) -> CsvReader<'a, R>

Set whether the CSV file has headers

pub fn with_separator(self, separator: u8) -> CsvReader<'a, R>

Set the CSV file’s column separator as a byte character

pub fn with_comment_prefix( self, comment_prefix: Option<&str> ) -> CsvReader<'a, R>

Set the comment prefix for this instance. Lines starting with this prefix will be ignored.

pub fn _with_comment_prefix( self, comment_prefix: Option<CommentPrefix> ) -> CsvReader<'a, R>

Sets the comment prefix from CsvParserOptions for internal initialization.

pub fn with_end_of_line_char(self, eol_char: u8) -> CsvReader<'a, R>

pub fn with_null_values( self, null_values: Option<NullValues> ) -> CsvReader<'a, R>

Set values that will be interpreted as missing/ null. Note that any value you set as null value will not be escaped, so if quotation marks are part of the null value you should include them.

pub fn with_missing_is_null(self, missing_is_null: bool) -> CsvReader<'a, R>

Treat missing fields as null.

pub fn with_dtypes(self, schema: Option<Arc<Schema>>) -> CsvReader<'a, R>

Overwrite the schema with the dtypes in this given Schema. The given schema may be a subset of the total schema.

pub fn with_dtypes_slice( self, dtypes: Option<&'a [DataType]> ) -> CsvReader<'a, R>

Overwrite the dtypes in the schema in the order of the slice that’s given. This is useful if you don’t know the column names beforehand

pub fn infer_schema(self, max_records: Option<usize>) -> CsvReader<'a, R>

Set the CSV reader to infer the schema of the file

Arguments

max_records - Maximum number of rows read for schema inference. Setting this to None will do a full table scan (slow).

pub fn with_projection(self, projection: Option<Vec<usize>>) -> CsvReader<'a, R>

Set the reader’s column projection. This counts from 0, meaning that vec![0, 4] would select the 1st and 5th column.

pub fn with_columns(self, columns: Option<Vec<String>>) -> CsvReader<'a, R>

Columns to select/ project

pub fn with_n_threads(self, n: Option<usize>) -> CsvReader<'a, R>

Set the number of threads used in CSV reading. The default uses the number of cores of your cpu.

Note that this only works if this is initialized with CsvReader::from_path. Note that the number of cores is the maximum allowed number of threads.

pub fn with_path<P>(self, path: Option<P>) -> CsvReader<'a, R>
where P: Into<PathBuf>,

The preferred way to initialize this builder. This allows the CSV file to be memory mapped and thereby greatly increases parsing performance.

pub fn sample_size(self, size: usize) -> CsvReader<'a, R>

Sets the size of the sample taken from the CSV file. The sample is used to get statistic about the file. These statistics are used to try to optimally allocate up front. Increasing this may improve performance.

pub fn raise_if_empty(self, toggle: bool) -> CsvReader<'a, R>

Raise an error if CSV is empty (otherwise return an empty frame)

pub fn low_memory(self, toggle: bool) -> CsvReader<'a, R>

Reduce memory consumption at the expense of performance

pub fn with_quote_char(self, quote_char: Option<u8>) -> CsvReader<'a, R>

Set the char used as quote char. The default is b'"'. If set to [None] quoting is disabled.

pub fn with_try_parse_dates(self, toggle: bool) -> CsvReader<'a, R>

Automatically try to parse dates/ datetimes and time. If parsing fails, columns remain of dtype [DataType::String].

pub fn with_predicate( self, predicate: Option<Arc<dyn PhysicalIoExpr>> ) -> CsvReader<'a, R>

pub fn truncate_ragged_lines(self, toggle: bool) -> CsvReader<'a, R>

Truncate lines that are longer than the schema.

impl<'a> CsvReader<'a, File>

pub fn from_path<P>(path: P) -> Result<CsvReader<'a, File>, PolarsError>
where P: Into<PathBuf>,

This is the recommended way to create a csv reader as this allows for fastest parsing.

impl<'a, R> CsvReader<'a, R>
where R: MmapBytesReader + 'a,

pub fn batched_borrowed_mmap( &'a mut self ) -> Result<BatchedCsvReaderMmap<'a>, PolarsError>

pub fn batched_borrowed_read( &'a mut self ) -> Result<BatchedCsvReaderRead<'a>, PolarsError>

impl<'a> CsvReader<'a, Box<dyn MmapBytesReader>>

pub fn batched_mmap( self, schema: Option<Arc<Schema>> ) -> Result<OwnedBatchedCsvReaderMmap, PolarsError>

pub fn batched_read( self, schema: Option<Arc<Schema>> ) -> Result<OwnedBatchedCsvReader, PolarsError>