Struct polars::prelude::JoinBuilder

pub struct JoinBuilder { /* private fields */ }

Available on crate feature lazy only.

Implementations

impl JoinBuilder

pub fn new(lf: LazyFrame) -> JoinBuilder

Create the JoinBuilder with the provided LazyFrame as the left table.

pub fn with(self, other: LazyFrame) -> JoinBuilder

The right table in the join.

pub fn how(self, how: JoinType) -> JoinBuilder

Select the join type.

pub fn validate(self, validation: JoinValidation) -> JoinBuilder

pub fn on<E>(self, on: E) -> JoinBuilder
where E: AsRef<[Expr]>,

The expressions you want to join both tables on.

The passed expressions must be valid in both LazyFrames in the join.

pub fn left_on<E>(self, on: E) -> JoinBuilder
where E: AsRef<[Expr]>,

The expressions you want to join the left table on.

The passed expressions must be valid in the left table.

pub fn right_on<E>(self, on: E) -> JoinBuilder
where E: AsRef<[Expr]>,

The expressions you want to join the right table on.

The passed expressions must be valid in the right table.

pub fn allow_parallel(self, allow: bool) -> JoinBuilder

Allow parallel table evaluation.

pub fn force_parallel(self, force: bool) -> JoinBuilder

Force parallel table evaluation.

pub fn join_nulls(self, join_nulls: bool) -> JoinBuilder

Join on null values. By default null values will never produce matches.

pub fn suffix<S>(self, suffix: S) -> JoinBuilder
where S: AsRef<str>,

Suffix to add duplicate column names in join. Defaults to "_right" if this method is never called.

pub fn finish(self) -> LazyFrame

Finish builder