polars::prelude

Struct LazyCsvReader

pub struct LazyCsvReader { /* private fields */ }

Available on crate feature lazy only.

Implementations

impl LazyCsvReader

pub fn map_parse_options<F>(self, map_func: F) -> LazyCsvReader
where F: Fn(CsvParseOptions) -> CsvParseOptions,

Re-export to shorten code.

pub fn new_paths(paths: Arc<[PlPath]>) -> LazyCsvReader

pub fn new_with_sources(sources: ScanSources) -> LazyCsvReader

pub fn new(path: PlPath) -> LazyCsvReader

pub fn with_skip_rows_after_header(self, offset: usize) -> LazyCsvReader

Skip this number of rows after the header location.

pub fn with_row_index(self, row_index: Option<RowIndex>) -> LazyCsvReader

Add a row index column.

pub fn with_n_rows(self, num_rows: Option<usize>) -> LazyCsvReader

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

pub fn with_infer_schema_length(self, num_rows: Option<usize>) -> LazyCsvReader

Set the number of rows to use when inferring the csv schema. The default is 100 rows. Setting to None will do a full table scan, which is very slow.

pub fn with_ignore_errors(self, ignore: bool) -> LazyCsvReader

Continue with next batch when a ParserError is encountered.

pub fn with_schema(self, schema: Option<Arc<Schema<DataType>>>) -> LazyCsvReader

Set the CSV file’s schema

pub fn with_skip_rows(self, skip_rows: usize) -> LazyCsvReader

Skip the first n rows during parsing. The header will be parsed at row n. Note that by row we mean valid CSV, encoding and comments are respected.

pub fn with_skip_lines(self, skip_lines: usize) -> LazyCsvReader

Skip the first n lines during parsing. The header will be parsed at line n. We don’t respect CSV escaping when skipping lines.

pub fn with_dtype_overwrite( self, schema: Option<Arc<Schema<DataType>>>, ) -> LazyCsvReader

Overwrite the schema with the dtypes in this given Schema. The given schema may be a subset of the total schema.

pub fn with_has_header(self, has_header: bool) -> LazyCsvReader

Set whether the CSV file has headers

pub fn with_chunk_size(self, chunk_size: usize) -> LazyCsvReader

Sets the chunk size used by the parser. This influences performance. This can be used as a way to reduce memory usage during the parsing at the cost of performance.

pub fn with_separator(self, separator: u8) -> LazyCsvReader

Set the CSV file’s column separator as a byte character

pub fn with_comment_prefix( self, comment_prefix: Option<PlSmallStr>, ) -> LazyCsvReader

Set the comment prefix for this instance. Lines starting with this prefix will be ignored.

pub fn with_quote_char(self, quote_char: Option<u8>) -> LazyCsvReader

Set the char used as quote char. The default is b'"'. If set to None quoting is disabled.

pub fn with_eol_char(self, eol_char: u8) -> LazyCsvReader

Set the char used as end of line. The default is b'\n'.

pub fn with_null_values(self, null_values: Option<NullValues>) -> LazyCsvReader

Set values that will be interpreted as missing/ null.

pub fn with_missing_is_null(self, missing_is_null: bool) -> LazyCsvReader

Treat missing fields as null.

pub fn with_cache(self, cache: bool) -> LazyCsvReader

Cache the DataFrame after reading.

pub fn with_low_memory(self, low_memory: bool) -> LazyCsvReader

Reduce memory usage at the expense of performance

pub fn with_encoding(self, encoding: CsvEncoding) -> LazyCsvReader

Set CsvEncoding

pub fn with_try_parse_dates(self, try_parse_dates: bool) -> LazyCsvReader

Automatically try to parse dates/datetimes and time. If parsing fails, columns remain of dtype DataType::String.

pub fn with_raise_if_empty(self, raise_if_empty: bool) -> LazyCsvReader

Raise an error if CSV is empty (otherwise return an empty frame)

pub fn with_truncate_ragged_lines( self, truncate_ragged_lines: bool, ) -> LazyCsvReader

Truncate lines that are longer than the schema.

pub fn with_decimal_comma(self, decimal_comma: bool) -> LazyCsvReader

pub fn with_glob(self, toggle: bool) -> LazyCsvReader

Expand path given via globbing rules.

pub fn with_cloud_options( self, cloud_options: Option<CloudOptions>, ) -> LazyCsvReader

pub fn with_schema_modify<F>(self, f: F) -> Result<LazyCsvReader, PolarsError>
where F: Fn(Schema<DataType>) -> Result<Schema<DataType>, PolarsError>,

Modify a schema before we run the lazy scanning.

Important! Run this function latest in the builder!

pub fn with_include_file_paths( self, include_file_paths: Option<PlSmallStr>, ) -> LazyCsvReader