Struct polars::prelude::Logical

pub struct Logical<Logical, Physical>(pub ChunkedArray<Physical>, _, pub Option<DataType>)
where
    Logical: PolarsDataType,
    Physical: PolarsDataType;

Maps a logical type to a a chunked array implementation of the physical type. This saves a lot of compiler bloat and allows us to reuse functionality.

Tuple Fields
0: ChunkedArray<Physical>2: Option<DataType>
Implementations

impl<K, T> Logical<K, T>where K: PolarsDataType, T: PolarsDataType, Logical<K, T>: LogicalType,

pub fn field(&self) -> Field

impl Logical<DateType, Int32Type>

pub fn as_date_iter( &self ) -> impl Iterator<Item = Option<NaiveDate>> + TrustedLen

pub fn from_naive_date<I>(name: &str, v: I) -> Logical<DateType, Int32Type>where I: IntoIterator<Item = NaiveDate>,

Construct a new DateChunked from an iterator over NaiveDate.

pub fn to_string(&self, format: &str) -> ChunkedArray<Utf8Type>

Convert from Date into Utf8 with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> ChunkedArray<Utf8Type>

Convert from Date into Utf8 with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_date_options<I>( name: &str, v: I ) -> Logical<DateType, Int32Type>where I: IntoIterator<Item = Option<NaiveDate>>,

Construct a new DateChunked from an iterator over optional NaiveDate.

impl Logical<DatetimeType, Int64Type>

pub fn as_datetime_iter( &self ) -> impl Iterator<Item = Option<NaiveDateTime>> + TrustedLen

pub fn time_unit(&self) -> TimeUnit

pub fn time_zone(&self) -> &Option<String>

pub fn to_string( &self, format: &str ) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Convert from Datetime into Utf8 with the given format. See chrono strftime/strptime.

pub fn strftime( &self, format: &str ) -> Result<ChunkedArray<Utf8Type>, PolarsError>

Convert from Datetime into Utf8 with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn from_naive_datetime<I>( name: &str, v: I, tu: TimeUnit ) -> Logical<DatetimeType, Int64Type>where I: IntoIterator<Item = NaiveDateTime>,

Construct a new DatetimeChunked from an iterator over NaiveDateTime.

pub fn from_naive_datetime_options<I>( name: &str, v: I, tu: TimeUnit ) -> Logical<DatetimeType, Int64Type>where I: IntoIterator<Item = Option<NaiveDateTime>>,

pub fn cast_time_unit(&self, tu: TimeUnit) -> Logical<DatetimeType, Int64Type>

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, tu: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.

impl Logical<DurationType, Int64Type>

pub fn time_unit(&self) -> TimeUnit

pub fn cast_time_unit(&self, tu: TimeUnit) -> Logical<DurationType, Int64Type>

Change the underlying TimeUnit. And update the data accordingly.

pub fn set_time_unit(&mut self, tu: TimeUnit)

Change the underlying TimeUnit. This does not modify the data.

pub fn from_duration<I>( name: &str, v: I, tu: TimeUnit ) -> Logical<DurationType, Int64Type>where I: IntoIterator<Item = Duration>,

Construct a new DurationChunked from an iterator over ChronoDuration.

pub fn from_duration_options<I>( name: &str, v: I, tu: TimeUnit ) -> Logical<DurationType, Int64Type>where I: IntoIterator<Item = Option<Duration>>,

Construct a new DurationChunked from an iterator over optional ChronoDuration.

impl Logical<TimeType, Int64Type>

pub fn to_string(&self, format: &str) -> ChunkedArray<Utf8Type>

Convert from Time into Utf8 with the given format. See chrono strftime/strptime.

pub fn strftime(&self, format: &str) -> ChunkedArray<Utf8Type>

Convert from Time into Utf8 with the given format. See chrono strftime/strptime.

Alias for to_string.

pub fn as_time_iter( &self ) -> impl Iterator<Item = Option<NaiveTime>> + TrustedLen

pub fn from_naive_time<I>(name: &str, v: I) -> Logical<TimeType, Int64Type>where I: IntoIterator<Item = NaiveTime>,

Construct a new TimeChunked from an iterator over NaiveTime.

pub fn from_naive_time_options<I>( name: &str, v: I ) -> Logical<TimeType, Int64Type>where I: IntoIterator<Item = Option<NaiveTime>>,

Construct a new TimeChunked from an iterator over optional NaiveTime.