Trait polars::prelude::AnonymousScan

pub trait AnonymousScan: Send + Sync {
    fn scan(
        &self,
        scan_opts: AnonymousScanOptions
    ) -> Result<DataFrame, PolarsError>;

    fn schema(
        &self,
        _infer_schema_length: Option<usize>
    ) -> Result<Schema, PolarsError> { ... }
    fn allows_predicate_pushdown(&self) -> bool { ... }
    fn allows_projection_pushdown(&self) -> bool { ... }
    fn allows_slice_pushdown(&self) -> bool { ... }
}

Required Methods

fn scan(&self, scan_opts: AnonymousScanOptions) -> Result<DataFrame, PolarsError>

Creates a dataframe from the supplied function & scan options.

Provided Methods

fn schema(
&self,
_infer_schema_length: Option<usize>
) -> Result<Schema, PolarsError>

function to supply the schema. Allows for an optional infer schema argument for data sources with dynamic schemas

fn allows_predicate_pushdown(&self) -> bool

specify if the scan provider should allow predicate pushdowns

Defaults to false

fn allows_projection_pushdown(&self) -> bool

specify if the scan provider should allow projection pushdowns

Defaults to false

fn allows_slice_pushdown(&self) -> bool

specify if the scan provider should allow slice pushdowns

Defaults to false