Trait polars::prelude::AsofJoin

pub trait AsofJoin: IntoDf {
    // Provided method
    fn join_asof(
        &self,
        other: &DataFrame,
        left_on: &str,
        right_on: &str,
        strategy: AsofStrategy,
        tolerance: Option<AnyValue<'static>>,
        suffix: Option<String>
    ) -> Result<DataFrame, PolarsError> { ... }
}

Available on crate feature polars-ops only.

Provided Methods

fn join_asof( &self, other: &DataFrame, left_on: &str, right_on: &str, strategy: AsofStrategy, tolerance: Option<AnyValue<'static>>, suffix: Option<String> ) -> Result<DataFrame, PolarsError>

This is similar to a left-join except that we match on nearest key rather than equal keys. The keys must be sorted to perform an asof join