polars::prelude

Trait ChunkAgg

pub trait ChunkAgg<T> {
    // Required method
    fn _sum_as_f64(&self) -> f64;

    // Provided methods
    fn sum(&self) -> Option<T> { ... }
    fn min(&self) -> Option<T> { ... }
    fn max(&self) -> Option<T> { ... }
    fn min_max(&self) -> Option<(T, T)> { ... }
    fn mean(&self) -> Option<f64> { ... }
}

Aggregation operations.

Required Methods

fn _sum_as_f64(&self) -> f64

Provided Methods

fn sum(&self) -> Option<T>

Aggregate the sum of the ChunkedArray. Returns None if not implemented for T. If the array is empty, 0 is returned

fn min(&self) -> Option<T>

fn max(&self) -> Option<T>

Returns the maximum value in the array, according to the natural order. Returns None if the array is empty or only contains null values.

fn min_max(&self) -> Option<(T, T)>

fn mean(&self) -> Option<f64>

Returns the mean value in the array. Returns None if the array is empty or only contains null values.