Trait polars::prelude::ChunkApply

pub trait ChunkApply<'a, A, B> {
    pub fn apply_cast_numeric<F, S>(&'a self, f: F) -> ChunkedArray<S>
    where
        F: Fn(A) -> <S as ArrowPrimitiveType>::Native + Copy,
        S: PolarsNumericType;

    pub fn branch_apply_cast_numeric_no_null<F, S>(
        &'a self, 
        f: F
    ) -> ChunkedArray<S>
    where
        F: Fn(Option<A>) -> <S as ArrowPrimitiveType>::Native + Copy,
        S: PolarsNumericType;

    pub fn apply<F>(&'a self, f: F) -> Self
    where
        F: Fn(A) -> B + Copy;

    pub fn apply_with_idx<F>(&'a self, f: F) -> Self
    where
        F: Fn((usize, A)) -> B + Copy;

    pub fn apply_with_idx_on_opt<F>(&'a self, f: F) -> Self
    where
        F: Fn((usize, Option<A>)) -> Option<B> + Copy;
}

Fastest way to do elementwise operations on a ChunkedArray when the operation is cheaper than branching due to null checking

Required methods

pub fn apply_cast_numeric<F, S>(&'a self, f: F) -> ChunkedArray<S> where
F: Fn(A) -> <S as ArrowPrimitiveType>::Native + Copy,
S: PolarsNumericType,

Apply a closure elementwise and cast to a Numeric ChunkedArray. This is fastest when the null check branching is more expensive than the closure application.

Null values remain null.

pub fn branch_apply_cast_numeric_no_null<F, S>(
&'a self,
f: F
) -> ChunkedArray<S> where
F: Fn(Option<A>) -> <S as ArrowPrimitiveType>::Native + Copy,
S: PolarsNumericType,

Apply a closure on optional values and cast to Numeric ChunkedArray without null values.

pub fn apply<F>(&'a self, f: F) -> Self where
F: Fn(A) -> B + Copy,

Apply a closure elementwise. This is fastest when the null check branching is more expensive than the closure application. Often it is.

Null values remain null.

Example

use polars_core::prelude::*;
fn double(ca: &UInt32Chunked) -> UInt32Chunked {
    ca.apply(|v| v * 2)
}

pub fn apply_with_idx<F>(&'a self, f: F) -> Self where
F: Fn((usize, A)) -> B + Copy,

Apply a closure elementwise. The closure gets the index of the element as first argument.

pub fn apply_with_idx_on_opt<F>(&'a self, f: F) -> Self where
F: Fn((usize, Option<A>)) -> Option<B> + Copy,

Apply a closure elementwise. The closure gets the index of the element as first argument.

Loading content...