Trait polars::prelude::ChunkOps

pub trait ChunkOps: ChunkLen {
    fn rechunk(&self) -> Self;

    fn slice(&self, offset: i64, length: usize) -> Self;

    fn limit(&self, num_elements: usize) -> Self { ... }

    fn head(&self, length: Option<usize>) -> Self { ... }

    fn tail(&self, length: Option<usize>) -> Self { ... }
}

Required methods

fn rechunk(&self) -> Self

Aggregate to contiguous memory.

fn slice(&self, offset: i64, length: usize) -> Self

Slice the array. The chunks are reallocated the underlying data slices are zero copy.

When offset is negative it will be counted from the end of the array. This method will never error, and will slice the best match when offset, or length is out of bounds

Provided methods

fn limit(&self, num_elements: usize) -> Self

Take a view of top n elements

fn head(&self, length: Option<usize>) -> Self

Get the head of the ChunkedArray

fn tail(&self, length: Option<usize>) -> Self

Get the tail of the ChunkedArray