Trait polars::prelude::Executor

pub trait Executor: Send + Sync {
    pub fn execute(
        &mut self, 
        cache: &ExecutionState
    ) -> Result<DataFrame, PolarsError>;
}

Executors will evaluate physical expressions and collect them in a DataFrame.

Executors have other executors as input. By having a tree of executors we can execute the physical plan until the last executor is evaluated.

Required methods

pub fn execute(
&mut self,
cache: &ExecutionState
) -> Result<DataFrame, PolarsError>

Loading content...