Trait polars::prelude::Executor

pub trait Executor: Send + Sync {
    fn execute(
        &mut self, 
        cache: &ExecutionState
    ) -> Result<DataFrame, PolarsError>;
}

Executors will evaluate physical expressions and collect them in a DataFrame.

Executors have other executors as input. By having a tree of executors we can execute the physical plan until the last executor is evaluated.

Required methods

fn execute(&mut self, cache: &ExecutionState) -> Result<DataFrame, PolarsError>