Trait polars::prelude::IntoGroupsProxy

pub trait IntoGroupsProxy {
    fn group_tuples(
        &self,
        _multithreaded: bool,
        _sorted: bool
    ) -> Result<GroupsProxy, PolarsError> { ... }
}

Used to create the tuples for a groupby operation.

Provided Methods

fn group_tuples(
&self,
_multithreaded: bool,
_sorted: bool
) -> Result<GroupsProxy, PolarsError>

Create the tuples need for a groupby operation. * The first value in the tuple is the first index of the group. * The second value in the tuple is are the indexes of the groups including the first value.