polars::prelude

Trait ListNameSpaceImpl

pub trait ListNameSpaceImpl: AsList {

    // Provided methods
    fn lst_join(
        &self,
        separator: &ChunkedArray<StringType>,
        ignore_nulls: bool,
    ) -> Result<ChunkedArray<StringType>, PolarsError> { ... }
    fn join_literal(
        &self,
        separator: &str,
        ignore_nulls: bool,
    ) -> Result<ChunkedArray<StringType>, PolarsError> { ... }
    fn join_many(
        &self,
        separator: &ChunkedArray<StringType>,
        ignore_nulls: bool,
    ) -> Result<ChunkedArray<StringType>, PolarsError> { ... }
    fn lst_max(&self) -> Result<Series, PolarsError> { ... }
    fn lst_min(&self) -> Result<Series, PolarsError> { ... }
    fn lst_sum(&self) -> Result<Series, PolarsError> { ... }
    fn lst_mean(&self) -> Series { ... }
    fn lst_median(&self) -> Series { ... }
    fn lst_std(&self, ddof: u8) -> Series { ... }
    fn lst_var(&self, ddof: u8) -> Result<Series, PolarsError> { ... }
    fn same_type(&self, out: ChunkedArray<ListType>) -> ChunkedArray<ListType> { ... }
    fn lst_sort(
        &self,
        options: SortOptions,
    ) -> Result<ChunkedArray<ListType>, PolarsError> { ... }
    fn lst_reverse(&self) -> ChunkedArray<ListType> { ... }
    fn lst_n_unique(&self) -> Result<ChunkedArray<UInt32Type>, PolarsError> { ... }
    fn lst_unique(&self) -> Result<ChunkedArray<ListType>, PolarsError> { ... }
    fn lst_unique_stable(&self) -> Result<ChunkedArray<ListType>, PolarsError> { ... }
    fn lst_arg_min(&self) -> ChunkedArray<UInt32Type> { ... }
    fn lst_arg_max(&self) -> ChunkedArray<UInt32Type> { ... }
    fn lst_diff(
        &self,
        n: i64,
        null_behavior: NullBehavior,
    ) -> Result<ChunkedArray<ListType>, PolarsError> { ... }
    fn lst_shift(
        &self,
        periods: &Column,
    ) -> Result<ChunkedArray<ListType>, PolarsError> { ... }
    fn lst_slice(&self, offset: i64, length: usize) -> ChunkedArray<ListType> { ... }
    fn lst_lengths(&self) -> ChunkedArray<UInt32Type> { ... }
    fn lst_get(
        &self,
        idx: i64,
        null_on_oob: bool,
    ) -> Result<Series, PolarsError> { ... }
    fn lst_concat(
        &self,
        other: &[Column],
    ) -> Result<ChunkedArray<ListType>, PolarsError> { ... }

}

Available on crate feature polars-ops only.

Provided Methods

fn lst_join( &self, separator: &ChunkedArray<StringType>, ignore_nulls: bool, ) -> Result<ChunkedArray<StringType>, PolarsError>

In case the inner dtype DataType::String, the individual items will be joined into a single string separated by separator.

fn join_literal( &self, separator: &str, ignore_nulls: bool, ) -> Result<ChunkedArray<StringType>, PolarsError>

fn join_many( &self, separator: &ChunkedArray<StringType>, ignore_nulls: bool, ) -> Result<ChunkedArray<StringType>, PolarsError>

fn lst_max(&self) -> Result<Series, PolarsError>

fn lst_min(&self) -> Result<Series, PolarsError>

fn lst_sum(&self) -> Result<Series, PolarsError>

fn lst_mean(&self) -> Series

fn lst_median(&self) -> Series

fn lst_std(&self, ddof: u8) -> Series

fn lst_var(&self, ddof: u8) -> Result<Series, PolarsError>

fn same_type(&self, out: ChunkedArray<ListType>) -> ChunkedArray<ListType>

fn lst_sort( &self, options: SortOptions, ) -> Result<ChunkedArray<ListType>, PolarsError>

fn lst_reverse(&self) -> ChunkedArray<ListType>

fn lst_n_unique(&self) -> Result<ChunkedArray<UInt32Type>, PolarsError>

fn lst_unique(&self) -> Result<ChunkedArray<ListType>, PolarsError>

fn lst_unique_stable(&self) -> Result<ChunkedArray<ListType>, PolarsError>

fn lst_arg_min(&self) -> ChunkedArray<UInt32Type>

fn lst_arg_max(&self) -> ChunkedArray<UInt32Type>

fn lst_diff( &self, n: i64, null_behavior: NullBehavior, ) -> Result<ChunkedArray<ListType>, PolarsError>

fn lst_shift( &self, periods: &Column, ) -> Result<ChunkedArray<ListType>, PolarsError>

fn lst_slice(&self, offset: i64, length: usize) -> ChunkedArray<ListType>

fn lst_lengths(&self) -> ChunkedArray<UInt32Type>

fn lst_get(&self, idx: i64, null_on_oob: bool) -> Result<Series, PolarsError>

Get the value by index in the sublists. So index 0 would return the first item of every sublist and index -1 would return the last item of every sublist if an index is out of bounds, it will return a None.

fn lst_concat( &self, other: &[Column], ) -> Result<ChunkedArray<ListType>, PolarsError>