Trait polars::prelude::PhysicalExpr

pub trait PhysicalExpr: Send + Sync {
    pub fn evaluate(
        &self, 
        df: &DataFrame, 
        _state: &ExecutionState
    ) -> Result<Series, PolarsError>;

    pub fn to_field(&self, input_schema: &Schema) -> Result<Field, PolarsError>;

    pub fn as_expression(&self) -> &Expr { ... }

    pub fn evaluate_on_groups(
        &self, 
        df: &DataFrame, 
        groups: &'a Vec<(u32, Vec<u32, Global>), Global>, 
        state: &ExecutionState
    ) -> Result<(Series, Cow<'a, Vec<(u32, Vec<u32, Global>), Global>>), PolarsError> { ... }

    pub fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation, PolarsError> { ... }
}

Take a DataFrame and evaluate the expressions. Implement this for Column, lt, eq, etc

Required methods

pub fn evaluate(
&self,
df: &DataFrame,
_state: &ExecutionState
) -> Result<Series, PolarsError>

Take a DataFrame and evaluate the expression.

pub fn to_field(&self, input_schema: &Schema) -> Result<Field, PolarsError>

Get the output field of this expr

Loading content...
Provided methods

pub fn as_expression(&self) -> &Expr

pub fn evaluate_on_groups(
&self,
df: &DataFrame,
groups: &'a Vec<(u32, Vec<u32, Global>), Global>,
state: &ExecutionState
) -> Result<(Series, Cow<'a, Vec<(u32, Vec<u32, Global>), Global>>), PolarsError>

Some expression that are not aggregations can be done per group Think of sort, slice, filter, etc. defaults to ignoring the group

This method is called by an aggregation function.

In case of a simple expr, like ‘column’, the groups are ignored and the column is returned. In case of an expr where group behavior makes sense, this method is called. For a filter operation for instance, a Series is created per groups and filtered.

This has some gotcha’s. An implementation may also change the group tuples instead of the Series.

pub fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation, PolarsError>

Convert to a aggregation expression. This can only be done for the final expressions that produce an aggregated result.

The expression sum, min, max etc can be called as evaluate in the standard context, or during a groupby execution, this method is called to convert them to an AggPhysicalExpr

Loading content...