Trait polars::prelude::PhysicalExpr

pub trait PhysicalExpr: Send + Sync {
    pub fn evaluate(&self, df: &DataFrame) -> Result<Series, PolarsError>;

    pub fn to_field(&self, input_schema: &Schema) -> Result<Field, PolarsError>;

    pub fn as_expression(&self) -> &Expr { ... }

    pub fn evaluate_on_groups(
        &self, 
        df: &DataFrame, 
        groups: &'a Vec<(u32, Vec<u32, Global>), Global>
    ) -> Result<(Series, Cow<'a, Vec<(u32, Vec<u32, Global>), Global>>), PolarsError> { ... }

    pub fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation, PolarsError> { ... }
}

Take a DataFrame and evaluate the expressions. Implement this for Column, lt, eq, etc

Required methods

pub fn evaluate(&self, df: &DataFrame) -> Result<Series, PolarsError>

Take a DataFrame and evaluate the expression.

pub fn to_field(&self, input_schema: &Schema) -> Result<Field, PolarsError>

Get the output field of this expr

Loading content...
Provided methods

pub fn as_expression(&self) -> &Expr

pub fn evaluate_on_groups(
&self,
df: &DataFrame,
groups: &'a Vec<(u32, Vec<u32, Global>), Global>
) -> Result<(Series, Cow<'a, Vec<(u32, Vec<u32, Global>), Global>>), PolarsError>

Some expression that are not aggregations can be done per group Think of sort, slice, etc.

defaults to ignoring the group

pub fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation, PolarsError>

Convert to a aggregation expression. This can only be done for the final expressions that produce an aggregated result.

The expression sum, min, max etc can be called as evaluate in the standard context, or during a groupby execution, this method is called to convert them to an AggPhysicalExpr

Loading content...