Enum polars::series::Series

pub enum Series {
    UInt32(ChunkedArray<UInt32Type>),
    Int32(ChunkedArray<Int32Type>),
    Int64(ChunkedArray<Int64Type>),
    Float32(ChunkedArray<Float32Type>),
    Float64(ChunkedArray<Float64Type>),
    Utf8(ChunkedArray<Utf8Type>),
    Bool(ChunkedArray<BooleanType>),
    Date32(ChunkedArray<Date32Type>),
    Date64(ChunkedArray<Date64Type>),
    Time64Ns(ChunkedArray<Time64NanosecondType>),
    DurationNs(ChunkedArray<DurationNanosecondType>),
}

Variants

UInt32(ChunkedArray<UInt32Type>)

Int32(ChunkedArray<Int32Type>)

Int64(ChunkedArray<Int64Type>)

Float32(ChunkedArray<Float32Type>)

Float64(ChunkedArray<Float64Type>)

Utf8(ChunkedArray<Utf8Type>)

Bool(ChunkedArray<BooleanType>)

Date32(ChunkedArray<Date32Type>)

Date64(ChunkedArray<Date64Type>)

Time64Ns(ChunkedArray<Time64NanosecondType>)

DurationNs(ChunkedArray<DurationNanosecondType>)

Implementations

impl Series

pub fn sum<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive,

Returns None if the array is empty or only contains null values.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.sum(), Some(6));

pub fn min<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive,

Returns the minimum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.min(), Some(1));

pub fn max<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive,

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.max(), Some(3));

pub fn mean<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive + Div<Output = T>,

impl Series

pub fn from_chunked_array<T: PolarsDataType>(ca: ChunkedArray<T>) -> Self

pub fn chunk_lengths(&self) -> &Vec<usize>

Get the lengths of the underlying chunks

pub fn name(&self) -> &str

Name of series.

pub fn rename(&mut self, name: &str)

Rename series.

pub fn field(&self) -> &Field

Get field (used in schema)

pub fn dtype(&self) -> &ArrowDataType

Get datatype of series.

pub fn chunks(&self) -> &Vec<ArrayRef>

Underlying chunks.

pub fn n_chunks(&self) -> usize

No. of chunks

pub fn i32(&self) -> Result<&Int32Chunked>

Unpack to ChunkedArray

let s: Series = [1, 2, 3].iter().collect();
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

pub fn i64(&self) -> Result<&Int64Chunked>

Unpack to ChunkedArray

pub fn f32(&self) -> Result<&Float32Chunked>

Unpack to ChunkedArray

pub fn f64(&self) -> Result<&Float64Chunked>

Unpack to ChunkedArray

pub fn u32(&self) -> Result<&UInt32Chunked>

Unpack to ChunkedArray

pub fn bool(&self) -> Result<&BooleanChunked>

Unpack to ChunkedArray

pub fn utf8(&self) -> Result<&Utf8Chunked>

Unpack to ChunkedArray

pub fn date32(&self) -> Result<&Date32Chunked>

Unpack to ChunkedArray

pub fn date64(&self) -> Result<&Date64Chunked>

Unpack to ChunkedArray

pub fn time64ns(&self) -> Result<&Time64NsChunked>

Unpack to ChunkedArray

pub fn duration_ns(&self) -> Result<&DurationNsChunked>

Unpack to ChunkedArray

pub fn append_array(&mut self, other: ArrayRef) -> Result<()>

pub fn limit(&self, num_elements: usize) -> Result<Self>

Take num_elements from the top as a zero copy view.

pub fn slice(&self, offset: usize, length: usize) -> Result<Self>

Get a zero copy view of the data.

pub fn append(&mut self, other: &Self) -> Result<()>

Append a Series of the same type in place.

pub fn filter<T: AsRef<BooleanChunked>>(&self, filter: T) -> Result<Self>

Filter by boolean mask. This operation clones data.

pub fn take_iter(
&self,
iter: impl Iterator<Item = usize>,
capacity: Option<usize>
) -> Result<Self>

Take by index from an iterator. This operation clones the data.

pub fn take_opt_iter(
&self,
iter: impl Iterator<Item = Option<usize>>,
capacity: Option<usize>
) -> Result<Self>

Take by index from an iterator. This operation clones the data.

pub fn take<T: TakeIndex>(&self, indices: &T) -> Result<Self>

Take by index. This operation is clone.

pub fn len(&self) -> usize

Get length of series.

pub fn rechunk(&self, chunk_lengths: Option<&[usize]>) -> Result<Self>

Aggregate all chunks to a contiguous array of memory.

pub fn head(&self, length: Option<usize>) -> Self

Get the head of the Series.

pub fn tail(&self, length: Option<usize>) -> Self

Get the tail of the Series.

pub fn cast<N>(&self) -> Result<Self> where
N: PolarsDataType,

Cast to an some primitive type.

pub fn get(&self, index: usize) -> AnyType

Get a single value by index. Don't use this operation for loops as a runtime cast is needed for every iteration.

pub fn sort_in_place(&mut self, reverse: bool)

Sort in place.

pub fn sort(&self, reverse: bool) -> Self

pub fn argsort(&self, reverse: bool) -> Vec<usize>

Retrieve the indexes needed for a sort.

pub fn null_count(&self) -> usize

Count the null values.

pub fn unique(&self) -> Self

Get unique values in the Series.

pub fn arg_unique(&self) -> Vec<usize>

Get first indexes of unique values.

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn null_bits(&self) -> Vec<(usize, Option<Buffer>)>

Get the bits that represent the null values of the underlying ChunkedArray

pub fn reverse(&self) -> Self

return a Series in reversed order

pub fn as_single_ptr(&mut self) -> usize

Rechunk and return a pointer to the start of the Series. Only implemented for numeric types

impl Series

pub fn series_equal(&self, other: &Series) -> bool