Enum polars::series::Series

pub enum Series {
    UInt8(ChunkedArray<UInt8Type>),
    UInt16(ChunkedArray<UInt16Type>),
    UInt32(ChunkedArray<UInt32Type>),
    UInt64(ChunkedArray<UInt64Type>),
    Int8(ChunkedArray<Int8Type>),
    Int16(ChunkedArray<Int16Type>),
    Int32(ChunkedArray<Int32Type>),
    Int64(ChunkedArray<Int64Type>),
    Float32(ChunkedArray<Float32Type>),
    Float64(ChunkedArray<Float64Type>),
    Utf8(ChunkedArray<Utf8Type>),
    Bool(ChunkedArray<BooleanType>),
    Date32(ChunkedArray<Date32Type>),
    Date64(ChunkedArray<Date64Type>),
    Time32Millisecond(Time32MillisecondChunked),
    Time32Second(Time32SecondChunked),
    Time64Nanosecond(ChunkedArray<Time64NanosecondType>),
    Time64Microsecond(ChunkedArray<Time64MicrosecondType>),
    DurationNanosecond(ChunkedArray<DurationNanosecondType>),
    DurationMicrosecond(DurationMicrosecondChunked),
    DurationMillisecond(DurationMillisecondChunked),
    DurationSecond(DurationSecondChunked),
    IntervalDayTime(IntervalDayTimeChunked),
    IntervalYearMonth(IntervalYearMonthChunked),
    TimestampNanosecond(TimestampNanosecondChunked),
    TimestampMicrosecond(TimestampMicrosecondChunked),
    TimestampMillisecond(TimestampMillisecondChunked),
    TimestampSecond(TimestampSecondChunked),
    List(ListChunked),
    Object(Box<dyn SeriesOps>),
}

Series

The columnar data type for a DataFrame. The Series enum consists of typed ChunkedArray's. To quickly cast a Series to a ChunkedArray you can call the method with the name of the type:

let s: Series = [1, 2, 3].iter().collect();
// Quickly obtain the ChunkedArray wrapped by the Series.
let chunked_array = s.i32().unwrap();

Arithmetic

You can do standard arithmetic on series.

let s: Series = [1, 2, 3].iter().collect();
let out_add = &s + &s;
let out_sub = &s - &s;
let out_div = &s / &s;
let out_mul = &s * &s;

Or with series and numbers.

let s: Series = (1..3).collect();
let out_add_one = &s + 1;
let out_multiply = &s * 10;

// Could not overload left hand side operator.
let out_divide = 1.div(&s);
let out_add = 1.add(&s);
let out_subtract = 1.sub(&s);
let out_multiply = 1.mul(&s);

Comparison

You can obtain boolean mask by comparing series.

use itertools::Itertools;
let s = Series::new("dollars", &[1, 2, 3]);
let mask = s.eq(1);
let valid = [true, false, false].iter();
assert!(mask
    .into_iter()
    .map(|opt_bool| opt_bool.unwrap()) // option, because series can be null
    .zip(valid)
    .all(|(a, b)| a == *b))

See all the comparison operators in the CmpOps trait

Iterators

The Series variants contain differently typed ChunkedArray's. These structs can be turned into iterators, making it possible to use any function/ closure you want on a Series.

These iterators return an Option<T> because the values of a series may be null.

use polars::prelude::*;
let pi = 3.14;
let s = Series::new("angle", [2f32 * pi, pi, 1.5 * pi].as_ref());
let s_cos: Series = s.f32()
                    .expect("series was not an f32 dtype")
                    .into_iter()
                    .map(|opt_angle| opt_angle.map(|angle| angle.cos()))
                    .collect();

Creation

Series can be create from different data structures. Below we'll show a few ways we can create a Series object.

// Series van be created from Vec's, slices and arrays
Series::new("boolean series", &vec![true, false, true]);
Series::new("int series", &[1, 2, 3]);
// And can be nullable
Series::new("got nulls", &[Some(1), None, Some(2)]);

// Series can also be collected from iterators
let from_iter: Series = (0..10)
    .into_iter()
    .collect();

Variants

UInt8(ChunkedArray<UInt8Type>)

UInt16(ChunkedArray<UInt16Type>)

UInt32(ChunkedArray<UInt32Type>)

UInt64(ChunkedArray<UInt64Type>)

Int8(ChunkedArray<Int8Type>)

Int16(ChunkedArray<Int16Type>)

Int32(ChunkedArray<Int32Type>)

Int64(ChunkedArray<Int64Type>)

Float32(ChunkedArray<Float32Type>)

Float64(ChunkedArray<Float64Type>)

Utf8(ChunkedArray<Utf8Type>)

Bool(ChunkedArray<BooleanType>)

Date32(ChunkedArray<Date32Type>)

Date64(ChunkedArray<Date64Type>)

Time32Millisecond(Time32MillisecondChunked)

Time32Second(Time32SecondChunked)

Time64Nanosecond(ChunkedArray<Time64NanosecondType>)

Time64Microsecond(ChunkedArray<Time64MicrosecondType>)

DurationNanosecond(ChunkedArray<DurationNanosecondType>)

DurationMicrosecond(DurationMicrosecondChunked)

DurationMillisecond(DurationMillisecondChunked)

DurationSecond(DurationSecondChunked)

IntervalDayTime(IntervalDayTimeChunked)

IntervalYearMonth(IntervalYearMonthChunked)

TimestampNanosecond(TimestampNanosecondChunked)

TimestampMicrosecond(TimestampMicrosecondChunked)

TimestampMillisecond(TimestampMillisecondChunked)

TimestampSecond(TimestampSecondChunked)

List(ListChunked)

Object(Box<dyn SeriesOps>)

Implementations

impl Series

pub fn sum<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive,

Returns None if the array is empty or only contains null values.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.sum(), Some(6));

pub fn min<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive,

Returns the minimum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.min(), Some(1));

pub fn max<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive,

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.max(), Some(3));

pub fn mean<T>(&self) -> Option<T> where
T: NumCast + Zero + ToPrimitive + Div<Output = T>,

impl Series

pub fn array_data(&self) -> Vec<ArrayDataRef>

Get Arrow ArrayData

pub fn from_chunked_array<T: PolarsDataType>(ca: ChunkedArray<T>) -> Self

pub fn chunk_lengths(&self) -> &Vec<usize>

Get the lengths of the underlying chunks

pub fn name(&self) -> &str

Name of series.

pub fn rename(&mut self, name: &str) -> &mut Self

Rename series.

pub fn field(&self) -> &Field

Get field (used in schema)

pub fn dtype(&self) -> &ArrowDataType

Get datatype of series.

pub fn chunks(&self) -> &Vec<ArrayRef>

Underlying chunks.

pub fn n_chunks(&self) -> usize

No. of chunks

pub fn i8(&self) -> Result<&Int8Chunked>

pub fn i16(&self) -> Result<&Int16Chunked>

pub fn i32(&self) -> Result<&Int32Chunked>

Unpack to ChunkedArray

let s: Series = [1, 2, 3].iter().collect();
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

pub fn i64(&self) -> Result<&Int64Chunked>

Unpack to ChunkedArray

pub fn f32(&self) -> Result<&Float32Chunked>

Unpack to ChunkedArray

pub fn f64(&self) -> Result<&Float64Chunked>

Unpack to ChunkedArray

pub fn u8(&self) -> Result<&UInt8Chunked>

Unpack to ChunkedArray

pub fn u16(&self) -> Result<&UInt16Chunked>

Unpack to ChunkedArray

pub fn u32(&self) -> Result<&UInt32Chunked>

Unpack to ChunkedArray

pub fn u64(&self) -> Result<&UInt64Chunked>

Unpack to ChunkedArray

pub fn bool(&self) -> Result<&BooleanChunked>

Unpack to ChunkedArray

pub fn utf8(&self) -> Result<&Utf8Chunked>

Unpack to ChunkedArray

pub fn date32(&self) -> Result<&Date32Chunked>

Unpack to ChunkedArray

pub fn date64(&self) -> Result<&Date64Chunked>

Unpack to ChunkedArray

pub fn time32_millisecond(&self) -> Result<&Time32MillisecondChunked>

Unpack to ChunkedArray

pub fn time32_second(&self) -> Result<&Time32SecondChunked>

Unpack to ChunkedArray

pub fn time64_nanosecond(&self) -> Result<&Time64NanosecondChunked>

Unpack to ChunkedArray

pub fn time64_microsecond(&self) -> Result<&Time64MicrosecondChunked>

Unpack to ChunkedArray

pub fn duration_nanosecond(&self) -> Result<&DurationNanosecondChunked>

Unpack to ChunkedArray

pub fn duration_microsecond(&self) -> Result<&DurationMicrosecondChunked>

Unpack to ChunkedArray

pub fn duration_millisecond(&self) -> Result<&DurationMillisecondChunked>

Unpack to ChunkedArray

pub fn duration_second(&self) -> Result<&DurationSecondChunked>

Unpack to ChunkedArray

pub fn timestamp_nanosecond(&self) -> Result<&TimestampNanosecondChunked>

Unpack to ChunkedArray

pub fn timestamp_microsecond(&self) -> Result<&TimestampMicrosecondChunked>

Unpack to ChunkedArray

pub fn timestamp_millisecond(&self) -> Result<&TimestampMillisecondChunked>

Unpack to ChunkedArray

pub fn timestamp_second(&self) -> Result<&TimestampSecondChunked>

Unpack to ChunkedArray

pub fn interval_daytime(&self) -> Result<&IntervalDayTimeChunked>

Unpack to ChunkedArray

pub fn interval_year_month(&self) -> Result<&IntervalYearMonthChunked>

Unpack to ChunkedArray

pub fn list(&self) -> Result<&ListChunked>

Unpack to ChunkedArray

pub fn append_array(&mut self, other: ArrayRef) -> Result<&mut Self>

pub fn limit(&self, num_elements: usize) -> Result<Self>

Take num_elements from the top as a zero copy view.

pub fn slice(&self, offset: usize, length: usize) -> Result<Self>

Get a zero copy view of the data.

pub fn append(&mut self, other: &Self) -> Result<&mut Self>

Append a Series of the same type in place.

pub fn filter<T: AsRef<BooleanChunked>>(&self, filter: T) -> Result<Self>

Filter by boolean mask. This operation clones data.

pub fn take_iter(
&self,
iter: impl Iterator<Item = usize>,
capacity: Option<usize>
) -> Self

Take by index from an iterator. This operation clones the data.

Safety

Out of bounds access doesn't Error but will return a Null value

pub unsafe fn take_iter_unchecked(
&self,
iter: impl Iterator<Item = usize>,
capacity: Option<usize>
) -> Self

Take by index from an iterator. This operation clones the data.

Safety

This doesn't check any bounds or null validity.

pub unsafe fn take_opt_iter_unchecked(
&self,
iter: impl Iterator<Item = Option<usize>>,
capacity: Option<usize>
) -> Self

Take by index from an iterator. This operation clones the data.

Safety

This doesn't check any bounds or null validity.

pub fn take_opt_iter(
&self,
iter: impl Iterator<Item = Option<usize>>,
capacity: Option<usize>
) -> Self

Take by index from an iterator. This operation clones the data.

Safety

Out of bounds access doesn't Error but will return a Null value

pub fn take<T: AsTakeIndex>(&self, indices: &T) -> Self

Take by index. This operation is clone.

Safety

Out of bounds access doesn't Error but will return a Null value

pub fn len(&self) -> usize

Get length of series.

pub fn is_empty(&self) -> bool

Check if Series is empty.

pub fn rechunk(&self, chunk_lengths: Option<&[usize]>) -> Result<Self>

Aggregate all chunks to a contiguous array of memory.

pub fn head(&self, length: Option<usize>) -> Self

Get the head of the Series.

pub fn tail(&self, length: Option<usize>) -> Self

Get the tail of the Series.

pub fn drop_nulls(&self) -> Self

Drop all null values and return a new Series.

pub fn expand_at_index(&self, index: usize, length: usize) -> Self

Create a new Series filled with values at that index.

Example

use polars::prelude::*;
let s = Series::new("a", [0i32, 1, 8]);
let expanded = s.expand_at_index(2, 4);
assert_eq!(Vec::from(expanded.i32().unwrap()), &[Some(8), Some(8), Some(8), Some(8)])

pub fn cast<N>(&self) -> Result<Self> where
N: PolarsDataType,

Cast to some primitive type.

pub fn cast_with_arrow_datatype(
&self,
data_type: &ArrowDataType
) -> Result<Self>

pub fn to_dummies(&self) -> Result<DataFrame>

Create dummy variables. See DataFrame

pub fn value_counts(&self) -> Result<DataFrame>

pub fn unpack<N>(&self) -> Result<&ChunkedArray<N>> where
N: PolarsDataType,

Get the ChunkedArray for some PolarsDataType

pub fn get(&self, index: usize) -> AnyType<'_>

Get a single value by index. Don't use this operation for loops as a runtime cast is needed for every iteration.

pub fn sort_in_place(&mut self, reverse: bool) -> &mut Self

Sort in place.

pub fn sort(&self, reverse: bool) -> Self

pub fn argsort(&self, reverse: bool) -> Vec<usize>

Retrieve the indexes needed for a sort.

pub fn null_count(&self) -> usize

Count the null values.

pub fn unique(&self) -> Result<Self>

Get unique values in the Series.

pub fn arg_unique(&self) -> Result<Vec<usize>>

Get first indexes of unique values.

pub fn is_null(&self) -> BooleanChunked

Get a mask of the null values.

pub fn is_not_null(&self) -> BooleanChunked

Get a mask of the non-null values.

pub fn is_unique(&self) -> Result<BooleanChunked>

Get a mask of all the unique values.

pub fn is_duplicated(&self) -> Result<BooleanChunked>

Get a mask of all the duplicated values.

pub fn null_bits(&self) -> Vec<(usize, Option<Buffer>)>

Get the bits that represent the null values of the underlying ChunkedArray

pub fn reverse(&self) -> Self

return a Series in reversed order

pub fn as_single_ptr(&mut self) -> usize

Rechunk and return a pointer to the start of the Series. Only implemented for numeric types

pub fn shift(&self, periods: i32) -> Result<Self>

Shift the values by a given period and fill the parts that will be empty due to this operation with Nones.

NOTE: If you want to fill the Nones with a value use the shift operation on ChunkedArray<T>.

Example

fn example() -> Result<()> {
    let s = Series::new("series", &[1, 2, 3]);

    let shifted = s.shift(1)?;
    assert_eq!(Vec::from(shifted.i32()?), &[None, Some(1), Some(2)]);

    let shifted = s.shift(-1)?;
    assert_eq!(Vec::from(shifted.i32()?), &[Some(2), Some(3), None]);

    let shifted = s.shift(2)?;
    assert_eq!(Vec::from(shifted.i32()?), &[None, None, Some(1)]);

    Ok(())
}
example();

pub fn fill_none(&self, strategy: FillNoneStrategy) -> Result<Self>

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)

NOTE: If you want to fill the Nones with a value use the fill_none operation on ChunkedArray<T>.

Example

fn example() -> Result<()> {
    let s = Series::new("some_missing", &[Some(1), None, Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Forward)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Backward)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Min)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Max)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Mean)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    Ok(())
}
example();

pub fn zip_with(&self, mask: &BooleanChunked, other: &Series) -> Result<Self>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

pub fn sum_as_series(&self) -> Series

Get the sum of the Series as a new Series of length 1.

pub fn max_as_series(&self) -> Series

Get the max of the Series as a new Series of length 1.

pub fn min_as_series(&self) -> Series

Get the min of the Series as a new Series of length 1.

pub fn mean_as_series(&self) -> Series

Get the mean of the Series as a new Series of length 1.

pub fn median_as_series(&self) -> Series

Get the median of the Series as a new Series of length 1.

pub fn var_as_series(&self) -> Series

Get the variance of the Series as a new Series of length 1.

pub fn std_as_series(&self) -> Series

Get the standard deviation of the Series as a new Series of length 1.

pub fn quantile_as_series(&self, quantile: f64) -> Result<Series>

Get the quantile of the ChunkedArray as a new Series of length 1.

pub fn rolling_mean(
&self,
window_size: usize,
weight: Option<&[f64]>,
ignore_null: bool
) -> Result<Self>

Apply a rolling mean to a Series. See: ChunkedArray::rolling_mean.

pub fn rolling_sum(
&self,
window_size: usize,
weight: Option<&[f64]>,
ignore_null: bool
) -> Result<Self>

Apply a rolling sum to a Series. See: ChunkedArray::rolling_mean.

pub fn rolling_min(
&self,
window_size: usize,
weight: Option<&[f64]>,
ignore_null: bool
) -> Result<Self>

Apply a rolling min to a Series. See: ChunkedArray::rolling_mean.

pub fn rolling_max(
&self,
window_size: usize,
weight: Option<&[f64]>,
ignore_null: bool
) -> Result<Self>

Apply a rolling max to a Series. See: ChunkedArray::rolling_mean.

pub fn hour(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract hour from underlying NaiveDateTime representation. Returns the hour number from 0 to 23.

pub fn minute(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract minute from underlying NaiveDateTime representation. Returns the minute number from 0 to 59.

pub fn second(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract second from underlying NaiveDateTime representation. Returns the second number from 0 to 59.

pub fn nanosecond(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract second from underlying NaiveDateTime representation. Returns the number of nanoseconds since the whole non-leap second. The range from 1,000,000,000 to 1,999,999,999 represents the leap second.

pub fn day(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract day from underlying NaiveDateTime representation. Returns the day of month starting from 1.

The return value ranges from 1 to 31. (The last day of month differs by months.)

pub fn ordinal_day(&self) -> Result<Self>

This is supported on crate feature temporal only.

Returns the day of year starting from 1.

The return value ranges from 1 to 366. (The last day of year differs by years.)

pub fn month(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract month from underlying NaiveDateTime representation. Returns the month number starting from 1.

The return value ranges from 1 to 12.

pub fn year(&self) -> Result<Self>

This is supported on crate feature temporal only.

Extract month from underlying NaiveDateTime representation. Returns the year number in the calendar date.

impl Series

pub fn series_equal(&self, other: &Series) -> bool

Check if series are equal. Note that None == None evaluates to false

pub fn series_equal_missing(&self, other: &Series) -> bool

Check if all values in series are equal where None == None evaluates to true.