Struct polars::series::Series

pub struct Series(pub Arc<dyn SeriesTrait + 'static>);

Series

The columnar data type for a DataFrame.

Most of the available functions are defined in the SeriesTrait trait.

The Series struct consists of typed ChunkedArray’s. To quickly cast a Series to a ChunkedArray you can call the method with the name of the type:

let s: Series = [1, 2, 3].iter().collect();
// Quickly obtain the ChunkedArray wrapped by the Series.
let chunked_array = s.i32().unwrap();

Arithmetic

You can do standard arithmetic on series.

let s = Series::new("a", [1 , 2, 3]);
let out_add = &s + &s;
let out_sub = &s - &s;
let out_div = &s / &s;
let out_mul = &s * &s;

Or with series and numbers.

let s: Series = (1..3).collect();
let out_add_one = &s + 1;
let out_multiply = &s * 10;

// Could not overload left hand side operator.
let out_divide = 1.div(&s);
let out_add = 1.add(&s);
let out_subtract = 1.sub(&s);
let out_multiply = 1.mul(&s);

Comparison

You can obtain boolean mask by comparing series.

let s = Series::new("dollars", &[1, 2, 3]);
let mask = s.equal(1).unwrap();
let valid = [true, false, false].iter();
assert!(mask
    .into_iter()
    .map(|opt_bool| opt_bool.unwrap()) // option, because series can be null
    .zip(valid)
    .all(|(a, b)| a == *b))

See all the comparison operators in the CmpOps trait

Iterators

The Series variants contain differently typed ChunkedArray’s. These structs can be turned into iterators, making it possible to use any function/ closure you want on a Series.

These iterators return an Option<T> because the values of a series may be null.

use polars_core::prelude::*;
let pi = 3.14;
let s = Series::new("angle", [2f32 * pi, pi, 1.5 * pi].as_ref());
let s_cos: Series = s.f32()
                    .expect("series was not an f32 dtype")
                    .into_iter()
                    .map(|opt_angle| opt_angle.map(|angle| angle.cos()))
                    .collect();

Creation

Series can be create from different data structures. Below we’ll show a few ways we can create a Series object.

// Series can be created from Vec's, slices and arrays
Series::new("boolean series", &[true, false, true]);
Series::new("int series", &[1, 2, 3]);
// And can be nullable
Series::new("got nulls", &[Some(1), None, Some(2)]);

// Series can also be collected from iterators
let from_iter: Series = (0..10)
    .into_iter()
    .collect();

Tuple Fields
0: Arc<dyn SeriesTrait + 'static>
Implementations

impl Series

pub fn sample_n(
&self,
n: usize,
with_replacement: bool,
shuffle: bool,
seed: Option<u64>
) -> Result<Series, PolarsError>

pub fn sample_frac(
&self,
frac: f64,
with_replacement: bool,
shuffle: bool,
seed: Option<u64>
) -> Result<Series, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

pub fn shuffle(&self, seed: Option<u64>) -> Series

impl Series

pub fn from_any_values_and_dtype(
name: &str,
av: &[AnyValue<'_>],
dtype: &DataType
) -> Result<Series, PolarsError>

pub fn from_any_values(
name: &str,
av: &[AnyValue<'_>]
) -> Result<Series, PolarsError>

impl Series

pub fn array_ref(&self, chunk_idx: usize) -> &Box<dyn Array + 'static, Global>

Returns a reference to the Arrow ArrayRef

pub fn to_arrow(&self, chunk_idx: usize) -> Box<dyn Array + 'static, Global>

Convert a chunk in the Series to the correct Arrow type. This conversion is needed because polars doesn’t use a 1 on 1 mapping for logical/ categoricals, etc.

impl Series

pub fn iter(&self) -> SeriesIter<'_>

iterate over Series as AnyValue.

Panics

This will panic if the array is not rechunked first.

pub fn phys_iter(
&self
) -> Box<dyn ExactSizeIterator<Item = AnyValue<'_>>, Global>

impl Series

pub fn i8(&self) -> Result<&ChunkedArray<Int8Type>, PolarsError>

Unpack to ChunkedArray of dtype i8

pub fn i16(&self) -> Result<&ChunkedArray<Int16Type>, PolarsError>

Unpack to ChunkedArray i16

pub fn i32(&self) -> Result<&ChunkedArray<Int32Type>, PolarsError>

Unpack to ChunkedArray

let s = Series::new("foo", [1i32 ,2, 3]);
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

pub fn i64(&self) -> Result<&ChunkedArray<Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype i64

pub fn f32(&self) -> Result<&ChunkedArray<Float32Type>, PolarsError>

Unpack to ChunkedArray of dtype f32

pub fn f64(&self) -> Result<&ChunkedArray<Float64Type>, PolarsError>

Unpack to ChunkedArray of dtype f64

pub fn u8(&self) -> Result<&ChunkedArray<UInt8Type>, PolarsError>

Unpack to ChunkedArray of dtype u8

pub fn u16(&self) -> Result<&ChunkedArray<UInt16Type>, PolarsError>

Unpack to ChunkedArray of dtype u16

pub fn u32(&self) -> Result<&ChunkedArray<UInt32Type>, PolarsError>

Unpack to ChunkedArray of dtype u32

pub fn u64(&self) -> Result<&ChunkedArray<UInt64Type>, PolarsError>

Unpack to ChunkedArray of dtype u64

pub fn bool(&self) -> Result<&ChunkedArray<BooleanType>, PolarsError>

Unpack to ChunkedArray of dtype bool

pub fn utf8(&self) -> Result<&ChunkedArray<Utf8Type>, PolarsError>

Unpack to ChunkedArray of dtype utf8

pub fn time(&self) -> Result<&Logical<TimeType, Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype Time

pub fn date(&self) -> Result<&Logical<DateType, Int32Type>, PolarsError>

Unpack to ChunkedArray of dtype Date

pub fn datetime(&self) -> Result<&Logical<DatetimeType, Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype datetime

pub fn duration(&self) -> Result<&Logical<DurationType, Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype duration

pub fn list(&self) -> Result<&ChunkedArray<ListType>, PolarsError>

Unpack to ChunkedArray of dtype list

impl Series

pub fn extend_constant(
&self,
value: AnyValue<'_>,
n: usize
) -> Result<Series, PolarsError>

Extend with a constant value.

impl Series

pub fn full_null(name: &str, size: usize, dtype: &DataType) -> Series

impl Series

pub fn to_list(&self) -> Result<ChunkedArray<ListType>, PolarsError>

Convert the values of this Series to a ListChunked with a length of 1, So a Series of: [1, 2, 3] becomes [[1, 2, 3]]

pub fn reshape(&self, dims: &[i64]) -> Result<Series, PolarsError>

impl Series

pub fn new_empty(name: &str, dtype: &DataType) -> Series

Create a new empty Series

pub fn set_sorted(&mut self, sorted: IsSorted)

pub fn into_frame(self) -> DataFrame

pub fn rename(&mut self, name: &str) -> &mut Series

Rename series.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn append(&mut self, other: &Series) -> Result<&mut Series, PolarsError>

Append in place. This is done by adding the chunks of other to this Series.

See ChunkedArray::append and ChunkedArray::extend.

pub fn extend(&mut self, other: &Series) -> Result<&mut Series, PolarsError>

Extend the memory backed by this array with the values from other.

See ChunkedArray::extend and ChunkedArray::append.

pub fn sort(&self, reverse: bool) -> Series

pub fn as_single_ptr(&mut self) -> Result<usize, PolarsError>

Only implemented for numeric types

pub fn cast(&self, dtype: &DataType) -> Result<Series, PolarsError>

Cast [Series] to another [DataType]

pub fn sum<T>(&self) -> Option<T>where
T: NumCast,

Compute the sum of all values in this Series. Returns Some(0) if the array is empty, and None if the array only contains null values.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

let s = Series::new("days", &[1, 2, 3]);
assert_eq!(s.sum(), Some(6));

pub fn min<T>(&self) -> Option<T>where
T: NumCast,

Returns the minimum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.min(), Some(1));

pub fn max<T>(&self) -> Option<T>where
T: NumCast,

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

let s = Series::new("days", [1, 2, 3].as_ref());
assert_eq!(s.max(), Some(3));

pub fn explode(&self) -> Result<Series, PolarsError>

Explode a list or utf8 Series. This expands every item to a new row..

pub fn is_nan(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if float value is NaN (note this is different than missing/ null)

pub fn is_not_nan(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if float value is NaN (note this is different than missing/ null)

pub fn is_finite(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if float value is finite

pub fn is_infinite(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if float value is infinite

pub fn zip_with(
&self,
mask: &ChunkedArray<BooleanType>,
other: &Series
) -> Result<Series, PolarsError>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

pub fn to_physical_repr(&self) -> Cow<'_, Series>

Cast a datelike Series to their physical representation. Primitives remain unchanged

Date -> Int32

Datetime-> Int64

Time -> Int64

Categorical -> UInt32

pub unsafe fn take_unchecked_threaded(
&self,
idx: &ChunkedArray<UInt32Type>,
rechunk: bool
) -> Result<Series, PolarsError>

Take by index if ChunkedArray contains a single chunk.

Safety

This doesn’t check any bounds. Null validity is checked.

pub fn take_threaded(
&self,
idx: &ChunkedArray<UInt32Type>,
rechunk: bool
) -> Result<Series, PolarsError>

Take by index. This operation is clone.

Notes

Out of bounds access doesn’t Error but will return a Null value

pub fn filter_threaded(
&self,
filter: &ChunkedArray<BooleanType>,
rechunk: bool
) -> Result<Series, PolarsError>

Filter by boolean mask. This operation clones data.

pub fn sum_as_series(&self) -> Series

Get the sum of the Series as a new Series of length 1. Returns a Series with a single zeroed entry if self is an empty numeric series.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn cummax(&self, _reverse: bool) -> Series

Get an array with the cumulative max computed at every element

pub fn cummin(&self, _reverse: bool) -> Series

Get an array with the cumulative min computed at every element

pub fn cumsum(&self, reverse: bool) -> Series

Get an array with the cumulative sum computed at every element

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn cumprod(&self, reverse: bool) -> Series

Get an array with the cumulative product computed at every element

If the DataType is one of {Int8, UInt8, Int16, UInt16, Int32, UInt32} the Series is first cast to Int64 to prevent overflow issues.

pub fn product(&self) -> Series

Get the product of an array.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn strict_cast(&self, data_type: &DataType) -> Result<Series, PolarsError>

Cast throws an error if conversion had overflows

pub fn str_value(&self, index: usize) -> Result<Cow<'_, str>, PolarsError>

pub fn head(&self, length: Option<usize>) -> Series

Get the head of the Series.

pub fn tail(&self, length: Option<usize>) -> Series

Get the tail of the Series.

pub fn mean_as_series(&self) -> Series

pub fn unique_stable(&self) -> Result<Series, PolarsError>

Compute the unique elements, but maintain order. This requires more work than a naive Series::unique.

pub fn idx(&self) -> Result<&ChunkedArray<UInt32Type>, PolarsError>

pub fn estimated_size(&self) -> usize

Returns an estimation of the total (heap) allocated size of the Series in bytes.

Implementation

This estimation is the sum of the size of its buffers, validity, including nested arrays. Multiple arrays may share buffers and bitmaps. Therefore, the size of 2 arrays is not the sum of the sizes computed from this function. In particular, StructArray’s size is an upper bound.

When an array is sliced, its allocated size remains constant because the buffer unchanged. However, this function will yield a smaller number. This is because this function returns the visible size of the buffer, not its total capacity.

FFI buffers are included in this estimation.

impl Series

pub fn series_equal(&self, other: &Series) -> bool

Check if series are equal. Note that None == None evaluates to false

pub fn series_equal_missing(&self, other: &Series) -> bool

Check if all values in series are equal where None == None evaluates to true. Two Datetime series are not equal if their timezones are different, regardless if they represent the same UTC time or not.

pub fn get_data_ptr(&self) -> usize

Get a pointer to the underlying data of this Series. Can be useful for fast comparisons.