polars::series

Struct Series

pub struct Series(pub Arc<dyn SeriesTrait>);

Series

The columnar data type for a DataFrame.

Most of the available functions are defined in the SeriesTrait trait.

The Series struct consists of typed ChunkedArray’s. To quickly cast a Series to a ChunkedArray you can call the method with the name of the type:

let s: Series = [1, 2, 3].iter().collect();
// Quickly obtain the ChunkedArray wrapped by the Series.
let chunked_array = s.i32().unwrap();

Arithmetic

You can do standard arithmetic on series.

let s = Series::new("a".into(), [1 , 2, 3]);
let out_add = &s + &s;
let out_sub = &s - &s;
let out_div = &s / &s;
let out_mul = &s * &s;

Or with series and numbers.

let s: Series = (1..3).collect();
let out_add_one = &s + 1;
let out_multiply = &s * 10;

// Could not overload left hand side operator.
let out_divide = 1.div(&s);
let out_add = 1.add(&s);
let out_subtract = 1.sub(&s);
let out_multiply = 1.mul(&s);

Comparison

You can obtain boolean mask by comparing series.

let s = Series::new("dollars".into(), &[1, 2, 3]);
let mask = s.equal(1).unwrap();
let valid = [true, false, false].iter();
assert!(mask
    .into_iter()
    .map(|opt_bool| opt_bool.unwrap()) // option, because series can be null
    .zip(valid)
    .all(|(a, b)| a == *b))

See all the comparison operators in the ChunkCompareEq trait and ChunkCompareIneq trait.

Iterators

The Series variants contain differently typed ChunkedArrays. These structs can be turned into iterators, making it possible to use any function/ closure you want on a Series.

These iterators return an Option<T> because the values of a series may be null.

use polars_core::prelude::*;
let pi = 3.14;
let s = Series::new("angle".into(), [2f32 * pi, pi, 1.5 * pi].as_ref());
let s_cos: Series = s.f32()
                    .expect("series was not an f32 dtype")
                    .into_iter()
                    .map(|opt_angle| opt_angle.map(|angle| angle.cos()))
                    .collect();

Creation

Series can be create from different data structures. Below we’ll show a few ways we can create a Series object.

// Series can be created from Vec's, slices and arrays
Series::new("boolean series".into(), &[true, false, true]);
Series::new("int series".into(), &[1, 2, 3]);
// And can be nullable
Series::new("got nulls".into(), &[Some(1), None, Some(2)]);

// Series can also be collected from iterators
let from_iter: Series = (0..10)
    .into_iter()
    .collect();

Tuple Fields
0: Arc<dyn SeriesTrait>
Implementations

impl Series

pub fn fill_null( &self, strategy: FillNullStrategy, ) -> Result<Series, PolarsError>

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)

Zero fill (replace None with the value zero)

One fill (replace None with the value one)

NOTE: If you want to fill the Nones with a value use the fill_null operation on ChunkedArray<T>.

Example

fn example() -> PolarsResult<()> {
    let s = Column::new("some_missing".into(), &[Some(1), None, Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Forward(None))?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Backward(None))?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Min)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Max)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Mean)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::Zero)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(0), Some(2)]);

    let filled = s.fill_null(FillNullStrategy::One)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    Ok(())
}
example();

impl Series

pub fn sample_n( &self, n: usize, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> Result<Series, PolarsError>

pub fn sample_frac( &self, frac: f64, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> Result<Series, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

pub fn shuffle(&self, seed: Option<u64>) -> Series

impl Series

pub fn fmt_list(&self) -> String

impl Series

pub fn serialize_into_writer( &self, writer: &mut dyn Write, ) -> Result<(), PolarsError>

pub fn serialize_to_bytes(&self) -> Result<Vec<u8>, PolarsError>

pub fn deserialize_from_reader( reader: &mut dyn Read, ) -> Result<Series, PolarsError>

impl Series

pub fn from_any_values( name: PlSmallStr, values: &[AnyValue<'_>], strict: bool, ) -> Result<Series, PolarsError>

Construct a new Series from a slice of AnyValues.

The data type of the resulting Series is determined by the values and the strict parameter:

If strict is true, the data type is equal to the data type of the first non-null value. If any other non-null values do not match this data type, an error is raised.

If strict is false, the data type is the supertype of the values. An error is returned if no supertype can be determined. WARNING: A full pass over the values is required to determine the supertype.

If no values were passed, the resulting data type is Null.

pub fn from_any_values_and_dtype( name: PlSmallStr, values: &[AnyValue<'_>], dtype: &DataType, strict: bool, ) -> Result<Series, PolarsError>

Construct a new Series with the given dtype from a slice of AnyValues.

If strict is true, an error is returned if the values do not match the given data type. If strict is false, values that do not match the given data type are cast. If casting is not possible, the values are set to null instead.

impl Series

pub fn wrapping_trunc_div_scalar<T>(&self, rhs: T) -> Series
where T: Num + NumCast,

impl Series

pub fn try_add_owned(self, other: Series) -> Result<Series, PolarsError>

pub fn try_sub_owned(self, other: Series) -> Result<Series, PolarsError>

pub fn try_mul_owned(self, other: Series) -> Result<Series, PolarsError>

impl Series

pub fn from_array<A>(name: PlSmallStr, array: A) -> Series
where A: ParameterFreeDtypeStaticArray,

pub fn from_chunk_and_dtype( name: PlSmallStr, chunk: Box<dyn Array>, dtype: &DataType, ) -> Result<Series, PolarsError>

pub unsafe fn from_chunks_and_dtype_unchecked( name: PlSmallStr, chunks: Vec<Box<dyn Array>>, dtype: &DataType, ) -> Series

Takes chunks and a polars datatype and constructs the Series This is faster than creating from chunks and an arrow datatype because there is no casting involved

Safety

The caller must ensure that the given dtype’s physical type matches all the ArrayRef dtypes.

pub unsafe fn _try_from_arrow_unchecked( name: PlSmallStr, chunks: Vec<Box<dyn Array>>, dtype: &ArrowDataType, ) -> Result<Series, PolarsError>

Safety

The caller must ensure that the given dtype matches all the ArrayRef dtypes.

pub unsafe fn _try_from_arrow_unchecked_with_md( name: PlSmallStr, chunks: Vec<Box<dyn Array>>, dtype: &ArrowDataType, md: Option<&BTreeMap<PlSmallStr, PlSmallStr>>, ) -> Result<Series, PolarsError>

Create a new Series without checking if the inner dtype of the chunks is correct

Safety

The caller must ensure that the given dtype matches all the ArrayRef dtypes.

impl Series

pub fn try_new<T>( name: PlSmallStr, data: T, ) -> Result<Series, <(PlSmallStr, T) as TryInto<Series>>::Error>
where (PlSmallStr, T): TryInto<Series>,

impl Series

pub fn new_null(name: PlSmallStr, len: usize) -> Series

impl Series

pub fn array_ref(&self, chunk_idx: usize) -> &Box<dyn Array>

Returns a reference to the Arrow ArrayRef

pub fn to_arrow( &self, chunk_idx: usize, compat_level: CompatLevel, ) -> Box<dyn Array>

Convert a chunk in the Series to the correct Arrow type. This conversion is needed because polars doesn’t use a 1 on 1 mapping for logical/categoricals, etc.

impl Series

pub fn iter(&self) -> SeriesIter<'_>

iterate over Series as AnyValue.

Panics

This will panic if the array is not rechunked first.

pub fn phys_iter(&self) -> Box<dyn ExactSizeIterator<Item = AnyValue<'_>> + '_>

impl Series

pub fn try_i8(&self) -> Option<&ChunkedArray<Int8Type>>

Unpack to ChunkedArray of dtype DataType::Int8

pub fn try_i16(&self) -> Option<&ChunkedArray<Int16Type>>

Unpack to ChunkedArray of dtype DataType::Int16

pub fn try_i32(&self) -> Option<&ChunkedArray<Int32Type>>

Unpack to ChunkedArray

let s = Series::new("foo".into(), [1i32 ,2, 3]);
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

Unpack to ChunkedArray of dtype DataType::Int32

pub fn try_i64(&self) -> Option<&ChunkedArray<Int64Type>>

Unpack to ChunkedArray of dtype DataType::Int64

pub fn try_i128(&self) -> Option<&ChunkedArray<Int128Type>>

Unpack to ChunkedArray of dtype DataType::Int128

pub fn try_f32(&self) -> Option<&ChunkedArray<Float32Type>>

Unpack to ChunkedArray of dtype DataType::Float32

pub fn try_f64(&self) -> Option<&ChunkedArray<Float64Type>>

Unpack to ChunkedArray of dtype DataType::Float64

pub fn try_u8(&self) -> Option<&ChunkedArray<UInt8Type>>

Unpack to ChunkedArray of dtype DataType::UInt8

pub fn try_u16(&self) -> Option<&ChunkedArray<UInt16Type>>

Unpack to ChunkedArray of dtype DataType::UInt16

pub fn try_u32(&self) -> Option<&ChunkedArray<UInt32Type>>

Unpack to ChunkedArray of dtype DataType::UInt32

pub fn try_u64(&self) -> Option<&ChunkedArray<UInt64Type>>

Unpack to ChunkedArray of dtype DataType::UInt64

pub fn try_bool(&self) -> Option<&ChunkedArray<BooleanType>>

Unpack to ChunkedArray of dtype DataType::Boolean

pub fn try_str(&self) -> Option<&ChunkedArray<StringType>>

Unpack to ChunkedArray of dtype DataType::String

pub fn try_binary(&self) -> Option<&ChunkedArray<BinaryType>>

Unpack to ChunkedArray of dtype DataType::Binary

pub fn try_binary_offset(&self) -> Option<&ChunkedArray<BinaryOffsetType>>

Unpack to ChunkedArray of dtype DataType::Binary

pub fn try_time(&self) -> Option<&Logical<TimeType, Int64Type>>

Unpack to ChunkedArray of dtype DataType::Time

pub fn try_date(&self) -> Option<&Logical<DateType, Int32Type>>

Unpack to ChunkedArray of dtype DataType::Date

pub fn try_datetime(&self) -> Option<&Logical<DatetimeType, Int64Type>>

Unpack to ChunkedArray of dtype DataType::Datetime

pub fn try_duration(&self) -> Option<&Logical<DurationType, Int64Type>>

Unpack to ChunkedArray of dtype DataType::Duration

pub fn try_decimal(&self) -> Option<&Logical<DecimalType, Int128Type>>

Unpack to ChunkedArray of dtype DataType::Decimal

pub fn try_list(&self) -> Option<&ChunkedArray<ListType>>

Unpack to ChunkedArray of dtype list

pub fn try_array(&self) -> Option<&ChunkedArray<FixedSizeListType>>

Unpack to ChunkedArray of dtype DataType::Array

pub fn try_cat<T>( &self, ) -> Option<&Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>>
where T: PolarsCategoricalType,

pub fn try_cat8( &self, ) -> Option<&Logical<Categorical8Type, <Categorical8Type as PolarsCategoricalType>::PolarsPhysical>>

Unpack to ChunkedArray of dtype DataType::Categorical or DataType::Enum with a physical type of UInt8.

pub fn try_cat16( &self, ) -> Option<&Logical<Categorical16Type, <Categorical16Type as PolarsCategoricalType>::PolarsPhysical>>

pub fn try_cat32( &self, ) -> Option<&Logical<Categorical32Type, <Categorical32Type as PolarsCategoricalType>::PolarsPhysical>>

pub fn try_struct(&self) -> Option<&ChunkedArray<StructType>>

Unpack to ChunkedArray of dtype DataType::Struct

pub fn try_null(&self) -> Option<&NullChunked>

Unpack to ChunkedArray of dtype DataType::Null

pub fn i8(&self) -> Result<&ChunkedArray<Int8Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Int8

pub fn i16(&self) -> Result<&ChunkedArray<Int16Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Int16

pub fn i32(&self) -> Result<&ChunkedArray<Int32Type>, PolarsError>

Unpack to ChunkedArray

let s = Series::new("foo".into(), [1i32 ,2, 3]);
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

Unpack to ChunkedArray of dtype DataType::Int32

pub fn i64(&self) -> Result<&ChunkedArray<Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Int64

pub fn i128(&self) -> Result<&ChunkedArray<Int128Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Int128

pub fn f32(&self) -> Result<&ChunkedArray<Float32Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Float32

pub fn f64(&self) -> Result<&ChunkedArray<Float64Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Float64

pub fn u8(&self) -> Result<&ChunkedArray<UInt8Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::UInt8

pub fn u16(&self) -> Result<&ChunkedArray<UInt16Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::UInt16

pub fn u32(&self) -> Result<&ChunkedArray<UInt32Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::UInt32

pub fn u64(&self) -> Result<&ChunkedArray<UInt64Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::UInt64

pub fn bool(&self) -> Result<&ChunkedArray<BooleanType>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Boolean

pub fn str(&self) -> Result<&ChunkedArray<StringType>, PolarsError>

Unpack to ChunkedArray of dtype DataType::String

pub fn binary(&self) -> Result<&ChunkedArray<BinaryType>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Binary

pub fn binary_offset( &self, ) -> Result<&ChunkedArray<BinaryOffsetType>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Binary

pub fn time(&self) -> Result<&Logical<TimeType, Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Time

pub fn date(&self) -> Result<&Logical<DateType, Int32Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Date

pub fn datetime(&self) -> Result<&Logical<DatetimeType, Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Datetime

pub fn duration(&self) -> Result<&Logical<DurationType, Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Duration

pub fn decimal(&self) -> Result<&Logical<DecimalType, Int128Type>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Decimal

pub fn list(&self) -> Result<&ChunkedArray<ListType>, PolarsError>

Unpack to ChunkedArray of dtype list

pub fn array(&self) -> Result<&ChunkedArray<FixedSizeListType>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Array

pub fn cat<T>( &self, ) -> Result<&Logical<T, <T as PolarsCategoricalType>::PolarsPhysical>, PolarsError>
where T: PolarsCategoricalType,

Unpack to ChunkedArray of dtype DataType::Categorical or DataType::Enum.

pub fn cat8( &self, ) -> Result<&Logical<Categorical8Type, <Categorical8Type as PolarsCategoricalType>::PolarsPhysical>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Categorical or DataType::Enum with a physical type of UInt8.

pub fn cat16( &self, ) -> Result<&Logical<Categorical16Type, <Categorical16Type as PolarsCategoricalType>::PolarsPhysical>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Categorical or DataType::Enum with a physical type of UInt16.

pub fn cat32( &self, ) -> Result<&Logical<Categorical32Type, <Categorical32Type as PolarsCategoricalType>::PolarsPhysical>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Categorical or DataType::Enum with a physical type of UInt32.

pub fn struct_(&self) -> Result<&ChunkedArray<StructType>, PolarsError>

Unpack to ChunkedArray of dtype DataType::Struct

pub fn null(&self) -> Result<&NullChunked, PolarsError>

Unpack to ChunkedArray of dtype DataType::Null

impl Series

pub fn extend_constant( &self, value: AnyValue<'_>, n: usize, ) -> Result<Series, PolarsError>

Extend with a constant value.

impl Series

pub fn full_null(name: PlSmallStr, size: usize, dtype: &DataType) -> Series

impl Series

pub fn get_leaf_array(&self) -> Series

Recurse nested types until we are at the leaf array.

pub fn list_offsets_and_validities_recursive( &self, ) -> (Vec<OffsetsBuffer<i64>>, Vec<Option<Bitmap>>)

TODO: Move this somewhere else?

pub fn implode(&self) -> Result<ChunkedArray<ListType>, PolarsError>

Convert the values of this Series to a ListChunked with a length of 1, so a Series of [1, 2, 3] becomes [[1, 2, 3]].

pub fn reshape_array( &self, dimensions: &[ReshapeDimension], ) -> Result<Series, PolarsError>

pub fn reshape_list( &self, dimensions: &[ReshapeDimension], ) -> Result<Series, PolarsError>

impl Series

pub fn new_empty(name: PlSmallStr, dtype: &DataType) -> Series

Create a new empty Series.

pub fn clear(&self) -> Series

pub fn take_inner<T>(self) -> ChunkedArray<T>
where T: PolarsPhysicalType,

Take or clone a owned copy of the inner ChunkedArray.

pub unsafe fn chunks_mut(&mut self) -> &mut Vec<Box<dyn Array>>

Safety

The caller must ensure the length and the data types of ArrayRef does not change. And that the null_count is updated (e.g. with a compute_len())

pub fn into_chunks(self) -> Vec<Box<dyn Array>>

pub fn select_chunk(&self, i: usize) -> Series

pub fn is_sorted_flag(&self) -> IsSorted

pub fn set_sorted_flag(&mut self, sorted: IsSorted)

pub fn get_flags(&self) -> StatisticsFlags

pub fn into_frame(self) -> DataFrame

pub fn rename(&mut self, name: PlSmallStr) -> &mut Series

Rename series.

pub fn with_name(self, name: PlSmallStr) -> Series

Return this Series with a new name.

pub fn from_arrow_chunks( name: PlSmallStr, arrays: Vec<Box<dyn Array>>, ) -> Result<Series, PolarsError>

pub fn from_arrow( name: PlSmallStr, array: Box<dyn Array>, ) -> Result<Series, PolarsError>

pub fn shrink_to_fit(&mut self)

Shrink the capacity of this array to fit its length.

pub fn append(&mut self, other: &Series) -> Result<&mut Series, PolarsError>

Append in place. This is done by adding the chunks of other to this Series.

See ChunkedArray::append and ChunkedArray::extend.

pub fn append_owned( &mut self, other: Series, ) -> Result<&mut Series, PolarsError>

Append in place. This is done by adding the chunks of other to this Series.

See ChunkedArray::append_owned and ChunkedArray::extend.

pub fn compute_len(&mut self)

Redo a length and null_count compute

pub fn extend(&mut self, other: &Series) -> Result<&mut Series, PolarsError>

Extend the memory backed by this array with the values from other.

See ChunkedArray::extend and ChunkedArray::append.

pub fn sort(&self, sort_options: SortOptions) -> Result<Series, PolarsError>

Sort the series with specific options.

Example

let s = Series::new("foo".into(), [2, 1, 3]);
let sorted = s.sort(SortOptions::default())?;
assert_eq!(sorted, Series::new("foo".into(), [1, 2, 3]));
}

See SortOptions for more options.

pub fn as_single_ptr(&mut self) -> Result<usize, PolarsError>

Only implemented for numeric types

pub fn cast(&self, dtype: &DataType) -> Result<Series, PolarsError>

pub fn cast_with_options( &self, dtype: &DataType, options: CastOptions, ) -> Result<Series, PolarsError>

Cast Series to another DataType.

pub unsafe fn cast_unchecked( &self, dtype: &DataType, ) -> Result<Series, PolarsError>

Cast from physical to logical types without any checks on the validity of the cast.

Safety

This can lead to invalid memory access in downstream code.

pub unsafe fn from_physical_unchecked( &self, dtype: &DataType, ) -> Result<Series, PolarsError>

Convert a non-logical series back into a logical series without casting.

Safety

This can lead to invalid memory access in downstream code.

pub fn to_float(&self) -> Result<Series, PolarsError>

Cast numerical types to f64, and keep floats as is.

pub fn sum<T>(&self) -> Result<T, PolarsError>
where T: NumCast,

Compute the sum of all values in this Series. Returns Some(0) if the array is empty, and None if the array only contains null values.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn min<T>(&self) -> Result<Option<T>, PolarsError>
where T: NumCast,

Returns the minimum value in the array, according to the natural order. Returns an option because the array is nullable.

pub fn max<T>(&self) -> Result<Option<T>, PolarsError>
where T: NumCast,

Returns the maximum value in the array, according to the natural order. Returns an option because the array is nullable.

pub fn explode(&self, skip_empty: bool) -> Result<Series, PolarsError>

Explode a list Series. This expands every item to a new row..

pub fn is_nan(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if numeric value is NaN (note this is different than missing/ null)

pub fn is_not_nan(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if numeric value is NaN (note this is different than missing/null)

pub fn is_finite(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if numeric value is finite

pub fn is_infinite(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Check if numeric value is infinite

pub fn zip_with( &self, mask: &ChunkedArray<BooleanType>, other: &Series, ) -> Result<Series, PolarsError>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false. This function automatically broadcasts unit length inputs.

pub fn to_physical_repr(&self) -> Cow<'_, Series>

Converts a Series to their physical representation, if they have one, otherwise the series is left unchanged.

Date -> Int32

Datetime -> Int64

Duration -> Int64

Decimal -> Int128

Time -> Int64

Categorical -> U8/U16/U32

List(inner) -> List(physical of inner)

Array(inner) -> Array(physical of inner)

Struct -> Struct with physical repr of each struct column

pub fn gather_every( &self, n: usize, offset: usize, ) -> Result<Series, PolarsError>

Traverse and collect every nth element in a new array.

pub fn sum_reduce(&self) -> Result<Scalar, PolarsError>

Get the sum of the Series as a new Series of length 1. Returns a Series with a single zeroed entry if self is an empty numeric series.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn product(&self) -> Result<Scalar, PolarsError>

Get the product of an array.

If the DataType is one of {Int8, UInt8, Int16, UInt16} the Series is first cast to Int64 to prevent overflow issues.

pub fn strict_cast(&self, dtype: &DataType) -> Result<Series, PolarsError>

Cast throws an error if conversion had overflows

pub fn str_value(&self, index: usize) -> Result<Cow<'_, str>, PolarsError>

pub fn head(&self, length: Option<usize>) -> Series

Get the head of the Series.

pub fn tail(&self, length: Option<usize>) -> Series

Get the tail of the Series.

pub fn mean_reduce(&self) -> Scalar

pub fn unique_stable(&self) -> Result<Series, PolarsError>

Compute the unique elements, but maintain order. This requires more work than a naive Series::unique.

pub fn try_idx(&self) -> Option<&ChunkedArray<UInt32Type>>

pub fn idx(&self) -> Result<&ChunkedArray<UInt32Type>, PolarsError>

pub fn estimated_size(&self) -> usize

Returns an estimation of the total (heap) allocated size of the Series in bytes.

Implementation

This estimation is the sum of the size of its buffers, validity, including nested arrays. Multiple arrays may share buffers and bitmaps. Therefore, the size of 2 arrays is not the sum of the sizes computed from this function. In particular, StructArray’s size is an upper bound.

When an array is sliced, its allocated size remains constant because the buffer unchanged. However, this function will yield a smaller number. This is because this function returns the visible size of the buffer, not its total capacity.

FFI buffers are included in this estimation.

pub fn as_list(&self) -> ChunkedArray<ListType>

Packs every element into a list.

impl Series

pub fn equals(&self, other: &Series) -> bool

Check if series are equal. Note that None == None evaluates to false

pub fn equals_missing(&self, other: &Series) -> bool

Check if all values in series are equal where None == None evaluates to true.