Trait polars::series::SeriesTrait

pub trait SeriesTrait: PrivateSeries + Send + Sync {
    pub fn rename(&mut self, name: &str);

    pub fn take_every(&self, n: usize) -> Series;

    pub fn sample_n(
        &self, 
        n: usize, 
        with_replacement: bool
    ) -> Result<Series, PolarsError>;

    pub fn sample_frac(
        &self, 
        frac: f64, 
        with_replacement: bool
    ) -> Result<Series, PolarsError>;

    pub fn cum_max(&self, _reverse: bool) -> Series { ... }

    pub fn cum_min(&self, _reverse: bool) -> Series { ... }

    pub fn cum_sum(&self, _reverse: bool) -> Series { ... }

    pub fn array_data(&self) -> Vec<Arc<ArrayData>, Global> { ... }

    pub fn chunk_lengths(&self) -> &Vec<usize, Global> { ... }

    pub fn name(&self) -> &str { ... }

    pub fn field(&self) -> &Field { ... }

    pub fn dtype(&self) -> &DataType { ... }

    pub fn chunks(&self) -> &Vec<Arc<dyn Array + 'static>, Global> { ... }

    pub fn n_chunks(&self) -> usize { ... }

    pub fn i8(&self) -> Result<&ChunkedArray<Int8Type>, PolarsError> { ... }

    pub fn i16(&self) -> Result<&ChunkedArray<Int16Type>, PolarsError> { ... }

    pub fn i32(&self) -> Result<&ChunkedArray<Int32Type>, PolarsError> { ... }

    pub fn i64(&self) -> Result<&ChunkedArray<Int64Type>, PolarsError> { ... }

    pub fn f32(&self) -> Result<&ChunkedArray<Float32Type>, PolarsError> { ... }

    pub fn f64(&self) -> Result<&ChunkedArray<Float64Type>, PolarsError> { ... }

    pub fn u8(&self) -> Result<&ChunkedArray<UInt8Type>, PolarsError> { ... }

    pub fn u16(&self) -> Result<&ChunkedArray<UInt16Type>, PolarsError> { ... }

    pub fn u32(&self) -> Result<&ChunkedArray<UInt32Type>, PolarsError> { ... }

    pub fn u64(&self) -> Result<&ChunkedArray<UInt64Type>, PolarsError> { ... }

    pub fn bool(&self) -> Result<&ChunkedArray<BooleanType>, PolarsError> { ... }

    pub fn utf8(&self) -> Result<&ChunkedArray<Utf8Type>, PolarsError> { ... }

    pub fn date32(&self) -> Result<&ChunkedArray<Date32Type>, PolarsError> { ... }

    pub fn date64(&self) -> Result<&ChunkedArray<Date64Type>, PolarsError> { ... }

    pub fn time64_nanosecond(
        &self
    ) -> Result<&ChunkedArray<Time64NanosecondType>, PolarsError> { ... }

    pub fn duration_nanosecond(
        &self
    ) -> Result<&ChunkedArray<DurationNanosecondType>, PolarsError> { ... }

    pub fn duration_millisecond(
        &self
    ) -> Result<&ChunkedArray<DurationMillisecondType>, PolarsError> { ... }

    pub fn list(&self) -> Result<&ChunkedArray<ListType>, PolarsError> { ... }

    pub fn categorical(
        &self
    ) -> Result<&ChunkedArray<CategoricalType>, PolarsError> { ... }

    pub fn append_array(
        &mut self, 
        _other: Arc<dyn Array + 'static>
    ) -> Result<(), PolarsError> { ... }

    pub fn limit(&self, num_elements: usize) -> Result<Series, PolarsError> { ... }

    pub fn slice(
        &self, 
        _offset: usize, 
        _length: usize
    ) -> Result<Series, PolarsError> { ... }

    pub fn append(&mut self, _other: &Series) -> Result<(), PolarsError> { ... }

    pub fn filter(
        &self, 
        _filter: &ChunkedArray<BooleanType>
    ) -> Result<Series, PolarsError> { ... }

    pub fn take_iter(&self, _iter: &mut dyn Iterator<Item = usize>) -> Series { ... }

    pub unsafe fn take_iter_unchecked(
        &self, 
        _iter: &mut dyn Iterator<Item = usize>
    ) -> Series { ... }

    pub unsafe fn take_unchecked(
        &self, 
        _idx: &ChunkedArray<UInt32Type>
    ) -> Result<Series, PolarsError> { ... }

    pub unsafe fn take_opt_iter_unchecked(
        &self, 
        _iter: &mut dyn Iterator<Item = Option<usize>>
    ) -> Series { ... }

    pub fn take_opt_iter(
        &self, 
        _iter: &mut dyn Iterator<Item = Option<usize>>
    ) -> Series { ... }

    pub fn take(&self, _indices: &ChunkedArray<UInt32Type>) -> Series { ... }

    pub fn len(&self) -> usize { ... }

    pub fn is_empty(&self) -> bool { ... }

    pub fn rechunk(&self) -> Series { ... }

    pub fn head(&self, _length: Option<usize>) -> Series { ... }

    pub fn tail(&self, _length: Option<usize>) -> Series { ... }

    pub fn drop_nulls(&self) -> Series { ... }

    pub fn expand_at_index(&self, _index: usize, _length: usize) -> Series { ... }

    pub fn cast_with_datatype(
        &self, 
        _data_type: &DataType
    ) -> Result<Series, PolarsError> { ... }

    pub fn to_dummies(&self) -> Result<DataFrame, PolarsError> { ... }

    pub fn value_counts(&self) -> Result<DataFrame, PolarsError> { ... }

    pub fn get(&self, _index: usize) -> AnyValue<'_> { ... }

    pub unsafe fn get_unchecked(&self, _index: usize) -> AnyValue<'_> { ... }

    pub fn sort_in_place(&mut self, _reverse: bool) { ... }

    pub fn sort(&self, _reverse: bool) -> Series { ... }

    pub fn argsort(&self, _reverse: bool) -> ChunkedArray<UInt32Type> { ... }

    pub fn null_count(&self) -> usize { ... }

    pub fn unique(&self) -> Result<Series, PolarsError> { ... }

    pub fn n_unique(&self) -> Result<usize, PolarsError> { ... }

    pub fn arg_unique(&self) -> Result<Vec<u32, Global>, PolarsError> { ... }

    pub fn arg_true(&self) -> Result<ChunkedArray<UInt32Type>, PolarsError> { ... }

    pub fn is_null(&self) -> ChunkedArray<BooleanType> { ... }

    pub fn is_not_null(&self) -> ChunkedArray<BooleanType> { ... }

    pub fn is_unique(&self) -> Result<ChunkedArray<BooleanType>, PolarsError> { ... }

    pub fn is_duplicated(
        &self
    ) -> Result<ChunkedArray<BooleanType>, PolarsError> { ... }

    pub fn null_bits(&self) -> Vec<(usize, Option<Buffer>), Global> { ... }

    pub fn reverse(&self) -> Series { ... }

    pub fn as_single_ptr(&mut self) -> Result<usize, PolarsError> { ... }

    pub fn shift(&self, _periods: i64) -> Series { ... }

    pub fn fill_none(
        &self, 
        _strategy: FillNoneStrategy
    ) -> Result<Series, PolarsError> { ... }

    pub fn zip_with(
        &self, 
        _mask: &ChunkedArray<BooleanType>, 
        _other: &Series
    ) -> Result<Series, PolarsError> { ... }

    pub fn sum_as_series(&self) -> Series { ... }

    pub fn max_as_series(&self) -> Series { ... }

    pub fn min_as_series(&self) -> Series { ... }

    pub fn mean_as_series(&self) -> Series { ... }

    pub fn median_as_series(&self) -> Series { ... }

    pub fn var_as_series(&self) -> Series { ... }

    pub fn std_as_series(&self) -> Series { ... }

    pub fn quantile_as_series(
        &self, 
        _quantile: f64
    ) -> Result<Series, PolarsError> { ... }

    pub fn rolling_mean(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series, PolarsError> { ... }

    pub fn rolling_sum(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series, PolarsError> { ... }

    pub fn rolling_min(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series, PolarsError> { ... }

    pub fn rolling_max(
        &self, 
        _window_size: usize, 
        _weight: Option<&[f64]>, 
        _ignore_null: bool
    ) -> Result<Series, PolarsError> { ... }

    pub fn fmt_list(&self) -> String { ... }

    pub fn hour(&self) -> Result<Series, PolarsError> { ... }

    pub fn minute(&self) -> Result<Series, PolarsError> { ... }

    pub fn second(&self) -> Result<Series, PolarsError> { ... }

    pub fn nanosecond(&self) -> Result<Series, PolarsError> { ... }

    pub fn day(&self) -> Result<Series, PolarsError> { ... }

    pub fn ordinal_day(&self) -> Result<Series, PolarsError> { ... }

    pub fn month(&self) -> Result<Series, PolarsError> { ... }

    pub fn year(&self) -> Result<Series, PolarsError> { ... }

    pub fn datetime_str_fmt(&self, fmt: &str) -> Result<Series, PolarsError> { ... }

    pub fn clone_inner(&self) -> Arc<dyn SeriesTrait + 'static> { ... }

    pub fn get_as_any(&self, _index: usize) -> &(dyn Any + 'static) { ... }

    pub fn pow(&self, _exponent: f64) -> Result<Series, PolarsError> { ... }
}

Required methods

pub fn rename(&mut self, name: &str)

Rename the Series.

pub fn take_every(&self, n: usize) -> Series

Take every nth value as a new Series

pub fn sample_n(
&self,
n: usize,
with_replacement: bool
) -> Result<Series, PolarsError>

Sample n datapoints from this Series.

pub fn sample_frac(
&self,
frac: f64,
with_replacement: bool
) -> Result<Series, PolarsError>

Sample a fraction between 0.0-1.0 of this ChunkedArray.

Loading content...
Provided methods

pub fn cum_max(&self, _reverse: bool) -> Series

Get an array with the cumulative max computed at every element

pub fn cum_min(&self, _reverse: bool) -> Series

Get an array with the cumulative min computed at every element

pub fn cum_sum(&self, _reverse: bool) -> Series

Get an array with the cumulative sum computed at every element

pub fn array_data(&self) -> Vec<Arc<ArrayData>, Global>

Get Arrow ArrayData

pub fn chunk_lengths(&self) -> &Vec<usize, Global>

Get the lengths of the underlying chunks

pub fn name(&self) -> &str

Name of series.

pub fn field(&self) -> &Field

Get field (used in schema)

pub fn dtype(&self) -> &DataType

Get datatype of series.

pub fn chunks(&self) -> &Vec<Arc<dyn Array + 'static>, Global>

Underlying chunks.

pub fn n_chunks(&self) -> usize

Number of chunks in this Series

pub fn i8(&self) -> Result<&ChunkedArray<Int8Type>, PolarsError>

Unpack to ChunkedArray of dtype i8

pub fn i16(&self) -> Result<&ChunkedArray<Int16Type>, PolarsError>

Unpack to ChunkedArray i16

pub fn i32(&self) -> Result<&ChunkedArray<Int32Type>, PolarsError>

Unpack to ChunkedArray

let s: Series = [1, 2, 3].iter().collect();
let s_squared: Series = s.i32()
    .unwrap()
    .into_iter()
    .map(|opt_v| {
        match opt_v {
            Some(v) => Some(v * v),
            None => None, // null value
        }
}).collect();

pub fn i64(&self) -> Result<&ChunkedArray<Int64Type>, PolarsError>

Unpack to ChunkedArray of dtype i64

pub fn f32(&self) -> Result<&ChunkedArray<Float32Type>, PolarsError>

Unpack to ChunkedArray of dtype f32

pub fn f64(&self) -> Result<&ChunkedArray<Float64Type>, PolarsError>

Unpack to ChunkedArray of dtype f64

pub fn u8(&self) -> Result<&ChunkedArray<UInt8Type>, PolarsError>

Unpack to ChunkedArray of dtype u8

pub fn u16(&self) -> Result<&ChunkedArray<UInt16Type>, PolarsError>

Unpack to ChunkedArray of dtype u16

pub fn u32(&self) -> Result<&ChunkedArray<UInt32Type>, PolarsError>

Unpack to ChunkedArray of dtype u32

pub fn u64(&self) -> Result<&ChunkedArray<UInt64Type>, PolarsError>

Unpack to ChunkedArray of dtype u64

pub fn bool(&self) -> Result<&ChunkedArray<BooleanType>, PolarsError>

Unpack to ChunkedArray of dtype bool

pub fn utf8(&self) -> Result<&ChunkedArray<Utf8Type>, PolarsError>

Unpack to ChunkedArray of dtype utf8

pub fn date32(&self) -> Result<&ChunkedArray<Date32Type>, PolarsError>

Unpack to ChunkedArray of dtype date32

pub fn date64(&self) -> Result<&ChunkedArray<Date64Type>, PolarsError>

Unpack to ChunkedArray of dtype date64

pub fn time64_nanosecond(
&self
) -> Result<&ChunkedArray<Time64NanosecondType>, PolarsError>

Unpack to ChunkedArray of dtype time64_nanosecond

pub fn duration_nanosecond(
&self
) -> Result<&ChunkedArray<DurationNanosecondType>, PolarsError>

Unpack to ChunkedArray of dtype duration_nanosecond

pub fn duration_millisecond(
&self
) -> Result<&ChunkedArray<DurationMillisecondType>, PolarsError>

Unpack to ChunkedArray of dtype duration_millisecond

pub fn list(&self) -> Result<&ChunkedArray<ListType>, PolarsError>

Unpack to ChunkedArray of dtype list

pub fn categorical(&self) -> Result<&ChunkedArray<CategoricalType>, PolarsError>

Unpack to ChunkedArray of dtype categorical

pub fn append_array(
&mut self,
_other: Arc<dyn Array + 'static>
) -> Result<(), PolarsError>

Append Arrow array of same dtype to this Series.

pub fn limit(&self, num_elements: usize) -> Result<Series, PolarsError>

Take num_elements from the top as a zero copy view.

pub fn slice(
&self,
_offset: usize,
_length: usize
) -> Result<Series, PolarsError>

Get a zero copy view of the data.

pub fn append(&mut self, _other: &Series) -> Result<(), PolarsError>

Append a Series of the same type in place.

pub fn filter(
&self,
_filter: &ChunkedArray<BooleanType>
) -> Result<Series, PolarsError>

Filter by boolean mask. This operation clones data.

pub fn take_iter(&self, _iter: &mut dyn Iterator<Item = usize>) -> Series

Take by index from an iterator. This operation clones the data.

Safety

Out of bounds access doesn't Error but will return a Null value

pub unsafe fn take_iter_unchecked(
&self,
_iter: &mut dyn Iterator<Item = usize>
) -> Series

Take by index from an iterator. This operation clones the data.

Safety

This doesn't check any bounds or null validity.

pub unsafe fn take_unchecked(
&self,
_idx: &ChunkedArray<UInt32Type>
) -> Result<Series, PolarsError>

Take by index if ChunkedArray contains a single chunk.

Safety

This doesn't check any bounds. Null validity is checked.

pub unsafe fn take_opt_iter_unchecked(
&self,
_iter: &mut dyn Iterator<Item = Option<usize>>
) -> Series

Take by index from an iterator. This operation clones the data.

Safety

This doesn't check any bounds or null validity.

pub fn take_opt_iter(
&self,
_iter: &mut dyn Iterator<Item = Option<usize>>
) -> Series

Take by index from an iterator. This operation clones the data.

Safety

Out of bounds access doesn't Error but will return a Null value

pub fn take(&self, _indices: &ChunkedArray<UInt32Type>) -> Series

Take by index. This operation is clone.

Safety

Out of bounds access doesn't Error but will return a Null value

pub fn len(&self) -> usize

Get length of series.

pub fn is_empty(&self) -> bool

Check if Series is empty.

pub fn rechunk(&self) -> Series

Aggregate all chunks to a contiguous array of memory.

pub fn head(&self, _length: Option<usize>) -> Series

Get the head of the Series.

pub fn tail(&self, _length: Option<usize>) -> Series

Get the tail of the Series.

pub fn drop_nulls(&self) -> Series

Drop all null values and return a new Series.

pub fn expand_at_index(&self, _index: usize, _length: usize) -> Series

Create a new Series filled with values at that index.

Example

use polars_core::prelude::*;
let s = Series::new("a", [0i32, 1, 8]);
let expanded = s.expand_at_index(2, 4);
assert_eq!(Vec::from(expanded.i32().unwrap()), &[Some(8), Some(8), Some(8), Some(8)])

pub fn cast_with_datatype(
&self,
_data_type: &DataType
) -> Result<Series, PolarsError>

pub fn to_dummies(&self) -> Result<DataFrame, PolarsError>

Create dummy variables. See DataFrame

pub fn value_counts(&self) -> Result<DataFrame, PolarsError>

pub fn get(&self, _index: usize) -> AnyValue<'_>

Get a single value by index. Don't use this operation for loops as a runtime cast is needed for every iteration.

pub unsafe fn get_unchecked(&self, _index: usize) -> AnyValue<'_>

Get a single value by index. Don't use this operation for loops as a runtime cast is needed for every iteration.

Safety

Does not do any bounds checking

pub fn sort_in_place(&mut self, _reverse: bool)

Sort in place.

pub fn sort(&self, _reverse: bool) -> Series

pub fn argsort(&self, _reverse: bool) -> ChunkedArray<UInt32Type>

Retrieve the indexes needed for a sort.

pub fn null_count(&self) -> usize

Count the null values.

pub fn unique(&self) -> Result<Series, PolarsError>

Get unique values in the Series.

pub fn n_unique(&self) -> Result<usize, PolarsError>

Get unique values in the Series.

pub fn arg_unique(&self) -> Result<Vec<u32, Global>, PolarsError>

Get first indexes of unique values.

pub fn arg_true(&self) -> Result<ChunkedArray<UInt32Type>, PolarsError>

Get indexes that evaluate true

pub fn is_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the null values.

pub fn is_not_null(&self) -> ChunkedArray<BooleanType>

Get a mask of the non-null values.

pub fn is_unique(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Get a mask of all the unique values.

pub fn is_duplicated(&self) -> Result<ChunkedArray<BooleanType>, PolarsError>

Get a mask of all the duplicated values.

pub fn null_bits(&self) -> Vec<(usize, Option<Buffer>), Global>

Get the bits that represent the null values of the underlying ChunkedArray

pub fn reverse(&self) -> Series

return a Series in reversed order

pub fn as_single_ptr(&mut self) -> Result<usize, PolarsError>

Rechunk and return a pointer to the start of the Series. Only implemented for numeric types

pub fn shift(&self, _periods: i64) -> Series

Shift the values by a given period and fill the parts that will be empty due to this operation with Nones.

NOTE: If you want to fill the Nones with a value use the shift operation on ChunkedArray<T>.

Example

fn example() -> Result<()> {
    let s = Series::new("series", &[1, 2, 3]);

    let shifted = s.shift(1);
    assert_eq!(Vec::from(shifted.i32()?), &[None, Some(1), Some(2)]);

    let shifted = s.shift(-1);
    assert_eq!(Vec::from(shifted.i32()?), &[Some(2), Some(3), None]);

    let shifted = s.shift(2);
    assert_eq!(Vec::from(shifted.i32()?), &[None, None, Some(1)]);

    Ok(())
}
example();

pub fn fill_none(
&self,
_strategy: FillNoneStrategy
) -> Result<Series, PolarsError>

Replace None values with one of the following strategies:

Forward fill (replace None with the previous value)

Backward fill (replace None with the next value)

Mean fill (replace None with the mean of the whole array)

Min fill (replace None with the minimum of the whole array)

Max fill (replace None with the maximum of the whole array)

NOTE: If you want to fill the Nones with a value use the fill_none operation on ChunkedArray<T>.

Example

fn example() -> Result<()> {
    let s = Series::new("some_missing", &[Some(1), None, Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Forward)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Backward)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Min)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Max)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(2), Some(2)]);

    let filled = s.fill_none(FillNoneStrategy::Mean)?;
    assert_eq!(Vec::from(filled.i32()?), &[Some(1), Some(1), Some(2)]);

    Ok(())
}
example();

pub fn zip_with(
&self,
_mask: &ChunkedArray<BooleanType>,
_other: &Series
) -> Result<Series, PolarsError>

Create a new ChunkedArray with values from self where the mask evaluates true and values from other where the mask evaluates false

pub fn sum_as_series(&self) -> Series

Get the sum of the Series as a new Series of length 1.

pub fn max_as_series(&self) -> Series

Get the max of the Series as a new Series of length 1.

pub fn min_as_series(&self) -> Series

Get the min of the Series as a new Series of length 1.

pub fn mean_as_series(&self) -> Series

Get the mean of the Series as a new Series of length 1.

pub fn median_as_series(&self) -> Series

Get the median of the Series as a new Series of length 1.

pub fn var_as_series(&self) -> Series

Get the variance of the Series as a new Series of length 1.

pub fn std_as_series(&self) -> Series

Get the standard deviation of the Series as a new Series of length 1.

pub fn quantile_as_series(&self, _quantile: f64) -> Result<Series, PolarsError>

Get the quantile of the ChunkedArray as a new Series of length 1.

pub fn rolling_mean(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series, PolarsError>

Apply a rolling mean to a Series. See: ChunkedArray::rolling_mean.

pub fn rolling_sum(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series, PolarsError>

Apply a rolling sum to a Series. See: ChunkedArray::rolling_mean.

pub fn rolling_min(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series, PolarsError>

Apply a rolling min to a Series. See: ChunkedArray::rolling_mean.

pub fn rolling_max(
&self,
_window_size: usize,
_weight: Option<&[f64]>,
_ignore_null: bool
) -> Result<Series, PolarsError>

Apply a rolling max to a Series. See: ChunkedArray::rolling_mean.

pub fn fmt_list(&self) -> String

pub fn hour(&self) -> Result<Series, PolarsError>

Extract hour from underlying NaiveDateTime representation. Returns the hour number from 0 to 23.

pub fn minute(&self) -> Result<Series, PolarsError>

Extract minute from underlying NaiveDateTime representation. Returns the minute number from 0 to 59.

pub fn second(&self) -> Result<Series, PolarsError>

Extract second from underlying NaiveDateTime representation. Returns the second number from 0 to 59.

pub fn nanosecond(&self) -> Result<Series, PolarsError>

Extract second from underlying NaiveDateTime representation. Returns the number of nanoseconds since the whole non-leap second. The range from 1,000,000,000 to 1,999,999,999 represents the leap second.

pub fn day(&self) -> Result<Series, PolarsError>

Extract day from underlying NaiveDateTime representation. Returns the day of month starting from 1.

The return value ranges from 1 to 31. (The last day of month differs by months.)

pub fn ordinal_day(&self) -> Result<Series, PolarsError>

Returns the day of year starting from 1.

The return value ranges from 1 to 366. (The last day of year differs by years.)

pub fn month(&self) -> Result<Series, PolarsError>

Extract month from underlying NaiveDateTime representation. Returns the month number starting from 1.

The return value ranges from 1 to 12.

pub fn year(&self) -> Result<Series, PolarsError>

Extract month from underlying NaiveDateTime representation. Returns the year number in the calendar date.

pub fn datetime_str_fmt(&self, fmt: &str) -> Result<Series, PolarsError>

Format Date32/Date64 with a fmt rule. See chrono strftime/strptime.

pub fn clone_inner(&self) -> Arc<dyn SeriesTrait + 'static>

Clone inner ChunkedArray and wrap in a new Arc

pub fn get_as_any(&self, _index: usize) -> &(dyn Any + 'static)

Get the value at this index as a downcastable Any trait ref.

pub fn pow(&self, _exponent: f64) -> Result<Series, PolarsError>

Raise a numeric series to the power of exponent.

Loading content...
Implementations

impl<'a> dyn SeriesTrait + 'a

pub fn unpack<N>(&self) -> Result<&ChunkedArray<N>, PolarsError> where
N: 'static + PolarsDataType,